% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\docType{data}
\name{ANY_ITEMSETS}
\alias{ANY_ITEMSETS}
\title{ANY_ITEMSETS}
\format{
An object of class \code{character} of length 1.
}
\usage{
ANY_ITEMSETS
}
\description{
Reference value for defining possible entities: any itemsets.
}
\keyword{internal}
