% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cap}
\alias{cap}
\title{Capitalize a character string}
\usage{
cap(s)
}
\arguments{
\item{s}{Character string to be capitalized.}
}
\value{
The character string corresponding to the argument \code{s}, starting with a capital letter.
}
\description{
Change the first letter of a string to an uppercase.
}
\keyword{internal}
