% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{category_tree_chart,TransactionAnalyzer-method}
\alias{category_tree_chart,TransactionAnalyzer-method}
\alias{category_tree_chart}
\title{Category tree}
\usage{
\S4method{category_tree_chart}{TransactionAnalyzer}(
  object,
  category = NULL,
  items = object["items"],
  use_names = TRUE,
  n.cutoff = NULL,
  c.cutoff = NULL,
  vertex_size = 4,
  vertex_alpha = 1,
  leaf_size = 3,
  leaf_alpha = 1,
  leaf_margin = 0,
  label_size = 3,
  label_margin = 0.05
)
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}

\item{category}{Name or number of the category to represent on the tree (numbering according to
the order of the columns of \code{object["items_categories"]}).}

\item{items}{Items to represent on the tree. Any subset of \code{object["items"]}.

\code{"items"} and \code{"i"} are special values for \code{object["items"]}.}

\item{use_names}{If \code{TRUE}, display item names if they are defined. Display their identification
codes otherwise.}

\item{n.cutoff}{If \code{use_names} is \code{TRUE}, limit number of characters to display concerning the names
of the represented items.}

\item{c.cutoff}{Limit number of characters to display in the legend for the category represented.}

\item{vertex_size}{Size of vertices at depth 1, representing the category values. One of the
following.
\describe{
\item{A single numeric value}{All vertices have the size defined by this value.}
\item{A named numeric vector}{The sizes are assigned to the category values identified by
the names of the vector.}
\item{An unnamed numeric vector}{Must be of length equal to the number of category values
represented on the tree. The sizes are directly assigned to these category values
in alphanumeric order.}
}}

\item{vertex_alpha}{Opacity of vertices at depth 1 (from 0 to 1).}

\item{leaf_size}{Size of the leaves (i.e., vertices at depth 2), representing the items.
One of the following.
\describe{
\item{A single numeric value}{All leaves have the size defined by this value.}
\item{A named numeric vector}{The sizes are assigned to the items identified by the names of
the vector.}
\item{An unnamed numeric vector}{Must be of length equal to the number of items specified
by the argument \code{items}. The sizes are directly assigned to these items in the order
they are given.}
}}

\item{leaf_alpha}{Opacity of the leaves (from 0 to 1).}

\item{leaf_margin}{Margin before the leaves (i.e. distance between the ends of the edges and the
centers of the leaves).}

\item{label_size}{Size of the labels associated with the leaves.}

\item{label_margin}{Margin before the labels (i.e. distance between the centers of the leaves and
the labels).}
}
\value{
Graph created with the packages \code{ggraph} and \code{ggplot2}.
}
\description{
Plot a tree representing the values of one category and the related items. Vertices at depth 1
represent the values of the chosen category and vertices at depth 2 represent items related to each
specific category value.
}
\details{
If \code{category} is \code{NULL}, items are sorted alphanumerically.
If \code{category} is not \code{NULL}, items are sorted according to the values of the category then in the
order they are given.

The chart being plotted with the packages \code{ggraph} and \code{ggplot2}, it can be modified or completed
afterwards using \code{\link[ggplot2:last_plot]{ggplot2::last_plot}} or the returned object.

Default colors associated with the values of each category represented are selected circularly
among the 20 colors of the palette \code{category20} from D3 (see \code{ggsci::pal_d3("category20")}).
Therefore, if the number of values exceeds \code{20}, some colors will be used more than once.
For example, the \out{22<sup>nd</sup>} value will share the color of the \out{2<sup>nd</sup>} one.
See attribute \code{categories_colors} of the \code{TransactionAnalyzer} object to reassign colors to the
category values.
}
\examples{
category_tree_chart(TA_instance, "family")
category_tree_chart(TA_instance, 1,
                    items = c(19, 25, 27, 77, 87, 163, 192, 1603, 3146, 3350))
category_tree_chart(TA_instance, 1, items = TA_instance["items"][2:11]) +
  ggplot2::expand_limits(x = c(-1.3, 1.3), y = c(-1.3, 1.3))

## Use of an indicator as leaf size
ratio <- complexity_ratio(TA_instance["transactions"])
category_tree_chart(TA_instance, "family",
                    items = names(ratio),
                    use_names = FALSE,
                    leaf_size = ratio * 4)

}
\seealso{
\code{\link{co_occurrence_chart}}.
}
\author{
Gauthier Magnin
}
