% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{category_values,TransactionAnalyzer-method}
\alias{category_values,TransactionAnalyzer-method}
\alias{category_values}
\title{Itemset category values}
\usage{
\S4method{category_values}{TransactionAnalyzer}(object, itemsets, as_character = FALSE, unique = TRUE)
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}

\item{itemsets}{List of itemsets whose category values are to be found. Can be any itemsets.

\code{"transactions"}, \code{"t"}, \code{"nodes"}, \code{"n"}, \code{"patterns"} and \code{"p"} are special values for the itemsets
corresponding to the transactions, the nodes or the patterns contained in \code{object}.}

\item{as_character}{If \code{FALSE}, category values are returned as factor, in a list.
If \code{TRUE}, they are returned as character, in a data frame.}

\item{unique}{If \code{TRUE}, sorted unique values are returned for each itemset. If \code{FALSE},
duplicated values are not removed and there is correspondence between the return values and the
items forming the given itemsets.}
}
\value{
List or data frame (depending on the value of \code{as_character}) contaning for each category
(associated with the items in \code{object["items_categories"]}) the values regarding each itemset.
}
\description{
Give the category values associated with specific itemsets.
}
\examples{
## Values associated with specific itemsets
itemset_list <- list(c("19", "25"),
                     c("3156", "3157", "3345"),
                     c("19", "163", "929"))

category_values(TA_instance, itemset_list, unique = TRUE)
category_values(TA_instance, itemset_list, unique = FALSE)

category_values(TA_instance, itemset_list, as_character = FALSE)
category_values(TA_instance, itemset_list, as_character = TRUE)

## Values associated with itemsets from transactions or patterns
category_values(TA_instance, "transactions")
category_values(TA_instance, TA_instance["patterns"]$pattern[17:18])

## Values associated with itemsets from association rules
rules <- extract_rules(TA_instance)
category_values(TA_instance, rules$antecedent)
category_values(TA_instance, rules$consequent)

}
\seealso{
\code{\link{get_nodes}}, \code{\link{get_patterns}}, \code{\link{get_complexes}}, \code{\link{get_trx_from_category}}
}
\author{
Gauthier Magnin
}
