% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{check_acces_for_category,SpectralAnalyzer-method}
\alias{check_acces_for_category,SpectralAnalyzer-method}
\title{Vérifie que les paramètres fournis correspondent à une catégorie existante.}
\usage{
\S4method{check_acces_for_category}{SpectralAnalyzer}(object, category, value)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{category}{Nom ou numéro de la catégorie à laquelle accéder (numérotée selon l'ordre des colonnes de \code{object["items_categories"]}).}

\item{value}{Valeur recherchée pour la catégorie spécifiée par l'argument \code{category}.}
}
\description{
Affiche un message d'erreur si ce n'est pas le cas.
}
\author{
Gauthier Magnin
}
