% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{check_access_for_category,SpectralAnalyzer-method}
\alias{check_access_for_category,SpectralAnalyzer-method}
\alias{check_access_for_category}
\title{Validation of parameters for search by category}
\usage{
\S4method{check_access_for_category}{SpectralAnalyzer}(object, category, value)
}
\arguments{
\item{object}{SpectralAnalyzer class object.}

\item{category}{Name or number of the category to access (numbering according to the order of the
columns of \code{object["items_categories"]}).}

\item{value}{Sought value for the category specified by the argument \code{category}, or NA.}
}
\description{
Check that the parameters provided match an existing category.
Print an error message if not.
}
\details{
If \code{value = NA}, only the parameter \code{category} is checked.
}
\seealso{
\code{\link{extract_patterns_from_category}}, \code{\link{extract_nodes_from_category}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
