% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{check_access_for_category,TransactionAnalyzer-method}
\alias{check_access_for_category,TransactionAnalyzer-method}
\alias{check_access_for_category}
\title{Validation of parameters for search by category}
\usage{
\S4method{check_access_for_category}{TransactionAnalyzer}(object, category, value, stop = TRUE)
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}

\item{category}{Name or number of the category to access (numbering according to the order of the
columns of \code{object["items_categories"]}).}

\item{value}{Sought value(s) for the category specified by the argument \code{category}, or \code{NA}.}

\item{stop}{If \code{TRUE}, stop the execution and print an error message if the parameters do not
allow access to a category and a category value. If \code{FALSE}, see 'Value' section.}
}
\value{
\code{TRUE} or \code{FALSE} whether the parameters allow access to a category and a category
 value.
}
\description{
Check that the parameters provided match an existing category and an existing value within this
 category. Stop the execution and print an error message if not.
}
\details{
If \code{value} is \code{NA}, it is not checked; only the parameter \code{category} is.

If \code{category} is \code{NA}, only existence of any category is checked.
}
\note{
This method is not exported from the package.
}
\seealso{
\code{\link{get_patterns}}, \code{\link{get_nodes}}.
         \code{\link{get_patterns_from_category}}, \code{\link{get_nodes_from_category}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
