% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixer.R
\name{check_data_for_mcr_by_class}
\alias{check_data_for_mcr_by_class}
\title{Check data for MCR approach according to classes}
\usage{
check_data_for_mcr_by_class(
  values,
  references = NULL,
  vector = TRUE,
  matrix = TRUE,
  list = TRUE
)
}
\arguments{
\item{values}{Numeric named vector or matrix, or list of numeric named vectors. Data structure
whose naming is to be checked.}

\item{references}{Numeric vector or list of numeric vectors. Reference values associated with the
\code{values}.}

\item{vector}{Should be \code{TRUE} if \code{values} can be a vector.}

\item{matrix}{Should be \code{TRUE} if \code{values} can be a matrix.}

\item{list}{Should be \code{TRUE} if \code{values} can be a list.}
}
\description{
Check the naming and structure of the data of values and references on which to apply the MCR approach
according to classes. Stop the execution and print an error message if the data do not satisfy the
naming and structure criteria.
}
\details{
The criteria about naming and structure are the following:
\itemize{
\item If \code{values} is a vector, its values must be named.
\item If \code{values} is a matrix, its rows must be named.
\item If \code{values} is a list, it must contain vectors of named values.
\item If \code{values} and \code{references} are two lists, lengths of the elements of these lists must match.
\item If \code{values} is a list and \code{references} is a vector, the values of the latter must be named.
}

If \code{references = NULL}, check the \code{values} only.
}
\seealso{
\code{\link{mcr_summary_by_class}} \code{\link{mcr_chart_by_class}}, \code{\link{thq_pairs_freq_by_class}},
\code{\link{thq_freq_by_group_by_class}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
