% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_extension}
\alias{check_extension}
\title{Extension de fichier}
\usage{
check_extension(filename, ext)
}
\arguments{
\item{filename}{Chaîne de caractères devant être considérée comme un nom de fichier.}

\item{ext}{Extension souhaitée.}
}
\value{
Chaîne de caractères, terminant par l'extension \code{ext}.
}
\description{
Vérifie si un nom de fichier possède l'extension désirée.
Ajoute cette extension si ce n'est pas le cas.
}
\keyword{internal}
