% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{check_init,SpectralAnalyzer-method}
\alias{check_init,SpectralAnalyzer-method}
\alias{check_init}
\title{Initialization validation}
\usage{
\S4method{check_init}{SpectralAnalyzer}(object, part = NULL, stop = TRUE, prefix = "", suffix = "")
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{part}{\code{NULL} or character value corresponding to the part of the attributes to test.
One or more of the following:
\describe{
\item{\code{"nodes"}, \code{"n"}}{Attributes related to nodes.}
\item{\code{"nodes_links"}, \code{"nl"}}{Attributes related to links between nodes.}
\item{\code{"patterns"}, \code{"p"}}{Attributes related to patterns.}
\item{\code{"patterns_links"}, \code{"pl"}}{Attributes related to links between patterns.}
}}

\item{stop}{If \code{TRUE}, stop the execution and print an error message if the attributes related to
\code{part} are note initialized. If \code{FALSE}, see 'Value' section.}

\item{prefix}{Text to be prefixed to the message.}

\item{suffix}{Text to be suffixed to the message.}
}
\value{
If \code{part} is not \code{NULL}, as many values as there are in \code{part}:
\itemize{
\item{\code{TRUE} if the related attributes are initialized.}
\item{\code{FALSE} if the related attributes are not initialized.}
}

If \code{part} is \code{NULL}, 4 logical values corresponding to the initialization related to \code{nodes},
\code{nodes_links}, \code{patterns}, \code{patterns_links} (in this order).
}
\description{
Check that a part of a \code{SpectralAnalyzer} object is initialized.
Stop the execution and print an error message if not.
}
\details{
The initialization of the nodes consists of the initialization of the attributes \code{n_links} and
\code{nodes_links}.

The initialization of the links between nodes consists of the initialization of the attributes
\code{nodes} and \code{nodes_per_year}.

The initialization of the patterns consists of the initialization of the attributes \code{patterns},
\code{patterns_per_year} and \code{nodes_patterns}.

The initialization of the links between patterns consists of the initialization of the attributes
\code{p_links} and \code{patterns_links}.
}
\seealso{
\code{\link{is_init}}, \code{\link{init}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
