% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_param}
\alias{check_param}
\title{Validation of a parameter}
\usage{
check_param(
  x,
  types = NULL,
  values = NULL,
  range = NULL,
  quotes = TRUE,
  stop = TRUE,
  prefix = "",
  suffix = ""
)
}
\arguments{
\item{x}{Parameter to validate.}

\item{types}{Types among which \code{x} must be.}

\item{values}{Values among which \code{x} must be.}

\item{range}{Two-element numeric vector. Range in which \code{x} must be. Ignored if \code{values} is not \code{NULL}.}

\item{quotes}{If \code{TRUE}, surround in quotes the possible values in the error message.}

\item{stop}{If \code{TRUE}, stop the execution and print an error message if the parameter is not valid.
If \code{FALSE}, see 'Value' section.}

\item{prefix}{Text to be prefixed to the message.}

\item{suffix}{Text to be suffixed to the message.}
}
\value{
\code{TRUE} or \code{FALSE} whether the parameter is valid.
}
\description{
Check that a parameter is of a specific type, is member of a set of values or is in a range of values.
Stop the execution and print an error message if not.
The expected type, range or possible values are detailed in the error message.
}
\details{
Related validation is ignored if \code{types}, \code{values} or \code{range} is \code{NULL}.

Parameter type is checked using the \code{\link[base:mode]{base::mode}} function.
}
\note{
This function is not exported from the package.
}
\author{
Gauthier Magnin
}
\keyword{internal}
