% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{check_params_for_RI,SpectralAnalyzer-method}
\alias{check_params_for_RI,SpectralAnalyzer-method}
\alias{check_params_for_RI}
\title{Validation of RI computation parameters}
\usage{
\S4method{check_params_for_RI}{SpectralAnalyzer}(object, t, period)
}
\arguments{
\item{object}{SpectralAnalyzer class object.}

\item{t}{Year of the end of the period, i.e. the date on which to characterize the patterns.
\code{NULL} specifies that the characterization must be done in relation to the last year covered
by the observations.}

\item{period}{Time interval over which to compute the reporting indexes (number of years).
For example, if \code{t = 2015} and \code{period = 2} then the computation is made over the
period [2014 - 2015].
\code{Inf} specifies that the period considered covers an interval starting on the date of the
oldest observation and ending in the year \code{t}.}
}
\value{
List containing the final values of \code{t} and \code{period}.
}
\description{
Check the validity of the values of the parameters given for the computation of reporting indexes.
Adapt their values if they do not fall within the correct range and print a warning message.
}
\seealso{
\code{\link{compute_reporting_indexes}}, \code{\link{compute_reporting_indexes_limits}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
