% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{cluster_chart,SpectralAnalyzer-method}
\alias{cluster_chart,SpectralAnalyzer-method}
\title{Identifie le cluster associé à l'item fourni en argument et en dessine un spectrosome.}
\usage{
\S4method{cluster_chart}{SpectralAnalyzer}(
  object,
  entities,
  item,
  vertex_size = "relative",
  path = getwd(),
  name = paste0(substr(entities, 1, nchar(entities) - 1), "_cluster_of_", item, ".png"),
  title = paste(cap(substr(entities, 1, nchar(entities) - 1)), "cluster of", item)
)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{entities}{Type d'élément pour rechercher un cluster (noeuds ou motifs).
Choix parmi \code{c("nodes", "patterns")}.}

\item{item}{Code de l'item dont le cluster est à identifier.}

\item{vertex_size}{Façon dont les tailles des sommets du graphe doivent être définies.
Choix parmi \code{c("relative", "grouped", "absolute", "equal")}. Cf. \code{\link{spectrosome_chart}}}

\item{path}{Chemin du dossier dans lequel enregistrer les graphiques.
Par défaut, les graphiques sont enregistrés dans le répertoire de travail.}

\item{name}{Nom (avec extension) du fichier dans lequel enregistrer le graphique.
Par défaut, le nom dépend des arguments \code{entities} et \code{item}.
Exemple de nom par défaut : \code{"node_cluster_of_25"}.}

\item{title}{Titre du graphique.
Par défaut, le titre dépend des arguments \code{entities} et \code{item}.
Exemple de titre par défaut : \code{"Node cluster of 25"}.}
}
\value{
\code{NULL} si aucun ou un seul noeud ou motif contient l'item recherché. Sinon, liste contenant :
 \itemize{
   \item{} {Data.frame des noeuds ou motifs et caractéristiques utilisées, associés aux identifiants des sommets du graphe.}
   \item{} {Data.frame des informations relatives aux arêtes du graphe.}
 }
}
\description{
Identifie le cluster associé à l'item fourni en argument et en dessine un spectrosome.
}
\seealso{
\code{\link{spectrosome_chart}}
}
\author{
Gauthier Magnin
}
