% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{cluster_text,TransactionAnalyzer-method}
\alias{cluster_text,TransactionAnalyzer-method}
\alias{cluster_text}
\title{Display of cluster names}
\usage{
\S4method{cluster_text}{TransactionAnalyzer}(
  object,
  graph,
  links,
  display = Inf,
  highlight = 3,
  use_names = TRUE,
  cutoff = NULL
)
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}

\item{graph}{Graph generated by the function \code{\link[sna:gplot]{gplot}} from the package
\code{sna}: "A two-column matrix containing the vertex positions as x,y coordinates".}

\item{links}{Links of nodes or patterns used to generate \code{graph}.}

\item{display}{Maximum number of clusters to name on the graph.
If the actual number of clusters is greater, the names of the smaller ones are not displayed.}

\item{highlight}{Number of clusters to highlight among those named on the graph.
The names of the largest clusters are displayed in bold.}

\item{use_names}{If \code{TRUE}, display item names if they are defined. Display their identification
codes otherwise.}

\item{cutoff}{If \code{use_names = TRUE}, limit number of characters to display concerning the names
of the represented items.}
}
\description{
Identify and display the names of the clusters of the graph provided, corresponding to nodes or
 patterns sharing one item.
The names of clusters confused because all of their links are mixed links, are not displayed.
Texts are written on the active graphics device.
}
\note{
This method is not exported from the package.
}
\seealso{
\code{\link{spectrosome_chart}}, \code{\link[sna:gplot]{sna::gplot}}.
}
\author{
Delphine Bosson-Rieutort, Gauthier Magnin
}
\keyword{internal}
