% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{co_occurrence_chart,SpectralAnalyzer-method}
\alias{co_occurrence_chart,SpectralAnalyzer-method}
\alias{co_occurrence_chart}
\title{Co-occurrence chart}
\usage{
\S4method{co_occurrence_chart}{SpectralAnalyzer}(
  object,
  items,
  category = NULL,
  min_occ = 1,
  max_occ = Inf,
  use_names = TRUE,
  n.cutoff = NULL,
  c.cutoff = NULL,
  sort_by = "category",
  vertex_size = 2,
  vertex_alpha = 0.5,
  vertex_margin = 0.05,
  label_size = 3,
  label_margin = 0.05,
  edge_tension = 0.7,
  edge_alpha = 1,
  palette = "Blues",
  palette_direction = 1
)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{items}{Items for which to count co-occurrences between pairs and to plot on the graph.
Any subset of \code{object["items"]}.\cr
\code{"items"} and \code{"i"} are specific values for \code{object["items"]}.}

\item{category}{Name or number of the category to represent on the graph (numbering according to
the order of the columns of \code{object["items_categories"]}).}

\item{min_occ}{Minimum number of co-occurrences to consider to plot a link between two items.}

\item{max_occ}{Maximum number of co-occurrences to consider to plot a link between two items.}

\item{use_names}{If \code{TRUE}, display item names if they are defined. Display their identification
codes otherwise.}

\item{n.cutoff}{If \code{use_names = TRUE}, limit number of characters to display concerning the names
of the represented items.}

\item{c.cutoff}{Limit number of characters to display in the legend for the category represented.}

\item{sort_by}{Sorting method of displayed items. One of \code{"category"}, \code{"item"}.}

\item{vertex_size}{Size of the vertices.}

\item{vertex_alpha}{Opacity of the vertices (from 0 to 1).}

\item{vertex_margin}{Margin before the vertices (i.e. distance between the ends of the edges and the
centers of the vertices).}

\item{label_size}{Size of the labels associated with the vertices.}

\item{label_margin}{Margin before the labels (i.e. distance between the centers of the vertices and
the labels).}

\item{edge_tension}{Looseness of the connecting lines (from 0 to 1).
The closer the value is to 1, the straighter the lines will be.
The closer the value is to 0, the more the lines will be curved.}

\item{edge_alpha}{Opacity of the lines connecting vertices (from 0 to 1).}

\item{palette}{Name (or number) of the sequential palette to use for coloring the edges.
One of \code{"Blues"}, \code{"BuGn"}, \code{"BuPu"}, \code{"GnBu"}, \code{"Greens"}, \code{"Greys"}, \code{"Oranges"}, \code{"OrRd"},
\code{"PuBu"}, \code{"PuBuGn"}, \code{"PuRd"}, \code{"Purples"}, \code{"RdPu"}, \code{"Reds"}, \code{"YlGn"}, \code{"YlGnBu"}, \code{"YlOrBr"},
\code{"YlOrRd"}.}

\item{palette_direction}{Direction in which to use the color palette.
If \code{1}, colors are in original order: from the lightest to the darkest.
If \code{-1}, the order of colors is reversed: from the darkest to the lightest.}
}
\value{
Graph created with the packages \code{ggraph} and \code{ggplot2}.
}
\description{
Plot a graph in which vertices are items and edges are their co-occurences in observations (i.e. for
each pair of items, the number of observations containing it).
}
\details{
The chart being plotted with the packages \code{ggraph} and \code{ggplot2}, it can be modified or completed
afterwards using \code{\link[ggplot2:last_plot]{ggplot2::last_plot}} or the returned object.

The colors associated with the values of the possible category represented are selected circularly
among the 20 colors of the palette \code{category20} from D3 (see \code{ggsci::pal_d3("category20")}).
Therefore, if the number of values exceeds \eqn{20}, some colors will be used more than once.
For example, the \out{22<sup>nd</sup>} value will share the color of the \out{2<sup>nd</sup>}
value.
See the attribute \code{categories_colors} of \code{object} to reassign colors to the category values.
}
\examples{
co_occurrence_chart(SA_instance, SA_instance["items"], "family")
co_occurrence_chart(SA_instance, "items", category = 1, n.cutoff = 20) +
  ggplot2::expand_limits(x = c(-1.3, 1.3), y = c(-1.3, 1.3))
co_occurrence_chart(SA_instance, "items", category = "family",
                    min_occ = 2, palette = "OrRd")
co_occurrence_chart(SA_instance, SA_instance["items"][2:13], "family")

}
\seealso{
\code{\link{co_occurrence_matrix}}, \code{\link{category_tree_chart}}.
}
\author{
Gauthier Magnin
}
