% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{co_occurrence_chart,TransactionAnalyzer-method}
\alias{co_occurrence_chart,TransactionAnalyzer-method}
\alias{co_occurrence_chart}
\alias{co_occurrence_chart,TransactionAnalyzer}
\title{Co-occurrence chart, for TransactionAnalyzer}
\usage{
\S4method{co_occurrence_chart}{TransactionAnalyzer}(
  object,
  items = object["items"],
  category = NULL,
  co_occ = NULL,
  proportions = FALSE,
  min_occ = if (proportions) .Machine$double.xmin else 1,
  max_occ = Inf,
  use_names = TRUE,
  n.cutoff = NULL,
  c.cutoff = NULL,
  sort_by = "category",
  vertex_size = 3,
  vertex_alpha = 1,
  vertex_margin = 0.05,
  label_size = 3,
  label_margin = 0.05,
  edge_looseness = 0.8,
  edge_alpha = 1,
  palette = "Blues",
  palette_direction = 1
)
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}

\item{items}{Items for which to plot co-occurrences between pairs.
Any subset of \code{object["items"]} containing at least two items.

\code{"items"} and \code{"i"} are special values for \code{object["items"]}.}

\item{category}{Name or number of the category to represent on the graph (numbering according to
the order of the columns of \code{object["items_categories"]}).}

\item{co_occ}{Matrix containing the co-occurrences (or their proportions) for at least the items
specified by the argument \code{items}. Is computed if \code{NULL}.}

\item{proportions}{\code{TRUE} if the proportions of co-occurrences are to be plotted (and computed, if
\code{co_occ} is \code{NULL}) instead of the co-occurrences themselves.}

\item{min_occ}{Minimum number of co-occurrences (or minimum proportion) to consider to plot a link
between two items.
Default value depends on the argument \code{proportions} and allows not to plot links between items
that never co-occur.}

\item{max_occ}{Maximum number of co-occurrences (or maximum proportion) to consider to plot a link
between two items.}

\item{use_names}{If \code{TRUE}, display item names if they are defined. Display their identification
codes otherwise.}

\item{n.cutoff}{If \code{use_names = TRUE}, limit number of characters to display concerning the names
of the represented items.}

\item{c.cutoff}{Limit number of characters to display in the legend for the category represented.}

\item{sort_by}{Sorting method of displayed items. One of \code{"category"}, \code{"item"}.}

\item{vertex_size}{Size of the vertices.}

\item{vertex_alpha}{Opacity of the vertices (from 0 to 1).}

\item{vertex_margin}{Margin before the vertices (i.e. distance between the ends of the edges and the
centers of the vertices).}

\item{label_size}{Size of the labels associated with the vertices.}

\item{label_margin}{Margin before the labels (i.e. distance between the centers of the vertices and
the labels).}

\item{edge_looseness}{Looseness of the connecting lines (from 0 to 1).
The closer the value is to 0, the straighter the lines will be.
The closer the value is to 1, the more the lines will be curved.}

\item{edge_alpha}{Opacity of the lines connecting vertices (from 0 to 1).}

\item{palette}{Name (or number) of the sequential palette to use for coloring the edges.
One of \code{"Blues"}, \code{"BuGn"}, \code{"BuPu"}, \code{"GnBu"}, \code{"Greens"}, \code{"Greys"}, \code{"Oranges"}, \code{"OrRd"},
\code{"PuBu"}, \code{"PuBuGn"}, \code{"PuRd"}, \code{"Purples"}, \code{"RdPu"}, \code{"Reds"}, \code{"YlGn"}, \code{"YlGnBu"}, \code{"YlOrBr"},
\code{"YlOrRd"}.}

\item{palette_direction}{Direction in which to use the color palette.
If \code{1}, colors are in original order: from the lightest to the darkest.
If \code{-1}, color order is reversed: from the darkest to the lightest.}
}
\value{
Graph created with the packages \code{ggraph} and \code{ggplot2}.
}
\description{
Plot a graph in which vertices are items and edges are their co-occurences in transactions (i.e. for
each pair of items, the number of transactions containing it). Edges can also represent the
proportions of these co-occurrences (i.e., the ratio between the number of transactions containing a
pair of items and the number of transactions containing at least one of them).
}
\details{
The chart being plotted with the packages \code{ggraph} and \code{ggplot2}, it can be modified or completed
afterwards using \code{\link[ggplot2:last_plot]{ggplot2::last_plot}} or the returned object.

If \code{category} is \code{NULL} or \code{sort_by = "item"}, items are sorted alphanumerically.
If \code{category} is not \code{NULL} and \code{sort_by = "category"}, items are sorted according to the values of
the category then alphanumerically.

Default colors associated with the values of each category represented are selected circularly
among the 20 colors of the palette \code{category20} from D3 (see \code{ggsci::pal_d3("category20")}).
Therefore, if the number of values exceeds \code{20}, some colors will be used more than once.
For example, the \out{22<sup>nd</sup>} value will share the color of the \out{2<sup>nd</sup>} one.
See attribute \code{categories_colors} of the \code{TransactionAnalyzer} object to reassign colors to the
category values.
}
\note{
If using the RStudio IDE and the value of the argument \code{edge_alpha} is not \eqn{1}, edges may not
be displayed in the RStudio "Plots" pane. However, they will be actually displayed in the "Plot Zoom"
window; while exporting the plot; or by using another graphics device.
}
\section{Warning}{


Several versions of the package \code{ggraph} experience some issues about
hierarchical edge bundling charts.
Such charts are created by the functions \code{co_occurrence_chart} and
\code{rules_chart}.

The main issue encountered is that wrong colors are assigned to edges.
Note that such an issue does not throw any error or warning messages.
One way to spot it is to notice light lines above dark lines.

Note also that some of the newer versions of \code{ggraph} have the same issues.
A simple way to know if the installed version is concerned is to run the
first instruction in section "4.2. Co-occurrence charts" of the vignette
titled "Transaction analyzes" and compare the resulting graph with the one
shown in the document.

The correct operation of these functions has been validated with the version
2.0.5 of \code{ggraph}. To install it, run the following instruction.
\preformatted{
remotes::install_version("ggraph", version = "2.0.5")
}
}

\examples{
co_occurrence_chart(TA_instance, category = "family")
co_occurrence_chart(TA_instance, category = 1, n.cutoff = 20) +
  ggplot2::expand_limits(x = c(-1.5, 1.5), y = c(-1.5, 1.5))
co_occurrence_chart(TA_instance, category = "family",
                    min_occ = 2, palette = "OrRd")

co_occurrence_chart(TA_instance, TA_instance["items"][2:13], "family")
co_occurrence_chart(TA_instance, TA_instance["items"][2:13], "family",
                    proportions = TRUE)

}
\seealso{
\code{\link{co_occurrence_matrix}}, \code{\link{category_tree_chart}}.

Method for signature \code{TransactionSet}:
\code{\link[=co_occurrence_chart,TransactionSet-method]{co_occurrence_chart,TransactionSet}}.
}
\author{
Gauthier Magnin
}
