% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_set.R
\name{co_occurrence_matrix,ObservationSet-method}
\alias{co_occurrence_matrix,ObservationSet-method}
\alias{co_occurrence_matrix}
\title{Co-occurrence matrix}
\usage{
\S4method{co_occurrence_matrix}{ObservationSet}(object, items = NULL)
}
\arguments{
\item{object}{S4 object of class \code{ObservationSet}.}

\item{items}{Items for which to count co-occurrences between pairs. The default \code{NULL} means to count
them considering each existing item.}
}
\value{
Co-occurrence matrix between each pair of items.
}
\description{
Build the co-occurrence matrix: the matrix of the number of observations containing a specific pair of
items, for each possible pair. The diagonal is the number of observations containing each specific
item.
}
\examples{
co_occurrence_matrix(OS_instance)
co_occurrence_matrix(OS_instance, items = c(19, 25, 148, 3146))

}
\seealso{
\code{\link{complexity_ratio}}, \code{\link{complexity_index}},
\code{\link{get_all_items}}, \code{\link{get_obs_from_items}}, \code{\link{get_complex_obs}}, \code{\link{get_simple_obs}}.
}
\author{
Gauthier Magnin
}
