% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_set.R
\name{co_occurrence_matrix,TransactionSet-method}
\alias{co_occurrence_matrix,TransactionSet-method}
\alias{co_occurrence_matrix}
\title{Co-occurrence matrix}
\usage{
\S4method{co_occurrence_matrix}{TransactionSet}(object, items = NULL, proportions = FALSE)
}
\arguments{
\item{object}{S4 object of class \code{TransactionSet}.}

\item{items}{Items for which to count co-occurrences between pairs. The default \code{NULL} means to count
them considering each existing item.}

\item{proportions}{If \code{TRUE}, proportions are computed instead of numbers of co-occurrences:
ratio between the number of transactions containing a pair of items and the number of transactions
containing at least one of them.}
}
\value{
Matrix of co-occurrences (or of proportions of co-occurrences, according to \code{proportions})
between each pair of items.
\code{NA} values are assigned in the diagonal to given items that do not exist in the transactions.
}
\description{
Build the co-occurrence matrix: the matrix of the number of transactions containing a specific pair of
items, for each possible pair. The diagonal is the number of transactions containing each specific
item.
}
\examples{
co_occurrence_matrix(TS_instance)
co_occurrence_matrix(TS_instance, items = c(497, 931, 3157, 3350))
co_occurrence_matrix(TS_instance, items = c(497, 931, 3157, 3350),
                     proportions = TRUE)

}
\seealso{
\code{\link{complexity_ratio}}, \code{\link{complexity_index}},
\code{\link{get_all_items}}, \code{\link{get_trx_from_items}}, \code{\link{get_complex_trx}}, \code{\link{get_simple_trx}}.
}
\author{
Gauthier Magnin
}
