% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_maker.R
\name{co_occurrence_matrix}
\alias{co_occurrence_matrix}
\title{Co-occurrence matrix}
\usage{
co_occurrence_matrix(observations, key = "CODE")
}
\arguments{
\item{observations}{List of observations.}

\item{key}{Access key to the items in an observation.}
}
\value{
Co-occurrence matrix between each pair of items.
}
\description{
Build the co-occurrence matrix: the matrix of the number of observations containing a specific pair of
items, for each possible pair.
}
\examples{
obs <- make_observations(oedb_sample, by = "ID", additional = c("CODE", "NAME"))
co_occurrence_matrix(obs)

}
\seealso{
\code{\link{complexity_ratio}}, \code{\link{complexity_index}},
\code{\link{get_all_items}}, \code{\link{get_obs_from_items}}, \code{\link{get_complex_obs}}, \code{\link{get_simple_obs}}.
}
\author{
Gauthier Magnin
}
