% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_maker.R
\name{co_occurrence_proportions}
\alias{co_occurrence_proportions}
\title{Proportions of co-occurrence of each item with any other item}
\usage{
co_occurrence_proportions(observations, key = "CODE")
}
\arguments{
\item{observations}{List of observations.}

\item{key}{Access key to the items in an observation.}
}
\value{
Vector of the proportions of occurrences of each item in complex observations among all
 observations.
}
\description{
For each item, compute the ratio between the number of occurrences of the item in complex observations
 and the number of occurrences of the item in all observations. In other words, compute the inverse of
 the ratio between the number of times the items appears alone and the number of times the item appears.
}
\examples{
obs <- make_observations(oedb_sample, by = "ID", additional = c("CODE", "NAME"))
co_occurrence_proportions(obs)

}
\seealso{
\code{\link{get_all_items}}, \code{\link{get_obs_from_items}},
         \code{\link{get_complex_obs}}, \code{\link{get_simple_obs}}.
}
\author{
Gauthier Magnin
}
