% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_manager.R
\name{coerce_list}
\alias{coerce_list}
\title{Turn a list into a matrix or a data frame}
\usage{
coerce_list(
  x,
  to,
  by_name = FALSE,
  stringsAsFactors = default.stringsAsFactors()
)
}
\arguments{
\item{x}{List of vectors to turn into a matrix or a data frame.}

\item{to}{Data structure into which to transform the list. One of \code{"logical matrix"} (or \code{"lm"} for
short), \code{"character matrix"} (or \code{"cm"}), \code{"data.frame"} (or \code{"df"}).}

\item{by_name}{If \code{FALSE}, use the values of the list. If \code{TRUE}, use the names of these values.}

\item{stringsAsFactors}{logical: should the character vectors be converted to factors?
Ignored if \code{to} doesn't refer to data frame.}
}
\value{
If \code{to} refers to logical matrix, logical matrix in which row names are the names of the list and
column names are the values of the list or the names of these values (according to \code{by_name}).

Otherwise, character matrix or data frame (according to \code{to}) in which the first column contains the
names of the list and the second one contains the related values or the names of these values
(according to \code{by_name}).
}
\description{
Turn a list into a logical matrix, a character matrix or a data frame, associating names of the list
with its values or with the names of these values.
}
\examples{
l <- list(e1 = c(v1 = "E", v2 = "F", v3 = "I"),
          e2 = c(v1 = "I"),
          e3 = c(v1 = "C", v2 = "I"),
          e4 = c(v1 = "A", v2 = "C", v3 = "D", v4 = "F"),
          e5 = c(v1 = "B", v2 = "D", v3 = "E", v4 = "I", v5 = "G"))

coerce_list(l, "logical matrix")
coerce_list(l, "lm", by_name = TRUE)

coerce_list(l, "character matrix")
coerce_list(l, "cm", by_name = TRUE)

coerce_list(l, "data.frame")
coerce_list(l, "df", by_name = TRUE)

}
\seealso{
\code{\link{invert_list}}, \code{\link{coerce_to_list}}.
}
\author{
Gauthier Magnin
}
