% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_manager.R
\name{coerce_to_list}
\alias{coerce_to_list}
\title{Turn a matrix or a data frame into a list}
\usage{
coerce_to_list(x, by_row = TRUE)
}
\arguments{
\item{x}{Matrix or data frame to turn into a list.}

\item{by_row}{Only considered if \code{x} is a logical matrix.
If \code{TRUE}, use row names as names of the list and column names as values.
If \code{FALSE}, use column names as names of the list and row names as values.}
}
\value{
If \code{x} is a logical matrix, list whose values correspond to the values \code{TRUE} of each row or of each
column (according to \code{by_row}).

Otherwise, list whose names correspond to the first column of \code{x} and whose values correspond to the
second column of \code{x}.
}
\description{
Turn a logical matrix, a two-column character matrix or a two-column data frame into a list.
}
\examples{
m1 <- matrix(c(FALSE, FALSE, FALSE, FALSE, TRUE,  TRUE,  FALSE, TRUE,
              FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, TRUE,
              FALSE, FALSE, TRUE,  FALSE, FALSE, FALSE, FALSE, TRUE,
              TRUE,  FALSE, TRUE,  TRUE,  FALSE, TRUE,  FALSE, FALSE,
              FALSE, TRUE,  FALSE, TRUE,  TRUE,  FALSE, TRUE,  TRUE),
            nrow = 5, byrow = TRUE,
            dimnames = list(c("e1", "e2", "e3", "e4", "e5"),
                            c("A", "B", "C", "D", "E", "F", "G", "I")))

coerce_to_list(m1)
coerce_to_list(m1, by_row = FALSE)


m2 <- matrix(c("e1", "e1", "e1", "e2", "e3", "e3", "e4",
               "e4", "e4", "e4", "e5", "e5", "e5", "e5", "e5",
               "E", "F", "I", "I", "C", "I", "A",
               "C", "D", "F", "B", "D", "E", "G", "I"),
             ncol = 2)

coerce_to_list(m2)


df <- data.frame(V1 = c("e1", "e1", "e1", "e2", "e3", "e3", "e4",
                        "e4", "e4", "e4", "e5", "e5", "e5", "e5", "e5"),
                 V2 = c("E", "F", "I", "I", "C", "I", "A",
                        "C", "D", "F", "B", "D", "E", "G", "I"),
                 stringsAsFactors = FALSE)

coerce_to_list(df)

}
\seealso{
\code{\link{invert_list}}, \code{\link{coerce_list}}.
}
\author{
Gauthier Magnin
}
