% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_set.R
\name{complexity_index,ObservationSet-method}
\alias{complexity_index,ObservationSet-method}
\alias{complexity_index}
\title{Complexity Index, by item}
\usage{
\S4method{complexity_index}{ObservationSet}(object, items = NULL)
}
\arguments{
\item{object}{S4 object of class \code{ObservationSet}.}

\item{items}{Items for which to calculate the complexity index. The default \code{NULL} means to calculate
it for each existing item.}
}
\value{
Named vector of complexity indexes: for each item, the number of complex observations
containing it.
}
\description{
For each item, count the number of complex observations containing the item.
}
\examples{
complexity_index(OS_instance)
complexity_index(OS_instance, items = c(19, 25, 148))

}
\seealso{
\code{\link{complexity_ratio}}, \code{\link{co_occurrence_matrix}},
\code{\link{get_all_items}}, \code{\link{get_obs_from_items}}, \code{\link{get_complex_obs}}, \code{\link{get_simple_obs}}.
}
\author{
Gauthier Magnin
}
