% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_set.R
\name{complexity_index,TransactionSet-method}
\alias{complexity_index,TransactionSet-method}
\alias{complexity_index}
\title{Complexity Index, by item}
\usage{
\S4method{complexity_index}{TransactionSet}(object, items = NULL)
}
\arguments{
\item{object}{S4 object of class \code{TransactionSet}.}

\item{items}{Items for which to calculate the complexity index. The default \code{NULL} means to calculate
it for each existing item.}
}
\value{
Complexity indexes: for each item, the number of complex transactions containing it.
Named vector if \code{items} contains more than one item. Single value otherwise.
\code{NA} values are assigned to given items that do not exist in the transations.
}
\description{
For each item, count the number of complex transactions containing the item.
}
\examples{
complexity_index(TS_instance)
complexity_index(TS_instance, items = c(19, 25, 148))
complexity_index(TS_instance, items = 3146)

}
\seealso{
\code{\link{complexity_ratio}}, \code{\link{co_occurrence_matrix}},
\code{\link{get_all_items}}, \code{\link{get_trx_from_items}}, \code{\link{get_complex_trx}}, \code{\link{get_simple_trx}}.
}
\author{
Gauthier Magnin
}
