% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_maker.R
\name{complexity_index}
\alias{complexity_index}
\title{Complexity Index, by item}
\usage{
complexity_index(observations, items = NULL, key = "CODE")
}
\arguments{
\item{observations}{List of observations.}

\item{items}{Items for which to calculate the complexity index. The default \code{NULL} means to calculate
it for each existing item.}

\item{key}{Access key to the items in an observation.}
}
\value{
Vector of complexity indexes: for each item, the number of complex observations containing it.
}
\description{
For each item, count the number of complex observations containing the item.
}
\examples{
obs <- make_observations(oedb_sample, by = "ID", additional = c("CODE", "NAME"))
complexity_index(obs)
complexity_index(obs, items = c(19, 25, 148))

}
\seealso{
\code{\link{complexity_ratio}}, \code{\link{co_occurrence_matrix}},
\code{\link{get_all_items}}, \code{\link{get_obs_from_items}}, \code{\link{get_complex_obs}}, \code{\link{get_simple_obs}}.
}
\author{
Gauthier Magnin
}
