% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_set.R
\name{complexity_ratio,ObservationSet-method}
\alias{complexity_ratio,ObservationSet-method}
\alias{complexity_ratio}
\title{Complexity Ratio, by item}
\usage{
\S4method{complexity_ratio}{ObservationSet}(object, items = NULL)
}
\arguments{
\item{object}{S4 object of class \code{ObservationSet}.}

\item{items}{Items for which to compute the complexity ratio. The default \code{NULL} means to compute
it for each existing item.}
}
\value{
Named vector of complexity ratios: for each item, the proportion of complex observations
containing it among all observations containing it.
}
\description{
For each item, compute the ratio between the number of complex observations containing the item and
the total number of observations containing the item. In other words, compute the complement of the
ratio between the number of times the item appears alone and the number of times the item appears.
}
\examples{
complexity_ratio(OS_instance)
complexity_ratio(OS_instance, items = c(25, 148, 3146))

}
\seealso{
\code{\link{complexity_index}}, \code{\link{co_occurrence_matrix}},
\code{\link{get_all_items}}, \code{\link{get_obs_from_items}}, \code{\link{get_complex_obs}}, \code{\link{get_simple_obs}}.
}
\author{
Gauthier Magnin
}
