% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_set.R
\name{complexity_ratio,TransactionSet-method}
\alias{complexity_ratio,TransactionSet-method}
\alias{complexity_ratio}
\title{Complexity Ratio, by item}
\usage{
\S4method{complexity_ratio}{TransactionSet}(object, items = NULL)
}
\arguments{
\item{object}{S4 object of class \code{TransactionSet}.}

\item{items}{Items for which to compute the complexity ratio. The default \code{NULL} means to compute
it for each existing item.}
}
\value{
Complexity ratios: for each item, the proportion of complex transactions containing it among
all transactions containing it.
Named vector if \code{items} contains more than one item. Single value otherwise.
\code{NA} values are assigned to given items that do not exist in the transations.
}
\description{
For each item, compute the ratio between the number of complex transactions containing the item and
the total number of transactions containing the item. In other words, compute the complement of the
ratio between the number of times the item appears alone and the number of times the item appears.
}
\examples{
complexity_ratio(TS_instance)
complexity_ratio(TS_instance, items = c(25, 148, 3146))
complexity_ratio(TS_instance, items = 3146)

}
\seealso{
\code{\link{complexity_index}}, \code{\link{co_occurrence_matrix}},
\code{\link{get_all_items}}, \code{\link{get_trx_from_items}}, \code{\link{get_complex_trx}}, \code{\link{get_simple_trx}}.
}
\author{
Gauthier Magnin
}
