% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_maker.R
\name{complexity_ratio}
\alias{complexity_ratio}
\title{Complexity Ratio, by item}
\usage{
complexity_ratio(observations, key = "CODE")
}
\arguments{
\item{observations}{List of observations.}

\item{key}{Access key to the items in an observation.}
}
\value{
Vector of complexity ratios: for each item, the proportion of complex observations containing
it among all observations containing it.
}
\description{
For each item, compute the ratio between the number of complex observations containing the item and
the total number of observations containing the item. In other words, compute the complement of the
ratio between the number of times the item appears alone and the number of times the item appears.
}
\examples{
obs <- make_observations(oedb_sample, by = "ID", additional = c("CODE", "NAME"))
complexity_ratio(obs)

}
\seealso{
\code{\link{complexity_index}}, \code{\link{co_occurrence_matrix}},
\code{\link{get_all_items}}, \code{\link{get_obs_from_items}}, \code{\link{get_complex_obs}}, \code{\link{get_simple_obs}}.
}
\author{
Gauthier Magnin
}
