% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{compute_additional_rule_indicators,TransactionAnalyzer-method}
\alias{compute_additional_rule_indicators,TransactionAnalyzer-method}
\alias{compute_additional_rule_indicators}
\title{Computation of indicators relating to association rules}
\usage{
\S4method{compute_additional_rule_indicators}{TransactionAnalyzer}(object, rules, transactions = NULL)
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}

\item{rules}{Object of class \code{rules} containing the rules whose indicators are to be computed.}

\item{transactions}{Object of class \code{transactions} used to generate \code{rules}. Needed if the \code{rules}
were generated using the \code{arules} function \code{ruleInduction} and not \code{apriori}.}
}
\value{
Matrix in which each column is one indicator and rows match the association rules.
}
\description{
Compute indicators relating to association rules excluding the main ones (which are support,
confidence and lift). Compute the specificity, accuracy and added value of each given rule.
}
\details{
\loadmathjax
Considering the notation \mjeqn{\overline{X}}{not X} denoting any itemset not containing \eqn{X},
the computed indicators of an association rule of the form \mjeqn{X \rightarrow Y}{X -> Y} are:
\itemize{
\item The \strong{specificity}: confidence of the rule
\mjeqn{\overline{X} \rightarrow \overline{Y}}{not X -> not Y}.
\item The \strong{accuracy}: sum of the support of the itemset \mjeqn{X \cup Y}{X union Y} and the
support of \mjeqn{\overline{X \cup Y}}{not(X union Y)}.
\item The \strong{added value}: difference between the confidence of the rule \mjeqn{X \rightarrow Y}{X -> Y}
and the support of the consequent \eqn{Y}.
}

As a reminder, the support and the confidence of \mjeqn{X \rightarrow Y}{X -> Y} are defined as
follows.
\itemize{
\item The \strong{support} of the rule is the support of the itemset \mjeqn{X \cup Y}{X union Y}, i.e. the
proportion of transactions containing \mjeqn{X \cup Y}{X union Y} among all transactions.
\item The \strong{confidence} of the rule is the quotient of the support of \mjeqn{X \cup Y}{X union Y} and
the support of \eqn{X}, i.e. the number of transactions in which the rule is correct
relative to the number of transactions containing the antecedent \eqn{X}.
}
}
\note{
This method is not exported from the package.
}
\seealso{
\code{\link{extract_rules}}, \code{\link[arules:rules-class]{arules::rules}},
\code{\link[arules:transactions-class]{arules::transactions}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
