% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{compute_ksi_threshold,SpectralAnalyzer-method}
\alias{compute_ksi_threshold,SpectralAnalyzer-method}
\alias{compute_ksi_threshold}
\title{Computation of threshold ksi}
\usage{
\S4method{compute_ksi_threshold}{SpectralAnalyzer}(object, reporting_indexes)
}
\arguments{
\item{object}{SpectralAnalyzer class object.}

\item{reporting_indexes}{Reporting indexes associated with the patterns.}
}
\value{
Computed threshold.
}
\description{
Compute the number of patterns allowing to explain the main part of the reporting indexes.
}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
            \emph{The spectrosome of occupational health problems}. PLoS ONE 13(1): e0190196.
            \url{https://doi.org/10.1371/journal.pone.0190196}.
}
\seealso{
\code{\link{compute_reporting_indexes}}, \code{\link{compute_reporting_indexes_limits}},
         \code{\link{compute_ri_threshold}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
