% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{compute_pattern_distribution_in_nodes,SpectralAnalyzer-method}
\alias{compute_pattern_distribution_in_nodes,SpectralAnalyzer-method}
\alias{compute_pattern_distribution_in_nodes}
\title{Distribution of patterns among nodes}
\usage{
\S4method{compute_pattern_distribution_in_nodes}{SpectralAnalyzer}(object, patterns)
}
\arguments{
\item{object}{SpectralAnalyzer class object.}

\item{patterns}{Patterns whose distributions are to be calculated.}
}
\value{
\describe{
   \item{\code{weight_distribution}}{The distribution, for each pattern, of the weights of nodes in
                                     which they are included.}
   \item{\code{length_distribution}}{The distribution, for each pattern, of the lengths of nodes in
                                     which they are included.}
 }
}
\description{
Compute the distributions of the weights and lenghts of the nodes in which each pattern is included.
}
\author{
Gauthier Magnin
}
\keyword{internal}
