% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{compute_reporting_indexes,TransactionAnalyzer-method}
\alias{compute_reporting_indexes,TransactionAnalyzer-method}
\alias{compute_reporting_indexes}
\title{Computation of the Reporting Index (RI)}
\usage{
\S4method{compute_reporting_indexes}{TransactionAnalyzer}(object, patterns, t = NULL, period = Inf)
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}

\item{patterns}{Patterns whose reporting indexes are to be computed.}

\item{t}{Year of the end of the period, i.e. the date on which to characterize the patterns.
\code{NULL} specifies that the characterization must be done in relation to the last year covered
by the transactions.}

\item{period}{Time interval over which to compute the reporting indexes (number of years).
For example, if \code{t = 2019} and \code{period = 9} then the computation is made over the
period [2011 - 2019].

\code{Inf} specifies that the period considered covers an interval starting on the date of the
oldest transaction and ending in the year \code{t}.}
}
\value{
Vector containing the reporting index of each pattern contained in \code{patterns}.
}
\description{
Compute the reporting index of each pattern for a given period.
This index provides information on the proportion and importance of the occurrences of a pattern,
 taking into account the occurrences of the other patterns.
}
\details{
\loadmathjax
The reporting index of the pattern \mjseqn{p} is given by:
 \mjdeqn{RI_p(t_1,t_0) = \frac{\sum_{t = t_0}^{t_1} F_{p,t}}{\sum_{q \in P} \sum_{t = t_0}^{t_1} F_{q,t}}}{RI_p(t_1,t_0) = sum F_pt from t = t_0 to t_1 / sum F_qt for q in P and from t = t_0 to t_1}
where \mjseqn{P} is the set of patterns, \mjeqn{F_{p,t}}{F_pt} is the frequency of the pattern
 \mjseqn{p} in the transactions of the year \mjseqn{t}, \mjseqn{t_0} and \mjseqn{t_1} are the first
 and last years defining the period on which to compute the reporting index.
}
\note{
This method is not exported from the package.
}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
            The spectrosome of occupational health problems. \emph{PLoS ONE} 13(1): e0190196.
            \url{https://doi.org/10.1371/journal.pone.0190196}.
}
\seealso{
\code{\link{compute_reporting_indexes_limits}}, \code{\link{compute_xi_threshold}},
         \code{\link{compute_ri_threshold}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
