% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{compute_reporting_indexes_limits,TransactionAnalyzer-method}
\alias{compute_reporting_indexes_limits,TransactionAnalyzer-method}
\alias{compute_reporting_indexes_limits}
\title{Computation of RI at temporal limits}
\usage{
\S4method{compute_reporting_indexes_limits}{TransactionAnalyzer}(
  object,
  patterns,
  end = NULL,
  overall_period = Inf,
  recent_period = object["status_limit"]
)
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}

\item{patterns}{Patterns whose limits are to be computed.}

\item{end}{Year of end of the periods, i.e. the date on which to characterize the patterns.
\code{NULL} specifies that the characterization must be done in relation to the last year covered
by the transactions.}

\item{overall_period}{Time interval over which to compute the overall reporting indexes (number of
years). For example, if \code{end = 2019} and \code{overall_period = 9} then the computation is made
over the period [2011 - 2019].

\code{Inf} specifies that the period considered covers the interval starting in the year of the
oldest transaction and ending in the year \code{end}.}

\item{recent_period}{Shorter time interval over which to compute the second reporting indexes
(number of years). For example, if \code{end = 2019} and \code{recent_period = 2} then the
computation is made over the period [2018 - 2019].}
}
\value{
Matrix containing the two reporting indexes of each pattern contained in \code{patterns}.
}
\description{
Compute the reporting indexes at the temporal limits used to characterize the patterns.
The first ones correspond to the reporting indexes computed over the period defined by the arguments
 \code{end} and \code{overall_period}.
The second ones correspond to the reporting indexes computed over the period defined by the arguments
 \code{end} and \code{recent_period}.
}
\details{
\loadmathjax
The reporting index of the pattern \mjseqn{p} at the temporal limit \mjseqn{l} is given by:
 \mjdeqn{RI_{l,p} = RI_p(t_1, t_1 - l + 1)}{RI_lp = RI_p(t_1, t_1 - l + 1)}
where \mjseqn{t_1} is the year of end of the period to consider and \mjseqn{RI_p(t_1,t_0)}
 is the reporting index of the pattern \mjseqn{p} given by:
 \mjdeqn{RI_p(t_1,t_0) = \frac{\sum_{t = t_0}^{t_1} F_{p,t}}{\sum_{q \in P} \sum_{t = t_0}^{t_1} F_{q,t}}}{RI_p(t_1,t_0) = sum F_pt from t = t_0 to t_1 / sum F_qt for q in P and from t = t_0 to t_1}
where \mjseqn{P} is the set of patterns, \mjeqn{F_{p,t}}{F_pt} is the frequency of the pattern
 \mjseqn{p} in the transactions of the year \mjseqn{t}, \mjseqn{t_0} and \mjseqn{t_1} are the first and
 last years defining the period on which to compute the reporting index.
}
\note{
This method is not exported from the package.
}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
            The spectrosome of occupational health problems. \emph{PLoS ONE} 13(1): e0190196.
            \url{https://doi.org/10.1371/journal.pone.0190196}.
}
\seealso{
\code{\link{compute_reporting_indexes}}, \code{\link{compute_xi_threshold}},
         \code{\link{compute_ri_threshold}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
