% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{compute_ri_threshold,SpectralAnalyzer-method}
\alias{compute_ri_threshold,SpectralAnalyzer-method}
\alias{compute_ri_threshold}
\title{RI threshold computation}
\usage{
\S4method{compute_ri_threshold}{SpectralAnalyzer}(object, reporting_indexes, ksi = NULL)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{reporting_indexes}{Reporting indexes associated with the patterns.}

\item{ksi}{Number of patterns to consider before setting the RI threshold.
Is computed if \code{NULL}.}
}
\value{
Computed threshold.
}
\description{
Compute the limit value separating two dynamic profiles with respect to the reporting indexes.
The patterns are ordered in descending order of their reporting index value and separated by
 the threshold \code{ksi}.
}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
            The spectrosome of occupational health problems. \emph{PLoS ONE} 13(1): e0190196.
            \url{https://doi.org/10.1371/journal.pone.0190196}.
}
\seealso{
\code{\link{compute_reporting_indexes}}, \code{\link{compute_reporting_indexes_limits}},
         \code{\link{compute_ksi_threshold}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
