% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_gunits}
\alias{convert_gunits}
\title{Graphical units converter}
\usage{
convert_gunits(measures, from, to = from, dim = "width", rotation = FALSE)
}
\arguments{
\item{measures}{Values to convert.}

\item{from}{Graphical unit in which the measures are. One of \code{"user"}, \code{"inches"}, \code{"figure"}.}

\item{to}{Graphical unit in which to convert the measures. One of \code{"user"}, \code{"inches"}, \code{"figure"}.}

\item{dim}{Dimension on which measures were made (\verb{"width} or \verb{"height}).}

\item{rotation}{Logical indicating whether to consider a rotation of the measured elements.}
}
\value{
Converted measure values.
}
\description{
Convert measures between graphical units considering the active graphics device.
}
\details{
A measure in inches may not correspond to the same value in user coordinate unit along the X axis
as along the Y axis.
For example, a measure of \eqn{1} inch corresponding to the width of a text that will be rotated
\eqn{90} degrees may correspond to a width of \eqn{0.09} in user unit before the rotation but a
height of \eqn{0.05} in user unit after the rotation.
The argument \code{rotation} allows to take this into account.

The \code{"user"} unit is only available after \code{\link[graphics:frame]{graphics::plot.new}} has been called.

The results of the functions \code{\link[graphics:strwidth]{graphics::strwidth}} and \code{\link[graphics:strwidth]{graphics::strheight}} applied to the
character \code{"1"} and each different unit serve as reference values for conversions.
}
\author{
Gauthier Magnin
}
\keyword{internal}
