% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{create_spectrum_chart,SpectralAnalyzer-method}
\alias{create_spectrum_chart,SpectralAnalyzer-method}
\title{Dessine un graphique de type spectre.}
\usage{
\S4method{create_spectrum_chart}{SpectralAnalyzer}(
  object,
  patterns_characteristics,
  weights_by_node_type,
  title = "Spectrum of patterns"
)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{patterns_characteristics}{Ensemble des caractéristiques des motifs dont le spectre est à tracer.}

\item{weights_by_node_type}{Data.frame contenant, pour chaque motif, son poids dans des noeuds complexes et son poids dans des noeuds simples.}

\item{title}{Titre du graphique.}
}
\description{
Dessine un graphique de type spectre.
}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
            \emph{The spectrosome of occupational health problems}. PLoS ONE 13(1): e0190196.
            \url{https://doi.org/10.1371/journal.pone.0190196}.
}
\author{
Delphine Bosson-Rieutort

Gauthier Magnin
}
