% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{degree,SpectralAnalyzer-method}
\alias{degree,SpectralAnalyzer-method}
\alias{degree}
\title{Degré d'un sommet}
\usage{
\S4method{degree}{SpectralAnalyzer}(object, ID, links)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{ID}{Identifiant du sommet (nœud ou motif) dont le degré est à calculer.}

\item{links}{Data frame des liens (ou arêtes) d'un graphe de type spectrosome.}
}
\value{
Degré du sommet.
}
\description{
Calcule le degré d'un sommet dans un graphe, c'est-à-dire, le nombre de sommets
 auxquels il est adjacent.
}
\author{
Gauthier Magnin
}
