% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{degree,TransactionAnalyzer-method}
\alias{degree,TransactionAnalyzer-method}
\alias{degree}
\title{Degree of a vertex}
\usage{
\S4method{degree}{TransactionAnalyzer}(object, ID, links)
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}

\item{ID}{Identifier of the vertex (node or pattern) whose degree is to be computed.}

\item{links}{Data frame of the links (or edges) of a spectrosome graph.}
}
\value{
Degree of the vertex.
}
\description{
Compute the degree of a vertex in a graph, i.e. the number of vertices to which it is adjacent.
}
\examples{
spectrosome <- spectrosome_chart(TA_instance, TA_instance["patterns"][1:15, ])
degree(TA_instance, 7, spectrosome[["edges"]])

}
\author{
Gauthier Magnin
}
