% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dhms}
\alias{dhms}
\title{Décomposition d'un temps}
\usage{
dhms(t)
}
\arguments{
\item{t}{Numeric. Temps à convertir, en secondes.}
}
\value{
Chaîne de caractère correspondant au temps converti.
}
\description{
Convertit en une chaîne de caractères une période définie en secondes.
Exemples du format de la chaîne retournée : 1d 02h 25m 47s ; 01m 05s ; 19s.
}
\references{
Stack Overflow topic :
 \href{https://stackoverflow.com/questions/27312292/convert-seconds-to-days-hoursminutesseconds}{Convert seconds to days:hours:minutes:seconds}
}
\keyword{internal}
