% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{export,TransactionAnalyzer-method}
\alias{export,TransactionAnalyzer-method}
\alias{export}
\title{Save nodes, patterns or association rules}
\usage{
\S4method{export}{TransactionAnalyzer}(object, nporc, ...)
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}

\item{nporc}{Data frame of \strong{n}odes, \strong{p}atterns \strong{o}r \strong{r}ules and their \strong{c}haracteristics.}

\item{...}{Further arguments to the function \code{\link[utils:write.table]{utils::write.csv2}}.}
}
\description{
Write to a CSV file a data frame of nodes, patterns or association rules.
}
\examples{
export(TA_instance, TA_instance["nodes"],
       file = "nodes.csv")
export(TA_instance, TA_instance["patterns"][1:15, ],
       file = "patterns.csv")

spectrosome <- spectrosome_chart(TA_instance, "patterns")
export(TA_instance, spectrosome[["vertices"]],
       file = "spectrosome_vertices.csv", row.names = FALSE)

rules <- extract_rules(TA_instance, from = "transactions")
export(TA_instance, rules,
       file = "rules.csv", row.names = FALSE)

}
\seealso{
\code{\link[utils:write.table]{utils::write.csv2}}.
}
\author{
Gauthier Magnin
}
