% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_set.R, R/transaction_analyzer.R
\name{export}
\alias{export}
\alias{export,TransactionSet-method}
\alias{export,TransactionSet}
\alias{export,TransactionAnalyzer-method}
\alias{export,TransactionAnalyzer}
\title{Save transactions, nodes, patterns or association rules}
\usage{
\S4method{export}{TransactionSet}(object, ...)

\S4method{export}{TransactionAnalyzer}(object, nporc, ...)
}
\arguments{
\item{object}{S4 object of class \code{TransactionSet} or of class \code{TransactionAnalyzer}.}

\item{...}{Further arguments to the function \code{\link[utils:write.table]{utils::write.csv2}}.}

\item{nporc}{Data frame of \strong{n}odes, \strong{p}atterns \strong{o}r \strong{r}ules and their \strong{c}haracteristics.}
}
\description{
Write in CSV format a data frame of transactions, nodes, patterns or association rules.
}
\examples{
## Saving a set of transactions
export(TS_instance file = "transactions.csv")

## Saving nodes, patterns or association rules
export(TA_instance, TA_instance["nodes"],
       file = "nodes.csv")
export(TA_instance, TA_instance["patterns"][1:15, ],
       file = "patterns.csv")

spectrosome <- spectrosome_chart(TA_instance, "patterns")
export(TA_instance, spectrosome[["vertices"]],
       file = "spectrosome_vertices.csv", row.names = FALSE)

rules <- extract_rules(TA_instance, from = "transactions")
export(TA_instance, rules,
       file = "rules.csv", row.names = FALSE)

}
\seealso{
\code{\link[utils:write.table]{utils::write.csv2}}.
}
\author{
Gauthier Magnin
}
