% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations_maker.R
\name{extract_information}
\alias{extract_information}
\title{Extrait les informations associées aux observations qui contiennent l'ensemble d'items recherché.}
\usage{
extract_information(observations, items, info_names)
}
\arguments{
\item{observations}{Liste des observations sur lesquels effectuer la recherche.}

\item{items}{Éléments recherchés (un ou plusieurs).}

\item{info_names}{Noms des informations à extraire des observations.}
}
\value{
Vecteur (si un seul type d'information à extraire) ou matrice des informations correspondant à la recherche.
}
\description{
Extrait les informations associées aux observations qui contiennent l'ensemble d'items recherché.
}
\examples{
extract_information(observations, 27, c("METIER", "TACHE"))
}
\author{
Gauthier Magnin
}
