% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{extract_links,SpectralAnalyzer-method}
\alias{extract_links,SpectralAnalyzer-method}
\alias{extract_links}
\title{Extraction of links}
\usage{
\S4method{extract_links}{SpectralAnalyzer}(object, entities, characteristics)
}
\arguments{
\item{object}{SpectralAnalyzer class object.}

\item{entities}{Type of entities to search for links.
One of \code{"nodes"}, \code{"patterns"}.}

\item{characteristics}{Data frame of the characteristics of the nodes or patterns whose links are
to be sought.}
}
\value{
Data frame associating the linked nodes or linked patterns.
}
\description{
Extract from the links those corresponding to the desired nodes or patterns.
}
\details{
If among the nodes or patterns some become isolated because the other entities to which they
 are normally linked are not part of \code{characteristics}, these nodes or patterns are placed
 at the end of the return data frame.
These possible \code{n} additional lines are numbered \code{"A1"..."An"}.
}
\author{
Gauthier Magnin
}
