% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{extract_nodes_from_items,SpectralAnalyzer-method}
\alias{extract_nodes_from_items,SpectralAnalyzer-method}
\title{Extrait les noeuds contenant un ou plusieurs items recherchés.}
\usage{
\S4method{extract_nodes_from_items}{SpectralAnalyzer}(object, nodes_characteristics, items, target = "all")
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{nodes_characteristics}{Data.frame des caractéristiques des noeuds.}

\item{items}{Éléments recherchés (un ou plusieurs).}

\item{target}{Condition pour qu'un noeud soit extrait. Choix parmi \code{c("all", "any")}.
\code{"all"} exprime que l'intégralité des éléments recherchés doivent faire partie d'un noeud pour que ce noeud soit extrait.
\code{"any"} exprime qu'au moins un des éléments recherchés doit faire partie d'un noeud pour que ce noeud soit extrait.}
}
\value{
Sous-ensemble de la data.frame fournie en argument pour les noeuds correspondant au critère de recherche.
}
\description{
Extrait les noeuds contenant un ou plusieurs items recherchés.
}
\author{
Gauthier Magnin
}
