% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations_maker.R
\name{extract_observations_from_information}
\alias{extract_observations_from_information}
\title{Extrait les observations correspondant à un ou plusieurs critères de recherche.}
\usage{
extract_observations_from_information(observations, info_names, info_values)
}
\arguments{
\item{observations}{Liste des observations sur lesquels effectuer la recherche.}

\item{info_names}{Noms des informations contenues dans les observations, sur lesquels la recherche doit être faite.}

\item{info_values}{Valeurs recherchées pour les informations décrite par l'argument \code{info_names}.}
}
\value{
Sous-ensemble de la liste d'observations correspondant aux critères de recherche.
}
\description{
Extrait les observations correspondant à un ou plusieurs critères de recherche.
}
\examples{
extract_observations_from_information(observations, c("METIER", "TACHE"), c(44143009, "A3410"))
}
\author{
Gauthier Magnin
}
