% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations_maker.R
\name{extract_observations_from_items}
\alias{extract_observations_from_items}
\title{Recherche d'observations}
\usage{
extract_observations_from_items(observations, items, target = "all")
}
\arguments{
\item{observations}{Liste des observations sur lesquels effectuer la recherche.}

\item{items}{Élément(s) recherché(s).}

\item{target}{Condition pour qu'une observation soit extraite. Choix parmi \code{"all"}, \code{"any"}.
\describe{
 \item{\code{"all"}}{L'intégralité des items recherchés doivent faire partie d'une observation pour que cette observation soit extraite.}
 \item{\code{"any"}}{Au moins un des items recherchés doit faire partie d'une observation pour que cette observation soit extraite.}
}}
}
\value{
Sous-ensemble de la liste d'observations correspondant aux critères de recherche.
}
\description{
Extrait les observations correspondant à un ou plusieurs items recherchés.
}
\author{
Gauthier Magnin
}
