% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{extract_patterns_from_category,SpectralAnalyzer-method}
\alias{extract_patterns_from_category,SpectralAnalyzer-method}
\alias{extract_patterns_from_category}
\title{Recherche de motifs par catégorie}
\usage{
\S4method{extract_patterns_from_category}{SpectralAnalyzer}(object, category, value, target)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{category}{Nom ou numéro de la catégorie sur laquelle effectuer la recherche
(numérotation selon l'ordre des colonnes de \code{object["items_categories"]}).}

\item{value}{Valeur recherchée pour la catégorie spécifiée par l'argument \code{category}.}

\item{target}{Condition pour qu'un motif soit extrait. Choix parmi \code{"vertices"}, \code{"edges"}.
\describe{
  \item{\code{"vertices"}}{Recherche des motifs contenant un item associé à la valeur de catégorie recherchée.}
  \item{\code{"edges"}}{Recherche des motifs générant un lien correspondant à la valeur de catégorie recherchée.}
}}
}
\value{
Data frame des motifs correspondant aux critères de recherche.
}
\description{
Extrait les motifs correspondant à une valeur de catégorie recherchée
}
\seealso{
\code{\link{extract_patterns_from_items}}, \code{\link{extract_patterns_from_characteristic}},
         \code{\link{extract_patterns_from_status}}.
}
\author{
Gauthier Magnin
}
