% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{extract_patterns_from_status,SpectralAnalyzer-method}
\alias{extract_patterns_from_status,SpectralAnalyzer-method}
\title{Extrait les motifs dont le statut correspond à une valeur recherchée.}
\usage{
\S4method{extract_patterns_from_status}{SpectralAnalyzer}(
  object,
  patterns_characteristics,
  value,
  condition = "EQ"
)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{patterns_characteristics}{Data.frame des caractéristiques des motifs.}

\item{value}{Valeur de statut recherché (une ou plusieurs).}

\item{condition}{Condition de la recherche.
Choix parmi \code{c("EQ", "NE")}.
\code{"EQ"} : Le statut du motif doit être l'une des valeurs recherchées (Equal).
\code{"NE"} : Le statut du motif doit être différente des valeurs recherchées (Not Equal).}
}
\value{
Sous-ensemble de la data.frame fournie en argument pour les motifs satisfaisant le critère de recherche.
}
\description{
Extrait les motifs dont le statut correspond à une valeur recherchée.
}
\author{
Gauthier Magnin
}
