% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{extract_rules,SpectralAnalyzer-method}
\alias{extract_rules,SpectralAnalyzer-method}
\alias{extract_rules}
\title{Rules extraction}
\usage{
\S4method{extract_rules}{SpectralAnalyzer}(object, from, pruning = FALSE, as_sets = FALSE, ...)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{from}{Character or list of item sets for which to extract the association rules.
\itemize{
  \item{If \code{"observations"}, look for all rules within the observations saved in \code{object}
        or for rules with specific items.}
  \item{If \code{"patterns"}, look for rules whose union of the antecedent and the consequent form
        an entire pattern among those of \code{object["patterns"]$pattern}.}
  \item{Otherwise, a list of item sets defined the same way as \code{object["patterns"]$pattern}.
        The search is then done the same way as for \code{"patterns"}.}
}}

\item{pruning}{If \code{TRUE}, remove redundant rules.}

\item{as_sets}{If \code{FALSE}, antecedents and consequents of the rules will be character vectors.
If \code{TRUE}, they will be factors written in mathematical notation (i.e. set notation).}

\item{...}{Additional arguments to configure the extraction. See Details.}
}
\value{
Data frame containing the extracted rules and their characteristics.
 If \code{from} is not \code{"observations"}, the column \code{"itemset"} refers to the index of the
 item set from which the rule was generated, in the list of patterns (if \code{from = "patterns"})
 or the given list (otherwise).
}
\description{
Extract association rules from the observations. Can be used to find all rules, rules relating to
 patterns (or other specific item sets) or relating to specific items.
}
\details{
Only creates rules with one item in the consequent. The reason is detailed on
 \href{https://borgelt.net/doc/apriori/apriori.html#conseq}{this web page} by the author of the
 algorithms used. Here are some extracts: \cr
 \emph{"There are usually already (many) more association rules than item sets if only a single
 item is allowed in the consequents."} \cr
 \emph{"Multiple items in the consequents of association rules therefore come at a considerable
 cost."} \cr
 \emph{"There is no true benefit."}

A rule is redundant if a more general rule with the same or higher confidence exists. A rule is more
 general if it has the same consequent but one or more items removed from the antecedent.

If \code{from = "observations"}, additional arguments are \code{parameter}, \code{appearance} and
 \code{control} of function \code{\link[arules:apriori]{apriori}} from the package \code{arules}.
 These arguments allow to specify minimum support (default \code{0.1}), minimum confidence (default
 \code{0.8}), minimum length (default \code{1}), maximum length (default \code{10}), specific items
 in antecedent or consequent, and some operating parameters of the rule extraction algorithm.

If \code{from} is \code{"patterns"} or a list, additional arguments are \code{confidence} and
 \code{control} of function \code{\link[arules:ruleInduction]{ruleInduction}} from the package
 \code{arules}. These arguments allow to specify minmum confidence (default \code{0.8}) and some
 operating parameters of the rule extraction algorithm.
}
\examples{
## Basic rule extraction
rules_1 <- extract_rules(SA_instance, from = "observations")
rules_2 <- extract_rules(SA_instance, from = "patterns")
rules_3 <- extract_rules(SA_instance, from = list(c("931", "3180"),
                                                  c("25", "192", "328")))

## Rule extraction with conditions on the antecedent and the consequent
params <- list(supp = 0.001, conf = 0.5, maxlen = 2)
rules_4 <- extract_rules(SA_instance, from = "observations",
                         parameter = params,
                         appearance = list(rhs = "328"))
rules_5 <- extract_rules(SA_instance, from = "observations",
                         parameter = params,
                         appearance = list(lhs = "497"))
rules_6 <- extract_rules(SA_instance, from = "observations",
                         parameter = list(supp = 0.001, conf = 0,
                                          minlen = 2, maxlen = 2),
                         appearance = list(lhs = "328", rhs = "3180"))

}
\author{
Gauthier Magnin
}
