% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_helper.R
\name{fig_in_usr_coords}
\alias{fig_in_usr_coords}
\title{User coordinates of the figure region}
\usage{
fig_in_usr_coords(n = 1:4)
}
\arguments{
\item{n}{Numbers of the values to compute between \code{1} and \code{4}.
The same numbering as for \code{par("fig")}, \code{par("plt")} and \code{par("usr")}: \code{x0}, \code{x1}, \code{y0}, \code{y1}.}
}
\value{
User coordinates of the figure region.
}
\description{
Compute the coordinates of the figure region in user coordinate units.
Only available after \code{\link[graphics:frame]{graphics::plot.new}} has been called.
}
\seealso{
\code{\link{convert_gunits}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
