% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{first_characters}
\alias{first_characters}
\title{Find the first characters}
\usage{
first_characters(x, sep = " _-")
}
\arguments{
\item{x}{Character string (one or more) from which the first characters are to be extracted.}

\item{sep}{Character string each of which is a word separator.}
}
\value{
Character vector of size equal to that of \code{x} containing the first concatenated characters
of all words.
}
\description{
Search for the first character of each word and concatenate them.
}
\seealso{
\code{\link{cap}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
