% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{frequency_by_complexity,TransactionAnalyzer-method}
\alias{frequency_by_complexity,TransactionAnalyzer-method}
\alias{frequency_by_complexity}
\title{Pattern frequency by transaction complexity}
\usage{
\S4method{frequency_by_complexity}{TransactionAnalyzer}(object, patterns)
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}

\item{patterns}{Patterns whose frequencies according to the complexity of the transactions containing
them are to be computed. Any subset of \code{object["patterns"]$pattern}.\cr
\code{"patterns"} and \code{"p"} are special values for \code{object["patterns"]$pattern}.}
}
\value{
Two-column matrix containing, for each pattern, its frequency related to complex transactions
and its frequency related to simple transactions.
}
\description{
For each pattern, compute its frequency related to complex transactions (i.e. the number of
transactions containing more than one item and containing the pattern) and its frequency related to
simple transactions (i.e. the number of transactions containing only one item and containing the
pattern).
}
\examples{
frequency_by_complexity(TA_instance, "patterns")
frequency_by_complexity(TA_instance, TA_instance["patterns"]$pattern[1:15])

}
\seealso{
\code{\link{get_complex_trx}}, \code{\link{get_simple_trx}}, \code{\link{get_complexes}}
}
\author{
Gauthier Magnin
}
