% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_set.R
\name{get_all_items,ObservationSet-method}
\alias{get_all_items,ObservationSet-method}
\alias{get_all_items}
\title{Search all items}
\usage{
\S4method{get_all_items}{ObservationSet}(object)
}
\arguments{
\item{object}{S4 object of class \code{ObservationSet}.}
}
\value{
Vector of all unique items.
}
\description{
Extract all the items contained in the observations.
}
\examples{
get_all_items(OS_instance)

}
\seealso{
\code{\link{get_itemsets}}.
}
\author{
Gauthier Magnin
}
