% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_set.R
\name{get_all_items,TransactionSet-method}
\alias{get_all_items,TransactionSet-method}
\alias{get_all_items}
\title{Search all items}
\usage{
\S4method{get_all_items}{TransactionSet}(object)
}
\arguments{
\item{object}{S4 object of class \code{TransactionSet}.}
}
\value{
Vector of all unique items.
}
\description{
Extract all the items contained in the transactions.
}
\examples{
get_all_items(TS_instance)

}
\seealso{
\code{\link{get_itemsets}}.
}
\author{
Gauthier Magnin
}
