% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_set.R
\name{get_all_years,TransactionSet-method}
\alias{get_all_years,TransactionSet-method}
\alias{get_all_years}
\title{Search all years}
\usage{
\S4method{get_all_years}{TransactionSet}(object)
}
\arguments{
\item{object}{S4 object of class \code{TransactionSet}.}
}
\value{
Vector of all unique temporal data.
}
\description{
Extract all the years associated with the transactions.
}
\examples{
get_all_years(TS_instance)

}
\seealso{
\code{\link{get_all_items}}, \code{\link{get_itemsets}}.
}
\author{
Gauthier Magnin
}
