% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_set.R
\name{get_complex_obs,ObservationSet-method}
\alias{get_complex_obs,ObservationSet-method}
\alias{get_complex_obs}
\title{Search for complex observations}
\usage{
\S4method{get_complex_obs}{ObservationSet}(object)
}
\arguments{
\item{object}{S4 object of class \code{ObservationSet}.}
}
\value{
S4 object of class \code{ObservationSet} containing the subset of observations that contain more
than one item.
}
\description{
Extract the observations containing more than one item.
}
\examples{
get_complex_obs(OS_instance)

}
\seealso{
\code{\link{get_simple_obs}}, \code{\link{get_obs_from_items}}, \code{\link{get_obs_from_info}}.
}
\author{
Gauthier Magnin
}
