% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_maker.R
\name{get_complex_obs}
\alias{get_complex_obs}
\title{Search for complex observations}
\usage{
get_complex_obs(observations, key = "CODE")
}
\arguments{
\item{observations}{List of observations on which to do the search.}

\item{key}{Access key to the items in an observation.}
}
\value{
Subset of the list of observations that contain more than one item.
}
\description{
Extract the observations containing more than one item.
}
\examples{
obs <- make_observations(oedb_sample, by = "ID", additional = c("CODE", "NAME"))
get_complex_obs(obs)

}
\seealso{
\code{\link{get_simple_obs}}, \code{\link{get_obs_from_items}}, \code{\link{get_obs_from_info}}.
}
\author{
Gauthier Magnin
}
