% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{get_complexes,SpectralAnalyzer-method}
\alias{get_complexes,SpectralAnalyzer-method}
\alias{get_complexes}
\title{Search for complex nodes or patterns}
\usage{
\S4method{get_complexes}{SpectralAnalyzer}(
  object,
  nopc,
  category = NULL,
  condition = NULL,
  min_nb_values = 2
)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{nopc}{Data frame of \strong{n}odes \strong{o}r \strong{p}atterns and their
\strong{c}haracteristics. Nodes or patterns whose complexes are to be sought. Any subset of
\code{object["nodes"]} or \code{object["patterns"]}.\cr
\code{"nodes"}, \code{"n"}, \code{"patterns"} and \code{"p"} are specific values for
\code{object["nodes"]} and \code{object["patterns"]}.}

\item{category}{Name or number of the category on which to search (numbering according to the order
of the columns of \code{object["items_categories"]}).}

\item{condition}{Condition for a node or a pattern to be extracted.
One of \code{"items"}, \code{"links"}, \code{"vertices"}, \code{"edges"}.
\describe{
  \item{\code{"items"}, \code{"vertices"}}{Search for nodes or patterns associated (via their items)
        with several values of the category \code{category}.}
  \item{\code{"links"}, \code{"edges"}}{Search for nodes or patterns generating links corresponding
        with several values of the category \code{category}.}
}}

\item{min_nb_values}{Minimum number of different values of the category \code{category} a node, a
pattern or a link must have to extract the related entity.}
}
\value{
Subset of the data frame that corresponds to the complex entities sought.
}
\description{
Extract from the given nodes or patterns those which are complex by the number of items they
 contain or by the number of values with which they are associated, with regard to one category.
}
\details{
If \code{category} and \code{condition} are \code{NULL}, entities with more than \code{min_nb_values}
 items are sought. Otherwise, the search is related to the \code{category} (see \code{condition}).
}
\examples{
get_complexes(SA_instance, "patterns")
get_complexes(SA_instance, SA_instance["patterns"][1:15, ],
              category = "family", condition = "items")
get_complexes(SA_instance, SA_instance["patterns"][1:15, ],
              category = 1, condition = "links")

}
\seealso{
\code{\link{get_isolates}}, \code{\link{get_non_isolates}}, \code{\link{get_links}}.
}
\author{
Gauthier Magnin
}
