% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_set.R
\name{get_info_from_items,ObservationSet-method}
\alias{get_info_from_items,ObservationSet-method}
\alias{get_info_from_items}
\title{Search for information by item}
\usage{
\S4method{get_info_from_items}{ObservationSet}(object, items, info_names, presence = "all")
}
\arguments{
\item{object}{\code{ObservationSet} class object.}

\item{items}{Sought items.}

\item{info_names}{Names of information to extract from observations.}

\item{presence}{Item presence condition for information to be extracted from an observation.
One of \code{"all"}, \code{"any"}, \code{"exact"}.
\describe{
\item{\code{"all"}}{All the sought items must be part of an observation for its information to be
extracted.}
\item{\code{"any"}}{At least one of the sought items must be part of an observation for its information
to be extracted.}
\item{\code{"exact"}}{The item set contained in an observation must be exactly the same as the sought
item set for this observation to be extracted.}
}}
}
\value{
Vector or list of information corresponding to the search.
Vector if only one type of information is to be extracted. List otherwise.
}
\description{
Retrieve information associated with observations that contain a set of sought items.
}
\examples{
get_info_from_items(OS_instance, items = 3146,
                                 info_names = c("JOB.TITLE", "JOB.TASK"))

}
\seealso{
\code{\link{get_items_from_info}}, \code{\link{get_all_items}}.
}
\author{
Gauthier Magnin
}
