% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_maker.R
\name{get_info_from_items}
\alias{get_info_from_items}
\title{Recherche d'informations par item}
\usage{
get_info_from_items(
  observations,
  items,
  info_names,
  target = "all",
  key = "CODE"
)
}
\arguments{
\item{observations}{Liste des observations sur lesquels effectuer la recherche.}

\item{items}{Élément·s recherché·s.}

\item{info_names}{Noms des informations à extraire des observations.}

\item{target}{Condition pour qu'une information soit extraite d'une observation.
Choix parmi \code{"all"}, \code{"any"}.
 \describe{
  \item{\code{"all"}}{L'intégralité des items recherchés doivent faire partie d'une observation
                      pour que ses informations soient extraites.}
  \item{\code{"any"}}{Au moins un des items recherchés doit faire partie d'une observation pour
                      que ses informations soient extraites.}
 }}

\item{key}{Clé d'accès aux items de chaque observation.}
}
\value{
Vecteur ou liste des informations correspondant à la recherche.
 Vecteur si un seul type d'information est à extraire. Liste sinon.
}
\description{
Extrait les informations associées aux observations qui contiennent un ensemble d'items recherchés.
}
\seealso{
\code{\link{get_items_from_info}}.
}
\author{
Gauthier Magnin
}
