% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{get_isolates,SpectralAnalyzer-method}
\alias{get_isolates,SpectralAnalyzer-method}
\alias{get_isolates}
\title{Search for isolated nodes or patterns}
\usage{
\S4method{get_isolates}{SpectralAnalyzer}(object, entities, characteristics)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{entities}{Type of entities to search for isolated.
One of \code{"nodes"}, \code{"patterns"}.}

\item{characteristics}{Data frame of the characteristics of the nodes or patterns whose isolated
are to be sought.}
}
\value{
Subset of the data frame that corresponds to isolated entities.
}
\description{
Extract from the given nodes or patterns those which are isolated.
}
\examples{
get_isolates(SA_instance, "patterns", SA_instance["patterns"])
get_isolates(SA_instance, "patterns", SA_instance["patterns"][1:10, ])

}
\seealso{
\code{\link{get_non_isolates}}, \code{\link{get_complexes}}, \code{\link{get_links}}.
}
\author{
Gauthier Magnin
}
