% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{get_item_colors,TransactionAnalyzer-method}
\alias{get_item_colors,TransactionAnalyzer-method}
\alias{get_item_colors}
\title{Get item colors}
\usage{
\S4method{get_item_colors}{TransactionAnalyzer}(object, category, items = object["items"])
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}

\item{category}{Name or number of the category whose associated colors are to be found
(numbering according to the order of the columns of \code{object["items_categories"]}).}

\item{items}{Items for which to find the colors associated with their values in the given category.
Any subset of \code{object["items"]}.

\code{"items"} and \code{"i"} are special values for \code{object["items"]}.}
}
\value{
Colors associated with the values of the category \code{category} that are
associated with the \code{items}.
}
\description{
Find the colors associated with the values of a category that are associated with items.
}
\examples{
get_item_colors(TA_instance, category = "family", items = c(19, 25, 27))

}
\author{
Gauthier Magnin
}
