% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{get_item_names,SpectralAnalyzer-method}
\alias{get_item_names,SpectralAnalyzer-method}
\alias{get_item_names}
\title{Get item names}
\usage{
\S4method{get_item_names}{SpectralAnalyzer}(object, items)
}
\arguments{
\item{object}{S4 object of class \code{SpectralAnalyzer}.}

\item{items}{Vector of items: unnamed subset of \code{object["items"]}.}
}
\value{
Names of the \code{items} in \code{object["items"]}.
}
\description{
Find the names associated with items given their identification codes.
}
\seealso{
\code{\link[=get_items,SpectralAnalyzer]{get_items}}, \code{\link{has_item_names}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
