% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{get_item_names,TransactionAnalyzer-method}
\alias{get_item_names,TransactionAnalyzer-method}
\alias{get_item_names}
\title{Get item names}
\usage{
\S4method{get_item_names}{TransactionAnalyzer}(object, items)
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}

\item{items}{Vector of items: unnamed subset of \code{object["items"]}.}
}
\value{
Names of the \code{items} in \code{object["items"]}.
}
\description{
Find the names associated with items given their identification codes.
}
\examples{
get_item_names(TA_instance, items = c("19", "163", "1603"))

}
\author{
Gauthier Magnin
}
