% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{get_items,SpectralAnalyzer-method}
\alias{get_items,SpectralAnalyzer-method}
\alias{get_items}
\title{Get items}
\usage{
\S4method{get_items}{SpectralAnalyzer}(object, items)
}
\arguments{
\item{object}{S4 object of class \code{SpectralAnalyzer}.}

\item{items}{Vector of items or one of the following character value: \code{"items"}, \code{"i"}.}
}
\value{
Named vector of items corresponding to the arguments.
}
\description{
Find and return the vector or subset of the vector corresponding to the items of the object of class
\code{SpectralAnalyzer}, or return the given vector.
}
\details{
If \code{items} is a named vector corresponding to a subset of \code{object["items"]}, it is returned.

If \code{items} is a vector corresponding to an unnamed subset of \code{object["items"]}, the corresponding
named subset of \code{object["items"]} is returned.

If \code{items} is a character value equal to \code{"items"} or \code{"i"}, \code{object["items"]} is returned.
}
\seealso{
\code{\link{get_item_names}}, \code{\link{get_nopc}}, \code{\link{which_entities}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
