% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_set.R
\name{get_items,TransactionSet-method}
\alias{get_items,TransactionSet-method}
\alias{get_items}
\alias{get_items,TransactionSet}
\title{Get items}
\usage{
\S4method{get_items}{TransactionSet}(object, items)
}
\arguments{
\item{object}{S4 object of class \code{TransactionSet}.}

\item{items}{Vector of items or one of the following character value: \code{"items"}, \code{"i"}.}
}
\value{
Vector of items corresponding to the arguments.
}
\description{
Find and return the vector corresponding to the items of the object of class \code{TransactionSet}
or return the given vector.
}
\details{
If \code{items} is a vector corresponding to a subset of items contained in the transactions,
it is returned.

If \code{items} is a character value equal to \code{"items"} or \code{"i"}, the vector of unique items contained
in \code{object} is returned.
}
\seealso{
Method for signature \code{TransactionAnalyzer}:
\code{\link[=get_items,TransactionAnalyzer-method]{get_items,TransactionAnalyzer}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
