% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{get_items_from_category,TransactionAnalyzer-method}
\alias{get_items_from_category,TransactionAnalyzer-method}
\alias{get_items_from_category}
\title{Search for items by category}
\usage{
\S4method{get_items_from_category}{TransactionAnalyzer}(object, category, value, force.character = FALSE)
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}

\item{category}{Name or number of the category on which to search (numbering according to the order
of the columns of \code{object["items_categories"]}).}

\item{value}{Sought value for the category specified by the argument \code{category}.}

\item{force.character}{If \code{TRUE}, items are returned as character values.
If \code{FALSE}, they are the same type as in \code{object["items"]} (numeric or character).}
}
\value{

}
\description{
Extract the items corresponding to a sought category value.
}
\author{
Gauthier Magnin
}
\keyword{internal}
