% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_set.R
\name{get_items_from_info,TransactionSet-method}
\alias{get_items_from_info,TransactionSet-method}
\alias{get_items_from_info}
\title{Search for items by specific information}
\usage{
\S4method{get_items_from_info}{TransactionSet}(object, info, presence = "all", additional = NULL)
}
\arguments{
\item{object}{S4 object of class \code{TransactionSet}.}

\item{info}{Named list of sought information. Element names must refer to the names of variables
contained in the transactions and values must correspond to the sought values for these variables.}

\item{presence}{Information presence condition for an item to be extracted from a transaction.
One of \code{"all"}, \code{"any"}.
\describe{
\item{\code{"all"}}{All the sought information must be part of a transaction for its items to be
extracted.}
\item{\code{"any"}}{At least one of the sought information must be part of a transaction for its items
to be extracted.}
}}

\item{additional}{Names of additional information to extract during the search.}
}
\value{
Vector of the item codes corresponding to the search if \code{additional} is equal to \code{NULL}.
List of codes and information requested otherwise.
}
\description{
Retrieve the items associated with transactions whose information matches to one or more sought
information.
}
\examples{
get_items_from_info(TS_instance, info = list(JOB.TITLE = 44132017))
get_items_from_info(TS_instance, info = list(JOB.TITLE = 44132017),
                                 additional = "SAMPLE.ID")

}
\seealso{
\code{\link{get_all_items}}, \code{\link{get_info_from_items}}.
}
\author{
Gauthier Magnin
}
