% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations_maker.R
\name{get_items_from_info}
\alias{get_items_from_info}
\title{Recherche d'items par information spécifique}
\usage{
get_items_from_info(
  observations,
  ...,
  target = "all",
  additional = NULL,
  key = "CODE"
)
}
\arguments{
\item{observations}{Liste des observations sur lesquels effectuer la recherche.}

\item{...}{Arguments de type \code{clé = valeur} où \code{clé} fait référence au nom d'une variable
contenue dans les observations et \code{valeur} correspond à la valeur recherchée pour cette variable.}

\item{target}{Condition pour qu'une observation soit extraite. Choix parmi \code{"all"}, \code{"any"}.
\describe{
 \item{\code{"all"}}{L'intégralité des informations recherchées doivent faire partie d'une observation
                     pour que cette observation soit extraite.}
 \item{\code{"any"}}{Au moins une des informations recherchées doit faire partie d'une observation pour
                     que cette observation soit extraite.}
}}

\item{additional}{Noms d'informations supplémentaires à extraire lors de la recherche.}

\item{key}{Clé d'accès aux items de chaque observation.}
}
\value{
Vecteur des codes des items correspondant à la recherche si \code{additional} vaut \code{NULL}.
 Liste des codes et des informations demandées sinon.
}
\description{
Extrait les items associés aux observations correspondant à un ou plusieurs critères de recherche.
}
\seealso{
\code{\link{get_info_from_items}}.
}
\author{
Gauthier Magnin
}
