% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_set.R
\name{get_itemsets,ObservationSet-method}
\alias{get_itemsets,ObservationSet-method}
\alias{get_itemsets}
\title{Extract the item sets}
\usage{
\S4method{get_itemsets}{ObservationSet}(object)
}
\arguments{
\item{object}{S4 object of class \code{ObservationSet}.}
}
\value{
List of the item sets of the observations.
}
\description{
Extract the item sets corresponding to the observations.
}
\examples{
get_itemsets(OS_instance)

}
\seealso{
\code{\link{get_all_items}}.
}
\author{
Gauthier Magnin
}
