% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_set.R
\name{get_itemsets,TransactionSet-method}
\alias{get_itemsets,TransactionSet-method}
\alias{get_itemsets}
\title{Extract the item sets}
\usage{
\S4method{get_itemsets}{TransactionSet}(object)
}
\arguments{
\item{object}{S4 object of class \code{TransactionSet}.}
}
\value{
List of the item sets of the transactions.
}
\description{
Extract the item sets corresponding to the transactions.
}
\examples{
get_itemsets(TS_instance)

}
\seealso{
\code{\link{get_all_items}}, \code{\link{get_all_years}}.
}
\author{
Gauthier Magnin
}
