% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{get_links,SpectralAnalyzer-method}
\alias{get_links,SpectralAnalyzer-method}
\alias{get_links}
\title{Get links between nodes or patterns}
\usage{
\S4method{get_links}{SpectralAnalyzer}(object, nopc)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{nopc}{Data frame of \strong{n}odes \strong{o}r \strong{p}atterns and their
\strong{c}haracteristics. Nodes or patterns whose links are be to sought. Any subset of
\code{object["nodes"]} or \code{object["patterns"]}.\cr
\code{"nodes"}, \code{"n"}, \code{"patterns"} and \code{"p"} are special values for
\code{object["nodes"]} and \code{object["patterns"]}.}
}
\value{
Data frame associating the linked nodes or linked patterns.
}
\description{
Extract from the data frame of links between nodes or patterns those corresponding to the given
 nodes or patterns.
}
\details{
If among the nodes or patterns for which the links are sought, some become isolated because the other
 entities to which they are normally linked are not part of the subset \code{nopc}, these nodes or
 patterns are placed at the end of the return data frame.
These possible \code{n} additional lines are numbered \code{"A1"..."An"}.
}
\examples{
get_links(SA_instance, "patterns")
get_links(SA_instance, SA_instance["patterns"][1:10, ])

}
\seealso{
\code{\link{get_isolates}}, \code{\link{get_non_isolates}}, \code{\link{get_complexes}}.
}
\author{
Gauthier Magnin
}
