% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{get_nodes,SpectralAnalyzer-method}
\alias{get_nodes,SpectralAnalyzer-method}
\alias{get_nodes}
\title{Search for nodes by item, characteristic or category}
\usage{
\S4method{get_nodes}{SpectralAnalyzer}(object, nc, element, value, condition = "default")
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{nc}{Data frame of \strong{n}odes and their \strong{c}haracteristics. Any subset of \code{object["nodes"]}.\cr
\code{"nodes"} and \code{"n"} are special values for \code{object["nodes"]}.}

\item{element}{Type of element on which to search.
One of \code{"items"}, \code{"length"}, \code{"weight"} or the name or number of a category on which to search
(numbering according to the order of the columns of \code{object["items_categories"]}).}

\item{value}{Sought value(s) for the element specified by the argument \code{element}.}

\item{condition}{Search condition, depending on \code{element}. See 'Details' section.}
}
\value{
Subset of the data frame of nodes that match the search criteria.
}
\description{
Extract the nodes satisfying search criteria according to items, characteristics or categories.
}
\details{
If \code{element = "items"} one or more items can be sought. The condition for a node to be extracted
is the presence of the sought items (argument \code{value}). The argument \code{condition} must be \code{"all"},
\code{"any"} or \code{"exact"} (default is \code{"all"}):
\itemize{
\item \code{"all"}: all the sought items must be part of the node.
\item \code{"any"}: at least one of the sought items must be part of the node.
\item \code{"exact"}: the item set contained in the node must be exactly the same as the sought item set.
}

If \code{element} refers to a characteristic (i.e. is \code{"length"} or \code{"weight"}), the condition for a node
to be extracted is a comparison of the \code{value} according to one of the comparison operators (default
is equality):
\itemize{
\item \code{"EQ"}, \code{"=="}: \strong{EQ}ual. The value of the characteristic must be equal to that sought.
\item \code{"NE"}, \code{"!="}: \strong{N}ot \strong{E}qual. The value of the characteristic must be different from that
sought.
\item \code{"LT"}, \code{"<"}: \strong{L}ess \strong{T}han. The value of the characteristic must be less than that sought.
\item \code{"GT"}, \code{">"}: \strong{G}reater \strong{T}han. The value of the characteristic must be greater than that
sought.
\item \code{"LE"}, \code{"<="}: \strong{L}ess than or \strong{E}qual. The value of the caracteristic must be less than or
equal to that sought.
\item \code{"GE"}, \code{">="}: \strong{G}reater than or \strong{E}qual. The value of the characteristic must be greater
than or equal to that sought.
}

If \code{element} is the name or the number of a category, the condition for a node to be extracted is
the correspondence to the sought category value (argument \code{value}). The argument \code{condition} must
be one of \code{"items"}, \code{"links"}, \code{"vertices"}, \code{"edges"} (no default).
\itemize{
\item \code{"items"}, \code{"vertices"}: search for nodes containing an item associated with the sought category
value.
\item \code{"links"}, \code{"edges"}: search for nodes generating links corresponding to the sought category value.
}
}
\examples{
## Search on items
get_nodes(SA_instance, SA_instance["nodes"], element = "items", value = 3146)
get_nodes(SA_instance, SA_instance["nodes"],
          element = "items", value = c(3146, 3180), condition = "all")
get_nodes(SA_instance, "nodes",
          element = "items", value = c(3146, 3180), condition = "any")
get_nodes(SA_instance, "nodes",
          element = "items", value = c(3146, 3180), condition = "exact")

## Search on characteristics
get_nodes(SA_instance, SA_instance["nodes"], element = "weight", value = 2)
get_nodes(SA_instance, "nodes",
          element = "weight", value = 2, condition = ">=")
get_nodes(SA_instance, "nodes",
          element = "length", value = 5, condition = "LT")

## Search on categories
get_nodes(SA_instance, SA_instance["nodes"],
          element = "family", value = "Chrome", condition = "items")
get_nodes(SA_instance, "nodes",
          element = 1, value = "Chrome", condition = "links")

}
\seealso{
\code{\link{get_patterns}}, \code{\link{get_complexes}}, \code{\link{get_isolates}}, \code{\link{get_non_isolates}}.
}
\author{
Gauthier Magnin
}
