% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{get_nodes_from_category,TransactionAnalyzer-method}
\alias{get_nodes_from_category,TransactionAnalyzer-method}
\alias{get_nodes_from_category}
\title{Search for nodes by category}
\usage{
\S4method{get_nodes_from_category}{TransactionAnalyzer}(object, nc, category, value, condition)
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}

\item{nc}{Data frame of \strong{n}odes and their \strong{c}haracteristics. Any subset of
\code{object["nodes"]}.

\code{"nodes"} and \code{"n"} are special values for \code{object["nodes"]}.}

\item{category}{Name or number of the category on which to search (numbering according to the order
of the columns of \code{object["items_categories"]}).}

\item{value}{Sought value for the category specified by the argument \code{category}.}

\item{condition}{Category value search condition for a node to be extracted.
One of \code{"items"}, \code{"links"}, \code{"vertices"}, \code{"edges"}.
\describe{
  \item{\code{"items"}, \code{"vertices"}}{Search for nodes containing an item associated with the
        sought category value.}
  \item{\code{"links"}, \code{"edges"}}{Search for nodes generating links corresponding to the
        sought category value.}
}}
}
\value{
Subset of the data frame of nodes that match the search criteria.
}
\description{
Extract the nodes corresponding to a sought category value.
}
\note{
This method is not exported from the package.
}
\examples{
get_nodes_from_category(TA_instance, TA_instance["nodes"],
                        category = "family", value = "Chrome",
                        condition = "items")

get_nodes_from_category(TA_instance, TA_instance["nodes"],
                        category = 1, value = "Chrome",
                        condition = "links")

}
\seealso{
\code{\link{get_nodes}}, \code{\link{get_nodes_from_items}},
         \code{\link{get_nodes_from_characteristic}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
