% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{get_nodes_from_characteristic,SpectralAnalyzer-method}
\alias{get_nodes_from_characteristic,SpectralAnalyzer-method}
\alias{get_nodes_from_characteristic}
\title{Search for nodes by characteristic}
\usage{
\S4method{get_nodes_from_characteristic}{SpectralAnalyzer}(
  object,
  nc,
  characteristic,
  value,
  condition = "EQ"
)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{nc}{Data frame of \strong{n}odes and their \strong{c}haracteristics. Any subset of
\code{object["nodes"]}.\cr
\code{"nodes"} and \code{"n"} are specific values for \code{object["nodes"]}.}

\item{characteristic}{Name of the characteristic on which to do the search.
One of \code{"length"}, \code{"weight"}.}

\item{value}{Sought value for the characteristic specified by the parameter \code{characteristic}.}

\item{condition}{Search condition.
One of \code{"EQ"}, \code{"NE"}, \code{"LT"}, \code{"GT"}, \code{"LE"}, \code{"GE"},
\code{"=="}, \code{"!="}, \code{"<"}, \code{">"}, \code{"<="}, \code{">="}.
\describe{
  \item{\code{"EQ", "=="}}{\strong{EQ}ual: the value of the characteristic must be equal to that
                           sought.}
  \item{\code{"NE", "!="}}{\strong{N}ot \strong{E}qual: the value of the characteristic must be
                           different from that sought.}
  \item{\code{"LT", "<"}}{\strong{L}ess \strong{T}han: the value of the characteristic must be less
                          than that sought.}
  \item{\code{"GT", ">"}}{\strong{G}reater \strong{T}han: the value of the characteristic must be
                          greater than that sought.}
  \item{\code{"LE", "<="}}{\strong{L}ess than or \strong{E}qual: the value of the caracteristic must
                           be less than or equal to that sought.}
  \item{\code{"GE", ">="}}{\strong{G}reater than or \strong{E}qual: the value of the characteristic
                           must be greater than or equal to that sought.}
}}
}
\value{
Subset of the data frame of nodes that match the search criteria.
}
\description{
Extract the nodes satisfying a search criterion according to one characteristic.
}
\examples{
get_nodes_from_characteristic(SA_instance, SA_instance["nodes"],
                              characteristic = "weight",
                              value = 2)
get_nodes_from_characteristic(SA_instance, SA_instance["nodes"],
                              characteristic = "length",
                              value = 2, condition = ">=")
get_nodes_from_characteristic(SA_instance, SA_instance["nodes"],
                              characteristic = "length",
                              value = 5, condition = "LT")

}
\seealso{
\code{\link{get_nodes}}, \code{\link{get_nodes_from_items}},
         \code{\link{get_nodes_from_category}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
