% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{get_nodes_from_items,TransactionAnalyzer-method}
\alias{get_nodes_from_items,TransactionAnalyzer-method}
\alias{get_nodes_from_items}
\title{Search for nodes by item}
\usage{
\S4method{get_nodes_from_items}{TransactionAnalyzer}(object, nc, items, condition = "all")
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}

\item{nc}{Data frame of \strong{n}odes and their \strong{c}haracteristics. Any subset of \code{object["nodes"]}.

\code{"nodes"} and \code{"n"} are special values for \code{object["nodes"]}.}

\item{items}{Sought items (one or more).}

\item{condition}{Item presence condition for a node to be extracted.
One of \code{"all"}, \code{"any"}, \code{"exact"}.
\describe{
\item{\code{"all"}}{All the sought items must be part of a node for this node to be extracted.}
\item{\code{"any"}}{At least one of the sought items must be part of a node for this node to be
extracted.}
\item{\code{"exact"}}{The item set contained in a node must be exactly the same as the
sought item set for this node to be extracted.}
}}
}
\value{
Subset of the data frame of nodes that match the search criteria.
}
\description{
Extract the nodes containing one or more sought items.
}
\examples{
get_nodes_from_items(TA_instance, TA_instance["nodes"], items = 3146)
get_nodes_from_items(TA_instance, TA_instance["nodes"],
                     items = c(3146, 3180), condition = "all")
get_nodes_from_items(TA_instance, TA_instance["nodes"],
                     items = c(3146, 3180), condition = "any")
get_nodes_from_items(TA_instance, TA_instance["nodes"],
                     items = c(3146, 3180), condition = "exact")

}
\seealso{
\code{\link{get_nodes}}, \code{\link{get_nodes_from_characteristic}}, \code{\link{get_nodes_from_category}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
