% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{get_nop,SpectralAnalyzer-method}
\alias{get_nop,SpectralAnalyzer-method}
\alias{get_nop}
\title{Get nodes or patterns}
\usage{
\S4method{get_nop}{SpectralAnalyzer}(object, nop, entities = SpectralAnalyzer.NODES_OR_PATTERNS)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{nop}{List of \strong{n}odes \strong{o}r \strong{p}atterns or one of the following character values:
\code{"nodes"}, \code{"n"}, \code{"patterns"}, \code{"p"}.}

\item{entities}{Type of the entities that the list may refer to (\code{NODES}, \code{PATTERNS} or
\code{NODES_OR_PATTERNS}).}
}
\value{
List of nodes or patterns corresponding to the arguments.
}
\description{
Find and return the list corresponding to the nodes or the patterns of the \code{SpectralAnalyzer}
object, or return the given list.
}
\details{
If \code{nop} is a list, it is returned.

If \code{nop} is a character value equal to:
\itemize{
\item \code{"nodes"} or \code{"n"}: \code{object["nodes"]$node} is returned.
\item \code{"patterns"} or \code{"p"}: \code{object["patterns"]$pattern} is returned.
}

The argument \code{entities} is only used to adapt a possible error message.
}
\seealso{
\code{\link{get_nopc}}, \code{\link{which_entities}}, \code{\link{get_items}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
