% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{get_nopc,SpectralAnalyzer-method}
\alias{get_nopc,SpectralAnalyzer-method}
\alias{get_nopc}
\title{Get nodes or patterns characteristics}
\usage{
\S4method{get_nopc}{SpectralAnalyzer}(object, nopc, entities = SpectralAnalyzer.NODES_OR_PATTERNS)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{nopc}{Data frame of \strong{n}odes \strong{o}r \strong{p}atterns and their \strong{c}haracteristics or one of the
following character values: \code{"nodes"}, \code{"n"}, \code{"patterns"}, \code{"p"}.}

\item{entities}{Type of the entities that the data frame may refer to (\code{NODES}, \code{PATTERNS} or
\code{NODES_OR_PATTERNS}).}
}
\value{
Data frame of nodes or patterns and their characteristics corresponding to the arguments.
}
\description{
Find and return the data frame corresponding to the nodes or the patterns of the \code{SpectralAnalyzer}
object, or return the given data frame.
}
\details{
If \code{nopc} is a data frame, it is returned.

If \code{nopc} is a character value equal to:
\itemize{
\item \code{"nodes"} or \code{"n"}: \code{object["nodes"]} is returned.
\item \code{"patterns"} or \code{"p"}: \code{object["patterns"]} is returned.
}

The argument \code{entities} is only used to adapt a possible error message.
}
\seealso{
\code{\link{which_entities}}, \code{\link{get_items}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
