% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_set.R
\name{get_obs_from_info,ObservationSet-method}
\alias{get_obs_from_info,ObservationSet-method}
\alias{get_obs_from_info}
\title{Search for observations by specific information}
\usage{
\S4method{get_obs_from_info}{ObservationSet}(object, info, presence = "all")
}
\arguments{
\item{object}{S4 object of class \code{ObservationSet}.}

\item{info}{Named list of sought information. Element names must refer to the names of variables
contained in the observations and values must correspond to the sought values for these variables.}

\item{presence}{Information presence condition for an observation to be extracted.
One of \code{"all"}, \code{"any"}.
\describe{
\item{\code{"all"}}{All the sought information must be part of an observation for this observation to
be extracted.}
\item{\code{"any"}}{At least one of the sought information must be part of an observation for this
observation to be extracted.}
}}
}
\value{
\code{ObservationSet} containing the subset of observations that match the search criteria.
}
\description{
Extract the observations whose information matches to one or more sought information.
}
\examples{
get_obs_from_info(OS_instance, info = list(JOB.TITLE = 44132001,
                                           JOB.TASK = "A8310"))

}
\seealso{
\code{\link{get_complex_obs}}, \code{\link{get_simple_obs}}, \code{\link{get_obs_from_items}}.
}
\author{
Gauthier Magnin
}
