% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_set.R
\name{get_obs_from_items,ObservationSet-method}
\alias{get_obs_from_items,ObservationSet-method}
\alias{get_obs_from_items}
\title{Search for observations by item}
\usage{
\S4method{get_obs_from_items}{ObservationSet}(object, items, presence = "all")
}
\arguments{
\item{object}{\code{ObservationSet} class object.}

\item{items}{Sought items.}

\item{presence}{Item presence condition for an observation to be extracted.
One of \code{"all"}, \code{"any"}, \code{"exact"}.
\describe{
\item{\code{"all"}}{All the sought items must be part of an observation for this observation to be
extracted.}
\item{\code{"any"}}{At least one of the sought items must be part of an observation for this observation
to be extracted.}
\item{\code{"exact"}}{The item set contained in an observation must be exactly the same as the sought
item set for this observation to be extracted.}
}}
}
\value{
\code{ObservationSet} object containing the subset of observations that match the search criteria.
}
\description{
Extract the observations containing one or more sought items.
}
\examples{
get_obs_from_items(OS_instance, items = c(25, 192), presence = "all")
get_obs_from_items(OS_instance, items = c(25, 192), presence = "any")
get_obs_from_items(OS_instance, items = c(25, 192), presence = "exact")

}
\seealso{
\code{\link{get_complex_obs}}, \code{\link{get_simple_obs}}, \code{\link{get_obs_from_info}}.
}
\author{
Gauthier Magnin
}
