% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_maker.R
\name{get_obs_from_items}
\alias{get_obs_from_items}
\title{Recherche d'observations par item}
\usage{
get_obs_from_items(observations, items, target = "all", key = "CODE")
}
\arguments{
\item{observations}{Liste des observations sur lesquels effectuer la recherche.}

\item{items}{Élément·s recherché·s.}

\item{target}{Condition pour qu'une observation soit extraite. Choix parmi \code{"all"}, \code{"any"}.
\describe{
 \item{\code{"all"}}{L'intégralité des items recherchés doivent faire partie d'une observation
                     pour que cette observation soit extraite.}
 \item{\code{"any"}}{Au moins un des items recherchés doit faire partie d'une observation pour
                     que cette observation soit extraite.}
}}

\item{key}{Clé d'accès aux items de chaque observation.}
}
\value{
Sous-ensemble de la liste d'observations correspondant aux critères de recherche.
}
\description{
Extrait les observations contenant un ou plusieurs items recherchés.
}
\seealso{
\code{\link{get_obs_from_info}}.
}
\author{
Gauthier Magnin
}
