% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{get_onp,SpectralAnalyzer-method}
\alias{get_onp,SpectralAnalyzer-method}
\alias{get_onp}
\title{Get observations or nodes or patterns and their characteristics}
\usage{
\S4method{get_onp}{SpectralAnalyzer}(object, onp, entities = SpectralAnalyzer.NODES_OR_PATTERNS)
}
\arguments{
\item{object}{S4 object of class \code{SpectralAnalyzer}.}

\item{onp}{Object of class \code{ObservationSet} (\strong{o}) or data frame of \strong{n}odes or \strong{p}atterns and
their characteristics or one of the following character values: \code{"observations"}, \code{"o"}, \code{"nodes"},
\code{"n"}, \code{"patterns"}, \code{"p"}.}

\item{entities}{Type of the entities that \code{onp} may refer to (\code{OBSERVATIONS}, \code{NODES}, \code{PATTERNS},
\code{NODES_OR_PATTERNS} or \code{NODES_PATTERNS_OR_OBSERVATIONS}).}
}
\value{
Object of class \code{ObservationSet} or data frame of nodes or patterns and their characteristics,
corresponding to the arguments.
}
\description{
Find and return the \code{ObservationSet} corresponding to the observations or the data frame corresponding
to the nodes or the patterns of the object of class \code{SpectralAnalyzer}, or return the given value.
}
\details{
If \code{onp} is an object of class \code{ObservationSet} or a data frame, it is returned.

If \code{onp} is a character value equal to:
\itemize{
\item \code{"observations"} or \code{"o"}: \code{object["observations"]} is returned.
\item \code{"nodes"} or \code{"n"}: \code{object["nodes"]} is returned.
\item \code{"patterns"} or \code{"p"}: \code{object["patterns"]} is returned.
}

The argument \code{entities} is only used to adapt a possible error message.
}
\seealso{
\code{\link{get_onp_itemsets}}, \code{\link{which_entities}}, \code{\link[=get_items,SpectralAnalyzer]{get_items}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
