% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{get_onp_itemsets,SpectralAnalyzer-method}
\alias{get_onp_itemsets,SpectralAnalyzer-method}
\alias{get_onp_itemsets}
\title{Get observation, node or pattern itemsets}
\usage{
\S4method{get_onp_itemsets}{SpectralAnalyzer}(object, onp, entities = SpectralAnalyzer.NODES_OR_PATTERNS)
}
\arguments{
\item{object}{S4 object of class \code{SpectralAnalyzer}.}

\item{onp}{List of \strong{o}bservation, \strong{n}ode or \strong{p}attern itemsets or one of the following character
values: \code{"observations"}, \code{"o"}, \code{"nodes"}, \code{"n"}, \code{"patterns"}, \code{"p"}.}

\item{entities}{Type of the entities that the list may refer to (\code{OBSERVATIONS}, \code{NODES}, \code{PATTERNS},
\code{NODES_OR_PATTERNS} or \code{NODES_PATTERNS_OR_OBSERVATIONS}).}
}
\value{
List of observation, node or pattern itemsets corresponding to the arguments.
}
\description{
Find and return the list corresponding to the observations, the nodes or the patterns of the object of
class \code{SpectralAnalyzer}, or return the given list.
}
\details{
If \code{onp} is a list, it is returned.

If \code{onp} is a character value equal to:
\itemize{
\item \code{"observations"} or \code{"o"}: \code{object["observations"][object["observations"]["item_key"]]} is returned.
\item \code{"nodes"} or \code{"n"}: \code{object["nodes"]$node} is returned.
\item \code{"patterns"} or \code{"p"}: \code{object["patterns"]$pattern} is returned.
}

The argument \code{entities} is only used to adapt a possible error message.
}
\seealso{
\code{\link{get_onp}}, \code{\link{which_entities}}, \code{\link[=get_items,SpectralAnalyzer-method]{get_items}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
