% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{get_patterns_from_category,SpectralAnalyzer-method}
\alias{get_patterns_from_category,SpectralAnalyzer-method}
\alias{get_patterns_from_category}
\title{Search for patterns by category}
\usage{
\S4method{get_patterns_from_category}{SpectralAnalyzer}(
  object,
  patterns_characteristics,
  category,
  value,
  condition
)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{patterns_characteristics}{Data frame of patterns and their characteristics.
Any subset of \code{object["patterns"]}.}

\item{category}{Name or number of the category on which to search (numbering according to the order
of the columns of \code{object["items_categories"]}).}

\item{value}{Sought value for the category specified by the argument \code{category}.}

\item{condition}{Category value search condition for a pattern to be extracted.
One of \code{"items"}, \code{"links"}, \code{"vertices"}, \code{"edges"}.
\describe{
  \item{\code{"items"}, \code{"vertices"}}{Search for patterns containing an item associated with
        the sought category value.}
  \item{\code{"links"}, \code{"edges"}}{Search for patterns generating links corresponding to the
        sought category value.}
}}
}
\value{
Subset of the data frame of patterns that match the search criteria.
}
\description{
Extract the patterns corresponding to a sought category value.
}
\examples{
get_patterns_from_category(SA_instance, SA_instance["patterns"],
                           category = "family", value = "Chrome",
                           condition = "items")

get_patterns_from_category(SA_instance, SA_instance["patterns"],
                           category = 1, value = "Chrome",
                           condition = "links")

}
\seealso{
\code{\link{get_patterns}}, \code{\link{get_patterns_from_items}},
         \code{\link{get_patterns_from_characteristic}}, \code{\link{get_patterns_from_status}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
