% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{get_patterns_from_characteristic,TransactionAnalyzer-method}
\alias{get_patterns_from_characteristic,TransactionAnalyzer-method}
\alias{get_patterns_from_characteristic}
\title{Search for patterns by characteristic}
\usage{
\S4method{get_patterns_from_characteristic}{TransactionAnalyzer}(
  object,
  pc,
  characteristic,
  value,
  condition = "EQ"
)
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}

\item{pc}{Data frame of \strong{p}atterns and their \strong{c}haracteristics. Any subset of
\code{object["patterns"]}.

\code{"patterns"} and \code{"p"} are special values for \code{object["patterns"]}.}

\item{characteristic}{Name of the characteristic on which to do the search.
One of \code{"year"}, \code{"length"}, \code{"frequency"}, \code{"weight"}, \code{"specificity"}
See \code{\link{get_patterns_from_status}} to search by \code{"status"}.}

\item{value}{Sought value for the characteristic specified by the parameter \code{characteristic}.}

\item{condition}{Search condition.
One of \code{"EQ"}, \code{"NE"}, \code{"LT"}, \code{"GT"}, \code{"LE"}, \code{"GE"},
\code{"=="}, \code{"!="}, \code{"<"}, \code{">"}, \code{"<="}, \code{">="}.
\describe{
  \item{\code{"EQ", "=="}}{\strong{EQ}ual: the value of the characteristic must be equal to that
                           sought.}
  \item{\code{"NE", "!="}}{\strong{N}ot \strong{E}qual: the value of the characteristic must be
                           different from that sought.}
  \item{\code{"LT", "<"}}{\strong{L}ess \strong{T}han: the value of the characteristic must be less
                          than that sought.}
  \item{\code{"GT", ">"}}{\strong{G}reater \strong{T}han: the value of the characteristic must be
                          greater than that sought.}
  \item{\code{"LE", "<="}}{\strong{L}ess than or \strong{E}qual: the value of the caracteristic must
                           be less than or equal to that sought.}
  \item{\code{"GE", ">="}}{\strong{G}reater than or \strong{E}qual: the value of the characteristic
                           must be greater than or equal to that sought.}
}}
}
\value{
Subset of the data frame of patterns that match the search criteria.
}
\description{
Extract the patterns satisfying a search criterion according to one characteristic.
}
\note{
This method is not exported from the package.
}
\examples{
get_patterns_from_characteristic(TA_instance, TA_instance["patterns"],
                                 characteristic = "frequency",
                                 value = 3)
get_patterns_from_characteristic(TA_instance, TA_instance["patterns"],
                                 characteristic = "frequency",
                                 value = 3, condition = ">=")
get_patterns_from_characteristic(TA_instance, TA_instance["patterns"],
                                 characteristic = "length",
                                 value = 3, condition = "LT")

}
\seealso{
\code{\link{get_patterns}}, \code{\link{get_patterns_from_items}},
         \code{\link{get_patterns_from_status}}, \code{\link{get_patterns_from_category}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
