% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{get_patterns_from_items,SpectralAnalyzer-method}
\alias{get_patterns_from_items,SpectralAnalyzer-method}
\alias{get_patterns_from_items}
\title{Search for patterns by item}
\usage{
\S4method{get_patterns_from_items}{SpectralAnalyzer}(
  object,
  patterns_characteristics,
  items,
  condition = "all"
)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{patterns_characteristics}{Data frame of patterns and their characteristics.
Any subset of \code{object["patterns"]}.}

\item{items}{Sought items (one or more).}

\item{condition}{Item presence condition for a pattern to be extracted.
One of \code{"all"}, \code{"any"}.
\describe{
  \item{\code{"all"}}{All the sought items must be part of a pattern for this pattern to be
                      extracted.}
  \item{\code{"any"}}{At least one of the sought items must be part of a pattern for this pattern
                      to be extracted.}
}}
}
\value{
Subset of the data frame of patterns that match the search criteria.
}
\description{
Extract the patterns containing one or more sought items.
}
\examples{
get_patterns_from_items(SA_instance, SA_instance["patterns"], items = 3146)
get_patterns_from_items(SA_instance, SA_instance["patterns"],
                        items = c(3146, 3180), condition = "all")
get_patterns_from_items(SA_instance, SA_instance["patterns"],
                        items = c(3146, 3180), condition = "any")

}
\seealso{
\code{\link{get_patterns}}, \code{\link{get_patterns_from_characteristic}},
         \code{\link{get_patterns_from_status}}, \code{\link{get_patterns_from_category}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
