% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{get_patterns_from_items,TransactionAnalyzer-method}
\alias{get_patterns_from_items,TransactionAnalyzer-method}
\alias{get_patterns_from_items}
\title{Search for patterns by item}
\usage{
\S4method{get_patterns_from_items}{TransactionAnalyzer}(object, pc, items, condition = "all")
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}

\item{pc}{Data frame of \strong{p}atterns and their \strong{c}haracteristics. Any subset of
\code{object["patterns"]}.\cr
\code{"patterns"} and \code{"p"} are special values for \code{object["patterns"]}.}

\item{items}{Sought items (one or more).}

\item{condition}{Item presence condition for a pattern to be extracted.
One of \code{"all"}, \code{"any"}, \code{"exact"}.
\describe{
  \item{\code{"all"}}{All the sought items must be part of a pattern for this pattern to be
                      extracted.}
  \item{\code{"any"}}{At least one of the sought items must be part of a pattern for this pattern
                      to be extracted.}
  \item{\code{"exact"}}{The item set contained in a pattern must be exactly the same as the
                        sought item set for this pattern to be extracted.}
}}
}
\value{
Subset of the data frame of patterns that match the search criteria.
}
\description{
Extract the patterns containing one or more sought items.
}
\examples{
get_patterns_from_items(TA_instance, TA_instance["patterns"], items = 3146)
get_patterns_from_items(TA_instance, TA_instance["patterns"],
                        items = c(3146, 3180), condition = "all")
get_patterns_from_items(TA_instance, TA_instance["patterns"],
                        items = c(3146, 3180), condition = "any")
get_patterns_from_items(TA_instance, TA_instance["patterns"],
                        items = c(3146, 3180), condition = "exact")

}
\seealso{
\code{\link{get_patterns}}, \code{\link{get_patterns_from_characteristic}},
         \code{\link{get_patterns_from_status}}, \code{\link{get_patterns_from_category}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
