% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{get_patterns_from_status,SpectralAnalyzer-method}
\alias{get_patterns_from_status,SpectralAnalyzer-method}
\alias{get_patterns_from_status}
\title{Search for patterns by status}
\usage{
\S4method{get_patterns_from_status}{SpectralAnalyzer}(object, pc, value, condition = "EQ")
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{pc}{Data frame of \strong{p}atterns and their \strong{c}haracteristics. Any subset of
\code{object["patterns"]}.\cr
\code{"patterns"} and \code{"p"} are specific values for \code{object["patterns"]}.}

\item{value}{Status value sought (one or more).}

\item{condition}{Search condition. One of \code{"EQ"}, \code{"NE"}, \code{"=="}, \code{"!="}.
\describe{
  \item{\code{"EQ"}, \code{"=="}}{\strong{EQ}ual: the status of the pattern must be one of the
                                  sought values.}
  \item{\code{"NE"}, \code{"!="}}{\strong{N}ot \strong{E}qual: the status of the pattern must be
                                  different from the sought values.}
}}
}
\value{
Subset of the data frame of patterns that match the search criteria.
}
\description{
Extract the patterns whose status match one or more sought values.
}
\examples{
get_patterns_from_status(SA_instance, SA_instance["patterns"],
                         value = SA_instance["Class"]$STATUS_PERSISTENT,
                         condition = "EQ")

get_patterns_from_status(SA_instance, SA_instance["patterns"],
                         value = c("Persistent", "Declining"),
                         condition = "!=")

}
\seealso{
\code{\link{get_patterns}}, \code{\link{get_patterns_from_items}},
         \code{\link{get_patterns_from_characteristic}}, \code{\link{get_patterns_from_category}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
