% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_set.R
\name{get_simple_obs,ObservationSet-method}
\alias{get_simple_obs,ObservationSet-method}
\alias{get_simple_obs}
\title{Search for simple observations}
\usage{
\S4method{get_simple_obs}{ObservationSet}(object)
}
\arguments{
\item{object}{S4 object of class \code{ObservationSet}.}
}
\value{
S4 object of class \code{ObservationSet} containing the subset of observations that contain exactly
one item.
}
\description{
Extract the observations containing exactly one item.
}
\examples{
get_simple_obs(OS_instance)

}
\seealso{
\code{\link{get_complex_obs}}, \code{\link{get_obs_from_items}}, \code{\link{get_obs_from_info}}.
}
\author{
Gauthier Magnin
}
