% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_set.R
\name{get_simple_trx,TransactionSet-method}
\alias{get_simple_trx,TransactionSet-method}
\alias{get_simple_trx}
\title{Search for simple transactions}
\usage{
\S4method{get_simple_trx}{TransactionSet}(object, as_indices = FALSE)
}
\arguments{
\item{object}{S4 object of class \code{TransactionSet}.}

\item{as_indices}{\code{TRUE} or \code{FALSE} whether to return transactions or only
their indices.}
}
\value{
S4 object of class \code{TransactionSet} containing the subset of
transactions that contain exactly one item, or indices of these
transactions (according to the argument \code{as_indices}). If the given
transactions are named (and \code{as_indices} is \code{TRUE}), the returned indices
are named as well.
}
\description{
Extract the transactions containing exactly one item.
}
\examples{
get_simple_trx(TS_instance)
get_simple_trx(TS_instance, as_indices = TRUE)

}
\seealso{
\code{\link{get_complex_trx}}, \code{\link{get_trx_from_items}}, \code{\link{get_trx_from_info}}.
}
\author{
Gauthier Magnin
}
