% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{get_tnp,TransactionAnalyzer-method}
\alias{get_tnp,TransactionAnalyzer-method}
\alias{get_tnp}
\title{Get transactions or nodes or patterns and their characteristics}
\usage{
\S4method{get_tnp}{TransactionAnalyzer}(object, tnp, entities = NODES_OR_PATTERNS)
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}

\item{tnp}{Object of class \code{TransactionSet} (\strong{t}) or data frame of \strong{n}odes or \strong{p}atterns and
their characteristics or one of the following character values: \code{"transactions"}, \code{"t"}, \code{"nodes"},
\code{"n"}, \code{"patterns"}, \code{"p"}.}

\item{entities}{Type of the entities that \code{tnp} may refer to (\code{TRANSACTIONS}, \code{NODES}, \code{PATTERNS},
\code{NODES_OR_PATTERNS} or \code{NODES_PATTERNS_OR_TRANSACTIONS}).}
}
\value{
Object of class \code{TransactionSet} or data frame of nodes or patterns and their characteristics,
corresponding to the arguments.
}
\description{
Find and return the \code{TransactionSet} corresponding to the transactions or the data frame corresponding
to the nodes or the patterns of the object of class \code{TransactionAnalyzer}, or return the given value.
}
\details{
If \code{tnp} is an object of class \code{TransactionSet} or a data frame, it is returned.

If \code{tnp} is a character value equal to:
\itemize{
\item \code{"transactions"} or \code{"t"}: \code{object["transactions"]} is returned.
\item \code{"nodes"} or \code{"n"}: \code{object["nodes"]} is returned.
\item \code{"patterns"} or \code{"p"}: \code{object["patterns"]} is returned.
}

The argument \code{entities} is only used to adapt a possible error message.
}
\seealso{
\code{\link{get_tnp_itemsets}}, \code{\link{which_entities}}, \code{\link[=get_items,TransactionAnalyzer-method]{get_items}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
