% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_set.R
\name{get_trx_from_items,TransactionSet-method}
\alias{get_trx_from_items,TransactionSet-method}
\alias{get_trx_from_items}
\title{Search for transactions by item}
\usage{
\S4method{get_trx_from_items}{TransactionSet}(object, items, presence = "all")
}
\arguments{
\item{object}{S4 object of class \code{TransactionSet}.}

\item{items}{Sought items.}

\item{presence}{Item presence condition for a transaction to be extracted.
One of \code{"all"}, \code{"any"}, \code{"exact"}.
\describe{
\item{\code{"all"}}{All the sought items must be part of a transaction for this transaction to be
extracted.}
\item{\code{"any"}}{At least one of the sought items must be part of a transaction for this transaction
to be extracted.}
\item{\code{"exact"}}{The item set contained in a transaction must be exactly the same as the sought
item set for this transaction to be extracted.}
}}
}
\value{
S4 object of class \code{TransactionSet} containing the subset of transactions that match the search
criteria.
}
\description{
Extract the transactions containing one or more sought items.
}
\examples{
get_trx_from_items(TS_instance, items = c(25, 192), presence = "all")
get_trx_from_items(TS_instance, items = c(25, 192), presence = "any")
get_trx_from_items(TS_instance, items = c(25, 192), presence = "exact")

}
\seealso{
\code{\link{get_complex_trx}}, \code{\link{get_simple_trx}}, \code{\link{get_trx_from_info}}.
}
\author{
Gauthier Magnin
}
