% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{has_item_names,SpectralAnalyzer-method}
\alias{has_item_names,SpectralAnalyzer-method}
\alias{has_item_names}
\title{Check if the items have names}
\usage{
\S4method{has_item_names}{SpectralAnalyzer}(object)
}
\arguments{
\item{object}{S4 object of class \code{SpectralAnalyzer}.}
}
\value{
\code{TRUE} if at least one of the names of \code{object["items"]} is different from the related item.
\code{FALSE} if all names of \code{object["items"]} are equal to the values of \code{object["items"]}.
}
\description{
Check if the \code{SpectralAnalyzer} has item names, i.e. if the current names associated with the items
are different from their identification codes.
}
\seealso{
\code{\link{get_item_names}}, \code{\link[=get_items,SpectralAnalyzer]{get_items}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
