% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_set.R
\name{has_temporal_data,ObservationSet-method}
\alias{has_temporal_data,ObservationSet-method}
\alias{has_temporal_data}
\title{Check if an ObservationSet contains temporal data}
\usage{
\S4method{has_temporal_data}{ObservationSet}(object)
}
\arguments{
\item{object}{S4 object of class \code{ObservationSet}.}
}
\value{
\code{TRUE} or \code{FALSE} whether \code{object} contains temporal data.
}
\description{
Equivalent to checking if the attribute \code{year_key} is not \code{NA}, but more explicit.
}
\examples{
has_temporal_data(OS_instance)

OS_instance["year_key"] <- NA
has_temporal_data(OS_instance)

}
\author{
Gauthier Magnin
}
\keyword{internal}
