% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_set.R
\name{has_temporal_data,TransactionSet-method}
\alias{has_temporal_data,TransactionSet-method}
\alias{has_temporal_data}
\title{Check if a TransactionSet contains temporal data}
\usage{
\S4method{has_temporal_data}{TransactionSet}(object)
}
\arguments{
\item{object}{S4 object of class \code{TransactionSet}.}
}
\value{
\code{TRUE} or \code{FALSE} whether \code{object} contains temporal data.
}
\description{
Equivalent to checking if the attribute \code{year_key} is not \code{NA}, but more explicit.
}
\examples{
has_temporal_data(TS_instance)

TS_instance["year_key"] <- NA
has_temporal_data(TS_instance)

}
\author{
Gauthier Magnin
}
\keyword{internal}
