% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixer.R
\name{hazard_quotient}
\alias{hazard_quotient}
\title{Hazard Quotient (HQ)}
\usage{
hazard_quotient(values, references)
}
\arguments{
\item{values}{Numeric vector or matrix. Values for which the hazard quotients are to be computed.}

\item{references}{Numeric vector. Reference values associated with the \code{values}.}
}
\value{
Vector or matrix (according to \code{values}) of computed hazard quotients.
}
\description{
Compute the hazard quotient as the ratio between a value and a reference value.
}
\details{
If \code{values} is a matrix, the reference values are applied once on each set of values, i.e. on each row.
Therefore, there must be one reference value for each column of the matrix.

\loadmathjax

The hazard quotient of the value \eqn{j} in the vector \eqn{i} is given by:
\mjdeqn{HQ_{i,j} = \frac{V_{i,j}}{RV_j}}{HQ_ij = V_ij / RV_j}
where \eqn{V} denotes the \code{values} and \eqn{RV} denotes the \code{references}.
}
\examples{
hazard_quotient(c(1,2,3,4,5), c(1,2,3,4,5))
hazard_quotient(c(A = 1, B = 2, C = 3, D = 4, E = 5), c(1,2,3,4,5))

hazard_quotient(
  values = matrix(c(.1, .2, .3, .4, .5, .6, .7, .8, .9, 1),
                  nrow = 2, byrow = TRUE,
                  dimnames = list(c("set.1", "set.2"), LETTERS[1:5])),
  references = c(1,2,3,4,5)
)

}
\references{
Price PS, Han X (2011).
Maximum cumulative ratio (MCR) as a tool for assessing the value of performing a cumulative risk assessment.
\emph{International Journal of Environmental Research and Public Health}. 8(6): 2212-2225.
\url{https://doi.org/10.3390/ijerph8062212}.
}
\seealso{
\code{\link{hazard_index}}, \code{\link{maximum_hazard_quotient}}.
}
\author{
Gauthier Magnin
}
