% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_helper.R
\name{heb_chart}
\alias{heb_chart}
\title{Hierarchical Edge Bundling chart}
\usage{
heb_chart(
  hierarchy,
  vertices,
  edges,
  limits,
  legend_name = NULL,
  legend_values = NULL,
  vertex_size = 3,
  vertex_alpha = 1,
  vertex_margin = 0.05,
  label_size = 3,
  label_margin = 0.05,
  edge_tension = 0.8,
  edge_alpha = 1,
  palette = "Blues",
  palette_direction = 1
)
}
\arguments{
\item{hierarchy}{A two-column data frame representing the hierarchical network structure. A tree-like
representation of the vertices with a root and any internal nodes (only leaves are plotted).}

\item{vertices}{Data frame containing the vertices to plot (i.e. the leaves of the tree
representation). Must contain the columns:
\describe{
\item{\code{"name"}}{Name of the vertices in the \code{hierarchy}.}
\item{\code{"label"}}{Labels to plot next to the vertices.}
\item{\code{"group"}}{Optional. Additional column used to color the vertices.}
}}

\item{edges}{A three-column data frame representing the edges of the graph (i.e. the connections
between the leaves of the tree representation). Must contain two columns contaning the names of the
vertices to connect and a third one giving the intensities of the connections.}

\item{limits}{The scale limits (related to the third column of \code{edges}).}

\item{legend_name, legend_values}{Name and values of the additional legend related to the column
\code{group} of the data frame \code{vertices}.

\code{NULL} if there is no group. Character for \code{legend_name} and named character vector for
\code{legend_values}: colors values named with the group values.}

\item{vertex_size}{Size of the vertices.}

\item{vertex_alpha}{Opacity of the vertices (from 0 to 1).}

\item{vertex_margin}{Margin before the vertices (i.e. distance between the ends of the edges and the
centers of the vertices).}

\item{label_size}{Size of the labels associated with the vertices.}

\item{label_margin}{Margin before the labels (i.e. distance between the centers of the vertices and
the labels).}

\item{edge_tension}{Looseness of the connecting lines (from 0 to 1).
The closer the value is to 0, the straighter the lines will be.
The closer the value is to 1, the more the lines will be curved.}

\item{edge_alpha}{Opacity of the lines connecting vertices (from 0 to 1).}

\item{palette}{Name (or number) of the sequential palette to use for coloring the edges.
One of \code{"Blues"}, \code{"BuGn"}, \code{"BuPu"}, \code{"GnBu"}, \code{"Greens"}, \code{"Greys"}, \code{"Oranges"}, \code{"OrRd"},
\code{"PuBu"}, \code{"PuBuGn"}, \code{"PuRd"}, \code{"Purples"}, \code{"RdPu"}, \code{"Reds"}, \code{"YlGn"}, \code{"YlGnBu"}, \code{"YlOrBr"},
\code{"YlOrRd"}.}

\item{palette_direction}{Direction in which to use the color palette.
If \code{1}, colors are in original order: from the lightest to the darkest.
If \code{-1}, color order is reversed: from the darkest to the lightest.}
}
\value{
Graph created with the packages \code{ggraph} and \code{ggplot2}.
}
\description{
Plot a hierarchical edge bundling chart: graph in which vertices are arranged circularly.
}
\details{
The chart being plotted with the packages \code{ggraph} and \code{ggplot2}, it can be modified or completed
afterwards using \code{\link[ggplot2:last_plot]{ggplot2::last_plot}} or the returned object.
}
\author{
Gauthier Magnin
}
\keyword{internal}
