% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{if_2.0}
\alias{if_2.0}
\title{Improved 'if' statement}
\usage{
if_2.0(x, operator, y, expr = NULL, alt.expr = NULL)
}
\arguments{
\item{x}{First element to be compared.}

\item{operator}{Relational operator to use to compare \code{x} and \code{y}.
One of \code{"=="}, \code{"!="}, \code{"<"}, \code{">"}, \code{"<="}, \code{">="}.}

\item{y}{Second element to be compared.}

\item{expr}{Expression to evaluate if the condition is met.}

\item{alt.expr}{Expression to evaluate if the condition is not met.}
}
\value{
If the condition expressed by \code{x}, \code{operator} and \code{y} is met and \code{expr} is not \code{NULL}:
result of the evaluation of \code{expr}.

If the condition is not met and \code{alt.expr} is not \code{NULL}:
result of the evaluation of \code{alt.expr}.

Logical value resulting of the comparison otherwise.
}
\description{
Perform a comparison between two elements and return the resulting logical value or evaluate
an expression.
}
\note{
This function is not exported from the package.
}
\seealso{
\code{\link[base:expression]{base::expression}}, \code{\link[base:Control]{base::if}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
