% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{init,SpectralAnalyzer-method}
\alias{init,SpectralAnalyzer-method}
\alias{init}
\title{Initialization of attributes}
\usage{
\S4method{init}{SpectralAnalyzer}(object, part = NULL, verbose = TRUE)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{part}{\code{NULL} or character value corresponding to the attributes to initialize.
One of the following:
\describe{
\item{\code{"nodes"}, \code{"n"}}{Attributes related to nodes.}
\item{\code{"nodes_links"}, \code{"nl"}}{Attributes related to links between nodes.}
\item{\code{"patterns"}, \code{"p"}}{Attributes related to patterns.}
\item{\code{"patterns_links"}, \code{"pl"}}{Attributes related to links between patterns.}
\item{\code{NULL}}{All attributes.}
}}

\item{verbose}{Logical value indicating whether to report progress.}
}
\value{
Invisible. \code{NULL} if \code{part} does not refer to patterns. If it does refer to patterns,
object of class \code{\link[arules:itemsets-class]{itemsets}} (from the package \code{arules}).
}
\description{
Initialize the attributes related to nodes, links between nodes, patterns and/or links between
patterns.
}
\details{
The initialization of the nodes consists of the initialization of the attributes \code{n_links} and
\code{nodes_links}.

The initialization of the links between nodes consists of the initialization of the attributes
\code{nodes} and \code{nodes_per_year}.

The initialization of the patterns consists of the initialization of the attributes \code{patterns},
\code{patterns_per_year} and \code{nodes_patterns}.

The initialization of the links between patterns consists of the initialization of the attributes
\code{p_links} and \code{patterns_links}.
}
\examples{
## Creating a SpectralAnalyzer and initialize some parts of it
obs <- make_observations(oedb_sample, by = "ID",
                         additional = c("CODE", "NAME", "YEAR"))
sa_object <- spectral.analyzer(obs, init = FALSE)
init(sa_object, "nodes")
init(sa_object, "nodes_links")

## Testing of its initialization
is_init(sa_object, "nodes")
is_init(sa_object, "patterns")
is_init(sa_object)

}
\seealso{
\code{\link{is_init}}, \code{\link{reset}}, \code{\link[arules:itemsets-class]{arules::itemsets}}.
}
\author{
Gauthier Magnin
}
