% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{init,TransactionAnalyzer-method}
\alias{init,TransactionAnalyzer-method}
\alias{init}
\title{Initialization of attributes}
\usage{
\S4method{init}{TransactionAnalyzer}(object, part = NULL, verbose = TRUE)
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}

\item{part}{\code{NULL} or character value corresponding to the attributes to initialize.
One of the following:
\describe{
\item{\code{"nodes"}, \code{"n"}}{Attributes related to nodes.}
\item{\code{"node_links"}, \code{"nl"}}{Attributes related to links between nodes.}
\item{\code{"patterns"}, \code{"p"}}{Attributes related to patterns.}
\item{\code{"pattern_links"}, \code{"pl"}}{Attributes related to links between patterns.}
\item{\code{NULL}}{All attributes.}
}}

\item{verbose}{Logical value indicating whether to report progress.}
}
\value{
Invisible. \code{NULL} if \code{part} does not refer to patterns. If it does refer to patterns,
object of class \code{\link[arules:itemsets-class]{itemsets}} (from the package \code{arules}).
}
\description{
Initialize the attributes related to nodes, links between nodes, patterns and/or links between
patterns.
}
\details{
The initialization of the nodes consists of the initialization of the attributes \code{nodes} and
\code{nodes_per_year}:
\enumerate{
\item{Enumeration of the transactions per year.}
\item{Enumeration of the nodes and calculation of the number of occurrences.}
}

The initialization of the links between nodes consists of the initialization of the attributes
\code{n_links} and \code{node_links}:
\enumerate{
\item{Counting links between nodes.}
\item{Elaboration of links between nodes.}
}

The initialization of the patterns consists of the initialization of the attributes \code{patterns},
\code{patterns_per_year} and \code{nodes_patterns}:
\enumerate{
\item{Mining for itemsets.}
\item{Linking nodes to patterns.}
\item{Enumeration of the patterns per year.}
\item{Computation of pattern characteristics.}
}

The initialization of the links between patterns consists of the initialization of the attributes
\code{p_links} and \code{pattern_links}:
\enumerate{
\item{Counting links between patterns.}
\item{Elaboration of links between patterns.}
}
}
\examples{
## Creating a TransactionAnalyzer and initialize some parts of it
trx <- make_transactions(oedb_sample, by = "ID",
                         additional = c("CODE", "NAME", "YEAR"))
trx_object <- transaction.set(data = trx, item_key = "CODE", year_key = "YEAR")
ta_object <- transaction.analyzer(trx_object, init = FALSE)
init(ta_object, "nodes")
init(ta_object, "node_links")

## Testing of its initialization
is_init(ta_object, "nodes")
is_init(ta_object, "patterns")
is_init(ta_object)

}
\seealso{
\code{\link{is_init}}, \code{\link{reset}}, \code{\link{transaction.analyzer}},
\code{\link[arules:itemsets-class]{arules::itemsets}}.
}
\author{
Gauthier Magnin
}
