% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interval}
\alias{interval}
\title{Mathematical interval as character string}
\usage{
interval(x1, x2, inc1 = TRUE, inc2 = TRUE)
}
\arguments{
\item{x1, x2}{Numeric values. Endpoints of the interval.}

\item{inc1, inc2}{If \code{TRUE}, the related endpoint is included in the interval.
If \code{FALSE}, it is excluded.}
}
\value{
String corresponding to the interval defined by the given arguments.
}
\description{
Create a string corresponding to the mathematical notation of an interval.
}
\note{
This function is not exported from the package.
}
\author{
Gauthier Magnin
}
\keyword{internal}
