% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_manager.R
\name{invert_list}
\alias{invert_list}
\title{Inversion of list}
\usage{
invert_list(x, by_name = FALSE)
}
\arguments{
\item{x}{List to invert.}

\item{by_name}{If \code{FALSE}, use the values of the elements of the list. If \code{TRUE}, use the names of
these values.}
}
\value{
List in which the values are the names of \code{x} and the names are the values found in the
elements of \code{x} or the names of these values (according to \code{by_name}).
}
\description{
Invert a list. Set the names of the list as the new values and set the values of the list as the
new names.
}
\examples{
l <- list(e1 = c(v1 = "E", v2 = "F", v3 = "I"),
          e2 = c(v1 = "I"),
          e3 = c(v1 = "C", v2 = "I"),
          e4 = c(v1 = "A", v2 = "C", v3 = "D", v4 = "F"),
          e5 = c(v1 = "B", v2 = "D", v3 = "E", v4 = "I", v5 = "G"))

invert_list(l, by_name = FALSE)
invert_list(l, by_name = TRUE)

}
\seealso{
\code{\link{turn_list_into_logical_matrix}}, \code{\link{turn_list_into_char_matrix}},
\code{\link{turn_list_into_data_frame}}, \code{\link{turn_logical_matrix_into_list}},
\code{\link{turn_char_matrix_into_list}}, \code{\link{turn_data_frame_into_list}}.
}
\author{
Gauthier Magnin
}
