% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_helper.R
\name{is_color}
\alias{is_color}
\title{Check color representation}
\usage{
is_color(x, int = TRUE)
}
\arguments{
\item{x}{Character vector.}

\item{int}{Logical indicating whether to consider integer values as valid representations.}
}
\value{
Logical vector giving for each value of \code{x} whether it is a valid color representation.
}
\description{
Check if a character value is a valid color representation.
The three kinds of R color specifications are considered as valid representations (color name,
hexadecimal string and positive integer).
}
\references{
Josh O'Brien. Stack Overflow topic:
\href{https://stackoverflow.com/questions/13289009/check-if-character-string-is-a-valid-color-representation}{Check if character string is a valid color representation}.
}
\seealso{
Color names: \code{\link{colors}}, hexadecimal values: \code{\link{rgb}}, graphics palette: \code{\link{palette}}.

Color to RGB conversion: \code{\link{col2rgb}}.
}
\keyword{internal}
