% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{is_init,SpectralAnalyzer-method}
\alias{is_init,SpectralAnalyzer-method}
\alias{is_init}
\title{Test of attribute initialization}
\usage{
\S4method{is_init}{SpectralAnalyzer}(object, part = NULL)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{part}{\code{NULL} or character value corresponding to the attributes to test. One of the following:
\describe{
\item{\code{"nodes"}, \code{"n"}}{Attributes related to nodes.}
\item{\code{"nodes_links"}, \code{"nl"}}{Attributes related to links between nodes.}
\item{\code{"patterns"}, \code{"p"}}{Attributes related to patterns.}
\item{\code{"patterns_links"}, \code{"pl"}}{Attributes related to links between patterns.}
}}
}
\value{
If \code{part} is not \code{NULL}, \code{TRUE} if the related attributes are initialized.
If \code{part} is \code{NULL}, 4 logical values corresponding to the initialization related to \code{nodes},
\code{nodes_links}, \code{patterns}, \code{patterns_links} (in this order).
}
\description{
Test if the attributes related to nodes, links between nodes, patterns or links between patterns
are initialized.
}
\details{
The initialization of the nodes consists of the initialization of the attributes \code{n_links} and
\code{nodes_links}.

The initialization of the links between nodes consists of the initialization of the attributes
\code{nodes} and \code{nodes_per_year}.

The initialization of the patterns consists of the initialization of the attributes \code{patterns},
\code{patterns_per_year} and \code{nodes_patterns}.

The initialization of the links between patterns consists of the initialization of the attributes
\code{p_links} and \code{patterns_links}.
}
\examples{
## Creating a SpectralAnalyzer and initialize some parts of it
obs <- make_observations(oedb_sample, by = "ID",
                         additional = c("CODE", "NAME", "YEAR"))
sa_object <- spectral.analyzer(obs, init = FALSE)
init(sa_object, "nodes")
init(sa_object, "nodes_links")

## Testing of its initialization
is_init(sa_object, "nodes")
is_init(sa_object, "patterns")
is_init(sa_object)

}
\seealso{
\code{\link{init}}.
}
\author{
Gauthier Magnin
}
