% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_named}
\alias{is_named}
\title{Check if a data structure is named}
\usage{
is_named(x)
}
\arguments{
\item{x}{A vector, matrix, data frame or list.}
}
\value{
Logical value or logical vector of length 2.
\itemize{
\item If \code{x} is a vector: \code{TRUE} or \code{FALSE} if the vector is named or not.
\item If \code{x} is a matrix or a data frame: logical vector of length 2.
The first value determines if the rows are named. The second one determines if the columns are named.
\item If \code{x} is a list: logical vector of length 2.
The first value determines if the list is named. The second one determines if the elements inside
the list are named.
}
}
\description{
Check if the elements of a data structure are named.
}
\note{
This function is not exported from the package.
}
\author{
Gauthier Magnin
}
\keyword{internal}
