% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{list_obs_per_year,SpectralAnalyzer-method}
\alias{list_obs_per_year,SpectralAnalyzer-method}
\alias{list_obs_per_year}
\title{Dénombrement des observations par année}
\usage{
\S4method{list_obs_per_year}{SpectralAnalyzer}(object)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}
}
\value{
Invisible. Matrice du nombre d'occurrences de chaque observation distincte, par année.
 Les lignes correspondent aux observations. Les colonnes correspondent aux années.
}
\description{
Identifie les observations distinctes par année et calcule le nombre de recrutements
de chacune de ces observations.
La matrice résultante est assignée à l'attribut \code{nodes_per_year}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
