% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{list_patterns_by_node,SpectralAnalyzer-method}
\alias{list_patterns_by_node,SpectralAnalyzer-method}
\alias{list_patterns_by_node}
\title{Linking nodes to patterns}
\usage{
\S4method{list_patterns_by_node}{SpectralAnalyzer}(object)
}
\arguments{
\item{object}{S4 object of class \code{SpectralAnalyzer}.}
}
\value{
Invisible. Logical matrix in which rows correspond to nodes and columns correspond to
 patterns. A value of \code{TRUE} means the pattern is included in the node.
}
\description{
Associate each separate observation (i.e. each node) with the patterns included in it.
The resulting matrix is assigned to the attribute \code{nodes_patterns} of \code{object}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
