% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{list_patterns_by_obs,SpectralAnalyzer-method}
\alias{list_patterns_by_obs,SpectralAnalyzer-method}
\title{Associe à chaque observation distincte les motifs qui y sont inclus.}
\usage{
\S4method{list_patterns_by_obs}{SpectralAnalyzer}(object)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}
}
\value{
Matrice de booléens dans laquelle les lignes correspondent aux observations et les colones aux motifs.
}
\description{
La matrice résultante est associée à l'attribut \code{obs_patterns}.
}
\author{
Gauthier Magnin
}
