% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{list_patterns_per_year,SpectralAnalyzer-method}
\alias{list_patterns_per_year,SpectralAnalyzer-method}
\alias{list_patterns_per_year}
\title{Dénombrement des motifs par année}
\usage{
\S4method{list_patterns_per_year}{SpectralAnalyzer}(object)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}
}
\value{
Invisible. Matrice des poids de chaque motif, par année.
 Les lignes correspondent aux motifs. Les colonnes correspondent aux années.
}
\description{
Compte le nombre d'apparitions de chaque motif selon l'année.
La matrice résultante est assignée à l'attribut \code{patterns_per_year}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
