% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{list_patterns_per_year,TransactionAnalyzer-method}
\alias{list_patterns_per_year,TransactionAnalyzer-method}
\alias{list_patterns_per_year}
\title{Enumeration of patterns per year}
\usage{
\S4method{list_patterns_per_year}{TransactionAnalyzer}(object)
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}
}
\value{
Invisible. Matrix of the number of occurrences of each pattern in the transactions, per year.
 The rows correspond to the patterns. The columns correspond to the years.
}
\description{
Count the number of occurrences of each pattern per year.
The resulting matrix is assigned to the attribute \code{patterns_per_year} of \code{object}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
