% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{list_separate_obs,SpectralAnalyzer-method}
\alias{list_separate_obs,SpectralAnalyzer-method}
\alias{list_separate_obs}
\title{Énumération des nœuds}
\usage{
\S4method{list_separate_obs}{SpectralAnalyzer}(object)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}
}
\value{
Invisible. Data frame des observations distinctes et de leurs caractéristiques
 (longueur et poids).
}
\description{
Identifie les observations distinctes et calcule la taille et le nombre de recrutements
de chacune de ces observations.
La data frame résultante est assignée à l'attribut \code{nodes}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
