% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{list_separate_trx,TransactionAnalyzer-method}
\alias{list_separate_trx,TransactionAnalyzer-method}
\alias{list_separate_trx}
\title{Enumeration of nodes}
\usage{
\S4method{list_separate_trx}{TransactionAnalyzer}(object)
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}
}
\value{
Invisible. Data frame of the separate transactions and their characteristics (length and
 weight).
}
\description{
Identify the separate transactions (considering only their respective items) and compute their number
 of items and number of occurrences.
The resulting data frame is assigned to the attribute \code{nodes} of \code{object}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
