% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{list_trx_per_year,TransactionAnalyzer-method}
\alias{list_trx_per_year,TransactionAnalyzer-method}
\alias{list_trx_per_year}
\title{Enumeration of the transactions per year}
\usage{
\S4method{list_trx_per_year}{TransactionAnalyzer}(object)
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}
}
\value{
Invisible. Matrix of the number of occurrences of each separate transaction, per year.
 The rows correspond to the transactions. The columns correspond to the years.
}
\description{
Identify the separate transactions per year (considering only their respective items) and count their
 number of occurrences for each one.
The resulting matrix is assigned to the attribute \code{nodes_per_year} of \code{object}.
}
\note{
This method is not exported from the package.
}
\author{
Gauthier Magnin
}
\keyword{internal}
