% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations_maker.R
\name{make_obs_from_unspecified_situations}
\alias{make_obs_from_unspecified_situations}
\title{Regroupement de prélèvements}
\usage{
make_obs_from_unspecified_situations(
  measures,
  variable_names,
  additional = NULL
)
}
\arguments{
\item{measures}{Ensemble de mesures de prélèvements.}

\item{variable_names}{Noms des variables à prendre en compte pour considérer une situation de travail.}

\item{additional}{Noms des informations à conserver lors de la constitution des observations.}
}
\value{
Liste des observations identifiées.
}
\description{
Regroupe des prélèvements sous forme d'observations.
Une observation correspond à un ensemble de substances retrouvées dans une même situation de travail.
Ces situations sont décrites par les combinaisons de variables dont les noms sont définis en paramètres.
}
\author{
Gauthier Magnin
}
\keyword{internal}
