% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations_maker.R
\name{make_obs_from_work_situations}
\alias{make_obs_from_work_situations}
\title{Regroupe des prélèvements sous forme d'observations.}
\usage{
make_obs_from_work_situations(measures, work_situations, additional = NULL)
}
\arguments{
\item{measures}{Ensemble de mesures de prélèvements.}

\item{work_situations}{Description de situations de travail.}

\item{additional}{Noms des informations à conserver lors de la constitution des observations.}
}
\value{
Liste contenant :
 \itemize{
   \item{} {La liste des observations identifiées.}
   \item{} {Un vecteur spécifiant pour chaque ligne de \code{measures} si elle a été intégrée dans une observation.}
 }
}
\description{
Une observation correspond à un ensemble de substances retrouvées dans une même situation de travail.
}
\author{
Gauthier Magnin
}
