% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_maker.R
\name{make_obs_from_ws}
\alias{make_obs_from_ws}
\title{Grouping samples according to specific work situations}
\usage{
make_obs_from_ws(
  measures,
  work_situations,
  additional = NULL,
  unique_values = TRUE
)
}
\arguments{
\item{measures}{Set of sampling measures. Data frame that must contain at least the following two
variables:
\describe{
  \item{\code{YEAR}}{Year of inspection.}
  \item{\code{CODE}}{Code identifying the substance being sampled.}
}}

\item{work_situations}{Description of work situations. Data frame that must contain the values
associated with certain variables from \code{measures} forming one work situation.
A \code{WS_ID} (\strong{W}ork \strong{S}ituation \strong{ID}entifier) column must make possible
to identify the different associations of values considered to be the same work situation.}

\item{additional}{Names of information to keep when creating observations: vector included in
\code{colnames(measures)}.}

\item{unique_values}{logical or character.
\itemize{
  \item{If \code{TRUE}, simplification of the values associated with the variables defined in
        \code{additional} in order to remove duplicates.}
  \item{If \code{FALSE}, keep the duplicates and the correspondence between the values of
        these variables for the same sample (as many values as there are samples).}
  \item{Otherwise, vector of variable names included in \code{colnames(measures)} for which the
        removal of duplicates must be performed.}
}}
}
\value{
\describe{
   \item{\code{observations}}{List of observations identified.
     Each observation is a list containing:
     \describe{
       \item{\code{CODE}}{The codes identifying the substances grouped together.}
       \item{\code{YEAR}}{The year of the inspection from which the samples are grouped.}
       \item{\code{...}}{Data related to the samples grouped, concerning the variables defined in
                         the argument \code{additional}.}
     }
   }
   \item{\code{processed}}{Vector specifying for each line of \code{measures} whether it has been
                           included in an observation.}
 }
}
\description{
Grouping samples in the form of observations: make observations from work situations.
An observation corresponds to a set of substances found in the same work situation.
}
\seealso{
\code{\link{make_observations}}, \code{\link{make_obs_from_unspecified_ws}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
