% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_maker.R
\name{make_OE_transactions}
\alias{make_OE_transactions}
\title{Grouping samples about Occupational Exposure}
\usage{
make_OE_transactions(
  measures,
  keys,
  mode,
  work_situations = NULL,
  variable_names = NULL,
  additional = NULL,
  unique_values = TRUE
)
}
\arguments{
\item{measures}{Set of sampling measures. Data frame that must contain at least three variables for:
\itemize{
\item An inspection identifier. The measures resulting from the same inspection must have the same
inspection identifier.
\item A code identifying the substance being sampled.
\item A year of inspection.
}}

\item{keys}{Length-three character vector identifying the names of the three required variables
in the following order: keys for accessing inspection identifiers; substance codes; years.}

\item{mode}{Mode of creation of transactions. One of \code{1}, \code{2}, \code{3}.
\describe{
\item{\code{1}}{Construction of transactions from the situations described by \code{work_situations}
and existing situations from combinations of values of variables defined by \code{variable_names}
that would not be described in \code{work_situations}.}
\item{\code{2}}{Construction of transactions only from the situations described by \code{work_situations}.
Remaining samples are ignored.}
\item{\code{3}}{Construction of transactions only from combinations of values of variables defined
by \code{variable_names}.}
}}

\item{work_situations}{Description of work situations. Data frame that must contain the values
associated with several variables from \code{measures} forming one work situation.
One column must make possible to identify the different associations of values considered to be the
same work situation. This column must be named \code{WS.ID} (for \strong{W}ork \strong{S}ituation \strong{ID}entifier)
otherwise the first column is considered to contain such identifiers.}

\item{variable_names}{Names of variables to be taken into account to consider a work situation,
i.e. vector of names of variables contained in \code{measures} whose different combinations of
values form the different work situations not described in \code{work_situations}.}

\item{additional}{Names of information to keep when creating transactions (in addition to those given
by \code{keys}): vector included in \code{colnames(measures)}.}

\item{unique_values}{logical or character.
\itemize{
\item If \code{TRUE}, simplification of the values associated with the variables defined in
\code{additional} in order to remove duplicates.
\item If \code{FALSE}, keep the duplicates and the correspondence between the values of these variables
for the same sample (as many values as there are samples).
\item Otherwise, vector of variable names included in \code{colnames(measures)} for which the removal of
duplicates must be performed.
}}
}
\value{
S4 object of class \code{TransactionSet} containing the list of transactions identified.
Each transaction is a list containing:
\itemize{
\item The codes identifying the substances grouped together.
\item The year of the inspection from which the samples are grouped.
\item Data related to the samples grouped, concerning the variables defined in the argument \code{additional}.
}
}
\description{
Grouping samples into a list of lists (called transactions) specifically for the occupational
exposure domain.
A transaction corresponds to a set of substances found in the same work situation during the same
inspection.
Specific situations are defined by the argument \code{work_situations}.
Non-specific situations are described by existing combinations of variables whose names are defined
by the argument \code{variable_names}.
}
\examples{
to_keep <- c("CODE", "NAME", "ACTIVITY", "JOB.TITLE", "SAMPLE.ID")
ws <- data.frame(WS.ID = c(1, 2, 2, 3, 3),
                 JOB.TITLE = c(44121004, 44142001, 44132032, 44132019, 44132030),
                 JOB.TASK = c("A5440", "A6410", "A5110", "A5260", "A5240"),
                 stringsAsFactors = FALSE)
ws_vars <- c("JOB.TITLE", "JOB.TASK")

## Make transactions from 3 specific work situations and as many other
## situations as JOB.TITLE, JOB.TASK pairs
trx_1 <- make_OE_transactions(oedb_sample, keys = c("ID", "CODE", "YEAR"),
                              mode = 1,
                              work_situations = ws,
                              variable_names = ws_vars,
                              additional = to_keep)

## Make transactions from specific work situations only
trx_2 <- make_OE_transactions(oedb_sample, keys = c("ID", "CODE", "YEAR"),
                              mode = 2,
                              work_situations = ws,
                              additional = to_keep)

## Make transactions where each pair of JOB.TITLE, JOB.TASK is one situation
trx_3 <- make_OE_transactions(oedb_sample, keys = c("ID", "CODE", "YEAR"),
                              mode = 3,
                              variable_names = ws_vars,
                              additional = to_keep)

}
\seealso{
For a more generic way to group data: \code{\link{make_transactions}}.
}
\author{
Gauthier Magnin
}
