% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_maker.R
\name{make_transactions}
\alias{make_transactions}
\title{Grouping data}
\usage{
make_transactions(
  data,
  by,
  additional = NULL,
  unique_values = TRUE,
  item_key = NA,
  year_key = NA
)
}
\arguments{
\item{data}{Data frame containing a set of data to be grouped.}

\item{by}{Variable names included in \code{colnames(data)} used to group the data.
Each combination of values of these variables generates one transaction.}

\item{additional}{Names of information to keep when creating transactions: vector included in
\code{colnames(data)}.}

\item{unique_values}{logical or character.
\itemize{
\item If \code{TRUE}, simplification of the values associated with the variables defined in \code{additional}
in order to remove duplicates.
\item If \code{FALSE}, keep the duplicates and the correspondence between the values of these variables for
the same row in \code{data} (as many values as there are rows grouped).
\item Otherwise, vector of variable names included in \code{additional} for which the removal of duplicates
must be performed.
}

Duplicates in the variables specified in \code{by} are necessarily removed in all cases.}

\item{item_key, year_key}{If the output have to be a \code{TransactionSet}, names of the variables of
\code{data} containing the items of the transactions and the year in which the transactions were made.}
}
\value{
List of transactions identified or S4 object of class \code{TransactionSet} containing this list
(depending on whether \code{item_key} is \code{NA} or not).

Each transaction is named by the combination of values of the variables named in \code{by} that generated
it. Each transaction is a list whose elements correspond to the values of the variables named in \code{by}
and \code{additional} which have been grouped together.
}
\description{
Grouping data into a list of lists called transactions.
}
\details{
Variables of type \code{factor} are coerced to \code{character}.
}
\examples{
to_keep <- c("CODE", "NAME", "ACTIVITY", "JOB.TITLE", "SAMPLE.ID")

## Make transactions without duplicate elements in the additional variables
trx_1 <- make_transactions(oedb_sample, by = "ID", additional = to_keep)

## Make transactions with duplicate elements correspondence
trx_2 <- make_transactions(oedb_sample, by = "ID", additional = to_keep,
                           unique_values = FALSE)
trx_3 <- make_transactions(oedb_sample, by = "ID", additional = to_keep,
                           unique_values = to_keep[1:2])

## Make transactions as TransactionSet
trx_object <- make_transactions(oedb_sample, by = "ID",
                                additional = c("CODE", "NAME", "YEAR"),
                                item_key = "CODE", year_key = "YEAR")

}
\seealso{
For the construction of well-targeted transactions specific to the field of occupational
exposure in relation to inspections and similar work situations: \code{\link{make_OE_transactions}}.
}
\author{
Gauthier Magnin
}
