% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_maker.R
\name{make_trx_from_ws}
\alias{make_trx_from_ws}
\title{Grouping samples according to specific work situations}
\usage{
make_trx_from_ws(
  measures,
  keys,
  work_situations,
  additional = NULL,
  unique_values = TRUE
)
}
\arguments{
\item{measures}{Set of sampling measures. Data frame that must contain at least two variables for:
\itemize{
\item A code identifying the substance being sampled.
\item A year of inspection.
}}

\item{keys}{Length-two character vector identifying the names of the two required variables
in the following order: keys for accessing substance codes; key for accessing years.}

\item{work_situations}{Description of work situations. Data frame that must contain the values
associated with several variables from \code{measures} forming one work situation.
One column must make possible to identify the different associations of values considered to be the
same work situation. This column must be named \code{WS.ID} (for \strong{W}ork \strong{S}ituation \strong{ID}entifier)
otherwise the first column is considered to contain such identifiers.}

\item{additional}{Names of information to keep when creating transactions (in addition to those given
by \code{keys}): vector included in \code{colnames(measures)}.}

\item{unique_values}{logical or character.
\itemize{
\item If \code{TRUE}, simplification of the values associated with the variables defined in
\code{additional} in order to remove duplicates.
\item If \code{FALSE}, keep the duplicates and the correspondence between the values of these variables
for the same sample (as many values as there are samples).
\item Otherwise, vector of variable names included in \code{colnames(measures)} for which the removal of
duplicates must be performed.
}}
}
\value{
\describe{
\item{\code{transactions}}{List of transactions identified. Each transaction is a list containing:
\itemize{
\item{The codes identifying the substances grouped together.}
\item{The year of the inspection from which the samples are grouped.}
\item{Data related to the samples grouped, concerning the variables defined in the argument
\code{additional}.}
}
}
\item{\code{processed}}{Vector specifying for each line of \code{measures} whether it has been included in
a transaction.}
}
}
\description{
Grouping samples in the form of transactions: make transactions from work situations.
A transaction corresponds to a set of substances found in the same work situation.
}
\seealso{
\code{\link{make_OE_transactions}}, \code{\link{make_trx_from_unspecified_ws}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
