% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixer.R
\name{maximum_cumulative_ratio}
\alias{maximum_cumulative_ratio}
\title{Maximum Cumulative Ratio (MCR)}
\usage{
maximum_cumulative_ratio(values, references)
maximum_cumulative_ratio(hi, mhq)
}
\arguments{
\item{values}{Numeric vector or matrix. Values for which the maximum cumulative ratio is to be computed.}

\item{references}{Numeric vector. Reference values associated with the values.}

\item{hi}{Numeric value or vector. \strong{H}azard \strong{i}ndex(es) on which to compute the maximum cumulative
ratio(s).}

\item{mhq}{Numeric value or vector. \strong{M}aximum \strong{h}azard \strong{q}uotient(s) associated with the hazard
index(es).}
}
\value{
Numeric value or vector (according to \code{values} or \code{hi} and \code{mhq}) of the maximum cumulative
ratio(s).
}
\description{
Compute the maximum cumulative ratio as the ratio between a hazard index and a maximum hazard quotient.
It represents the magnitude of the hazard that is underestimated by not performing a cumulative risk
assessment, given the values and references generating the hazard index and maximum hazard quotient.
}
\details{
If \code{values} is a matrix, the reference values are applied once on each column to compute the hazard
indexes (i.e. it must have one reference value for each row of the matrix) before searching for the
maximum hazard quotient then computing the maximum cumulative ratio. Thus, call the function with
the arguments \code{hi} and \code{mhq} is faster.

If \code{values} is a matrix (or \code{hi} and \code{mhq} are vectors larger than 1), one maximum cumulative ratio
is computed for each column (or value, respectively).

\loadmathjax
The maximum cumulative ratio of the vector \eqn{i} is given by:
\mjdeqn{MCR_i = \frac{HI_i}{MHQ_i}}{MCR_i = HI_i / MHQ_i}
where \eqn{HI} denotes the hazard index \code{hi} and \eqn{MHQ} denotes the maximum hazard quotient \code{mhq}.

The hazard index of the vector \eqn{i} is given by:
\mjdeqn{HI_i = \sum_{j = 1}^N HQ_{i,j}}{HI_i = sum(HQ_ij) from j = 1 to N}
where \eqn{HQ} denotes the hazard quotients and \eqn{N} denotes the number of hazard quotients.

The maximum hazard quotient of the vector \eqn{i} is given by:
\mjdeqn{MHQ_i = HQ_{M,i} = \max_{j \in \lbrace 1,...,N\rbrace} HQ_{i,j}}{MHQ_i = HQ_Mi = max HQ_i}
where \eqn{HQ} denotes the hazard quotients and \eqn{N} denotes the number of hazard quotients.

The hazard quotient of the value \eqn{j} in the vector \eqn{i} is given by:
\mjdeqn{HQ_{i,j} = \frac{V_{i,j}}{RV_j}}{HQ_ij = V_ij / RV_j}
where \eqn{V} denotes the \code{values} and \eqn{RV} denotes the \code{references}.
}
\note{
The multiple usages proposed implies the arguments \code{hi} and \code{mhq} to be explicitly named in the
function call.
}
\examples{
maximum_cumulative_ratio(c(1,2,3,4,5), c(1,2,3,4,5))
maximum_cumulative_ratio(hi = hazard_index(c(1,2,3,4,5), c(1,2,3,4,5)),
                         mhq = maximum_hazard_quotient(c(1,2,3,4,5), c(1,2,3,4,5)))
maximum_cumulative_ratio(values = matrix(c(.1, .2, .3, .4, .5, .6, .7, .8, .9, 1),
                                         ncol = 2),
                         references = c(1,2,3,4,5))

}
\references{
Price PS, Han X (2011).
Maximum cumulative ratio (MCR) as a tool for assessing the value of performing a cumulative risk assessment.
\emph{International Journal of Environmental Research and Public Health}. 8(6): 2212-2225.
\url{https://doi.org/10.3390/ijerph8062212}.
}
\seealso{
\code{\link{hazard_index}}, \code{\link{maximum_hazard_quotient}}, \code{\link{reciprocal_of_mcr}}.
}
\author{
Gauthier Magnin
}
