% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixer.R
\name{maximum_hazard_quotient}
\alias{maximum_hazard_quotient}
\title{Maximum Hazard Quotient (MHQ)}
\usage{
maximum_hazard_quotient(values, references)
maximum_hazard_quotient(hq)
}
\arguments{
\item{values}{Numeric vector or matrix. Values for which to search for the maximum hazard quotient.}

\item{references}{Numeric vector. Reference values associated with the \code{values}.}

\item{hq}{Numeric vector or matrix. \strong{H}azard \strong{q}uotients on which to search for the maximum.}
}
\value{
Numeric value or vector (according to \code{values} or \code{hq}) of the maximum hazard quotient(s).
}
\description{
Search for the maximum of hazard quotients, also called the primary hazard quotient.
}
\details{
Arguments \code{values} and \code{references} are used to compute the hazard quotients before searching for
the maximum hazard quotient. Thus, call the function with the argument \code{hq} is faster (if it is
already computed).

If \code{values} is a matrix, the reference values are applied once on each column (i.e. it must have one
reference value for each row of the matrix).

If \code{values} or \code{hq} is a matrix, one maximum hazard quotient is searched for each column.

\loadmathjax
The maximum hazard quotient of the vector \eqn{i} is given by:
\mjdeqn{MHQ_i = HQ_{M,i} = \max_{j \in \lbrace 1,...,n\rbrace} HQ_{i,j}}{MHQ_i = HQ_Mi = max HQ_i}
where \eqn{HQ} denotes the hazard quotients and \eqn{n} denotes the number of hazard quotients.

The hazard quotient of the value \eqn{j} in the vector \eqn{i} is given by:
\mjdeqn{HQ_{i,j} = \frac{V_{i,j}}{RV_j}}{HQ_ij = V_ij / RV_j}
where \eqn{V} denotes the \code{values} and \eqn{RV} denotes the \code{references}.
}
\note{
Due to the multiple possible usages, the argument \code{hq} must be explicitly named in the function call.
}
\examples{
maximum_hazard_quotient(c(1,2,3,4,5), c(1,2,3,4,5))
maximum_hazard_quotient(hq = hazard_quotient(c(1,2,3,4,5), c(1,2,3,4,5)))
maximum_hazard_quotient(values = matrix(c(.1, .2, .3, .4, .5, .6, .7, .8, .9, 1),
                                        ncol = 2),
                        references = c(1,2,3,4,5))

}
\references{
Price PS, Han X (2011).
Maximum cumulative ratio (MCR) as a tool for assessing the value of performing a cumulative risk assessment.
\emph{International Journal of Environmental Research and Public Health}. 8(6): 2212-2225.
\url{https://doi.org/10.3390/ijerph8062212}.
}
\seealso{
\code{\link{hazard_quotient}}, \code{\link{top_hazard_quotient}}.
}
\author{
Gauthier Magnin
}
