% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixer.R
\name{mcr_approach_by_class}
\alias{mcr_approach_by_class}
\title{MCR approach by class}
\usage{
mcr_approach_by_class(values, references, classes, FUN, ...)
}
\arguments{
\item{values}{Numeric named vector or matrix, or list of numeric named vectors.
Values on which to perform the MCR approach according to classes.}

\item{references}{Numeric vector or list of numeric vectors. Reference values associated with the
\code{values}. See 'Details' to know the way it is associated with \code{values}.}

\item{classes}{List or logical matrix associating the \code{values} names with classes.
If list, its names are those present in \code{values} and the elements are vectors of associated classes.
If logical matrix, its columns are named according to the classes and the row names
contain the names associated with the \code{values}. A \code{TRUE} value indicates that a specific name
is part of a specific class.}

\item{FUN}{Either a function or a non-empty character string naming the function to apply on each
class, among \code{mcr_summary}, \code{mcr_chart}, \code{thq_pairs_freq}, \code{thq_freq_by_group}.}

\item{...}{Further arguments to the function \code{FUN}.}
}
\value{
See 'Value' of the corresponding help page:
\itemize{
\item \code{\link{mcr_summary_by_class}} if \code{FUN} is \code{mcr_summary}.
\item \code{\link{mcr_chart_by_class}} if \code{FUN} is \code{mcr_chart}.
\item \code{\link{thq_pairs_freq_by_class}} if \code{FUN} is \code{thq_pairs_freq}.
\item \code{\link{thq_freq_by_group_by_class}} if \code{FUN} is \code{thq_freq_by_group}.
}
}
\description{
Perform the MCR approach according to classes, given values and references. Wrapper of the four
functions allowing to perform the MCR pproach according to classes: \code{mcr_summary_by_class},
\code{mcr_chart_by_class}, \code{thq_pairs_freq_by_class}, \code{thq_freq_by_group_by_class}.
}
\details{
If \code{values} is a vector, the reference values are directly associated with these values.

If \code{values} is a matrix, the reference values are applied once on each column (i.e. it must have one
reference value for each row of the matrix).

If \code{values} is a list, the reference values can be a vector of named values or a list. In this case,
if \code{references} is a vector, there must be one reference for each name present in \code{values}.
Otherwise, \code{references} is a list of vectors having the same lengths as those present in \code{values}
so that \code{values} and \code{references} can be matched.
}
\examples{
## Creating a matrix of 5*50 values, one reference value for each of the 5
## elements (A, B, C, D and E) and association of classes (C1 to C8) with
## these elements.
v <- matrix(sample(seq(0.1, 2.1, by = 0.1), 250, replace = TRUE),
            ncol = 50, dimnames = list(LETTERS[1:5]))
r <- sample(seq(1,5), 5, replace = TRUE)
classes <- list(A = c("C5", "C6", "C8"),
                B = "C8",
                C = c("C3", "C8"),
                D = c("C1", "C3", "C4", "C6"),
                E = c("C2", "C4", "C5", "C7", "C8"))

## Application to the 4 possible functions
mcr_approach_by_class(values = c(A = 1, B = 2, C = 3, D = 4, E = 5),
                      references = r,
                      classes,
                      mcr_summary)

mcr_approach_by_class(v, r, classes, mcr_chart,
                      regions_lab = c(TRUE, TRUE, FALSE, TRUE),
                      regression = TRUE,
                      log_transform = FALSE)$C8

mcr_approach_by_class(values = list(V1 = c(A = 1, B = 5),
                                    V2 = c(A = 2),
                                    V3 = c(B = 3, C = 4)),
                      references = list(c(1, 2),
                                        1,
                                        c(2, 3)),
                      classes,
                      FUN = "thq_pairs_freq", levels = names(classes))

mcr_approach_by_class(values = list(V1 = c(A = 1, B = 5),
                                    V2 = c(A = 2),
                                    V3 = c(B = 3, C = 4)),
                      references = c(A = 1, B = 2, C = 3),
                      classes,
                      FUN = "thq_freq_by_group", levels = names(classes))

}
\seealso{
\code{\link{mcr_summary_by_class}}, \code{\link{mcr_chart_by_class}}, \code{\link{thq_pairs_freq_by_class}},
\code{\link{thq_freq_by_group_by_class}}.
}
\author{
Gauthier Magnin
}
