% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixer.R
\name{mcr_chart_by_class}
\alias{mcr_chart_by_class}
\title{MCR approach scatter plot by class}
\usage{
mcr_chart_by_class(
  values,
  references,
  classes,
  thq_col = NULL,
  regions = FALSE,
  regions_col = c("#b3cde3", "#edf8fb", "#8c96c6", "#88419d"),
  regions_alpha = 0.2,
  regions_lab = !regions,
  regression = FALSE,
  log_transform = TRUE,
  plot = FALSE
)
}
\arguments{
\item{values}{Numeric named matrix or list of numeric named vectors. Vectors of values for which the
chart are to be created, according to classes.}

\item{references}{Numeric vector or list of numeric vectors. Reference values associated with the
\code{values}. See 'Details' to know the way it is associated with \code{values}.}

\item{classes}{List or logical matrix associating the \code{values} names with classes.
If list, its names are those present in \code{values} and the elements are vectors of associated classes.
If logical matrix, its columns are named according to the classes and the row names
contain the names associated with the \code{values}. A \code{TRUE} value indicates that a specific name
is part of a specific class.}

\item{thq_col}{Character named vector. Colors to assign to the \strong{t}op \strong{h}azard \strong{q}uotients
elements.}

\item{regions}{If \code{TRUE}, the regions corresponding to the MIAT groups are filled with the colors
defined by \code{regions_col}.}

\item{regions_col}{Character vector of length 4. Define the colors for the regions of the MIAT groups
(in order: I, II, IIIA and IIIB).}

\item{regions_alpha}{Value between 0 and 1. Opacity of the regions filled with \code{regions_col}.}

\item{regions_lab}{Logical value or vector of length 4. Define if labels for the MIAT groups should
be displayed (in order: I, II, IIIA, IIIB). \code{TRUE} and \code{FALSE} are special values for all \code{TRUE} or
all \code{FALSE}.}

\item{regression}{If \code{TRUE}, the linear regression between the X and Y coordinates of the points
is represented with 95\% confidence interval.}

\item{log_transform}{If \code{TRUE}, the log version of the charts are created (i.e.
\mjeqn{log_{10}(HI)}{log10(HI)} versus \mjeqn{log_{10}(MCR - 1)}{log10(MCR - 1)}). If \code{FALSE},
the standard version of the charts are created (i.e. \eqn{HI} versus \eqn{MCR}).}

\item{plot}{If \code{FALSE} and the returned list is assigned, the charts are not plotted.
Otherwise, the charts are all plotted in the active graphics device.
But if \code{TRUE} and the returned list is not assigned, the charts are plotted twice.}
}
\value{
List of charts created with the \code{ggplot2} package or \code{NULL} if no points can be plotted
(see 'Details'). The length of the list corresponds to the number of classes encountered.
}
\description{
Create charts of \mjeqn{log_{10}(HI)}{log10(HI)} versus \mjeqn{log_{10}(MCR - 1)}{log10(MCR - 1)}
with the percentage of the reciprocal of the maximum cumulative ratio, the elements producing the
top hazard quotients and the associated MIAT groups, according to classes. For each class, one chart
is created from the subset of values corresponding to this class.
}
\details{
If \code{values} is a matrix, the reference values are applied once on each set of values, i.e. on each row.
Therefore, there must be one reference value for each column of the matrix.

If \code{values} is a list, the reference values can be a vector of named values or a list. In this case,
if \code{references} is a vector, there must be one reference for each name present in \code{values}.
Otherwise, \code{references} is a list of vectors having the same lengths as those present in \code{values}
so that \code{values} and \code{references} can be matched.

If \code{classes} is a list, it will be turned into a logical matrix before processing. Thus, call the
function with such a matrix is slightly faster.

The charts being created with the \code{ggplot2} package, they can be modified or completed afterwards
using the returned objects.

Color specification can be done using the R predefined color names or hexadecimal values.

In the standard version of the chart, the grey area represents the region in which no point can be
plotted because \eqn{MCR} cannot be lower than 1. In the log version, such a region does not exist.
However, in the latter, points having \eqn{MCR} equal to 1 have an ordinate equal to \code{-Inf} and
therefore cannot be plotted. Moreover, whatever the version, sets of values containing only values
equal to \eqn{0} (i.e., having \eqn{MCR} values equal to \eqn{0}) are not plotted.

Points that are on the boundaries of the region of the group I belong to this group.
Points that are on the boundaries of the region of the group II belong to this group.
Points that are on the boundary between the regions of the groups IIIA and IIIB belong to the
group IIIB.

\loadmathjax

The mixtures are assigned to the groups according the following conditions:
\itemize{
\item Group I: \mjeqn{MHQ_i \ge 1}{MHQ_i >= 1}
\item Group II: \mjeqn{MHQ_i < 1, HI_i \le 1}{MHQ_i < 1, HI_i <= 1}
\item Group IIIA: \mjeqn{MHQ_i < 1, HI_i > 1, MCR_i < 2}{MHQ_i < 1, HI_i > 1, MCR_i < 2}
\item Group IIIB: \mjeqn{MHQ_i < 1, HI_i > 1, MCR_i \ge 2}{MHQ_i < 1, HI_i > 1, MCR_i >= 2}
}

The maximum cumulative ratio of the vector \eqn{i} is given by:
\mjdeqn{MCR_i = \frac{HI_i}{MHQ_i}}{MCR_i = HI_i / MHQ_i}
where \eqn{HI} denotes the hazard index and \eqn{MHQ} denotes the maximum hazard quotient.
\cr
In the specific case where all values of the vector (or the corresponding \eqn{HI} and \eqn{MHQ})
are equal to \eqn{0}, the resulting \eqn{MCR} is \eqn{0}.

The reciprocal of the maximum cumulative ratio of the vector \eqn{i} is given by:
\mjdeqn{Reciprocal~of~MCR_i = \frac{1}{MCR_i} = \frac{MHQ_i}{HI_i}}{Reciprocal of MCR_i = 1 / MCR_i = MHQ_i / HI_i}
where \eqn{MCR}, \eqn{MHQ} and {HI} denotes the maximum cumulative ratio, the maximum hazard
quotient and the hazard index respectively.
\cr
In the specific case where all values of the vector (or the corresponding \eqn{MHQ} and \eqn{HI})
are equal to \eqn{0}, the resulting reciprocal of \eqn{MCR} is \eqn{0}.

The hazard index of the vector \eqn{i} is given by:
\mjdeqn{HI_i = \sum_{j = 1}^n HQ_{i,j}}{HI_i = sum(HQ_ij) from j = 1 to n}
where \eqn{HQ} denotes the hazard quotients and \eqn{n} denotes the number of hazard quotients.

The maximum hazard quotient of the vector \eqn{i} is given by:
\mjdeqn{MHQ_i = HQ_{M,i} = \max_{j \in \lbrace 1,...,n\rbrace} HQ_{i,j}}{MHQ_i = HQ_Mi = max HQ_i}
where \eqn{HQ} denotes the hazard quotients and \eqn{n} denotes the number of hazard quotients.

The hazard quotient of the value \eqn{j} in the vector \eqn{i} is given by:
\mjdeqn{HQ_{i,j} = \frac{V_{i,j}}{RV_j}}{HQ_ij = V_ij / RV_j}
where \eqn{V} denotes the \code{values} and \eqn{RV} denotes the \code{references}.

If several values are equal to the maximum hazard quotient, the name retained as the top hazard
quotient is the first one considering the given order.
}
\examples{
## Creating a matrix of 50*5 values and one reference value for each of the 5
## elements (A, B, C, D and E) and association of classes (C1 to C8) with
## these elements.
v <- matrix(sample(seq(0.1, 1.1, by = 0.1), 250, replace = TRUE),
            nrow = 50, byrow = TRUE,
            dimnames = list(paste0("set.", 1:50), LETTERS[1:5]))
r <- sample(seq(1,5), 5, replace = TRUE)
classes <- list(A = c("C5", "C6", "C8"),
                B = "C8",
                C = c("C3", "C8"),
                D = c("C1", "C3", "C4", "C6"),
                E = c("C2", "C4", "C5", "C7", "C8"))

## MCR charts on matrices
charts1 <- mcr_chart_by_class(v, r, classes, regions = TRUE)
View(charts1)
plot(charts1$C3)

charts2 <- mcr_chart_by_class(v, r, classes,
                              regions = TRUE, log_transform = FALSE)
View(charts2)
plot(charts2$C3)

## MCR charts on lists
charts3 <- mcr_chart_by_class(values = list(set.1 = c(A = 0.1, B = 0.5),
                                            set.2 = c(A = 0.2),
                                            set.3 = c(B = 0.3, C = 0.4)),
                              references = list(c(1, 2),
                                                1,
                                                c(2, 3)),
                              classes,
                              log_transform = TRUE)
View(charts3)
plot(charts3$C8)

charts4 <- mcr_chart_by_class(values = list(set.1 = c(A = 0.1, B = 0.5),
                                            set.2 = c(A = 0.2),
                                            set.3 = c(B = 0.3, C = 0.4)),
                              references = c(A = 1, B = 2, C = 3),
                              classes,
                              log_transform = FALSE)
View(charts4)
plot(charts4$C8)

}
\references{
Reyes JM, Price PS (2018).
An analysis of cumulative risks based on biomonitoring data for six phthalates using the Maximum Cumulative Ratio.
\emph{Environment International}, 112, 77-84.
\url{https://doi.org/10.1016/j.envint.2017.12.008}.

De Brouwere K, et al. (2014).
Application of the maximum cumulative ratio (MCR) as a screening tool for the evaluation of mixtures in residential indoor air.
\emph{The Science of the Total Environment}, 479-480, 267-276.
\url{https://doi.org/10.1016/j.scitotenv.2014.01.083}.
}
\seealso{
Chart independent of classes: \code{\link{mcr_chart}}.

Generic function to apply the MCR approach according to classes: \code{\link{mcr_approach_by_class}}.

Other functions of the MCR approach applying according to classes: \code{\link{mcr_summary_by_class}},
\code{\link{thq_pairs_by_class}}, \code{\link{thq_by_group_by_class}}.

Specific indicators: \code{\link{hazard_index}}, \code{\link{maximum_cumulative_ratio}}, \code{\link{reciprocal_of_mcr}},
\code{\link{top_hazard_quotient}}, \code{\link{classify_mixture}}.
}
\author{
Gauthier Magnin
}
