% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixer.R
\name{mcr_summary_by_class}
\alias{mcr_summary_by_class}
\title{Summary by class of indicators of the MCR approach}
\usage{
mcr_summary_by_class(values, references, classes)
}
\arguments{
\item{values}{Numeric named vector or matrix, or list of numeric named vectors.
Values whose indicators of the MCR approach are to be computed according to classes.}

\item{references}{Numeric vector or list of numeric vectors. Reference values associated with the
\code{values}. See 'Details' to know the way it is associated with \code{values}.}

\item{classes}{List or logical matrix associating the \code{values} names with classes.
If list, its names are those present in \code{values} and the elements are vectors of associated classes.
If logical matrix, its columns are named according to the classes and the row names
contain the names associated with the \code{values}. A \code{TRUE} value indicates that a specific name
is part of a specific class.}
}
\value{
Data frame or list of data frames (according to \code{values}) of the main indicators of the MCR
approach, computed on the given \code{values} and for each class encountered:
\itemize{
\item \strong{HI}: Hazard Index.
\item \strong{MCR}: Maximum Cumulative Ratio.
\item \strong{Reciprocal}: Reciprocal of the maximum cumulative ratio.
\item \strong{Group}: MIAT group.
\item \strong{THQ}: Top Hazard Quotient.
\item \strong{MHQ}: Maximum Hazard Quotient.
\item \strong{Missed}: Hazard missed if a cumulative risk assessment is not performed.
}
}
\description{
Compute a set of indicators of the MCR approach according to classes, given values and references.
Indicators are computed for each set of values and for each class. For each class, only values
corresponding to this class are considered.
}
\details{
If \code{values} is a vector, the reference values are directly associated with these values.

If \code{values} is a matrix, the reference values are applied once on each column (i.e. it must have one
reference value for each row of the matrix).

If \code{values} is a list, the reference values can be a vector of named values or a list. In this case,
if \code{references} is a vector, there must be one reference for each name present in \code{values}.
Otherwise, \code{references} is a list of vectors having the same lengths as those present in \code{values}
so that \code{values} and \code{references} can be matched.

\loadmathjax
The hazard quotient of the value \eqn{j} in the vector \eqn{i} is given by:
\mjdeqn{HQ_{i,j} = \frac{V_{i,j}}{RV_j}}{HQ_ij = V_ij / RV_j}
where \eqn{V} denotes the \code{values} and \eqn{RV} denotes the \code{references}.

The maximum hazard quotient of the vector \eqn{i} is given by:
\mjdeqn{MHQ_i = HQ_{M,i} = \max_{j \in \lbrace 1,...,N\rbrace} HQ_{i,j}}{MHQ_i = HQ_Mi = max HQ_i}
where \eqn{N} denotes the number of hazard quotients.

The hazard index of the vector \eqn{i} is given by:
\mjdeqn{HI_i = \sum_{j = 1}^N HQ_{i,j}}{HI_i = sum(HQ_ij) from j = 1 to N}

The maximum cumulative ratio of the vector \eqn{i} is given by:
\mjdeqn{MCR_i = \frac{HI_i}{MHQ_i}}{MCR_i = HI_i / MHQ_i}

The reciprocal of the maximum cumulative ratio of the vector \eqn{i} is given by:
\mjdeqn{Reciprocal~of~MCR_i = \frac{1}{MCR_i} = \frac{MHQ_i}{HI_i}}{Reciprocal of MCR_i = 1 / MCR_i = MHQ_i / HI_i}

The missed toxicity of the vector \eqn{i} is given by:
\mjdeqn{Missed~toxicity_i = 1 - \frac{1}{MCR_i}}{Missed toxiciy_i = 1 - 1 / MCR_i}

The mixtures are assigned to the groups according the following conditions:
\itemize{
\item Group I: \mjeqn{MHQ_i \ge 1}{MHQ_i >= 1}
\item Group II: \mjeqn{MHQ_i < 1, HI_i \le 1}{MHQ_i < 1, HI_i <= 1}
\item Group IIIA: \mjeqn{MHQ_i < 1, HI_i > 1, MCR_i < 2}{MHQ_i < 1, HI_i > 1, MCR_i < 2}
\item Group IIIB: \mjeqn{MHQ_i < 1, HI_i > 1, MCR_i \ge 2}{MHQ_i < 1, HI_i > 1, MCR_i >= 2}
}
}
\examples{
## Association of classes (C1 to C8) with elements A, B, C, D and E
classes <- list(A = c("C5", "C6", "C8"),
                B = "C8",
                C = c("C3", "C8"),
                D = c("C1", "C3", "C4", "C6"),
                E = c("C2", "C4", "C5", "C7", "C8"))

## MCR summary by class on vectors
mcr_summary_by_class(values = c(A = 1, B = 2, C = 3, D = 4, E = 5),
                     references = c(1,2,3,4,5),
                     classes = classes)
mcr_summary_by_class(values = c(A = 1, B = 2, C = 3, D = 4),
                     references = c(1, 2, 3, 4),
                     classes)

## MCR summary by class on matrices
mcr_summary_by_class(values = matrix(c(0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1.),
                                     ncol = 2,
                                     dimnames = list(LETTERS[1:5],
                                                     c("V1", "V2"))),
                     references = c(1,2,3,4,5),
                     classes)
mcr_summary_by_class(values = matrix(c(0.1,0.2,0.3,0.4,0.6,0.7,0.8,0.9),
                                     ncol = 2,
                                     dimnames = list(LETTERS[1:4],
                                                     c("V1", "V2"))),
                     references = c(1,2,3,4),
                     classes)

## MCR summary by class on lists
mcr_summary_by_class(values = list(V1 = c(A = 0.1, B = 0.5),
                                   V2 = c(A = 0.2),
                                   V3 = c(B = 0.3, C = 0.4)),
                     references = c(A = 1, B = 2, C = 3),
                     classes)
mcr_summary_by_class(values = list(V1 = c(A = 0.1, B = 0.5),
                                   V2 = c(A = 0.2),
                                   V3 = c(B = 0.3, C = 0.4)),
                     references = list(c(1, 2),
                                       1,
                                       c(2, 3)),
                     classes)

}
\references{
Price PS, Han X (2011).
Maximum cumulative ratio (MCR) as a tool for assessing the value of performing a cumulative risk assessment.
\emph{International Journal of Environmental Research and Public Health}. 8(6): 2212-2225.
\url{https://doi.org/10.3390/ijerph8062212}.

Reyes JM, Price PS (2018).
An analysis of cumulative risks based on biomonitoring data for six phthalates using the Maximum Cumulative Ratio.
\emph{Environment International}, 112, 77-84.
\url{https://doi.org/10.1016/j.envint.2017.12.008}.

De Brouwere K, et al. (2014).
Application of the maximum cumulative ratio (MCR) as a screening tool for the evaluation of mixtures in residential indoor air.
\emph{The Science of the Total Environment}, 479-480, 267-276.
\url{https://doi.org/10.1016/j.scitotenv.2014.01.083}.
}
\seealso{
\code{\link{mcr_summary}}, \code{\link{mcr_chart_by_class}}, \code{\link{thq_pairs_freq_by_class}},
\code{\link{thq_freq_by_group_by_class}}.
}
\author{
Gauthier Magnin
}
