% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixer.R
\name{mcr_summary_for_list}
\alias{mcr_summary_for_list}
\title{Summary of indicators of the MCR approach, on list}
\usage{
mcr_summary_for_list(values, references)
}
\arguments{
\item{values}{List of numeric vectors. Values whose indicators of the MCR approach are to be
computed.}

\item{references}{Numeric named vector or list of numeric vectors. Reference values associated with
the \code{values}. See 'Details' to know the way it is associated with \code{values}.}
}
\value{
Data frame containing the main indicators of the MCR approach computed on the given \code{values}:
\itemize{
\item \strong{n}: number of values different from 0.
\item \strong{HI}: Hazard Index.
\item \strong{MCR}: Maximum Cumulative Ratio.
\item \strong{Reciprocal}: Reciprocal of the maximum cumulative ratio.
\item \strong{Group}: MIAT group.
\item \strong{THQ}: Top Hazard Quotient.
\item \strong{MHQ}: Maximum Hazard Quotient.
\item \strong{Missed}: Hazard missed if a cumulative risk assessment is not performed.
}
}
\description{
Compute a set of indicators of the MCR approach, given values and references.
}
\details{
If \code{references} is a vector, there must be one reference for each name present in \code{values}.
Otherwise, \code{references} is a list of vectors having the same lengths as those present in \code{values}
so that \code{values} and \code{references} can be matched.

If several values are equal to the maximum hazard quotient, the name retained as the top hazard
quotient is the first one considering the given order.

\loadmathjax
The hazard quotient of the value \eqn{j} in the vector \eqn{i} is given by:
\mjdeqn{HQ_{i,j} = \frac{V_{i,j}}{RV_j}}{HQ_ij = V_ij / RV_j}
where \eqn{V} denotes the \code{values} and \eqn{RV} denotes the \code{references}.

The maximum hazard quotient of the vector \eqn{i} is given by:
\mjdeqn{MHQ_i = HQ_{M,i} = \max_{j \in \lbrace 1,...,n\rbrace} HQ_{i,j}}{MHQ_i = HQ_Mi = max HQ_i}
where \eqn{n} denotes the number of hazard quotients.

The hazard index of the vector \eqn{i} is given by:
\mjdeqn{HI_i = \sum_{j = 1}^n HQ_{i,j}}{HI_i = sum(HQ_ij) from j = 1 to n}

The maximum cumulative ratio of the vector \eqn{i} is given by:
\mjdeqn{MCR_i = \frac{HI_i}{MHQ_i}}{MCR_i = HI_i / MHQ_i}

The reciprocal of the maximum cumulative ratio of the vector \eqn{i} is given by:
\mjdeqn{Reciprocal~of~MCR_i = \frac{1}{MCR_i} = \frac{MHQ_i}{HI_i}}{Reciprocal of MCR_i = 1 / MCR_i = MHQ_i / HI_i}

The missed toxicity of the vector \eqn{i} is given by:
\mjdeqn{Missed~toxicity_i = 1 - \frac{1}{MCR_i}}{Missed toxiciy_i = 1 - 1 / MCR_i}

The mixtures are assigned to the groups according the following conditions:
\itemize{
\item Group I: \mjeqn{MHQ_i \ge 1}{MHQ_i >= 1}
\item Group II: \mjeqn{MHQ_i < 1, HI_i \le 1}{MHQ_i < 1, HI_i <= 1}
\item Group IIIA: \mjeqn{MHQ_i < 1, HI_i > 1, MCR_i < 2}{MHQ_i < 1, HI_i > 1, MCR_i < 2}
\item Group IIIB: \mjeqn{MHQ_i < 1, HI_i > 1, MCR_i \ge 2}{MHQ_i < 1, HI_i > 1, MCR_i >= 2}
}
}
\note{
This function is not exported from the package.
}
\references{
Price PS, Han X (2011).
Maximum cumulative ratio (MCR) as a tool for assessing the value of performing a cumulative risk assessment.
\emph{International Journal of Environmental Research and Public Health}. 8(6): 2212-2225.
\url{https://doi.org/10.3390/ijerph8062212}.

Reyes JM, Price PS (2018).
An analysis of cumulative risks based on biomonitoring data for six phthalates using the Maximum Cumulative Ratio.
\emph{Environment International}, 112, 77-84.
\url{https://doi.org/10.1016/j.envint.2017.12.008}.

De Brouwere K, et al. (2014).
Application of the maximum cumulative ratio (MCR) as a screening tool for the evaluation of mixtures in residential indoor air.
\emph{The Science of the Total Environment}, 479-480, 267-276.
\url{https://doi.org/10.1016/j.scitotenv.2014.01.083}.
}
\seealso{
\code{\link{mcr_summary}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
