% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixer.R
\name{missed_toxicity}
\alias{missed_toxicity}
\title{Missed Toxicity}
\usage{
missed_toxicity(values, references)
missed_toxicity(mcr)
}
\arguments{
\item{values}{Numeric vector or matrix. Values for which the missed toxicity is to be computed.}

\item{references}{Numeric vector. Reference values associated with the \code{values}.}

\item{mcr}{Numeric value or vector. \strong{M}aximum \strong{c}umulative \strong{r}atio(s) for which to compute the
missed toxicity.}
}
\value{
Numeric value or vector (according to \code{values} or \code{mcr}) of the missed toxicity.
}
\description{
Compute the hazard missed if a cumulative risk assessment is not performed, given the values and
references generating the maximum cumulative ratio.
}
\details{
Arguments \code{values} and \code{references} are used to compute the hazard quotients and the hazard index
before searching for the maximum hazard quotient, computing the maximum cumulative ratio then
computing the missed toxicity. Thus, call the function with the argument \code{mcr} is faster (if it is
already computed).

If \code{values} is a matrix, the reference values are applied once on each set of values, i.e. on each row.
Therefore, there must be one reference value for each column of the matrix.

If \code{values} is a matrix (or \code{mcr} is a vector larger than 1), one missed toxicity value is computed
for each row (or value, respectively).

\loadmathjax

The missed toxicity of the vector \eqn{i} is given by:
\mjdeqn{Missed~toxicity_i = 1 - \frac{1}{MCR_i}}{Missed toxiciy_i = 1 - 1 / MCR_i}
where \eqn{MCR} denotes the maximum cumulative ratio.

The maximum cumulative ratio of the vector \eqn{i} is given by:
\mjdeqn{MCR_i = \frac{HI_i}{MHQ_i}}{MCR_i = HI_i / MHQ_i}
where \eqn{HI} denotes the hazard index and \eqn{MHQ} denotes the maximum hazard quotient.

The hazard index of the vector \eqn{i} is given by:
\mjdeqn{HI_i = \sum_{j = 1}^n HQ_{i,j}}{HI_i = sum(HQ_ij) from j = 1 to n}
where \eqn{HQ} denotes the hazard quotients and \eqn{n} denotes the number of hazard quotients.

The maximum hazard quotient of the vector \eqn{i} is given by:
\mjdeqn{MHQ_i = HQ_{M,i} = \max_{j \in \lbrace 1,...,n\rbrace} HQ_{i,j}}{MHQ_i = HQ_Mi = max HQ_i}
where \eqn{HQ} denotes the hazard quotients and \eqn{n} denotes the number of hazard quotients.

The hazard quotient of the value \eqn{j} in the vector \eqn{i} is given by:
\mjdeqn{HQ_{i,j} = \frac{V_{i,j}}{RV_j}}{HQ_ij = V_ij / RV_j}
where \eqn{V} denotes the \code{values} and \eqn{RV} denotes the \code{references}.
}
\note{
Due to the multiple possible usages, the argument \code{mcr} must be explicitly named in the function call.
}
\examples{
missed_toxicity(c(1,2,3,4,5), c(1,2,3,4,5))
missed_toxicity(mcr = maximum_cumulative_ratio(c(1,2,3,4,5), c(1,2,3,4,5)))

missed_toxicity(
  values = matrix(c(.1, .2, .3, .4, .5, .6, .7, .8, .9, 1),
                  nrow = 2, byrow = TRUE,
                  dimnames = list(c("set.1", "set.2"), LETTERS[1:5])),
  references = c(1,2,3,4,5)
)

}
\references{
Price PS, Han X (2011).
Maximum cumulative ratio (MCR) as a tool for assessing the value of performing a cumulative risk assessment.
\emph{International Journal of Environmental Research and Public Health}. 8(6): 2212-2225.
\url{https://doi.org/10.3390/ijerph8062212}.
}
\seealso{
\code{\link{maximum_cumulative_ratio}}.
}
\author{
Gauthier Magnin
}
