% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{network_density,SpectralAnalyzer-method}
\alias{network_density,SpectralAnalyzer-method}
\alias{network_density}
\title{Densité d'un réseau}
\usage{
\S4method{network_density}{SpectralAnalyzer}(object, links)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{links}{Data frame des liens (ou arêtes) d'un graphe de type spectrosome.}
}
\value{
Densité du réseau.
}
\description{
Calcule la densité du graphe comme étant le ratio entre le nombre de liens identifiés
 et le nombre maximal de liens possibles.
}
\author{
Gauthier Magnin
}
