% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{network_density,TransactionAnalyzer-method}
\alias{network_density,TransactionAnalyzer-method}
\alias{network_density}
\title{Network density}
\usage{
\S4method{network_density}{TransactionAnalyzer}(object, links)
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}

\item{links}{Data frame of the links (or edges) of a spectrosome graph.}
}
\value{
Density of the network.
}
\description{
Compute the density of a graph as the ratio between the number of links identified and the
 number of links there would be if it was a complete graph (i.e. if all the vertices of the graph
 were connected to each other).
}
\examples{
spectrosome <- spectrosome_chart(TA_instance, TA_instance["patterns"][1:15, ])
network_density(TA_instance, spectrosome[["edges"]])

}
\author{
Gauthier Magnin
}
