% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\docType{data}
\name{NODE_LINKS}
\alias{NODE_LINKS}
\title{NODE_LINKS}
\format{
An object of class \code{character} of length 1.
}
\usage{
NODE_LINKS
}
\description{
Reference value for naming links between entities: links between nodes.
}
\keyword{internal}
