% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\docType{data}
\name{NODES}
\alias{NODES}
\title{NODES}
\format{
An object of class \code{character} of length 1.
}
\usage{
NODES
}
\description{
Reference value for naming entities: nodes.
}
\keyword{internal}
