% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\docType{data}
\name{NODES_PATTERNS_OR_OBSERVATIONS}
\alias{NODES_PATTERNS_OR_OBSERVATIONS}
\title{NODES_PATTERNS_OR_OBSERVATIONS}
\format{
An object of class \code{character} of length 1.
}
\usage{
NODES_PATTERNS_OR_OBSERVATIONS
}
\description{
Reference value for defining possible entities: nodes, patterns or observations.
}
\keyword{internal}
