% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\docType{data}
\name{NODES_PATTERNS_OR_RULES}
\alias{NODES_PATTERNS_OR_RULES}
\title{NODES_PATTERNS_OR_RULES}
\format{
An object of class \code{character} of length 1.
}
\usage{
NODES_PATTERNS_OR_RULES
}
\description{
Reference value for defining possible entities: nodes, patterns or rules.
}
\keyword{internal}
