% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_manager.R
\name{nth_values}
\alias{nth_values}
\title{Extraction of first, last or \out{n<sup>th</sup>} values from a list}
\usage{
nth_values(x, n)
}
\arguments{
\item{x}{List of vectors whose \code{n}\out{<sup>th</sup>} values are to be extracted.}

\item{n}{Position of the elements to extract in each vector of \code{x}. Negative values start from the
end.\cr
\code{"first"} and \code{"last"} are special values for the first and last values of each vector.}
}
\value{
Vector of the values at positions \code{n}.
}
\description{
Extract the first, last or \out{n<sup>th</sup>} values of the vectors of a list.
}
\examples{
l <- list(c(1), c(1,2), c(1,2,3))
nth_values(l, "first")
nth_values(l, "last")
nth_values(l, 2)
nth_values(l, -2)

}
\author{
Gauthier Magnin
}
