% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_set.R
\name{observation.set}
\alias{observation.set}
\title{Observation Set constructor}
\usage{
observation.set(data, item_key, year_key = NA, names = NULL)
}
\arguments{
\item{data}{List of observations containing items and any additional data.
Each observation is itself a list. All observations must share the same element names.}

\item{item_key}{Name of the element containing the items of an observation.}

\item{year_key}{Name of the element containing the year in which an observation was made.}

\item{names}{Names of the elements present in each observation (i.e. names of the elements contained in
the list corresponding to an observation).}
}
\value{
New object of class \code{ObservationSet}.
}
\description{
Create and initialize an object of class \code{ObservationSet}.
}
\details{
Observations do not necessarily require temporal data since \code{year_key} is not required.

If names are not specified using the argument \code{names}, they are automatically extracted from
\code{data}. Specifying \code{names} allows to create \code{ObservationSet} (or subset of existing \code{ObservationSet})
without any observations.
}
\examples{
## Creating a list of observations then an ObservationSet
obs <- list(O1 = list(items = "A",
                      date = 2018,
                      labels = c("l1", "l2")),
            
            O2 = list(items = c("A", "B", "D"),
                      date = 2018,
                      labels = "l3"),
            
            O3 = list(items = c("B", "C", "D"),
                      date = 2019,
                      labels = c("l2", "l3")),
            
            O4 = list(items = c("A", "B", "C", "D"),
                      date = 2020,
                      labels = c("l1", "l2", "l3"))
            )
obs_object <- observation.set(obs, item_key = "items", year_key = "date")
print(obs_object)

}
\seealso{
The class: \code{\link{ObservationSet}}.
}
\author{
Gauthier Magnin
}
