% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_set.R
\docType{class}
\name{ObservationSet-class}
\alias{ObservationSet-class}
\alias{ObservationSet}
\title{Observation Set}
\description{
S4 object class for representing observations.
An observation is defined as an itemset associated with any other data.
}
\details{
Since \code{year_key} is not required, observations do not need to contain temporal data. In such a case,
\code{year_key} is \code{NA}.
}
\section{Slots}{

\describe{
\item{\code{data}}{List of observations containing items and any additional data.
Each observation is itself a list. All observations share the same element names (corresponding to
the names of the various data).}

\item{\code{names}}{Names of the elements present in each observation (i.e. names of the elements contained in
the list corresponding to an observation).}

\item{\code{item_key}}{Name of the element containing the items of an observation.}

\item{\code{year_key}}{Name of the element containing the year in which an observation was made.}
}}

\seealso{
The \code{ObservationSet} constructor: \code{\link{observation.set}}.

An example object of class \code{ObservationSet}: \code{\link{OS_instance}}.
}
\author{
Gauthier Magnin
}
