% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OS_instance}
\alias{OS_instance}
\title{Example of an ObservationSet object}
\format{
An object of class \code{ObservationSet} containing 14 observations of 5 elements:
 \describe{
   \item{CODE}{Codes identifying the items corresponding to the observation.}
   \item{YEAR}{Year in which the observation was made.}
   \item{JOB.TITLE, JOB.TASK, SAMPLE.ID}{Additional data related to the observation.}
 }

For more about the attributes, see \code{\link{ObservationSet}}.
}
\usage{
OS_instance
}
\description{
An example of an object of class \code{ObservationSet}.
}
\details{
Here is the way it was created using the dataset \code{\link{oedb_sample}}:
\preformatted{
## Making a list of observations by grouping data
obs <- make_observations(oedb_sample,
                         by = "ID",
                         additional = c("CODE", "YEAR",
                                        "JOB.TITLE", "JOB.TASK", "SAMPLE.ID"))

## Creation of the ObservationSet
OS_instance <- observation.set(data = obs, item_key = "CODE", year_key = "YEAR")
OS_instance["names"] <- OS_instance["names"][-1]
}
}
\keyword{datasets}
