% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{pattern_chart,SpectralAnalyzer-method}
\alias{pattern_chart,SpectralAnalyzer-method}
\alias{pattern_chart}
\title{Pattern chart}
\usage{
\S4method{pattern_chart}{SpectralAnalyzer}(
  object,
  pc,
  identifiers = "original",
  length_one = FALSE,
  jitter = TRUE,
  display_status = TRUE,
  display_text = "ID",
  use_names = TRUE,
  n.cutoff = NULL,
  category = NULL,
  c.cutoff = NULL,
  sort_by = "category",
  title = "Pattern itemsets",
  path = NULL,
  name = NULL
)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{pc}{Data frame of \strong{p}atterns and their \strong{c}haracteristics. Patterns whose chart
is to be plotted. Any subset of \code{object["patterns"]}.\cr
\code{"patterns"} and \code{"p"} are special values for \code{object["patterns"]}.}

\item{identifiers}{Which IDs to use to identify the patterns on the chart and in the return data frame.
One of \code{"original"}, \code{"new"}.
\describe{
  \item{\code{"original"}}{Use of the original identifiers.}
  \item{\code{"new"}}{Use of new identifiers based on pattern sorting (see 'Details' section to learn
                      more about the sort that is performed).}
}}

\item{length_one}{If \code{FALSE}, patterns of order \eqn{1} are not plotted. If \code{TRUE}, all
patterns are plotted.}

\item{jitter}{If \code{FALSE} patterns of order \eqn{1} are aligned vertically.
If \code{TRUE}, they are spread over two vertical lines to avoid overplotting.
Ignored if \code{length_one} is \code{FALSE}.}

\item{display_status}{If \code{TRUE}, display pattern status.}

\item{display_text}{Text to display on the chart next to the patterns.
Pattern identifiers (\code{"ID"}) or one of the other characteristics (\code{"weight"},
\code{"frequency"}, \code{"specificity"}, \code{"year"}).
\code{NULL} value specifies to display no information.}

\item{use_names}{If \code{TRUE}, display item names if they are defined. Display their identification
codes otherwise.}

\item{n.cutoff}{If \code{use_names = TRUE}, limit number of characters to display concerning the names
of the represented items.}

\item{category}{Name or number of the category to represent on the chart (numbering according to
the order of the columns of `object["items_categories"]`).}

\item{c.cutoff}{Limit number of characters to display in the legend for the category represented.}

\item{sort_by}{Sorting method of displayed items. One of \code{"category"}, \code{"item"}.}

\item{title}{Chart title.}

\item{path}{Path of the directory in which to save the chart as a PDF file. To be ignored to plot
the chart in the active device.}

\item{name}{Name of the PDF file in which to save the chart. To be ignored to plot the chart in the
active device.}
}
\value{
Data frame of the patterns represented on the chart, associated with their characteristics
 and identifiers (visible on the chart if \code{display_text = "ID"}).
}
\description{
Plot a chart of the pattern itemsets. It can be automatically saved as a PDF file.
}
\details{
The patterns are sorted according to their order values, then to their weights.

If the arguments \code{path} and \code{name} are not \code{NULL}, the chart is plotted in a PDF file
 of A4 landscape paper size. If they are \code{NULL}, the chart is plotted in the active device.

The colors associated with the values of the possible category represented are selected circularly
 among the 20 colors of the palette \code{category20} from D3 (see \code{ggsci::pal_d3("category20")}).
Therefore, if the number of values exceeds \code{20}, some colors will be used more than once.
 For example, the \out{22<sup>nd</sup>} value will share the color of the \out{2<sup>nd</sup>}
 value.
See the attribute \code{categories_colors} of \code{object} to reassign colors to the category values.
}
\examples{
patterns_1 <- pattern_chart(SA_instance, "patterns", category = "family",
                            n.cutoff = 20, c.cutoff = 10)
patterns_2 <- pattern_chart(SA_instance, SA_instance["patterns"][1:15, ])

patterns_1 <- pattern_chart(SA_instance, "patterns", category = "family",
                            n.cutoff = 20, c.cutoff = 17,
                            path = getwd(), name = "pattern_itemsets")

}
\references{
Bosson-Rieutort D, Sarazin P, Bicout DJ, Ho V, Lavoué J (2020).
            Occupational Co-exposures to Multiple Chemical Agents from Workplace Measurements by the US Occupational Safety and Health Administration.
            \emph{Annals of Work Exposures and Health}, Volume 64, Issue 4, May 2020, Pages 402–415.
            \url{https://doi.org/10.1093/annweh/wxaa008}.
}
\author{
Delphine Bosson-Rieutort, Gauthier Magnin
}
