% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{pattern_node_characteristics,SpectralAnalyzer-method}
\alias{pattern_node_characteristics,SpectralAnalyzer-method}
\alias{pattern_node_characteristics}
\title{Pattern node characteristics}
\usage{
\S4method{pattern_node_characteristics}{SpectralAnalyzer}(object, patterns)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{patterns}{Patterns whose characteristics of the nodes that contain them are to be found.
Any subset of `object["patterns"]$pattern`.}
}
\value{
\describe{
   \item{\code{weights}}{For each pattern, the weights of the nodes in which it is included.}
   \item{\code{lengths}}{For each pattern, the lengths of the nodes in which it is included.}
 }
}
\description{
For each pattern, extract the weights and lengths of the nodes in which it is included.
}
\seealso{
\code{\link{weight_by_node_complexity}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
