% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{pattern_node_characteristics,TransactionAnalyzer-method}
\alias{pattern_node_characteristics,TransactionAnalyzer-method}
\alias{pattern_node_characteristics}
\title{Pattern node characteristics}
\usage{
\S4method{pattern_node_characteristics}{TransactionAnalyzer}(object, patterns)
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}

\item{patterns}{Patterns whose characteristics of the nodes that contain them are to be found.
Any subset of \code{object["patterns"]$pattern}.}
}
\value{
\describe{
   \item{\code{frequencies}}{For each pattern, the frequencies of the nodes in which it is included.}
   \item{\code{lengths}}{For each pattern, the lengths of the nodes in which it is included.}
 }
}
\description{
For each pattern, extract the frequencies and lengths of the nodes in which it is included.
}
\note{
This method is not exported from the package.
}
\seealso{
\code{\link{frequency_by_complexity}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
