% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\docType{data}
\name{PATTERNS}
\alias{PATTERNS}
\title{PATTERNS}
\format{
An object of class \code{character} of length 1.
}
\usage{
PATTERNS
}
\description{
Reference value for naming entities: patterns.
}
\keyword{internal}
