% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{plot_itemset_chart,SpectralAnalyzer-method}
\alias{plot_itemset_chart,SpectralAnalyzer-method}
\alias{plot_itemset_chart}
\title{Itemset chart plotting}
\usage{
\S4method{plot_itemset_chart}{SpectralAnalyzer}(
  object,
  nopc,
  items_category,
  category = NULL,
  c.cutoff = NULL,
  use_names = TRUE,
  n.cutoff = NULL,
  jitter = TRUE,
  under = "ID",
  over = NULL,
  title = "Itemsets"
)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{nopc}{Data frame of \strong{n}odes \strong{o}r \strong{p}atterns and their
\strong{c}haracteristics. Nodes or patterns whose chart is to be plotted. Any subset of
\code{object["nodes"]} or \code{object["patterns"]}.}

\item{items_category}{Data frame of items and one associated category.}

\item{category}{Name of the category to represent on the chart, used as the legend title.}

\item{c.cutoff}{Limit number of characters to display in the legend for the category represented.}

\item{use_names}{If \code{TRUE}, display item names if they are defined. Display their identification
codes otherwise.}

\item{n.cutoff}{If \code{use_names = TRUE}, limit number of characters to display concerning the names
of the represented items.}

\item{jitter}{If \code{FALSE} itemsets of length \eqn{1} are aligned vertically.
If \code{TRUE}, they are spread over two vertical lines to avoid overplotting.}

\item{under, over}{Text to display on the chart under and over the itemsets.
Identifiers (\code{"ID"}) or one of the characteristics (\code{"weight"}, \code{"frequency"},
\code{"specificity"}, \code{"year"}, \code{"status"}).

\code{"status"} can only be used for the argument \code{"over"}.
\code{NULL} value specifies to display no data.}

\item{title}{Chart title.}
}
\description{
Plot a chart of the node or pattern itemsets.
}
\details{
The colors associated with the values of the possible category represented are selected circularly
 among the 20 colors of the palette \code{category20} from D3 (see \code{ggsci::pal_d3("category20")}).
Therefore, if the number of values exceeds \code{20}, some colors will be used more than once.
 For example, the \out{22<sup>nd</sup>} value will share the color of the \out{2<sup>nd</sup>}
 value.
See the attribute \code{categories_colors} of \code{object} to reassign colors to the category values.
}
\references{
Bosson-Rieutort D, Sarazin P, Bicout DJ, Ho V, Lavoué J (2020).
            Occupational Co-exposures to Multiple Chemical Agents from Workplace Measurements by the US Occupational Safety and Health Administration.
            \emph{Annals of Work Exposures and Health}, Volume 64, Issue 4, May 2020, Pages 402–415.
            \url{https://doi.org/10.1093/annweh/wxaa008}.
}
\seealso{
\code{\link{itemset_chart}}.
}
\author{
Delphine Bosson-Rieutort, Gauthier Magnin
}
\keyword{internal}
