% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_helper.R
\name{plot_itemset_chart}
\alias{plot_itemset_chart}
\title{Itemset chart}
\usage{
plot_itemset_chart(
  itemsets,
  items,
  category = NULL,
  jitter = TRUE,
  under = NULL,
  over = NULL,
  over_legend = NULL,
  title = "Itemsets"
)
}
\arguments{
\item{itemsets}{List of character vectors. Itemsets to plot.}

\item{items}{Data frame associating each item of \code{itemsets} (column \code{"item"}) with a label
(column \code{"label"}) and possibly with a category (column named by the name of the category).}

\item{category}{Data frame associating each value of the category associated with the items (column
\code{"value"}) with a color (column \code{"col"}) and with a label (column \code{"label"}).
If not \code{NULL}, the items are colored according to the category values and a legend is displayed.}

\item{jitter}{If \code{FALSE}, non-equivalent itemsets of length \eqn{1} are aligned vertically.
If \code{TRUE}, they are spread over several vertical lines to avoid overplotting.}

\item{under}{Character vector. Text to display on the chart under the itemsets.}

\item{over}{Character vector. Text to display on the chart over the itemsets.
Color values (color names or hexadecimal values) can be specified to plot colored squared instead of
text.}

\item{over_legend}{If color values are given in the argument \code{over}, named character vector in which
values are the colored used and names are associated labels to plot a specific legend.}

\item{title}{Chart title.}
}
\description{
Plot a chart of itemsets. A color scale can be associated to the items and additional data
can be plotted under and over each itemset.
}
\references{
Bosson-Rieutort D, Sarazin P, Bicout DJ, Ho V, Lavoué J (2020).
Occupational Co-exposures to Multiple Chemical Agents from Workplace Measurements by the US Occupational Safety and Health Administration.
\emph{Annals of Work Exposures and Health}, Volume 64, Issue 4, May 2020, Pages 402–415.
\url{https://doi.org/10.1093/annweh/wxaa008}.
}
\author{
Delphine Bosson-Rieutort, Gauthier Magnin
}
\keyword{internal}
