% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixer.R
\name{plot_mcr_log_part}
\alias{plot_mcr_log_part}
\title{MCR approach scatter plot (log part)}
\usage{
plot_mcr_log_part(
  chart,
  xlim,
  ylim,
  regions,
  regions_col,
  regions_alpha,
  regions_lab
)
}
\arguments{
\item{chart}{\code{ggplot} object on which to add the log part.}

\item{xlim}{Lower and uper limits of the abscissa of the graph.}

\item{ylim}{Lower and uper limits of the ordinate of the graph.}

\item{regions}{If \code{TRUE}, the regions corresponding to the MIAT groups are filled with the colors
defined by \code{regions_col}.}

\item{regions_col}{Character vector of length 4. Define the colors for the regions of the MIAT groups
(in order: I, II, IIIA and IIIB).}

\item{regions_alpha}{Value between 0 and 1. Opacity of the regions filled with \code{regions_col}.}

\item{regions_lab}{Logical value or vector of length 4. Define if labels for the MIAT groups should
be displayed (in order: I, II, IIIA, IIIB). \code{TRUE} and \code{FALSE} are special values for all \code{TRUE} or
all \code{FALSE}.}
}
\value{
\code{ggplot} object of the chart after plotting the log part.
}
\description{
Plot \mjeqn{log_{10}(HI)}{log10(HI)} versus \mjeqn{log_{10}(MCR - 1)}{log10(MCR - 1)} with the
percentage of the reciprocal of the maximum cumulative ratio, the elements producing the top hazard
quotients and the associated MIAT groups. More precisely, plot the part of the chart that depends
on the logarithm transformation (regions: filling, boundaries, labels; points and axes).
\loadmathjax
}
\details{
The limits of the graph (\code{xlim} and \code{ylim}) are considered without taking into account the expand
limits (see \code{\link[ggplot2:expand_limits]{ggplot2::expand_limits}}).

The chart being plotted with the \code{ggplot2} package, it can be modified or completed afterwards using
\code{\link[ggplot2:last_plot]{ggplot2::last_plot}} or the returned object.

Color specification can be done using the R predefined color names or hexadecimal values.
}
\note{
This function is not exported from the package.
}
\references{
Reyes JM, Price PS (2018).
An analysis of cumulative risks based on biomonitoring data for six phthalates using the Maximum Cumulative Ratio.
\emph{Environment International}, 112, 77-84.
\url{https://doi.org/10.1016/j.envint.2017.12.008}.
}
\seealso{
\code{\link{mcr_chart}}, \code{\link{plot_mcr_standard_part}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
