% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixer.R
\name{plot_mcr_standard_part}
\alias{plot_mcr_standard_part}
\title{MCR approach scatter plot (standard part)}
\usage{
plot_mcr_standard_part(
  chart,
  xlim,
  ylim,
  regions,
  regions_col,
  regions_alpha,
  regions_lab,
  group_text
)
}
\arguments{
\item{chart}{\code{ggplot} object on which to add the standard part.}

\item{xlim}{Lower and uper limits of the abscissa of the graph.}

\item{ylim}{Lower and uper limits of the ordinate of the graph.}

\item{regions}{If \code{TRUE}, the regions corresponding to the MIAT groups are filled with the colors
defined by \code{regions_col}.}

\item{regions_col}{Character vector of length 4. Define the colors for the regions of the MIAT groups
(in order: I, II, IIIA and IIIB).}

\item{regions_alpha}{Value between 0 and 1. Opacity of the regions filled with \code{regions_col}.}

\item{regions_lab}{Logical value or vector of length 4. Define if labels for the MIAT groups should
be displayed (in order: I, II, IIIA, IIIB). \code{TRUE} and \code{FALSE} are special values for all \code{TRUE} or
all \code{FALSE}.}

\item{group_text}{Since MIAT groups can be considered as risk classes,
\code{"group"} or \code{"class"} to refer to them as "MIAT groups" or as "Risk classes".}
}
\value{
\code{ggplot} object of the chart after plotting the standard part.
}
\description{
Plot \eqn{HI} versus \eqn{MCR} with the percentage of the reciprocal of the maximum cumulative ratio,
the elements producing the top hazard quotients and the associated MIAT groups. More precisely, plot
the part of the chart that depends on the non-transformation of the axes (regions: filling,
boundaries, labels; points and axes).
}
\details{
The limits of the graph (\code{xlim} and \code{ylim}) are considered without taking into account the expand
limits (see \code{\link[ggplot2:expand_limits]{ggplot2::expand_limits}}).

The chart being plotted with the \code{ggplot2} package, it can be modified or completed afterwards using
\code{\link[ggplot2:last_plot]{ggplot2::last_plot}} or the returned object.

Color specification can be done using the R predefined color names or hexadecimal values.

The grey area represents the region in which no point can be plotted because MCR cannot be lower
than 1.
}
\note{
This function is not exported from the package.
}
\references{
De Brouwere K, et al. (2014).
Application of the maximum cumulative ratio (MCR) as a screening tool for the evaluation of mixtures in residential indoor air.
\emph{The Science of the Total Environment}, 479-480, 267-276.
\url{https://doi.org/10.1016/j.scitotenv.2014.01.083}.
}
\seealso{
\code{\link{mcr_chart}}, \code{\link{plot_mcr_log_part}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
