% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{plot_spectrum_chart,TransactionAnalyzer-method}
\alias{plot_spectrum_chart,TransactionAnalyzer-method}
\alias{plot_spectrum_chart}
\title{Pattern spectrum plotting}
\usage{
\S4method{plot_spectrum_chart}{TransactionAnalyzer}(object, pc, frequencies, title = "Spectrum of patterns")
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}

\item{pc}{Data frame of \strong{p}atterns and their \strong{c}haracteristics. Patterns whose spectrum
is to be plotted. Any subset of \code{object["patterns"]}.}

\item{frequencies}{Two-column matrix containing, for each pattern, its frequency related to complex
nodes and its frequency related to simple nodes.}

\item{title}{Chart title.}
}
\description{
Plot a spectrum chart.
}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
            The spectrosome of occupational health problems. \emph{PLoS ONE} 13(1): e0190196.
            \url{https://doi.org/10.1371/journal.pone.0190196}.
}
\seealso{
\code{\link{spectrum_chart}}, \code{\link{frequency_by_complexity}},
         \code{\link{pattern_node_characteristics}}.
}
\author{
Delphine Bosson-Rieutort, Gauthier Magnin
}
\keyword{internal}
