% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{plot_tree_chart,SpectralAnalyzer-method}
\alias{plot_tree_chart,SpectralAnalyzer-method}
\alias{plot_tree_chart}
\title{Arbre de la multi-association}
\usage{
\S4method{plot_tree_chart}{SpectralAnalyzer}(
  object,
  patterns_characteristics,
  items_category,
  category = NULL,
  cutoff = NULL,
  display_text = NULL,
  title = "Multi-association tree"
)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{patterns_characteristics}{Ensemble des caractéristiques des motifs dont l'arbre est à tracer.}

\item{items_category}{Data frame des items et d'une catégorie associée.}

\item{category}{Nom de la catégorie à représenter sur l'arbre, utilisé comme titre de légende.}

\item{cutoff}{Nombre limite de caractères à afficher dans la légende concernant la catégorie représentée.}

\item{display_text}{Texte à afficher sur le graphique à côté des motifs.
Choix entre les identifiants \code{"ID"} des motifs ou l'une des caractéristiques (\code{"weight"},
\code{"frequency"}, \code{"specificity"}, \code{"year"}, \code{"status"}).
Par défaut (valeur \code{NULL}), aucun texte n'est affiché.}

\item{title}{Titre du graphique.}
}
\description{
Dessine un graphique de type arbre de la multi-association.
}
\seealso{
\code{\link{tree_chart}}.
}
\author{
Delphine Bosson-Rieutort, Gauthier Magnin
}
\keyword{internal}
