% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pluralize}
\alias{pluralize}
\title{Pluzalize a character string}
\usage{
pluralize(x, var)
}
\arguments{
\item{x}{Character string to be pluralized.}

\item{var}{Variable to consider whether to pluralize the string.}
}
\value{
The character string corresponding to the argument \code{x}, possibly in its plural form.
}
\description{
Give the plural form of the string if the given variable is numeric and is greater than \eqn{1} or
has a length greater than \eqn{1}. Add an "s" at the end of the string or turn "y" into "ies".
}
\note{
This function is not exported from the package.
}
\seealso{
\code{\link{cap}}, \code{\link{standardize_nchar}}, \code{\link{first_characters}}, \code{\link{substr2}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
