% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_set.R
\name{prepare_itemset_chart,TransactionSet-method}
\alias{prepare_itemset_chart,TransactionSet-method}
\alias{prepare_itemset_chart}
\title{Prepare data for plotting itemset chart}
\usage{
\S4method{prepare_itemset_chart}{TransactionSet}(object, identifiers, length_one, under, over)
}
\arguments{
\item{object}{S4 object of class \code{TransactionSet}.}

\item{identifiers}{Which IDs to use to identify the transactions on the chart and in the
return object. One of \code{"original"}, \code{"new"}.
\describe{
\item{\code{"original"}}{Use of the original identifiers.}
\item{\code{"new"}}{Use of new identifiers based on sorting (see 'Details' section to learn more
about the sort that is performed).}
}}

\item{length_one}{If \code{FALSE}, itemsets of length \eqn{1} are not plotted. If \code{TRUE}, all itemsets
are plotted.}

\item{under, over}{Data to display on the chart under and over the itemsets.
Identifiers (\code{"ID"}) or one of the elements of the transactions (i.e. one of the values of
\code{object["names"]}).}
}
\value{
\describe{
\item{\code{transactions}}{Subset of \code{object} to plot.}
\item{\code{itemsets}}{List of itemsets to plot.}
\item{\code{items}}{Data frame containing the items contained in the itemsets to plot, duplicated in two
colonnes: item and label.}
\item{\code{under}}{Data to display under the itemsets.}
\item{\code{over}}{Data to display over the itemsets.}
}
}
\description{
Prepare transactions, itemsets, items and text for itemset chart.
}
\note{
This method is not exported from the package.
}
\seealso{
\code{\link[=itemset_chart,TransactionSet-method]{itemset_chart,TransactionSet}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
