% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixer.R
\name{reciprocal_of_mcr}
\alias{reciprocal_of_mcr}
\title{Reciprocal of Maximum Cumulative Ratio}
\usage{
reciprocal_of_mcr(values, references)
reciprocal_of_mcr(hi, mhq)
reciprocal_of_mcr(mcr)
}
\arguments{
\item{values}{Numeric vector or matrix. Values for which the reciprocal of maximum cumulative ratio
is to be computed.}

\item{references}{Numeric vector. Reference values associated with the \code{values}.}

\item{hi}{Numeric value or vector. \strong{H}azard \strong{i}ndex(es) for which to compute the reciprocal of
maximum cumulative ratio(s).}

\item{mhq}{Numeric value or vector. \strong{M}aximum \strong{h}azard \strong{q}uotient(s) associated with the hazard
index(es) \code{hi}.}

\item{mcr}{Numeric value or vector. \strong{M}aximum \strong{c}umulative \strong{r}atio(s) on which to compute the
reciprocal.}
}
\value{
Numeric value or vector (according to \code{values}, to \code{hi} and \code{mhq} or to \code{mcr}) of the
reciprocal of maximum cumulative ratio(s).
}
\description{
Compute the reciprocal of maximum cumulative ratio, which measures the percentage contribution of
the maximum hazard quotient to the total hazard.
}
\details{
Arguments \code{values} and \code{references} are used to compute the hazard quotients and the hazard index
before searching for the maximum hazard quotient, computing the maximum cumulative ratio then
computing the reciprocal. Thus, call the function with the arguments \code{hi} and \code{mhq} is faster and
call it with the argument \code{mcr} is even faster (if they are already computed).

If \code{values} is a matrix, the reference values are applied once on each column (i.e. it must have one
reference value for each row of the matrix).

If \code{values} is a matrix (or \code{hi} and \code{mhq} are vectors larger than 1, or \code{mcr} is such a vector),
one reciprocal of maximum cumulative ratio is computed for each column (or value, respectively).

\loadmathjax
The reciprocal of the maximum cumulative ratio of the vector \eqn{i} is given by:
\mjdeqn{Reciprocal~of~MCR_i = \frac{1}{MCR_i} = \frac{MHQ_i}{HI_i}}{Reciprocal of MCR_i = 1 / MCR_i = MHQ_i / HI_i}
where \eqn{MCR}, \eqn{MHQ} and {HI} denotes the maximum cumulative ratio, the maximum hazard
quotient and the hazard index respectively.

The maximum cumulative ratio of the vector \eqn{i} is given by:
\mjdeqn{MCR_i = \frac{HI_i}{MHQ_i}}{MCR_i = HI_i / MHQ_i}
where \eqn{HI} denotes the hazard index and \eqn{MHQ} denotes the maximum hazard quotient.

The hazard index of the vector \eqn{i} is given by:
\mjdeqn{HI_i = \sum_{j = 1}^N HQ_{i,j}}{HI_i = sum(HQ_ij) from j = 1 to N}
where \eqn{HQ} denotes the hazard quotients and \eqn{N} denotes the number of hazard quotients.

The maximum hazard quotient of the vector \eqn{i} is given by:
\mjdeqn{MHQ_i = HQ_{M,i} = \max_{j \in \lbrace 1,...,N\rbrace} HQ_{i,j}}{MHQ_i = HQ_Mi = max HQ_i}
where \eqn{HQ} denotes the hazard quotients and \eqn{N} denotes the number of hazard quotients.

The hazard quotient of the value \eqn{j} in the vector \eqn{i} is given by:
\mjdeqn{HQ_{i,j} = \frac{V_{i,j}}{RV_j}}{HQ_ij = V_ij / RV_j}
where \eqn{V} denotes the \code{values} and \eqn{RV} denotes the \code{references}.
}
\note{
Due to the multiple possible usages, the arguments \code{hi}, \code{mhq} and \code{mcr} must be explicitly named in
the function call.
}
\examples{
reciprocal_of_mcr(c(1,2,3,4,5), c(1,2,3,4,5))
reciprocal_of_mcr(hi = hazard_index(c(1,2,3,4,5), c(1,2,3,4,5)),
                  mhq = maximum_hazard_quotient(c(1,2,3,4,5), c(1,2,3,4,5)))
reciprocal_of_mcr(mcr = maximum_cumulative_ratio(c(1,2,3,4,5), c(1,2,3,4,5)))
reciprocal_of_mcr(values = matrix(c(.1, .2, .3, .4, .5, .6, .7, .8, .9, 1),
                                  ncol = 2),
                  references = c(1,2,3,4,5))

}
\references{
Reyes JM, Price PS (2018).
An analysis of cumulative risks based on biomonitoring data for six phthalates using the Maximum Cumulative Ratio.
\emph{Environment International}, 112, 77-84.
\url{https://doi.org/10.1016/j.envint.2017.12.008}.
}
\seealso{
\code{\link{maximum_cumulative_ratio}}, \code{\link{hazard_index}}, \code{\link{maximum_hazard_quotient}}.
}
\author{
Gauthier Magnin
}
