% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixer.R
\name{reduce_sets}
\alias{reduce_sets}
\title{Reduce sets according to the names of their components}
\usage{
reduce_sets(values, references = NULL, FUN, ignore_zero = TRUE, ...)
}
\arguments{
\item{values}{Numeric named vector or matrix, or list of numeric named vectors.
Sets of values to reduce.}

\item{references}{Optional. Numeric vector or list of numeric vectors. Reference values associated
with the \code{values}.}

\item{FUN}{Function to apply on each subset of values corresponding to each different name to reduce
it to a single value.}

\item{ignore_zero}{\code{TRUE} or \code{FALSE} whether to ignore values equal to 0.}

\item{...}{Further arguments to the function \code{FUN}.}
}
\value{
If \code{references} is \code{NULL}, sets of values resulting from the application of the function given as
argument on each subset corresponding to each different name associated with the values.\cr
If not, list containing:
\describe{
\item{\code{values}}{Sets of values resulting from the application of the function given as argument
on each subset corresponding to each different name associated with the values.}
\item{\code{references}}{Subset of references corresponding to the resulting values.}
}
}
\description{
For each set of values, apply a function to merge the ones having the same names or to choose one
of them, for each different name found in the set of values. This results in sets of values having
only one value per different name found in the original sets, instead of several ones.
}
\details{
If \code{values} is a vector, the reference values are directly associated with these values.

If \code{values} is a matrix, the reference values are applied once on each set of values, i.e. on each
row. Therefore, there must be one reference value for each column of the matrix.

If \code{values} is a list, the reference values can be a vector of named values or a list. In this case,
if \code{references} is a vector, it is not processed and is returned as is. Otherwise \code{references} must
be a list of vectors having the same lengths as those present in \code{values} so that \code{values} and
\code{references} can be matched.

It is assumed that the same reference values are given for same value names of the argument \code{values}.

The components of each resulting set of values are ordered in the order in which the names are found
in the original set. This is also the case for the references.
}
\examples{
## Reduce a set of values given as a vector
reduce_sets(values = c(A = 1, A = 3, B = 2, C = 1, C = 5, C = 4),
            FUN = max)
reduce_sets(values = c(C = 1, C = 5, C = 4, B = 2, A = 1, A = 3),
            FUN = function(x) min(x) + 0.5)

## Reduce sets of values given as a matrix
v <- matrix(sample(seq(0.1, 1, by = 0.1), 50, replace = TRUE),
            nrow = 10, byrow = TRUE,
            dimnames = list(paste0("set.", 1:10), c("A", "B", "A", "C", "C")))
reduce_sets(values = v, FUN = mean)

## Reduce sets of values and associated reference values given as lists
reduce_sets(
  values = list(set.1 = c(A = 0.1, A = 0.3, B = 0.5),
                set.2 = c(A = 0.2),
                set.3 = c(B = 0.3, B = 0.4, B = 0.7, C = 0.4, C = 0.1)),
  references = list(c(1, 1, 2),
                    1,
                    c(2, 2, 2, 3, 3)),
  FUN = median
)

}
\author{
Gauthier Magnin
}
