% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_set.R
\name{remove_empty_trx,TransactionSet-method}
\alias{remove_empty_trx,TransactionSet-method}
\alias{remove_empty_trx}
\title{Remove empty transactions}
\usage{
\S4method{remove_empty_trx}{TransactionSet}(object, keep_names = TRUE)
}
\arguments{
\item{object}{S4 object of class \code{TransactionSet} to be subsetted.}

\item{keep_names}{If \code{TRUE}, transactions of the returned object keep the
names from the initial \code{object}. If \code{FALSE}, they are not named.}
}
\value{
S4 object of class \code{TransactionSet} having a subset of transactions
from \code{object} (i.e. having a subset of the attribute \code{data} from \code{object}).
}
\description{
Return a \code{TransactionSet} in which transactions are the subset of those
containing at least one item from another \code{TransactionSet}.
}
\details{
If the transactions from \code{object} are not named and \code{keep_names = TRUE}, the
transactions of the resulting  object are named according to the indices of
the initial transactions.
}
\examples{
trx_object <- remove_items(TS_instance, 3146, remove_empty = FALSE)
remove_empty_trx(trx_object)

}
\seealso{
\code{\link{subset}}, \code{\link{reorder}}, \code{\link{remove_items}}.
}
\author{
Gauthier Magnin
}
