% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_set.R
\name{remove_items,TransactionSet-method}
\alias{remove_items,TransactionSet-method}
\alias{remove_items}
\title{Remove items in transactions}
\usage{
\S4method{remove_items}{TransactionSet}(object, items, remove_empty = FALSE, keep_names = TRUE)
}
\arguments{
\item{object}{S4 object of class \code{TransactionSet}.}

\item{items}{Items to be removed from the transactions.}

\item{remove_empty}{Logical indicating whether to remove transactions that no
longer have any items after the removal of the given \code{items}.}

\item{keep_names}{Ignored if \code{remove_empty} is \code{FALSE}.
If \code{TRUE}, transactions of the returned object keep the names from the
initial \code{object}. If \code{FALSE}, they are not named.}
}
\value{
S4 object of class \code{TransactionSet} in which the given \code{items} no
longer exist in any transaction.
}
\description{
Remove items in the transactions of a \code{TransactionSet} object.
}
\details{
If the transactions from \code{object} are not named and \code{keep_names = TRUE}, the
transactions of the resulting  object are named according to the indices of
the initial transactions.
}
\examples{
remove_items(TS_instance, 3146, remove_empty = FALSE)
remove_items(TS_instance, 3146, remove_empty = TRUE)

}
\seealso{
\code{\link{subset}}, \code{\link{reorder}}, \code{\link{remove_empty_trx}}.
}
\author{
Gauthier Magnin
}
