% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_set.R
\name{reorder,ObservationSet-method}
\alias{reorder,ObservationSet-method}
\alias{reorder}
\title{Reorder Observation Set}
\usage{
\S4method{reorder}{ObservationSet}(object, permutation)
}
\arguments{
\item{object}{S4 object of class \code{ObservationSet}.}

\item{permutation}{Numeric or character vector. Permutation to use to rearrange the observations.}
}
\value{
S4 object of class \code{ObservationSet} containing the observations of \code{object} in the order
defined by \code{permutation}.
}
\description{
Return a copy of an \code{ObservationSet} in which the observations are in another order.
}
\examples{
reorder(OS_instance, c(2, 1, 3:14))
reorder(OS_instance, names(OS_instance["data"])[c(2, 1, 3:14)])
reorder(OS_instance, 14:1)

}
\author{
Gauthier Magnin
}
