% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_set.R
\name{reorder,TransactionSet-method}
\alias{reorder,TransactionSet-method}
\alias{reorder}
\title{Reorder Transaction Set}
\usage{
\S4method{reorder}{TransactionSet}(x, permutation)
}
\arguments{
\item{x}{S4 object of class \code{TransactionSet}.}

\item{permutation}{Numeric or character vector. Permutation to use to rearrange the transactions.}
}
\value{
S4 object of class \code{TransactionSet} containing the transactions of \code{x} in the order
defined by \code{permutation}.
}
\description{
Return a copy of an \code{TransactionSet} in which the transactions are in another order.
}
\examples{
reorder(TS_instance, c(2, 1, 3:14))
reorder(TS_instance, names(TS_instance["data"])[c(2, 1, 3:14)])
reorder(TS_instance, 14:1)

}
\author{
Gauthier Magnin
}
