% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resave_with_best_compression}
\alias{resave_with_best_compression}
\title{Resave data with the best compression method}
\usage{
resave_with_best_compression(paths)
}
\arguments{
\item{paths}{A character vector of paths to found data and save files.}
}
\description{
Search for the best compression method to save existing '.RData' or '.rda' files and resave them
with this method.
}
\details{
Use of the maximum compression level (9).
}
\seealso{
\code{\link[tools:resaveRdaFiles]{tools::resaveRdaFiles}}, \code{\link[tools:checkRdaFiles]{tools::checkRdaFiles}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
