% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{reset,SpectralAnalyzer-method}
\alias{reset,SpectralAnalyzer-method}
\alias{reset}
\title{Reconstruit en partie un analyseur spectral}
\usage{
\S4method{reset}{SpectralAnalyzer}(object, from = 1)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{from}{Étape à partir de laquelle recalculer les attributs.}
}
\description{
Redéfinit les attributs d'un analyseur à partir d'une étape spécifique.
}
\author{
Gauthier Magnin
}
