% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\docType{data}
\name{RULES}
\alias{RULES}
\title{RULES}
\format{
An object of class \code{character} of length 1.
}
\usage{
RULES
}
\description{
Reference value for naming entities: rules.
}
\keyword{internal}
