% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{rules_chart,TransactionAnalyzer-method}
\alias{rules_chart,TransactionAnalyzer-method}
\alias{rules_chart}
\title{Length-two association rule visualization}
\usage{
\S4method{rules_chart}{TransactionAnalyzer}(
  object,
  rules = NULL,
  items = NULL,
  parameter = list(supp = 0.001, conf = 0),
  display = "highest confidence",
  threshold = NULL,
  direction = FALSE,
  use_names = TRUE,
  n.cutoff = NULL,
  category = NULL,
  c.cutoff = NULL,
  sort_by = "category",
  vertex_size = 3,
  vertex_alpha = 1,
  vertex_margin = 0.05,
  label_size = 3,
  label_margin = 0.05,
  edge_looseness = 0.8,
  edge_alpha = 1,
  palette = "default",
  palette_direction = 1,
  plot = FALSE
)
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}

\item{rules}{Data frame of association rules to plot (given by the \code{\link{extract_rules}} method).
Only those of length 2 are considered. If \code{NULL}, rules of length 2 are extracted from
\code{object["transactions"]} using the mining parameters given as \code{parameter}.}

\item{items}{Items to consider in the given or extracted rules. If \code{NULL}, items considered are all
those present in the given or extracted rules. Any subset of \code{object["items"]}.

\code{"items"} and \code{"i"} are special values for \code{object["items"]}.}

\item{parameter}{List of mining parameters specifying minimum support and minimum confidence of
association rules to extract. Ignored if \code{rules} is not \code{NULL}. See
\link[arules:ASparameter-classes]{APparameter} for more.}

\item{display}{Rule characteristic to visualize. One of the following.
\describe{
\item{\code{"support"} (or \code{"supp"})}{The support of the two rules existing between two items is
represented.}
\item{\code{"confidence"} or (\code{"conf"})}{The two confidence values of the two rules existing between two
items are both represented.}
\item{\code{"lift"}}{The lift of the two rules existing between two items is represented.}
\item{\code{"frequency"} (or \code{"freq"})}{The frequency of the two rules existing between two items is
represented.}
\item{\code{"specificity"} (or \code{"spec"})}{The two specifity values of the two rules existing between two
items are both represented.}
\item{\code{"accuracy"} (or \code{"accu"})}{The accuracy of the two rules existing between two items is
represented.}
\item{\code{"added.value"} (or \code{"adva"})}{The two added values of the two rules existing between two
items are both represented.}
}
For the confidence, the specificy and the added value, this can be preceded by \code{"highest "} or by
\code{"lowest "} to choose between visualizing a rule or its reciprocal instead of visualizing both
rules.
\describe{
\item{\code{"highest "} (or \code{"hi."})}{Only the rule having the highest value of the choosen
characteristic is represented among the two rules existing between two items.}
\item{\code{"lowest "} (or \code{"lo."})}{Only the rule having the lowest value of the choosen
characteristic is represented among the two rules existing between two items.}
}}

\item{threshold}{Threshold from which the characteristic referred by \code{display} must be for a rule to
be considered.}

\item{direction}{Ignored if \code{display} does not refer to highest or lowest values.
If \code{FALSE}, the opacity of the edges representing the rules is set by the argument \code{edge_alpha}.
If \code{TRUE}, the opacity increases gradually according to the direction of the rule represented.
Always \code{FALSE} if \code{display} refers to the support, the lift, the frequency or the accuracy.
Always \code{TRUE} if \code{display} refers to the confidence, the specificity or the added value without
referring to highest or lowest values.}

\item{use_names}{If \code{TRUE}, display item names if they are defined. Display their identification
codes otherwise.}

\item{n.cutoff}{If \code{use_names = TRUE}, limit number of characters to display concerning the names
of the represented items.}

\item{category}{Name or number of the category to represent on the graph (numbering according to
the order of the columns of \code{object["items_categories"]}).}

\item{c.cutoff}{Limit number of characters to display in the legend for the category represented.}

\item{sort_by}{Sorting method of displayed items. One of \code{"category"}, \code{"item"}.}

\item{vertex_size}{Size of the vertices.}

\item{vertex_alpha}{Opacity of the vertices (from 0 to 1).}

\item{vertex_margin}{Margin before the vertices (i.e. distance between the ends of the edges and the
centers of the vertices).}

\item{label_size}{Size of the labels associated with the vertices.}

\item{label_margin}{Margin before the labels (i.e. distance between the centers of the vertices and
the labels).}

\item{edge_looseness}{Looseness of the connecting lines (from 0 to 1).
The closer the value is to 0, the straighter the lines will be.
The closer the value is to 1, the more the lines will be curved.}

\item{edge_alpha}{Opacity of the lines connecting vertices (from 0 to 1).
Ignored if \code{display} refers to the confidence, the specificity or the added value and \code{direction}
is \code{TRUE}.}

\item{palette}{Name of the palette to use for coloring the edges.

If \code{display} refers to the added value, one of \code{"Spectral"}, \code{"RdYlGn"}, \code{"RdYlBu"}, \code{"RdGy"},
\code{"RdBu"}, \code{"PuOr"}, \code{"PRGn"}, \code{"PiYG"}, \code{"BrBG"}. Default is \code{"RdBu"}.

If \code{display} does not refer to the added value, one of \code{"Blues"}, \code{"BuGn"}, \code{"BuPu"}, \code{"GnBu"},
\code{"Greens"}, \code{"Greys"}, \code{"Oranges"}, \code{"OrRd"}, \code{"PuBu"}, \code{"PuBuGn"}, \code{"PuRd"}, \code{"Purples"},
\code{"RdPu"}, \code{"Reds"}, \code{"YlGn"}, \code{"YlGnBu"}, \code{"YlOrBr"}, \code{"YlOrRd"}. Default is \code{"Blues"}.}

\item{palette_direction}{Direction in which to use the color palette.
If \code{1}, colors are in original order (e.g. from the lightest to the darkest).
If \code{-1}, color order is reversed (e.g. from the darkest to the lightest).}

\item{plot}{If \code{TRUE}, the chart is plotted in the active graphics device before the return.}
}
\value{
\describe{
\item{\code{graph}}{Graph created with the packages \code{ggraph} and \code{ggplot2}.}
\item{\code{rules}}{Association rules represented on the graph (i.e. of length 2 and considering
the arguments \code{items}, \code{parameter} and \code{threshold}).}
}
}
\description{
Plot a graph in which vertices are items and edges are association rules in which the antecedent is
an item and the consequent is another item.
}
\details{
\loadmathjax
For an association rule \mjeqn{X \rightarrow Y}{X -> Y}, the reciprocal rule is the rule
\mjeqn{Y \rightarrow X}{Y -> X}. These two rules have the same support, lift, frequency and
accuracy but may have two different confidence, specificity and added values.

If \code{display} refers to the selection of rules having the highest or lowest specified characteristic
and if an association rule and its reciprocal have the same value for this characteristic, both
rules are represented.

The chart being plotted with the packages \code{ggraph} and \code{ggplot2}, it can be modified or completed
afterwards using \code{\link[ggplot2:last_plot]{ggplot2::last_plot}} or the returned object.

If \code{category} is \code{NULL} or \code{sort_by = "item"}, items are sorted alphanumerically.
If \code{category} is not \code{NULL} and \code{sort_by = "category"}, items are sorted according to the values of
the category then alphanumerically.

Since other values are returned besides the graph and a graph is automatically plotted if
the return is not assigned to a variable but is not plotted if the return is assigned,
the argument \code{plot} allows to counter this natural effect. The graph can thus be plotted despite
an assignment thanks to this argument. However, if \code{plot = TRUE} and the return is not assigned
to a variable, the graph will be plotted twice.

Default colors associated with the values of each category represented are selected circularly
among the 20 colors of the palette \code{category20} from D3 (see \code{ggsci::pal_d3("category20")}).
Therefore, if the number of values exceeds \code{20}, some colors will be used more than once.
For example, the \out{22<sup>nd</sup>} value will share the color of the \out{2<sup>nd</sup>} one.
See attribute \code{categories_colors} of the \code{TransactionAnalyzer} object to reassign colors to the
category values.
}
\note{
If using the RStudio IDE and if the argument \code{display} refers to the confidence, the specificity or
the added value and \code{direction} is \code{TRUE}, edges may not be displayed in the RStudio "Plots" pane.
However, they will be actually displayed in the "Plot Zoom" window; while exporting the plot; or by
using another graphics device.
Moreover, such plotting may take a while.
}
\examples{
## All rules of length 2
result <- rules_chart(TA_instance, category = "family")
plot(result$graph)
result$rules

## Rules from a data frame
rules_chart(TA_instance, rules = result$rules[11:20, ], category = 1)
rules_chart(TA_instance, rules = result$rules[11:20, ], items = "items")

## Rules relating to specific items
## Display of confidence or display rules of highest or of lowest confidence
rules_chart(TA_instance, items = c(497, 930, 402), category = 1,
            display = "confidence")
rules_chart(TA_instance, items = c(497, 930, 402), category = 1,
            display = "confidence", palette = "category10")
rules_chart(TA_instance, items = c(497, 930, 402), category = 1,
            display = "highest confidence", palette_direction = 1)
rules_chart(TA_instance, items = c(497, 930, 402), category = 1,
            display = "lowest confidence", palette_direction = -1)

## Display of support or lift
rules_chart(TA_instance, category = 1, display = "support")
rules_chart(TA_instance, category = 1, display = "lift")
rules_chart(TA_instance, category = 1, display = "lift",
            threshold = 5, n.cutoff = 20)$graph +
  ggplot2::expand_limits(x = c(-1.5, 1.5), y = c(-1.5, 1.5))

}
\seealso{
\code{\link{extract_rules}}.
}
\author{
Gauthier Magnin
}
