% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{save_characteristics,SpectralAnalyzer-method}
\alias{save_characteristics,SpectralAnalyzer-method}
\title{Enregistre au format CSV un ensemble de caractéristiques de noeuds ou motifs.}
\usage{
\S4method{save_characteristics}{SpectralAnalyzer}(object, entities, characteristics, ...)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{entities}{Type d'élément associé aux caractéristiques (noeuds ou motifs).
Choix parmi \code{c("nodes", "patterns")}.}

\item{characteristics}{Data.frame des caractéristiques des noeuds ou des motifs.}

\item{...}{Arguments fournis à la fonction \code{write.csv2}.}
}
\description{
Enregistre au format CSV un ensemble de caractéristiques de noeuds ou motifs.
}
\seealso{
\code{\link[utils]{write.csv2}}
}
\author{
Gauthier Magnin
}
