% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{save_characteristics,TransactionAnalyzer-method}
\alias{save_characteristics,TransactionAnalyzer-method}
\alias{save_characteristics}
\title{Saving nodes, patterns or association rules}
\usage{
\S4method{save_characteristics}{TransactionAnalyzer}(object, characteristics, ...)
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}

\item{characteristics}{Data frame of the characteristics of nodes, patterns or rules.}

\item{...}{Further arguments to the function \code{\link[utils:write.table]{utils::write.csv2}}.}
}
\description{
Save in CSV format a set of nodes, patterns or association rules as well as their characteristics.
}
\examples{
save_characteristics(TA_instance, TA_instance["nodes"],
                     file = "nodes.csv")
save_characteristics(TA_instance, TA_instance["patterns"][1:15, ],
                     file = "patterns.csv")

spectrosome <- spectrosome_chart(TA_instance, "patterns")
save_characteristics(TA_instance, spectrosome[["vertices"]],
                     file = "spectrosome_vertices.csv", row.names = FALSE)

rules <- extract_rules(TA_instance, from = "transactions")
save_characteristics(TA_instance, rules,
                     file = "rules.csv", row.names = FALSE)

}
\seealso{
\code{\link[utils:write.table]{utils::write.csv2}}.
}
\author{
Gauthier Magnin
}
