% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_notation}
\alias{set_notation}
\title{Turn a vector notation into a set notation}
\usage{
set_notation(sets, type = "character")
}
\arguments{
\item{sets}{Itemsets to write in mathematical notation.}

\item{type}{Type of the return variable. One of \code{"character"}, \code{"factor"}.}
}
\value{
Character vector or factor of the itemsets \code{sets} written in mathematical notation.
}
\description{
Convert a list of character vectors or a character vector into itemsets written in mathematical
 notation.
Example of conversion: \code{c("item i", "item j")} becomes \code{{item i,item j}}.
}
\examples{
sets = list("A", c("B", "C"), c("A", "B"), "A")
set_notation(sets, type = "character")
set_notation(sets, type = "factor")

}
\seealso{
\code{\link{vector_notation}}.
}
\author{
Gauthier Magnin
}
