% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_manager.R
\name{set_operations_on_list}
\alias{set_operations_on_list}
\alias{intersect_on_list}
\alias{union_on_list}
\title{Set operations on list}
\usage{
intersect_on_list(list, indices = seq_along(list))

union_on_list(list, indices = seq_along(list))
}
\arguments{
\item{list}{List on which to perform the operation.}

\item{indices}{Indices or names of the elements on which to perform the operation.}
}
\value{
Vector of values corresponding to the intersection or the union of all specified elements
of the list.
}
\description{
Perform an intersection or union operation on the elements of a list.
}
\examples{
l <- list("e1" = c("A", "B", "C"),
          "e2" = "B",
          "e3" = c("B", "C", "D"))

intersect_on_list(l)
intersect_on_list(l, indices = c(1, 3))

union_on_list(l)
union_on_list(l, indices = c("e1", "e2"))

}
\seealso{
\code{\link{table_on_list}}.
}
\author{
Gauthier Magnin
}
