% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shadowtext}
\alias{shadowtext}
\title{Ombrage de texte}
\usage{
shadowtext(
  x,
  y = NULL,
  labels,
  col = "black",
  bg = "white",
  theta = seq(0, 2 * pi, length.out = 32),
  r = 0.1,
  ...
)
}
\arguments{
\item{x, y}{Coordonnées du texte à afficher.}

\item{labels}{Texte à afficher.}

\item{col}{Couleur du texte.}

\item{bg}{Couleur du contour.}

\item{theta}{Angles à utiliser pour créer le contour.}

\item{r}{Taille du contour.}
}
\description{
Affiche du texte en y ajoutant un contour d'une autre couleur.
Le contour est créé en affichant d'abord le texte avec la seconde couleur, plusieurs fois
 et avec de légers décalages.
}
\references{
Greg Snow. Stack Overflow topic :
 \href{https://stackoverflow.com/questions/25631216/r-plots-is-there-any-way-to-draw-border-shadow-or-buffer-around-text-labels}{Any way to draw border, shadow or buffer around text labels}.
}
\keyword{internal}
