% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{simplify_naming,SpectralAnalyzer-method}
\alias{simplify_naming,SpectralAnalyzer-method}
\alias{simplify_naming}
\title{Simplify names referring to entities or links}
\usage{
\S4method{simplify_naming}{SpectralAnalyzer}(object, name)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{name}{Character corresponding to \code{NODES}, \code{PATTERNS}, \code{RULES}, \code{NODES_LINKS}, \code{PATTERN_LINKS} or
their simplifications. One or more.}
}
\value{
Simplified form of the characters given by \code{name}.
}
\description{
Simplify names referring to entities or links by extracting only the first letter of each word
(each word being separated by the caracter \code{"_"}).
}
\seealso{
\code{\link{which_entities}}, \code{\link{which_associated_links}}, \code{\link{which_name}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
