% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{specific_init,SpectralAnalyzer-method}
\alias{specific_init,SpectralAnalyzer-method}
\alias{init_nodes,SpectralAnalyzer-method}
\alias{init_nodes}
\alias{init_nodes_links,SpectralAnalyzer-method}
\alias{init_nodes_links}
\alias{init_patterns,SpectralAnalyzer-method}
\alias{init_patterns}
\alias{init_patterns_links,SpectralAnalyzer-method}
\alias{init_patterns_links}
\title{Initialization of specific attributes}
\usage{
\S4method{init_nodes}{SpectralAnalyzer}(object, verbose = TRUE)

\S4method{init_nodes_links}{SpectralAnalyzer}(object, verbose = TRUE)

\S4method{init_patterns}{SpectralAnalyzer}(object, verbose = TRUE)

\S4method{init_patterns_links}{SpectralAnalyzer}(object, verbose = TRUE)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{verbose}{Logical value indicating whether to report progress.}
}
\description{
Initialize the attributes related to nodes, links between nodes, patterns and/or links between
patterns.
}
\details{
The initialization of the nodes consists of the initialization of the attributes \code{n_links} and
\code{nodes_links}.

The initialization of the links between nodes consists of the initialization of the attributes
\code{nodes} and \code{nodes_per_year}.

The initialization of the patterns consists of the initialization of the attributes \code{patterns},
\code{patterns_per_year} and \code{nodes_patterns}.

The initialization of the links between patterns consists of the initialization of the attributes
\code{p_links} and \code{patterns_links}.
}
\seealso{
\code{\link{init}}, \code{\link{is_init}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
