% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{specific_init,TransactionAnalyzer-method}
\alias{specific_init,TransactionAnalyzer-method}
\alias{init_nodes,TransactionAnalyzer-method}
\alias{init_nodes}
\alias{init_node_links,TransactionAnalyzer-method}
\alias{init_node_links}
\alias{init_patterns,TransactionAnalyzer-method}
\alias{init_patterns}
\alias{init_pattern_links,TransactionAnalyzer-method}
\alias{init_pattern_links}
\title{Initialization of specific attributes}
\usage{
\S4method{init_nodes}{TransactionAnalyzer}(object, verbose = TRUE)

\S4method{init_node_links}{TransactionAnalyzer}(object, verbose = TRUE)

\S4method{init_patterns}{TransactionAnalyzer}(object, verbose = TRUE)

\S4method{init_pattern_links}{TransactionAnalyzer}(object, verbose = TRUE)
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}

\item{verbose}{Logical value indicating whether to report progress.}
}
\value{
Invisible. Object of class \code{\link[arules:itemsets-class]{itemsets}} (from the package \code{arules})
for the method \code{init_patterns}. \code{NULL} otherwise.
}
\description{
Initialize the attributes related to nodes, links between nodes, patterns and/or links between
patterns.
}
\details{
The initialization of the nodes consists of the initialization of the attributes \code{nodes} and
\code{nodes_per_year}:
\enumerate{
\item{Enumeration of the transactions per year.}
\item{Enumeration of the nodes and calculation of the number of occurrences.}
}

The initialization of the links between nodes consists of the initialization of the attributes
\code{n_links} and \code{node_links}:
\enumerate{
\item{Counting links between nodes.}
\item{Elaboration of links between nodes.}
}

The initialization of the patterns consists of the initialization of the attributes \code{patterns},
\code{patterns_per_year} and \code{nodes_patterns}:
\enumerate{
\item{Mining for itemsets.}
\item{Linking nodes to patterns.}
\item{Enumeration of the patterns per year.}
\item{Computation of pattern characteristics.}
}

The initialization of the links between patterns consists of the initialization of the attributes
\code{p_links} and \code{pattern_links}:
\enumerate{
\item{Counting links between patterns.}
\item{Elaboration of links between patterns.}
}
}
\note{
These methods are not exported from the package.
}
\seealso{
\code{\link{init}}, \code{\link{is_init}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
