% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{specific_is_init,TransactionAnalyzer-method}
\alias{specific_is_init,TransactionAnalyzer-method}
\alias{is_init_nodes,TransactionAnalyzer-method}
\alias{is_init_nodes}
\alias{is_init_node_links,TransactionAnalyzer-method}
\alias{is_init_node_links}
\alias{is_init_patterns,TransactionAnalyzer-method}
\alias{is_init_patterns}
\alias{is_init_pattern_links,TransactionAnalyzer-method}
\alias{is_init_pattern_links}
\title{Test of specific attribute initialization}
\usage{
\S4method{is_init_nodes}{TransactionAnalyzer}(object)

\S4method{is_init_node_links}{TransactionAnalyzer}(object)

\S4method{is_init_patterns}{TransactionAnalyzer}(object)

\S4method{is_init_pattern_links}{TransactionAnalyzer}(object)
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}
}
\value{
\code{TRUE} if the related attributes are initialized. \code{FALSE} otherwise.
}
\description{
Test if the attributes related to nodes, links between nodes, patterns or links between patterns
are initialized.
}
\details{
The initialization of the nodes consists of the initialization of the attributes \code{n_links} and
\code{node_links}.

The initialization of the links between nodes consists of the initialization of the attributes
\code{nodes} and \code{nodes_per_year}.

The initialization of the patterns consists of the initialization of the attributes \code{patterns},
\code{patterns_per_year} and \code{nodes_patterns}.

The initialization of the links between patterns consists of the initialization of the attributes
\code{p_links} and \code{pattern_links}.
}
\seealso{
\code{\link{is_init}}, \code{\link{init}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
