% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{spectral.analyzer}
\alias{spectral.analyzer}
\title{Constructeur d'Analyseur Spectral}
\usage{
spectral.analyzer(
  observations,
  items = NULL,
  target = "closed frequent itemsets",
  count = 1,
  min_length = 1,
  max_length = Inf,
  status_limit = 2
)
}
\arguments{
\item{observations}{Liste des éléments retrouvés pour chaque observation.
Chaque observation est elle-même une liste sous la forme \code{list( CODE = character(), YEAR = numeric )}.
Les valeurs de \code{CODE} ne doivent pas contenir le caractère "/".
Une observation peut contenir des informations supplémentaires quelconques.}

\item{items}{Data frame associant une ou plusieurs catégories à un élément (colonne \code{item}).
La valeur \code{NULL} par défaut précise qu'aucune catégorie n'est définie.}

\item{target}{Type de motifs à énumérer.
Choix parmi \code{"frequent itemsets"}, \code{"closed frequent itemsets"}, \code{"maximally frequent itemsets"}.
Par défaut, \code{"closed frequent itemsets"}, fournissant une synthèse des motifs fréquents
de sorte à économiser l'espace mémoire nécessaire.
Pour énumérer l'intégralité des motifs possibles, utiliser \code{"frequent itemsets"}.}

\item{count}{Nombre minimal d'apparition d'un motif pour être considéré comme "fréquent", et par conséquent, conservé.}

\item{min_length}{Taille minimale qu'un motif doit avoir pour être conservé lors de l'énumération.}

\item{max_length}{Taille maximale qu'un motif doit avoir pour être conservé lors de l'énumération.
La valeur \code{Inf} par défaut correspond à une recherche de motifs sans limite maximale de taille.}

\item{status_limit}{Intervalle de temps pour lequel caractériser le statut des motifs par rapport à la période totale d'observations (nombre d'années).}
}
\value{
Nouvel objet de classe \code{\link{SpectralAnalyzer}}.
}
\description{
Construit un objet de classe SpectralAnalyzer.
}
\seealso{
\code{\link[arules:ASparameter-class]{arules::APparameter}}.
}
\author{
Gauthier Magnin
}
