% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\docType{data}
\name{SpectralAnalyzer.NODES_LINKS}
\alias{SpectralAnalyzer.NODES_LINKS}
\alias{NODES_LINKS}
\title{NODES_LINKS}
\format{
An object of class \code{character} of length 1.
}
\usage{
SpectralAnalyzer.NODES_LINKS
}
\description{
Reference value for naming links between entities: links between nodes.
}
\keyword{internal}
