% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\docType{data}
\name{SpectralAnalyzer.NODES_OR_PATTERNS}
\alias{SpectralAnalyzer.NODES_OR_PATTERNS}
\alias{NODES_OR_PATTERNS}
\title{NODES_OR_PATTERNS}
\format{
An object of class \code{character} of length 1.
}
\usage{
SpectralAnalyzer.NODES_OR_PATTERNS
}
\description{
Reference value for defining possible entities: nodes or patterns.
}
\keyword{internal}
