% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\docType{data}
\name{SpectralAnalyzer.OBSERVATIONS}
\alias{SpectralAnalyzer.OBSERVATIONS}
\alias{OBSERVATIONS}
\title{OBSERVATIONS}
\format{
An object of class \code{character} of length 1.
}
\usage{
SpectralAnalyzer.OBSERVATIONS
}
\description{
Reference value for naming entities: observations.
}
\keyword{internal}
