% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\docType{data}
\name{SpectralAnalyzer.PATTERN_LINKS}
\alias{SpectralAnalyzer.PATTERN_LINKS}
\alias{PATTERN_LINKS}
\title{PATTERN_LINKS}
\format{
An object of class \code{character} of length 1.
}
\usage{
SpectralAnalyzer.PATTERN_LINKS
}
\description{
Reference value for naming links between entities: links between patterns.
}
\keyword{internal}
