% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\docType{data}
\name{SpectralAnalyzer.PATTERNS_LINKS}
\alias{SpectralAnalyzer.PATTERNS_LINKS}
\alias{PATTERNS_LINKS}
\title{PATTERNS_LINKS}
\format{
An object of class \code{character} of length 1.
}
\usage{
SpectralAnalyzer.PATTERNS_LINKS
}
\description{
Reference value for naming links between entities: links between patterns.
}
\keyword{internal}
