% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\docType{data}
\name{SpectralAnalyzer.RULES}
\alias{SpectralAnalyzer.RULES}
\alias{RULES}
\title{RULES}
\format{
An object of class \code{character} of length 1.
}
\usage{
SpectralAnalyzer.RULES
}
\description{
Reference value for naming entities: rules.
}
\keyword{internal}
