% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\docType{data}
\name{SpectralAnalyzer.STATUS_EMERGENT}
\alias{SpectralAnalyzer.STATUS_EMERGENT}
\alias{STATUS_EMERGENT}
\title{STATUS_EMERGENT}
\format{
An object of class \code{character} of length 1.
}
\usage{
SpectralAnalyzer.STATUS_EMERGENT
}
\description{
Reference value for pattern dynamic status: emergent.
}
\keyword{internal}
