% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\docType{data}
\name{SpectralAnalyzer.STATUS_LATENT}
\alias{SpectralAnalyzer.STATUS_LATENT}
\alias{STATUS_LATENT}
\title{STATUS_LATENT}
\format{
An object of class \code{character} of length 1.
}
\usage{
SpectralAnalyzer.STATUS_LATENT
}
\description{
Reference value for pattern dynamic status: latent.
}
\keyword{internal}
