% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\docType{data}
\name{SpectralAnalyzer.STATUS_PERSISTENT}
\alias{SpectralAnalyzer.STATUS_PERSISTENT}
\alias{STATUS_PERSISTENT}
\title{STATUS_PERSISTENT}
\format{
An object of class \code{character} of length 1.
}
\usage{
SpectralAnalyzer.STATUS_PERSISTENT
}
\description{
Reference value for pattern dynamic status: persistent.
}
\keyword{internal}
