% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{spectrosome_chart,SpectralAnalyzer-method}
\alias{spectrosome_chart,SpectralAnalyzer-method}
\title{Construit un ou plusieurs graphiques de type spectrosome et les enregistre au format PNG.}
\usage{
\S4method{spectrosome_chart}{SpectralAnalyzer}(
  object,
  entities,
  characteristics,
  nb_graph = 1,
  vertex_size = "relative",
  path = getwd(),
  name = paste0("spectrosome_of_", entities, ".png"),
  title = paste0("Network of ", entities)
)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{entities}{Type d'élément pour lequel construire le spectrosome (noeuds ou motifs).
Choix parmi \code{c("nodes", "patterns")}.}

\item{characteristics}{Ensemble des caractéristiques des noeuds ou motifs dont le spectrosome est à tracer.}

\item{nb_graph}{Nombre de graphes à générer et enregistrer.}

\item{vertex_size}{Façon dont les tailles des sommets du graphe doivent être définies.
Choix parmi \code{c("relative", "grouped", "absolute", "equal")}.
\code{"relative"} : la taille d'un sommet dépend de l'intervalle de valeurs des différents poids.
\code{"grouped"} : les poids des motifs sont regroupés selon des intervalles. À chaque intervalle correspond une taille.
\code{"absolute"} : la taille d'un sommet est définie directement en fonction du poids du motif.
\code{"equal"} : les sommets ont tous la même taille.}

\item{path}{Chemin du dossier dans lequel enregistrer les graphiques.
Par défaut, les graphiques sont enregistrés dans le répertoire de travail.}

\item{name}{Nom (avec extension) du fichier dans lequel enregistrer le graphique.
Si \code{nb_graph} est supérieur à \code{1}, un numéro est ajouté automatiquement à la fin du nom du fichier.}

\item{title}{Titre du graphique.}
}
\value{
Liste contenant :
 \itemize{
   \item{} {Data.frame des noeuds ou motifs et caractéristiques utilisées, associés aux identifiants des sommets du graphe.}
   \item{} {Data.frame des informations relatives aux arêtes du graphe.}
 }
}
\description{
Construit un ou plusieurs graphiques de type spectrosome et les enregistre au format PNG.
}
\details{
Si des catégories sont associées aux items, chaque catégorie génère un spectrosome.
 Le nom de la catégorie est ajouté à la fin du nom du fichier.
}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
            \emph{The spectrosome of occupational health problems}. PLoS ONE 13(1): e0190196.
            \url{https://doi.org/10.1371/journal.pone.0190196}.
}
\author{
Delphine Bosson-Rieutort

Gauthier Magnin
}
