% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{spectrosome_chart,TransactionAnalyzer-method}
\alias{spectrosome_chart,TransactionAnalyzer-method}
\alias{spectrosome_chart}
\title{Spectrosome}
\usage{
\S4method{spectrosome_chart}{TransactionAnalyzer}(
  object,
  nopc,
  identifiers = "original",
  nb_graphs = 1,
  min_link_weight = 1,
  vertex_size = "relative",
  size_range = c(0.5, 2.5),
  vertex_col = "status",
  clusters = Inf,
  highlight = 3,
  use_names = TRUE,
  n.cutoff = NULL,
  c.cutoff = NULL,
  display_mixt = TRUE,
  title = NULL,
  path = NULL,
  name = NULL,
  ...
)
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}

\item{nopc}{Data frame of \strong{n}odes \strong{o}r \strong{p}atterns and their
\strong{c}haracteristics. Nodes or patterns whose spectrosome is to be plotted. Any subset of
\code{object["nodes"]} or \code{object["patterns"]}.

\code{"nodes"}, \code{"n"}, \code{"patterns"} and \code{"p"} are special values for
\code{object["nodes"]} and \code{object["patterns"]}.}

\item{identifiers}{Which IDs to use to identify the nodes or patterns on the chart and in the
return data frames. One of \code{"original"}, \code{"new"}.
\describe{
  \item{\code{"original"}}{Use of the original identifiers.}
  \item{\code{"new"}}{Use of new identifiers ordered according to the subset corresponding to
                      \code{nopc}.}
}}

\item{nb_graphs}{Number of graphs to generate. The position of the vertices differs between each copy.}

\item{min_link_weight}{Minimum number of items in common between two entities to plot their link on
the chart.}

\item{vertex_size}{Way how the sizes of the vertices of the graph should be defined.
One of \code{"relative"}, \code{"grouped"}, \code{"absolute"}, \code{"equal"} or a numeric value
or vector of length equal to the number of nodes or patterns to plot.
\describe{
  \item{\code{"relative"}}{The sizes are defined by a linear interpolation of the frequencies of the
                           entities in \code{size_range}.}
  \item{\code{"grouped"}}{The frequencies of the entities are grouped according to 5 intervals
                          defined by quantiles. The 5 corresponding size values are taken in a
                          regular sequence bounded by \code{size_range}.}
  \item{\code{"absolute"}}{The size of a vertex is defined directly according to the frequency of
                           the entity.}
  \item{\code{"equal"}}{The vertices are all the same size of 1.}
  \item{A single numeric value}{The vertices are all the size defined by this value.}
  \item{A longer numeric vector}{The sizes defined in this vector are directly assigned to the
                                 nodes of patterns to plot.}
}}

\item{size_range}{If \code{vertex_size} is \code{"relative"} or \code{"grouped"},
range of vertex size values (given as expansion factors). Ignored otherwise.}

\item{vertex_col}{Way how the colors of the vertices of the graph should be defined.
One of \code{"status"}, \code{"categories"}, \code{"none"} or a single character value or vector
of length equal to the number of nodes or patterns to plot (characters corresponding to R
predefined color names or hexadecimal values).

If \code{"status"} and \code{nopc} refers to patterns, coloring according to the status of the
patterns. If \code{"categories"}, coloring according to the categories associated with the items of
the entities represented. If one specific color, all vertices are colored with this color.
If a vector of size equal to the number of nodes or patterns to plot, the colors are directly
assigned to these entities. If \code{"none"}, the vertices are colored gray.}

\item{clusters}{Maximum number of clusters to name on the graph.
If the actual number of clusters is greater, the names of the smaller clusters are not displayed.}

\item{highlight}{Number of clusters to highlight among those named on the graph.
The names of the largest clusters are displayed in bold.}

\item{use_names}{If \code{TRUE}, display item names if they are defined. Display their identification
codes otherwise.}

\item{n.cutoff}{If \code{use_names = TRUE}, limit number of characters to display concerning the names
of the represented items.}

\item{c.cutoff}{Limit number of characters to display in the legend for the categories represented.}

\item{display_mixt}{If \code{TRUE}, also display in the legend the category values included only in
mixed links (or in mixed vertices, if \code{vertex_col = "categories"}).}

\item{title}{Chart title. Default title depends on the type of entities contained in \code{nopc}.
Example of default title: \code{"Spectrosome of nodes"} if \code{nopc} contains nodes.}

\item{path}{Path of the directory in which to save the charts as PNG files. Default is the working
directory.}

\item{name}{Single name of the PNG files in which to save the charts. To be ignored to plot the charts
in the active device.}

\item{...}{Additional arguments to the function \code{\link[sna:gplot]{gplot}} from the package
\code{sna} for plotting the graph. See Details section.}
}
\value{
\describe{
   \item{\code{vertices}}{Data frame of the nodes or patterns and characteristics used,
                          associated with the identifiers of the vertices of the graph and their
                          degrees in the graph.}
   \item{\code{edges}}{Data frame of information relating to the edges of the graph.}
   \item{\code{coords}}{List containing the coordinate matrices of the vertices of the graph.
                        As many matrices as there are charts (\code{nb_graphs}), which can be
                        reused via the argument \code{coord} (see \code{...}).}
 }
}
\description{
Plot one or more spectrosome charts: graph in which vertices are nodes or patterns and edges are the
 items they have in common. Clusters corresponding to nodes or patterns sharing one item are
 mentioned by displaying the name or code identifying the item between these nodes or patterns.
 It can be automatically saved as a PNG file.
}
\details{
If the argument \code{name} is not \code{NULL}, charts are plotted in PNG files of size 950 x 700
 pixels. If it is \code{NULL}, charts are plotted in the active device, one behind the other.

If \code{nb_graphs} is greater than \code{1} and \code{name} is not \code{NULL} a number is
 automatically added to the end of the file names to distinguish the different files.
 
If categories are associated with the items, each category generates a spectrosome. If there is more
 than one category and \code{name} is not \code{NULL}, the category name is appended to the end of
 the file name (after the possible addition of the number of the graph).

If mixed links are relative to category values that are not represented by single links, these
 values are present in the legend below "Mixt", with no associated color.
The same is true for mixed vertices if the argument \code{vertex_col} is \code{"categories"}.

If \code{min_link_weight} is greater than 1 or the chart is to be plotted from a subset of
 all nodes or patterns, some of them may become isolated because their links to the other
 entities may no longer be considered.
These new isolated vertices are moved to the end of the return data frame \code{edges}.
The \code{n} related rows are numbered \code{"A1"..."An"}.

Default colors associated with the values of each category represented are selected circularly
among the 20 colors of the palette \code{category20} from D3 (see \code{ggsci::pal_d3("category20")}).
Therefore, if the number of values exceeds \code{20}, some colors will be used more than once.
For example, the \out{22<sup>nd</sup>} value will share the color of the \out{2<sup>nd</sup>} one.
See attribute \code{categories_colors} of the \code{TransactionAnalyzer} object to reassign colors to the
category values.

The names of clusters confused because all of their links are mixed links, are not displayed.

Additional arguments can be supplied to the function in charge of plotting the graph.
 See the list of parameters: \code{\link[sna:gplot]{sna::gplot}}.
Among them, the following parameters are already defined and cannot be modified: \code{dat},
 \code{gmode}, \code{vertex.sides}, \code{vertex.cex}, \code{vertex.col}, \code{edge.col}.
The following parameters, which are among those that can be redefined, have the following values:
 \itemize{
   \item{\code{mode = "fruchtermanreingold"}}
   \item{\code{displaylabels = TRUE}}
   \item{\code{label.pos = 0}}
   \item{\code{boxed.labels = TRUE}}
   \item{\code{displayisolates = TRUE}}
 }
}
\examples{
spectrosome_1 <- spectrosome_chart(TA_instance, "nodes")
spectrosome_2 <- spectrosome_chart(TA_instance, TA_instance["patterns"])
spectrosome_3 <- spectrosome_chart(TA_instance, TA_instance["patterns"][1:15, ])

spectrosome_2 <- spectrosome_chart(TA_instance, "patterns",
                                   path = getwd(),
                                   name = "spectrosome_of_patterns")

}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
            The spectrosome of occupational health problems. \emph{PLoS ONE} 13(1): e0190196.
            \url{https://doi.org/10.1371/journal.pone.0190196}.
}
\seealso{
\code{\link{degree}}, \code{\link{network_density}}, \code{\link[sna:gplot]{sna::gplot}}.
}
\author{
Delphine Bosson-Rieutort, Gauthier Magnin
}
