% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{spectrum_chart,SpectralAnalyzer-method}
\alias{spectrum_chart,SpectralAnalyzer-method}
\alias{spectrum_chart}
\title{Spectre des motifs}
\usage{
\S4method{spectrum_chart}{SpectralAnalyzer}(
  object,
  patterns_characteristics,
  path = getwd(),
  name = "spectrum_of_patterns.pdf",
  title = "Spectrum of patterns"
)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{patterns_characteristics}{Ensemble des caractéristiques des motifs dont le spectre est à tracer.}

\item{path}{Chemin du dossier dans lequel enregistrer le graphique.
Par défaut, le graphique est enregistré dans le répertoire de travail.}

\item{name}{Nom du fichier dans lequel enregistrer le graphique.}

\item{title}{Titre du graphique.}
}
\value{
Data frame des motifs et caractéristiques utilisées, associés aux identifiants visibles sur le graphique.
}
\description{
Construit un graphique de type spectre et l'enregistre dans un fichier au format PDF.
}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
            \emph{The spectrosome of occupational health problems}. PLoS ONE 13(1): e0190196.
            \url{https://doi.org/10.1371/journal.pone.0190196}.
}
\author{
Delphine Bosson-Rieutort, Gauthier Magnin
}
