% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\docType{data}
\name{STATUS_DECLINING}
\alias{STATUS_DECLINING}
\title{STATUS_DECLINING}
\format{
An object of class \code{character} of length 1.
}
\usage{
STATUS_DECLINING
}
\description{
Reference value for pattern dynamic status: declining.
}
\keyword{internal}
