% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{[,SpectralAnalyzer,ANY,ANY,ANY-method}
\alias{[,SpectralAnalyzer,ANY,ANY,ANY-method}
\alias{[,SpectralAnalyzer-method}
\alias{[<-,SpectralAnalyzer,ANY,ANY,ANY-method}
\alias{[<-,SpectralAnalyzer-method}
\title{Extract or replace parts of a \code{SpectralAnalyzer} object}
\usage{
\S4method{[}{SpectralAnalyzer,ANY,ANY,ANY}(x, i, j, drop)

\S4method{[}{SpectralAnalyzer,ANY,ANY,ANY}(x, i, j) <- value
}
\arguments{
\item{x}{
    object from which to extract element(s) or in which to replace element(s).
  }

\item{i}{
    indices specifying elements to extract or replace.  Indices are
    \code{numeric} or \code{character} vectors or empty (missing) or
    \code{NULL}.  Numeric values are coerced to integer as by
    \code{\link[base]{as.integer}} (and hence truncated towards zero).
    Character vectors will be matched to the \code{\link[base]{names}} of the
    object (or for matrices/arrays, the \code{\link[base]{dimnames}}):
    see \sQuote{Character indices} below for further details.

    For \code{[}-indexing only: \code{i}, \code{j}, \code{\dots} can be
    logical vectors, indicating elements/slices to select.  Such vectors
    are recycled if necessary to match the corresponding extent.
    \code{i}, \code{j}, \code{\dots} can also be negative integers,
    indicating elements/slices to leave out of the selection.

    When indexing arrays by \code{[} a single argument \code{i} can be a
    matrix with as many columns as there are dimensions of \code{x}; the
    result is then a vector with elements corresponding to the sets of
    indices in each row of \code{i}.

    An index value of \code{NULL} is treated as if it were \code{integer(0)}.
  }

\item{j}{
    indices specifying elements to extract or replace.  Indices are
    \code{numeric} or \code{character} vectors or empty (missing) or
    \code{NULL}.  Numeric values are coerced to integer as by
    \code{\link[base]{as.integer}} (and hence truncated towards zero).
    Character vectors will be matched to the \code{\link[base]{names}} of the
    object (or for matrices/arrays, the \code{\link[base]{dimnames}}):
    see \sQuote{Character indices} below for further details.

    For \code{[}-indexing only: \code{i}, \code{j}, \code{\dots} can be
    logical vectors, indicating elements/slices to select.  Such vectors
    are recycled if necessary to match the corresponding extent.
    \code{i}, \code{j}, \code{\dots} can also be negative integers,
    indicating elements/slices to leave out of the selection.

    When indexing arrays by \code{[} a single argument \code{i} can be a
    matrix with as many columns as there are dimensions of \code{x}; the
    result is then a vector with elements corresponding to the sets of
    indices in each row of \code{i}.

    An index value of \code{NULL} is treated as if it were \code{integer(0)}.
  }

\item{drop}{For matrices and arrays.  If \code{TRUE} the result is
    coerced to the lowest possible dimension (see the examples).  This
    only works for extracting elements, not for the replacement.  See
    \code{\link[base]{drop}} for further details.
  }

\item{value}{typically an array-like \R object of a similar class as
    \code{x}.}
}
\description{
General selector and mutator to access the attributes of an object of class \code{SpectralAnalyzer}.
Extraction and replacement can be done by using an attribute name or its numeric value in the order
 of the attributes.
}
\examples{
SA_instance["items"]
SA_instance["patterns"]
SA_instance[1]
SA_instance["target"]

SA_instance["target"] <- "maximally frequent itemsets"
SA_instance["count"] <- 2
SA_instance["max_length"] <- 3

}
\author{
Gauthier Magnin
}
