% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{[,TransactionAnalyzer,ANY,ANY,ANY-method}
\alias{[,TransactionAnalyzer,ANY,ANY,ANY-method}
\alias{[,TransactionAnalyzer-method}
\alias{[<-,TransactionAnalyzer,ANY,ANY,ANY-method}
\alias{[<-,TransactionAnalyzer-method}
\title{Extract or replace parts of an object of class TransactionAnalyzer}
\usage{
\S4method{[}{TransactionAnalyzer,ANY,ANY,ANY}(x, i)

\S4method{[}{TransactionAnalyzer,ANY,ANY,ANY}(x, i) <- value
}
\arguments{
\item{x}{Object from which to extract element(s) or in which to replace
element(s).}

\item{i}{Numeric or character values. Indices specifying elements to extract
or replace. See 'Details' section.}

\item{value}{Value of type similar to the element to be replaced.}
}
\description{
General selector and mutator to access the attributes of an object of class
 \code{TransactionAnalyzer}.
}
\details{
Character values can be used to access attributes as well as elements of the
 attribute \code{parameters}.

Numeric values can also be used to access attributes considering their order.
}
\examples{
TA_instance["items"]
TA_instance["patterns"]
TA_instance[1]
TA_instance["target"]

TA_instance["target"] <- "maximally frequent itemsets"
TA_instance["min_frequency"] <- 2
TA_instance["max_length"] <- 3

}
\author{
Gauthier Magnin
}
