% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_set.R
\name{[,TransactionSet,ANY,ANY,ANY-method}
\alias{[,TransactionSet,ANY,ANY,ANY-method}
\alias{[,TransactionSet-method}
\alias{[<-,TransactionSet,ANY,ANY,ANY-method}
\alias{[<-,TransactionSet-method}
\title{Extract or replace parts of an object of class TransactionSet}
\usage{
\S4method{[}{TransactionSet,ANY,ANY,ANY}(x, i, j, drop)

\S4method{[}{TransactionSet,ANY,ANY,ANY}(x, i, j) <- value
}
\arguments{
\item{x}{
    object from which to extract element(s) or in which to replace element(s).
  }

\item{i}{
    indices specifying elements to extract or replace.  Indices are
    \code{numeric} or \code{character} vectors or empty (missing) or
    \code{NULL}.  Numeric values are coerced to integer as by
    \code{\link[base]{as.integer}} (and hence truncated towards zero).
    Character vectors will be matched to the \code{\link[base]{names}} of the
    object (or for matrices/arrays, the \code{\link[base]{dimnames}}):
    see \sQuote{Character indices} below for further details.

    For \code{[}-indexing only: \code{i}, \code{j}, \code{\dots} can be
    logical vectors, indicating elements/slices to select.  Such vectors
    are recycled if necessary to match the corresponding extent.
    \code{i}, \code{j}, \code{\dots} can also be negative integers,
    indicating elements/slices to leave out of the selection.

    When indexing arrays by \code{[} a single argument \code{i} can be a
    matrix with as many columns as there are dimensions of \code{x}; the
    result is then a vector with elements corresponding to the sets of
    indices in each row of \code{i}.

    An index value of \code{NULL} is treated as if it were \code{integer(0)}.
  }

\item{j}{
    indices specifying elements to extract or replace.  Indices are
    \code{numeric} or \code{character} vectors or empty (missing) or
    \code{NULL}.  Numeric values are coerced to integer as by
    \code{\link[base]{as.integer}} (and hence truncated towards zero).
    Character vectors will be matched to the \code{\link[base]{names}} of the
    object (or for matrices/arrays, the \code{\link[base]{dimnames}}):
    see \sQuote{Character indices} below for further details.

    For \code{[}-indexing only: \code{i}, \code{j}, \code{\dots} can be
    logical vectors, indicating elements/slices to select.  Such vectors
    are recycled if necessary to match the corresponding extent.
    \code{i}, \code{j}, \code{\dots} can also be negative integers,
    indicating elements/slices to leave out of the selection.

    When indexing arrays by \code{[} a single argument \code{i} can be a
    matrix with as many columns as there are dimensions of \code{x}; the
    result is then a vector with elements corresponding to the sets of
    indices in each row of \code{i}.

    An index value of \code{NULL} is treated as if it were \code{integer(0)}.
  }

\item{drop}{For matrices and arrays.  If \code{TRUE} the result is
    coerced to the lowest possible dimension (see the examples).  This
    only works for extracting elements, not for the replacement.  See
    \code{\link[base]{drop}} for further details.
  }

\item{value}{typically an array-like \R object of a similar class as
    \code{x}.}
}
\description{
General selector and mutator to access the attributes of an object of class \code{TransactionSet}.
Extraction and replacement can be done by using an attribute name.
}
\details{
Sub-elements of the attribute \code{data} can be accessed this way as well.

Numeric values can be used to access elements of the attribute \code{data}. Replacing one element of
\code{data} will not change its name.

Replacing the attribute \code{names}:
\itemize{
\item reorders the elements contained in each transaction if \code{value} is a
reordered equivalent of \code{names};
\item renames the elements contained in each transaction if \code{value} is the same
length as \code{names} but is not equivalent. In this case, the attributes
\code{item_key} and \code{year_key} are updated (unless \code{year_key} is \code{NA});
\item removes missing elements of \code{value} in each transaction if it is smaller
than \code{names}.
}
}
\examples{
TS_instance["data"]

TS_instance["item_key"]
TS_instance["CODE"]

TS_instance[3]
TS_instance[3:5]

TS_instance["year_key"] <- NA
TS_instance[3] <- list(CODE = 1,
                       YEAR = 2000,
                       JOB.TITLE = 45454545,
                       JOB.TASK = "A3000",
                       SAMPLE.ID = c(1, 3, 5, 8))
TS_instance["names"] <- c("CODE", "YEAR")

}
\author{
Gauthier Magnin
}
