% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_set.R
\name{subset,ObservationSet-method}
\alias{subset,ObservationSet-method}
\alias{subset}
\title{Subsetting Observation Set}
\usage{
\S4method{subset}{ObservationSet}(x, indexes, keep_names = TRUE)
}
\arguments{
\item{x}{Object of class \code{ObservationSet} to be subsetted.}

\item{indexes}{Numeric vector indicating which observations from \code{x["data"]} to keep,
or logical vector indicating for each observation whether to keep it.}

\item{keep_names}{If \code{TRUE}, observations of the returned object keep the names from the initial object
\code{x}. If \code{FALSE}, they are not named.}
}
\value{
S4 object of class \code{ObservationSet} having a subset of observations from \code{x} (i.e. a subset of
the attribute \code{data} from \code{x}).
}
\description{
Return an \code{ObservationSet} in which observations are a subset of the ones of another \code{ObservationSet}.
}
\details{
If the observations from \code{x} are not named and \code{keep_names = TRUE}, the observations of the resulting
object are named according to the indexes of the initial observations.
}
\examples{
subset(OS_instance, c(1, 7, 8, 9))
subset(OS_instance, c(TRUE, rep(FALSE, 12), TRUE))

}
\author{
Gauthier Magnin
}
