% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixer.R
\name{subset_from_class}
\alias{subset_from_class}
\title{Subsets of values and references by class}
\usage{
subset_from_class(values, references, classes, class_name)
}
\arguments{
\item{values}{Numeric named matrix or list of numeric named vectors. Values from which to extract a
subset according to one specific class.}

\item{references}{Numeric vector or list of numeric vectors. Reference values associated with the
\code{values}. See 'Details' to know the way it is associated with \code{values}.}

\item{classes}{Logical matrix whose columns are named according to the classes and the row names
contain the names associated with the \code{values}. A \code{TRUE} value indicates that a specific name
is part of a specific class.}

\item{class_name}{Name of the class of interest, the one for which to extract the corresponding
subsets of \code{values} and \code{references}.}
}
\value{
\describe{
\item{\code{values}}{Subset of values that correspond to the class of interest.}
\item{\code{references}}{Subset of references that correspond to the class of interest.}
}
}
\description{
Extract from values and references those corresponding to one specific class.
}
\details{
If \code{values} is a matrix, it must have one reference value for each row.

If \code{values} is a list, the reference values can be a vector of named values or a list. In this case,
if \code{references} is a vector, there must be one reference for each name present in \code{values}.
Otherwise, \code{references} is a list of vectors having the same lengths as those present in \code{values}
so that \code{values} and \code{references} can be matched.
}
\seealso{
\code{\link{mcr_summary_by_class}}, \code{\link{mcr_chart_by_class}}, \code{\link{thq_pairs_freq_by_class}},
\code{\link{thq_freq_by_group_by_class}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
