% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{substances_classification}
\alias{substances_classification}
\title{Classification of substances by family}
\format{
A data frame with 1,151 rows and 4 variables:
 \describe{
   \item{CODE}{Code of a substance.}
   \item{NAME}{Name of the substance corresponding to the code.}
   \item{FAMILY}{Family of the substance.}
   \item{SUBFAMILY}{Subfamily of the substance.}
 }
}
\usage{
substances_classification
}
\description{
Classification associating each substance with a unique code, a family and a subfamily.
}
\keyword{datasets}
