% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{substances_information}
\alias{substances_information}
\title{Information about substances}
\format{
A data frame with 1,151 rows and 6 variables:
 \describe{
   \item{CODE}{Unique identifier of a substance.}
   \item{NAME}{Name of the substance corresponding to the code.}
   \item{FAMILY}{Family the substance belongs to.}
   \item{SUBFAMILY}{Subfamily the substance belongs to.}
   \item{TOXICITY}{Toxicity classes associated with the substance.}
   \item{LIMIT}{Exposure limit value.}
 }
}
\source{
Data from the \href{http://en.inrs.fr/}{INRS}.
}
\usage{
substances_information
}
\description{
Information about substances such as family of substances, toxicity classes...
}
\keyword{datasets}
