% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{substr2}
\alias{substr2}
\title{Substrings of a character vector}
\usage{
substr2(x, start = 1, stop = NULL)
}
\arguments{
\item{x}{A character vector.}

\item{start}{Indice of the first character to extract from each element of \code{x}.}

\item{stop}{Indice of the last character to extract from each element of \code{x}.
If \code{NULL}, all characters from \code{start} to the end of the element are extracted.}
}
\value{
A character vector of the same length as \code{x} containing the extracted parts.
}
\description{
Extract substrings in a character vector. Wrapper of the \code{\link{substr}} function allowing to use
\code{NULL} value.
}
\seealso{
\code{\link{cap}}, \code{\link{pluralize}}, \code{\link{first_characters}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
