% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_manager.R
\name{table_on_list}
\alias{table_on_list}
\title{Frequency table on list}
\usage{
table_on_list(list, indices = seq_along(list), with_zero = FALSE)
}
\arguments{
\item{list}{List for which to build the frequency table.}

\item{indices}{Indices or names of the elements on which to perform the operation. Can be a vector,
a matrix or a list.}

\item{with_zero}{If \code{FALSE}, count only occurrences of effective values. If \code{TRUE}, add
the frequencies of the values which are in \code{list} but not in the elements specified by \code{indices}.}
}
\value{
Table, matrix or list of tables (according to \code{indices}) of the frequencies of the values of
the specified elements of the list.
}
\description{
Build a frequency table of the counts of each value in a list.
}
\examples{
l <- list(e1 = c("E", "F", "I"),
          e2 = "I",
          e3 = c("C", "I"),
          e4 = c("A", "C", "D", "F"),
          e5 = c("B", "D", "E", "I", "G"))

## With indices as a vector
table_on_list(l)
table_on_list(l, indices = c(1,2,3))
table_on_list(l, indices = c("e1", "e2", "e3"), with_zero = TRUE)

## With indices as a matrix
table_on_list(l, indices = matrix(c("e1", "e2", "e3",
                                    "e1", "e2", "e4",
                                    "e2", "e3", "e5"),
                                  nrow = 3, byrow = TRUE))
table_on_list(l, indices = matrix(c(1, 2, 3,
                                    1, 2, 4,
                                    2, 3, 5),
                                  nrow = 3, byrow = TRUE,
                                  dimnames = list(c("v1", "v2", "v3"))),
              with_zero = TRUE)

## With indices as a list
table_on_list(l, indices = list(c("e1", "e2", "e3"),
                                c("e1", "e2"),
                                "e1"))
table_on_list(l, indices = list(v1 = c(1, 2, 3),
                                v2 = c(1, 2),
                                v3 = 1),
              with_zero = TRUE)

}
\seealso{
\code{\link{intersect_on_list}}, \code{\link{union_on_list}}.
}
\author{
Gauthier Magnin
}
