% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixer.R
\name{thq_by_group}
\alias{thq_by_group}
\title{Top Hazard Quotients frequency by group}
\usage{
thq_by_group(values, references, levels = NULL)
thq_by_group(hq, groups, levels = NULL)
thq_by_group(thq, groups, levels = NULL)
}
\arguments{
\item{values}{Numeric named matrix or list of numeric named vectors. Vectors of values for which the
table is to be built.}

\item{references}{Numeric vector or list of numeric vectors. Reference values associated with the
\code{values}. See 'Details' to know the way it is associated with \code{values}.}

\item{hq}{Numeric named matrix or list of numeric named vectors. \strong{H}azard \strong{q}uotients for which
the table is to be built.}

\item{thq}{Numeric named vector or list of numeric named vectors. \strong{T}op \strong{h}azard \strong{q}uotients
to use to the count. If list, only the first named value of each element of the list is considered.}

\item{groups}{Character vector. MIAT groups associated with the hazard quotients \code{hq} or with the
top hazard quotients \code{thq}.}

\item{levels}{Levels to consider in the output table. If \code{NULL}, only use of those that appear in the
top hazard quotients.}
}
\value{
Contingency table. Frequencies of elements that produce the top hazard quotient with their
associated groups.
}
\description{
Build a contingency table of the counts of each combination of name of the element producing the
top hazard quotient with the associated MIAT group. The groups describe the following situations:
\itemize{
\item Group I: the mixture presents a potential risk already based on individual components.
\item Group II: the assessment does not identify a concern.
\item Group IIIA: the majority of the risk offered by the mixture is driven by one component.
\item Group IIIB: the potential risk is driven by multiple components.
}
}
\details{
If \code{values} is a matrix, the reference values are applied once on each column (i.e. it must have one
reference value for each row of the matrix).

If \code{values} is a list, the reference values can be a vector of named values or a list. In this case,
if \code{references} is a vector, there must be one reference for each name present in \code{values}.
Otherwise, \code{references} is a list of vectors having the same lengths as those present in \code{values}
so that \code{values} and \code{references} can be matched.

Arguments \code{values} and \code{references} are used to compute the hazard quotients and the hazard indexes
before searching for the maximum hazard quotients, computing the maximum cumulative ratios, performing
the classification then build the table. Thus, call the function with the arguments \code{hq} and \code{groups}
is faster and call it with the arguments \code{thq} and \code{groups} is even faster (if they are already
computed).

\loadmathjax
The mixtures are assigned to the groups according the following conditions:
\itemize{
\item Group I: \mjeqn{MHQ_i \ge 1}{MHQ_i >= 1}
\item Group II: \mjeqn{MHQ_i < 1, HI_i \le 1}{MHQ_i < 1, HI_i <= 1}
\item Group IIIA: \mjeqn{MHQ_i < 1, HI_i > 1, MCR_i < 2}{MHQ_i < 1, HI_i > 1, MCR_i < 2}
\item Group IIIB: \mjeqn{MHQ_i < 1, HI_i > 1, MCR_i \ge 2}{MHQ_i < 1, HI_i > 1, MCR_i >= 2}
}

The maximum cumulative ratio of the vector \eqn{i} is given by:
\mjdeqn{MCR_i = \frac{HI_i}{MHQ_i}}{MCR_i = HI_i / MHQ_i}
where \eqn{HI} denotes the hazard index and \eqn{MHQ} denotes the maximum hazard quotient.

The hazard index of the vector \eqn{i} is given by:
\mjdeqn{HI_i = \sum_{j = 1}^n HQ_{i,j}}{HI_i = sum(HQ_ij) from j = 1 to n}
where \eqn{HQ} denotes the hazard quotients and \eqn{n} denotes the number of hazard quotients.

The maximum hazard quotient of the vector \eqn{i} is given by:
\mjdeqn{MHQ_i = HQ_{M,i} = \max_{j \in \lbrace 1,...,n\rbrace} HQ_{i,j}}{MHQ_i = HQ_Mi = max HQ_i}
where \eqn{HQ} denotes the hazard quotients and \eqn{n} denotes the number of hazard quotients.

The hazard quotient of the value \eqn{j} in the vector \eqn{i} is given by:
\mjdeqn{HQ_{i,j} = \frac{V_{i,j}}{RV_j}}{HQ_ij = V_ij / RV_j}
where \eqn{V} denotes the \code{values} and \eqn{RV} denotes the \code{references}.
}
\note{
Due to the multiple possible usages, the arguments \code{hq}, \code{thq}, \code{groups} and \code{levels} must be
explicitly named in the function call.
}
\examples{
## Creating a matrix of 4*3 values and one reference value for each of the 4
## elements (A, B, C, and D).
v <- matrix(c(.1, .2, 1, .4, .5, .6, .7, .8, 3, 1, 1, 1),
            ncol = 3, dimnames = list(LETTERS[1:4]))
r <- c(1, 2, 3, 0.5)

## Without levels parameter and with the different usages
thq_by_group(v, r)
thq_by_group(hq = hazard_quotient(v, r),
             groups = classify_mixture(v, r))
thq_by_group(thq = top_hazard_quotient(v, r),
             groups = classify_mixture(v, r))

## With levels parameter
thq_by_group(values = v, references = r, levels = LETTERS[1:4])

## Building contingency table from a list
thq_by_group(values = list(c(A = 0.1, B = 0.5),
                           c(A = 0.2),
                           c(B = 0.3, C = 0.4)),
             references = c(A = 1, B = 2, C = 3))
thq_by_group(values = list(c(A = 0.1, B = 0.5),
                           c(A = 0.2),
                           c(B = 0.3, C = 0.4)),
             references = list(c(1, 2),
                               1,
                               c(2, 3)))

}
\references{
Reyes JM, Price PS (2018).
An analysis of cumulative risks based on biomonitoring data for six phthalates using the Maximum Cumulative Ratio.
\emph{Environment International}, 112, 77-84.
\url{https://doi.org/10.1016/j.envint.2017.12.008}.
}
\seealso{
\code{\link{thq_pairs}}, \code{\link{classify_mixture}}, \code{\link{top_hazard_quotient}}, \code{\link{hazard_quotient}}.
}
\author{
Gauthier Magnin
}
