% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixer.R
\name{thq_by_group_by_class}
\alias{thq_by_group_by_class}
\title{Top Hazard Quotients frequency by group and by class}
\usage{
thq_by_group_by_class(values, references, classes, levels = NULL)
}
\arguments{
\item{values}{Numeric named matrix or list of numeric named vectors. Vectors of values for which the
tables are to be built, according to classes.}

\item{references}{Numeric vector or list of numeric vectors. Reference values associated with the
\code{values}. See 'Details' to know the way it is associated with \code{values}.}

\item{classes}{List or logical matrix associating the \code{values} names with classes.
If list, its names are those present in \code{values} and the elements are vectors of associated classes.
If logical matrix, its columns are named according to the classes and the row names
contain the names associated with the \code{values}. A \code{TRUE} value indicates that a specific name
is part of a specific class.}

\item{levels}{Levels to consider in the output tables. If \code{NULL}, only use of those that appear in the
top hazard quotients.}
}
\value{
List of contingency tables: frequencies of elements that produce the top hazard quotient with
their associated groups. The length of the list corresponds to the number of classes encountered.
}
\description{
Build contingency tables of the counts of each combination of name of the element producing the
top hazard quotient with the associated MIAT group, according to classes. For each class, one table
is built from the subset of values corresponding to this class. The groups describe the following
situations:
\itemize{
\item Group I: the mixture presents a potential risk already based on individual components.
\item Group II: the assessment does not identify a concern.
\item Group IIIA: the majority of the risk offered by the mixture is driven by one component.
\item Group IIIB: the potential risk is driven by multiple components.
}
}
\details{
If \code{values} is a matrix, the reference values are applied once on each column (i.e. it must have one
reference value for each row of the matrix).

If \code{values} is a list, the reference values can be a vector of named values or a list. In this case,
if \code{references} is a vector, there must be one reference for each name present in \code{values}.
Otherwise, \code{references} is a list of vectors having the same lengths as those present in \code{values}
so that \code{values} and \code{references} can be matched.

If \code{classes} is a list, it will be turned into a logical matrix before processing. Thus, call the
function with such a matrix is slightly faster.

\loadmathjax
The mixtures are assigned to the groups according the following conditions:
\itemize{
\item Group I: \mjeqn{MHQ_i \ge 1}{MHQ_i >= 1}
\item Group II: \mjeqn{MHQ_i < 1, HI_i \le 1}{MHQ_i < 1, HI_i <= 1}
\item Group IIIA: \mjeqn{MHQ_i < 1, HI_i > 1, MCR_i < 2}{MHQ_i < 1, HI_i > 1, MCR_i < 2}
\item Group IIIB: \mjeqn{MHQ_i < 1, HI_i > 1, MCR_i \ge 2}{MHQ_i < 1, HI_i > 1, MCR_i >= 2}
}

The maximum cumulative ratio of the vector \eqn{i} is given by:
\mjdeqn{MCR_i = \frac{HI_i}{MHQ_i}}{MCR_i = HI_i / MHQ_i}
where \eqn{HI} denotes the hazard index and \eqn{MHQ} denotes the maximum hazard quotient.

The hazard index of the vector \eqn{i} is given by:
\mjdeqn{HI_i = \sum_{j = 1}^n HQ_{i,j}}{HI_i = sum(HQ_ij) from j = 1 to n}
where \eqn{HQ} denotes the hazard quotients and \eqn{n} denotes the number of hazard quotients.

The maximum hazard quotient of the vector \eqn{i} is given by:
\mjdeqn{MHQ_i = HQ_{M,i} = \max_{j \in \lbrace 1,...,n\rbrace} HQ_{i,j}}{MHQ_i = HQ_Mi = max HQ_i}
where \eqn{HQ} denotes the hazard quotients and \eqn{n} denotes the number of hazard quotients.

The hazard quotient of the value \eqn{j} in the vector \eqn{i} is given by:
\mjdeqn{HQ_{i,j} = \frac{V_{i,j}}{RV_j}}{HQ_ij = V_ij / RV_j}
where \eqn{V} denotes the \code{values} and \eqn{RV} denotes the \code{references}.

If several values are equal to the maximum hazard quotient, the name retained as the top hazard
quotient is the first one considering the given order.
}
\examples{
## Creating a matrix of 5*50 values, one reference value for each of the 5
## elements (A, B, C, D and E) and association of classes (C1 to C8) with
## these elements.
v <- matrix(sample(seq(0.1, 2.1, by = 0.1), 250, replace = TRUE),
            ncol = 50, dimnames = list(LETTERS[1:5]))
r <- sample(seq(1,5), 5, replace = TRUE)
classes <- list(A = c("C5", "C6", "C8"),
                B = "C8",
                C = c("C3", "C8"),
                D = c("C1", "C3", "C4", "C6"),
                E = c("C2", "C4", "C5", "C7", "C8"))

## Building contingency table from matrix, without and with levels parameter
thq_by_group_by_class(v, r, classes)
thq_by_group_by_class(v, r, classes, levels = names(classes))

## Building contingency table from list
thq_by_group_by_class(values = list(V1 = c(A = 1, B = 5),
                                    V2 = c(A = 2),
                                    V3 = c(B = 3, C = 4)),
                      references = list(c(1, 2),
                                        1,
                                        c(2, 3)),
                      classes)
thq_by_group_by_class(values = list(V1 = c(A = 1, B = 5),
                                    V2 = c(A = 2),
                                    V3 = c(B = 3, C = 4)),
                      references = c(A = 1, B = 2, C = 3),
                      classes,
                      levels = LETTERS[1:3])

}
\references{
Reyes JM, Price PS (2018).
An analysis of cumulative risks based on biomonitoring data for six phthalates using the Maximum Cumulative Ratio.
\emph{Environment International}, 112, 77-84.
\url{https://doi.org/10.1016/j.envint.2017.12.008}.
}
\seealso{
Contingency table independent of classes: \code{\link{thq_by_group}}.

Generic function to apply the MCR approach according to classes: \code{\link{mcr_approach_by_class}}.

Other functions of the MCR approach applying according to classes: \code{\link{mcr_summary_by_class}},
\code{\link{mcr_chart_by_class}}, \code{\link{thq_pairs_by_class}}.

Specific indicators: \code{\link{classify_mixture}}, \code{\link{top_hazard_quotient}}, \code{\link{hazard_quotient}}.
}
\author{
Gauthier Magnin
}
