% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixer.R
\name{thq_freq_by_group}
\alias{thq_freq_by_group}
\title{Top Hazard Quotients frequency by group}
\usage{
thq_freq_by_group(values, references, levels = NULL)
thq_freq_by_group(hq, groups, levels = NULL)
thq_freq_by_group(thq, groups, levels = NULL)
}
\arguments{
\item{values}{Numeric named matrix. Vectors of values for which the table is to be build.}

\item{references}{Numeric vector. Reference values associated with the \code{values}.}

\item{hq}{Numeric named matrix. \strong{H}azard \strong{q}uotients for which the table is to be build.}

\item{thq}{Numeric named vector. \strong{T}op \strong{h}azard \strong{q}uotients to use to the count.}

\item{groups}{Character vector. MIAT groups associated with the hazard quotients \code{hq} or with the
top hazard quotients \code{thq}.}

\item{levels}{Levels to consider in the output table. If \code{NULL}, only use of those that appear in the
top hazard quotients.}
}
\value{
Contingency table. Frequency of element that produced the top hazard quotient with its
associated group.
}
\description{
Build a contingency table of the counts of each combination of name of the element producing the
top hazard quotient with the associated MIAT group.
}
\details{
Arguments \code{values} and \code{references} are used to compute the hazard quotients and the hazard indexes
before searching for the maximum hazard quotients, computing the maximum cumulative ratios, performing
the classification then build the table. Thus, call the function with the arguments \code{hq} and \code{groups}
is faster and call it with the arguments \code{thq} and \code{groups} is even faster (if they are already
computed).

\loadmathjax
The mixtures are assigned to the groups according the following conditions:
\itemize{
\item Group I: \mjeqn{MHQ_i \ge 1}{MHQ_i >= 1}
\item Group II: \mjeqn{MHQ_i < 1, HI_i \le 1}{MHQ_i < 1, HI_i <= 1}
\item Group IIIA: \mjeqn{MHQ_i < 1, HI_i > 1, MCR_i < 2}{MHQ_i < 1, HI_i > 1, MCR_i < 2}
\item Group IIIB: \mjeqn{MHQ_i < 1, HI_i > 1, MCR_i \ge 2}{MHQ_i < 1, HI_i > 1, MCR_i >= 2}
}

The maximum cumulative ratio of the vector \eqn{i} is given by:
\mjdeqn{MCR_i = \frac{HI_i}{MHQ_i}}{MCR_i = HI_i / MHQ_i}
where \eqn{HI} denotes the hazard index and \eqn{MHQ} denotes the maximum hazard quotient.

The hazard index of the vector \eqn{i} is given by:
\mjdeqn{HI_i = \sum_{j = 1}^N HQ_{i,j}}{HI_i = sum(HQ_ij) from j = 1 to N}
where \eqn{HQ} denotes the hazard quotients and \eqn{N} denotes the number of hazard quotients.

The maximum hazard quotient of the vector \eqn{i} is given by:
\mjdeqn{MHQ_i = HQ_{M,i} = \max_{j \in \lbrace 1,...,N\rbrace} HQ_{i,j}}{MHQ_i = HQ_Mi = max HQ_i}
where \eqn{HQ} denotes the hazard quotients and \eqn{N} denotes the number of hazard quotients.

The hazard quotient of the value \eqn{j} in the vector \eqn{i} is given by:
\mjdeqn{HQ_{i,j} = \frac{V_{i,j}}{RV_j}}{HQ_ij = V_ij / RV_j}
where \eqn{V} denotes the \code{values} and \eqn{RV} denotes the \code{references}.
}
\note{
Due to the multiple possible usages, the arguments \code{hq}, \code{thq}, \code{groups} and \code{levels} must be
explicitly named in the function call.
}
\examples{
thq_freq_by_group(c(a = 1, b = 2, c = 3, d = 4, e = 5), c(5,4,3,2,1))
thq_freq_by_group(hq = hazard_quotient(c(a = 1, b = 2, c = 3, d = 4, e = 5),
                                       c(5,4,3,2,1)),
                  groups = classify_mixtures(c(a = 1, b = 2, c = 3, d = 4, e = 5),
                                             c(5,4,3,2,1)))
thq_freq_by_group(thq = top_hazard_quotient(c(a = 1, b = 2, c = 3, d = 4, e = 5),
                                            c(5,4,3,2,1)),
                  groups = classify_mixtures(c(a = 1, b = 2, c = 3, d = 4, e = 5),
                                             c(5,4,3,2,1)))

## With and without levels parameter
thq_freq_by_group(values = matrix(c(.1, .2, 1, .4, .5, .6, .7, .8, 3, 1, 1, 1),
                                  ncol = 3, dimnames = list(letters[1:4])),
                  references = c(1,2,3,0.5))
thq_freq_by_group(values = matrix(c(.1, .2, 1, .4, .5, .6, .7, .8, 3, 1, 1, 1),
                                  ncol = 3, dimnames = list(letters[1:4])),
                  references = c(1,2,3,0.5),
                  levels = letters[1:4])

}
\references{
Reyes JM, Price PS (2018).
An analysis of cumulative risks based on biomonitoring data for six phthalates using the Maximum Cumulative Ratio.
\emph{Environment International}, 112, 77-84.
\url{https://doi.org/10.1016/j.envint.2017.12.008}.
}
\seealso{
\code{\link{thq_pairs_freq}}, \code{\link{classify_mixture}}, \code{\link{top_hazard_quotient}}, \code{\link{hazard_quotient}}.
}
\author{
Gauthier Magnin
}
