% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixer.R
\name{thq_pairs}
\alias{thq_pairs}
\title{Top Hazard Quotient pairs frequency}
\usage{
thq_pairs(values, references,
          levels = NULL,
          threshold = TRUE,
          alone = FALSE)
thq_pairs(hq, hi,
          levels = NULL,
          threshold = TRUE,
          alone = FALSE)
}
\arguments{
\item{values}{Numeric named matrix or list of numeric named vectors. Vectors of values for which the
top two hazard quotients are to be identified.}

\item{references}{Numeric vector or list of numeric vectors. Reference values associated with the
\code{values}. See 'Details' to know the way it is associated with \code{values}.}

\item{hq}{Numeric named matrix or list of numeric named vectors. \strong{H}azard \strong{q}uotients for which
the top two pairs are to be identified.}

\item{hi}{Numeric vector. \strong{H}azard \strong{i}ndexes associated with the hazard quotients \code{hq}.}

\item{levels}{Levels to consider in the output table. If \code{NULL}, only use of those that appear in the
pairs.}

\item{threshold}{If \code{TRUE}, only values or hazard quotients associated with hazard indexes greater
than 1 are considered.}

\item{alone}{If \code{TRUE}, take into account single top hazard quotients (i.e. sets of values of length
1). If so, a level named \code{"NULL"} is added as combination with such top hazard quotients.}
}
\value{
\code{NULL} if (1) \code{alone = FALSE} and no set of \code{values} (or \code{hq}) has more than one element different
from 0
or (2) \code{threshold = TRUE} and no related hazard index is greater than 1
or (3) \code{threshold = TRUE}, \code{alone = FALSE} and no set of values meets the two conditions.

Contingency table otherwise. Frequencies of pairs that produce the top two hazard quotients,
considering all sets of values or not (accordingly to the arguments \code{threshold} and \code{alone}).
}
\description{
Build a contingency table of the counts of each combination of the top two hazard quotients pairs for
which the associated hazard index is greater than 1.
}
\details{
If \code{values} is a matrix, the reference values are applied once on each column (i.e. it must have one
reference value for each row of the matrix).

If \code{values} is a list, the reference values can be a vector of named values or a list. In this case,
if \code{references} is a vector, there must be one reference for each name present in \code{values}.
Otherwise, \code{references} is a list of vectors having the same lengths as those present in \code{values}
so that \code{values} and \code{references} can be matched.

Values and hazard quotients equal to 0 are ignored.

Arguments \code{values} and \code{references} are used to compute the hazard quotients and the hazard indexes
before identifying the highest hazard quotients then building the contingency table. Thus, call the
function with the arguments \code{hq} and \code{hi} is faster (if they are already computed).

\loadmathjax
The hazard index of the vector \eqn{i} is given by:
\mjdeqn{HI_i = \sum_{j = 1}^n HQ_{i,j}}{HI_i = sum(HQ_ij) from j = 1 to n}
where \eqn{HQ} denotes the hazard quotients and \eqn{n} denotes the number of hazard quotients.

The hazard quotient of the value \eqn{j} in the vector \eqn{i} is given by:
\mjdeqn{HQ_{i,j} = \frac{V_{i,j}}{RV_j}}{HQ_ij = V_ij / RV_j}
where \eqn{V} denotes the \code{values} and \eqn{RV} denotes the \code{references}.

If the number of hazard quotients that are greater than or equal to the second greater hazard
quotient is greater than 2, only the first two values are considered and in the order given.
For example, if the hazard quotients are \code{D = 5}, \code{B = 1}, \code{C = 3} and \code{A = 3}, the top two hazard
quotients considered are \code{D} and \code{C}.
}
\note{
Due to the multiple possible usages, all arguments except \code{values} and \code{references} must be explicitly
named in the function call.
}
\examples{
thq_pairs(values = matrix(c(1, .2, .3, .4, .5, .6, .7, .8, .9, 1),
                          ncol = 2, dimnames = list(LETTERS[1:5])),
          references = c(1,2,3,4,5))
thq_pairs(hq = hazard_quotient(matrix(c(1, .2, .3, .4, .5, .6, .7, .8, .9, 1),
                                      ncol = 2, dimnames = list(LETTERS[1:5])),
                               c(1,2,3,4,5)),
          hi = hazard_index(matrix(c(1, .2, .3, .4, .5, .6, .7, .8, .9, 1),
                                   ncol = 2, dimnames = list(LETTERS[1:5])),
                            c(1,2,3,4,5)))

## With and without levels parameter
thq_pairs(values = matrix(c(.1, .2, 1, .4, .5, .6, .7, .8, 3, 1, 1, 1),
                          ncol = 3, dimnames = list(LETTERS[1:4])),
          references = c(1, 2, 3, .5),
          levels = LETTERS[1:4])
thq_pairs(values = matrix(c(.1, .2, 1, .4, .5, .6, .7, .8, 3, 1, 1, 1),
                          ncol = 3, dimnames = list(LETTERS[1:4])),
          references = c(1,2,3,0.5))

## NULL because all HI are lower than or equal to 1
thq_pairs(values = matrix(c(.1, .2, .3, .4),
                          ncol = 2, dimnames = list(c("A","B"))),
          references = c(5,5))
hazard_index(values = matrix(c(.1, .2, .3, .4),
                             ncol = 2, dimnames = list(c("A","B"))),
             references = c(5,5))
thq_pairs(values = matrix(c(.1, .2, .3, .4),
                          ncol = 2, dimnames = list(c("A","B"))),
          references = c(5,5),
          threshold = FALSE)

## Building contingency table from a list
thq_pairs(values = list(c(A = 0.5, B = 0.5),
                        c(A = 1),
                        c(B = 0.5, C = 0.5)),
          references = list(c(0.3, 0.6),
                            0.3,
                            c(0.6, 1)))
thq_pairs(values = list(c(A = 0.5, B = 0.5),
                        c(A = 1),
                        c(B = 0.5, C = 0.5)),
          references = c(A = 0.3, B = 0.6, C = 1))
thq_pairs(values = list(c(A = 0.5, B = 0.5),
                        c(A = 1),
                        c(B = 0.5, C = 0.5)),
          references = c(A = 0.3, B = 0.6, C = 1),
          alone = TRUE)

}
\references{
Reyes JM, Price PS (2018).
An analysis of cumulative risks based on biomonitoring data for six phthalates using the Maximum Cumulative Ratio.
\emph{Environment International}, 112, 77-84.
\url{https://doi.org/10.1016/j.envint.2017.12.008}.
}
\seealso{
\code{\link{thq_by_group}}, \code{\link{top_hazard_quotient}}, \code{\link{hazard_quotient}}, \code{\link{hazard_index}}.
}
\author{
Gauthier Magnin
}
