% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixer.R
\name{thq_pairs_by_class}
\alias{thq_pairs_by_class}
\title{Top Hazard Quotient pairs frequency by class}
\usage{
thq_pairs_by_class(
  values,
  references,
  classes,
  levels = NULL,
  threshold = TRUE,
  alone = FALSE
)
}
\arguments{
\item{values}{Numeric named matrix or list of numeric named vectors. Vectors of values for which the
top two hazard quotients are to be identified, according to classes.}

\item{references}{Numeric vector or list of numeric vectors. Reference values associated with the
\code{values}. See 'Details' to know the way it is associated with \code{values}.}

\item{classes}{List or logical matrix associating the \code{values} names with classes.
If list, its names are those present in \code{values} and the elements are vectors of associated classes.
If logical matrix, its columns are named according to the classes and the row names
contain the names associated with the \code{values}. A \code{TRUE} value indicates that a specific name
is part of a specific class.}

\item{levels}{Levels to consider in the output tables. If \code{NULL}, only use of those that appear in the
pairs.}

\item{threshold}{If \code{TRUE}, only values associated with hazard indexes greater than 1 are considered.}

\item{alone}{If \code{TRUE}, take into account single top hazard quotients (i.e. sets of values of length
1). If so, a level named \code{"NULL"} is added for such top hazard quotients.}
}
\value{
List whose length corresponds to the number of classes encountered, containing for each class:
\itemize{
\item \code{NULL} if among the subsets of values corresponding to the class, no one has more than one element
different from 0 or a hazard index greater than 1 (accordingly to the arguments \code{threshold} and
\code{alone}).
\item Contingency table otherwise. Frequencies of pairs that produce the top two hazard quotients.
}
}
\description{
Build contingency tables of the counts of each combination of the top two hazard quotients pairs for
which the associated hazard index is greater than 1, according to classes. For each class, one
table is built from the subset of values corresponding to this class.
}
\details{
If \code{values} is a matrix, the reference values are applied once on each set of values, i.e. on each row.
Therefore, there must be one reference value for each column of the matrix.

If \code{values} is a list, the reference values can be a vector of named values or a list. In this case,
if \code{references} is a vector, there must be one reference for each name present in \code{values}.
Otherwise, \code{references} is a list of vectors having the same lengths as those present in \code{values}
so that \code{values} and \code{references} can be matched.

Values equal to 0 are ignored.

If \code{classes} is a list, it will be turned into a logical matrix before processing. Thus, call the
function with such a matrix is slightly faster.

\loadmathjax
The hazard index of the vector \eqn{i} is given by:
\mjdeqn{HI_i = \sum_{j = 1}^n HQ_{i,j}}{HI_i = sum(HQ_ij) from j = 1 to n}
where \eqn{HQ} denotes the hazard quotients and \eqn{n} denotes the number of hazard quotients.

The hazard quotient of the value \eqn{j} in the vector \eqn{i} is given by:
\mjdeqn{HQ_{i,j} = \frac{V_{i,j}}{RV_j}}{HQ_ij = V_ij / RV_j}
where \eqn{V} denotes the \code{values} and \eqn{RV} denotes the \code{references}.

If the number of hazard quotients that are greater than or equal to the second greater hazard
quotient is greater than 2, only the first two values are considered and in the order given.
For example, if the hazard quotients are \code{D = 5}, \code{B = 1}, \code{C = 3} and \code{A = 3}, the top two hazard
quotients considered are \code{D} and \code{C}.
}
\examples{
## Creating a matrix of 50*5 values, one reference value for each of the 5
## elements (A, B, C, D and E) and association of classes (C1 to C8) with
## these elements.
v <- matrix(sample(seq(0.1, 2.1, by = 0.1), 250, replace = TRUE),
            nrow = 50, byrow = TRUE,
            dimnames = list(paste0("set.", 1:50), LETTERS[1:5]))
r <- sample(seq(1,5), 5, replace = TRUE)
classes <- list(A = c("C5", "C6", "C8"),
                B = "C8",
                C = c("C3", "C8"),
                D = c("C1", "C3", "C4", "C6"),
                E = c("C2", "C4", "C5", "C7", "C8"))

## Building contingency table from matrix, without and with levels parameter
thq_pairs_by_class(v, r, classes)
thq_pairs_by_class(v, r, classes, levels = names(classes))

## Building contingency table from list
thq_pairs_by_class(values = list(set.1 = c(A = 1, B = 5),
                                 set.2 = c(A = 2),
                                 set.3 = c(B = 3, C = 4)),
                   references = list(c(1, 2),
                                     1,
                                     c(2, 3)),
                   classes)
thq_pairs_by_class(values = list(set.1 = c(A = 1, B = 5),
                                 set.2 = c(A = 2),
                                 set.3 = c(B = 3, C = 4)),
                   references = c(A = 1, B = 2, C = 3),
                   classes,
                   levels = LETTERS[1:3])

# Use of the parameters alone and threshold
thq_pairs_by_class(values = list(set.1 = c(A = 1, B = 5),
                                 set.2 = c(A = 2),
                                 set.3 = c(B = 3, C = 4)),
                   references = c(A = 1, B = 2, C = 3),
                   classes,
                   levels = LETTERS[1:3],
                   alone = TRUE)
thq_pairs_by_class(values = list(set.1 = c(A = 1, B = 5),
                                 set.2 = c(A = 2),
                                 set.3 = c(B = 3, C = 4)),
                   references = c(A = 1, B = 2, C = 3),
                   classes,
                   levels = LETTERS[1:3],
                   threshold = FALSE, alone = TRUE)

}
\references{
Reyes JM, Price PS (2018).
An analysis of cumulative risks based on biomonitoring data for six phthalates using the Maximum Cumulative Ratio.
\emph{Environment International}, 112, 77-84.
\url{https://doi.org/10.1016/j.envint.2017.12.008}.
}
\seealso{
Contingency table independent of classes: \code{\link{thq_pairs}}.

Generic function to apply the MCR approach according to classes: \code{\link{mcr_approach_by_class}}.

Other functions of the MCR approach applying according to classes: \code{\link{mcr_summary_by_class}},
\code{\link{mcr_chart_by_class}}, \code{\link{thq_by_group_by_class}}.

Specific indicators: \code{\link{top_hazard_quotient}}, \code{\link{hazard_quotient}}.
}
\author{
Gauthier Magnin
}
