% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixer.R
\name{thq_pairs_freq}
\alias{thq_pairs_freq}
\title{Top Hazard Quotient pairs frequency}
\usage{
thq_pairs_freq(values, references, levels = NULL)
thq_pairs_freq(hq, hi, levels = NULL)
}
\arguments{
\item{values}{Numeric named matrix. Vectors of values for which the top two hazard quotients are to be
identified.}

\item{references}{Numeric vector. Reference values associated with the \code{values}.}

\item{hq}{Numeric named matrix. \strong{H}azard \strong{q}uotients for which the top two pairs are be to
identified.}

\item{hi}{Numeric vector. \strong{H}azard \strong{i}ndexes associated with the hazard quotients \code{hq}.}

\item{levels}{Levels to consider in the output table. If \code{NULL}, only use of those that appear in the
pairs.}
}
\value{
\code{NULL} if no hazard index is greater than 1.

Contingency table otherwise. Frequency of pairs that produced the top two hazard quotients while hazard
index is greater than 1.
}
\description{
Build a contingency table of the counts of each combination of the top two hazard quotients pairs for
which the associated hazard index is greater than 1.
}
\details{
Arguments \code{values} and \code{references} are used to compute the hazard quotients and the hazard indexes
before identifying the highest hazard quotients then building the contingency table. Thus, call the
function with the arguments \code{hq} and \code{hi} is faster (if they are already computed).

The reference values are applied once on each column of \code{values} to compute the hazard indexes
(i.e. it must have one reference value for each row of the matrix).

\loadmathjax
The hazard index of the vector \eqn{i} is given by:
\mjdeqn{HI_i = \sum_{j = 1}^N HQ_{i,j}}{HI_i = sum(HQ_ij) from j = 1 to N}
where \eqn{HQ} denotes the hazard quotients and \eqn{N} denotes the number of hazard quotients.

The hazard quotient of the value \eqn{j} in the vector \eqn{i} is given by:
\mjdeqn{HQ_{i,j} = \frac{V_{i,j}}{RV_j}}{HQ_ij = V_ij / RV_j}
where \eqn{V} denotes the \code{values} and \eqn{RV} denotes the \code{references}.
}
\note{
Due to the multiple possible usages, the arguments \code{hi}, \code{hq} and \code{levels} must be explicitly named in
the function call.
}
\examples{
thq_pairs_freq(values = matrix(c(1, .2, .3, .4, .5, .6, .7, .8, .9, 1),
                               ncol = 2, dimnames = list(letters[1:5])),
               references = c(1,2,3,4,5))
thq_pairs_freq(hq = hazard_quotient(matrix(c(1, .2, .3, .4, .5, .6, .7, .8, .9, 1),
                                           ncol = 2, dimnames = list(letters[1:5])),
                                    c(1,2,3,4,5)),
               hi = hazard_index(matrix(c(1, .2, .3, .4, .5, .6, .7, .8, .9, 1),
                                        ncol = 2, dimnames = list(letters[1:5])),
                                 c(1,2,3,4,5)))

## With and without levels parameter
thq_pairs_freq(values = matrix(c(.1, .2, 1, .4, .5, .6, .7, .8, 3, 1, 1, 1),
                               ncol = 3, dimnames = list(letters[1:4])),
               references = c(1, 2, 3, .5),
               levels = letters[1:4])
thq_pairs_freq(values = matrix(c(.1, .2, 1, .4, .5, .6, .7, .8, 3, 1, 1, 1),
                               ncol = 3, dimnames = list(letters[1:4])),
               references = c(1,2,3,0.5))

## NULL because all HI are lower than or equal to 1
thq_pairs_freq(values = matrix(c(.1, .2, .3, .4),
                               ncol = 2, dimnames = list(c("a","b"))),
               references = c(5,5))
hazard_index(values = matrix(c(.1, .2, .3, .4),
                             ncol = 2, dimnames = list(c("a","b"))),
             references = c(5,5))

}
\references{
Reyes JM, Price PS (2018).
An analysis of cumulative risks based on biomonitoring data for six phthalates using the Maximum Cumulative Ratio.
\emph{Environment International}, 112, 77-84.
\url{https://doi.org/10.1016/j.envint.2017.12.008}.
}
\seealso{
\code{\link{thq_freq_by_group}}, \code{\link{top_hazard_quotient}}, \code{\link{hazard_quotient}}, \code{\link{hazard_index}}.
}
\author{
Gauthier Magnin
}
