% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixer.R
\name{top_hazard_quotient}
\alias{top_hazard_quotient}
\title{Top Hazard Quotients (THQ)}
\usage{
top_hazard_quotient(values, references, k)
top_hazard_quotient(hq, k)
}
\arguments{
\item{values}{Numeric named vector or matrix. Values for which the top hazard quotients are to be
identify.}

\item{references}{Numeric vector. Reference values associated with the \code{values}.}

\item{hq}{Numeric named vector or matrix. \strong{H}azard \strong{q}uotients whose highest values are to be
identify.}

\item{k}{Number of hazard quotients to highlight. Default is the integer part of the maximum cumulative
ratio computed from \code{values} and \code{references}.}
}
\value{
Vector or list of vectors (according to \code{values} or \code{hq}) of the \code{k} highest hazard quotients.
}
\description{
Identify the top \code{k} hazard quotients, i.e. the named values which produce the \code{k} highest hazard
quotients.
}
\details{
Arguments \code{values} and \code{references} are used to compute the hazard quotients before identifying the
highest ones. Thus, call the function with the argument \code{hq} is faster (if it is already computed).

If \code{values} is a matrix, the reference values are applied once on each column (i.e. it must have one
reference value for each row of the matrix).

If \code{values} or \code{hq} is a matrix, \code{k} hazard quotients are highlight for each column.

If the number of hazard quotients that are greater than or equal to the \eqn{k-th} greater hazard
quotient is greater than \code{k}, only the first \code{k} values are considered and in the order given.
For example, if \code{hq = c(d = 5, b = 1, c = 3, a = 3)} and \code{k = 2}, the return is \code{c(d = 5, c = 3)}.

\loadmathjax
The hazard quotient of the value \eqn{j} in the vector \eqn{i} is given by:
\mjdeqn{HQ_{i,j} = \frac{V_{i,j}}{RV_j}}{HQ_ij = V_ij / RV_j}
where \eqn{V} denotes the \code{values} and \eqn{RV} denotes the \code{references}.
}
\note{
Due to the multiple possible usages, the arguments \code{hq} and \code{k} must be explicitly named in the
function call.
}
\examples{
top_hazard_quotient(c(a = 1, b = 2, c = 3, d = 4, e = 5), c(5,4,3,2,1),
                    k = 3)
top_hazard_quotient(hq = hazard_quotient(c(a = 1, b = 2, c = 3, d = 4, e = 5),
                                         c(5,4,3,2,1)),
                    k = 3)
top_hazard_quotient(values = matrix(c(.1, .2, .3, .4, .5, .6, .7, .8, .9, 1),
                                    ncol = 2, dimnames = list(letters[1:5])),
                    references = c(1,2,3,4,5))

}
\references{
Reyes JM, Price PS (2018).
An analysis of cumulative risks based on biomonitoring data for six phthalates using the Maximum Cumulative Ratio.
\emph{Environment International}, 112, 77-84.
\url{https://doi.org/10.1016/j.envint.2017.12.008}.
}
\seealso{
\code{\link{hazard_quotient}}, \code{\link{maximum_hazard_quotient}}.
}
\author{
Gauthier Magnin
}
