% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_set.R
\name{transaction.set}
\alias{transaction.set}
\title{Transaction Set constructor}
\usage{
transaction.set(data, item_key, year_key = NA, names = NULL)
}
\arguments{
\item{data}{List of transactions containing items and any additional data.
Each transaction is itself a list. All transactions must share the same element names.}

\item{item_key}{Name of the element containing the items of a transaction.}

\item{year_key}{Name of the element containing the year in which a transaction was made.}

\item{names}{Names of the elements present in each transaction (i.e. names of the elements contained in
the list corresponding to a transaction).}
}
\value{
New object of class \code{TransactionSet}.
}
\description{
Create and initialize an S4 object of class \code{TransactionSet}.
}
\details{
Transactions do not necessarily require temporal data since \code{year_key} is not required.

If names are not specified using the argument \code{names}, they are automatically extracted from
\code{data}. Specifying \code{names} allows to create \code{TransactionSet} (or subset of existing \code{TransactionSet})
without any transaction.
}
\examples{
## Creating a list of transactions then a TransactionSet
trx <- list(T1 = list(items = "A",
                      date = 2018,
                      labels = c("l1", "l2")),
            
            T2 = list(items = c("A", "B", "D"),
                      date = 2018,
                      labels = "l3"),
            
            T3 = list(items = c("B", "C", "D"),
                      date = 2019,
                      labels = c("l2", "l3")),
            
            T4 = list(items = c("A", "B", "C", "D"),
                      date = 2020,
                      labels = c("l1", "l2", "l3"))
            )
trx_object <- transaction.set(trx, item_key = "items", year_key = "date")
print(trx_object)

}
\seealso{
The class: \code{\link{TransactionSet}}.

Selectors and mutators to access the attributes:
\code{\link[=[,TransactionSet-method]{[}},
\code{\link[=[,TransactionSet-method]{[[}},
\code{\link[=[,TransactionSet-method]{$}}.
}
\author{
Gauthier Magnin
}
