% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\docType{class}
\name{TransactionAnalyzer-class}
\alias{TransactionAnalyzer-class}
\alias{TransactionAnalyzer}
\alias{print,TransactionAnalyzer-method}
\alias{summary,TransactionAnalyzer-method}
\title{Transaction Analyzer}
\description{
S4 object class allowing transaction analysis.
 It consists of mining frequent itemsets from data, compute characteristics of these itemsets
 (specificity, dynamic status...), plot them on charts (spectrum, spectrosome...) and extract
 association rules.
 Items from itemsets can be associated with categories used in charts and computations.
}
\details{
Default colors are assigned to the values of each category associated with the items.
 These colors are selected circularly among the 20 colors of the palette \code{category20}
 from D3 (see \code{ggsci::pal_d3("category20")}).
Therefore, if the number of values exceeds \code{20}, some colors will be used more than once.
 For example, the \out{22<sup>nd</sup>} value will share the color of the \out{2<sup>nd</sup>}
 value.
See the attribute \code{categories_colors} to reassign colors to the category values.
}
\section{Slots}{

\describe{
\item{\code{transactions}}{S4 object of class \code{TransactionSet}: list of transactions containing the items
corresponding to each one. It represents the dataset in which frequent itemsets are to be mined.}

\item{\code{items}}{Set of codes identifying the items found in the transactions.}

\item{\code{items_categories}}{Categories associated with the items. Each item is associated with one value
for each category.}

\item{\code{categories_colors}}{Colors associated with the values of the categories associated with the items.}

\item{\code{status_colors}}{Colors associated with the values of the status the patterns can have.}

\item{\code{parameters}}{List of parameters for pattern search and characterization:
\describe{
  \item{\code{target}}{Type of patterns to mine during the analysis.}
  \item{\code{min_frequency}}{Minimum number of occurrences that a pattern must appear to be kept
                              when mining patterns.}
  \item{\code{min_length}}{Minimum number of items that a pattern must have to be kept when mining
                           patterns.}
  \item{\code{max_length}}{Maximum number of items that a pattern must have to be kept when mining
                           patterns.}
  \item{\code{status_limit}}{Time interval for which to characterize the status of the patterns in
                             relation to the total period of transactions (number of years).}
}}

\item{\code{nodes}}{Set of nodes (separate transactions considering only their items) and characteristics of
these nodes. Data frame of 3 variables:
\describe{
  \item{\code{node}}{Set of items composing the node.}
  \item{\code{length}}{Number of items composing the node.}
  \item{\code{frequency}}{Number of transactions for which the set of items matches exactly.}
}}

\item{\code{nodes_per_year}}{Number of occurrences of each node in the transactions, per year.}

\item{\code{n_links}}{Set of weights of the links between the nodes. Adjacency matrix containing the number
of items in common between each pair of nodes.}

\item{\code{node_links}}{Set of links between the nodes and characteristics of these links.
Isolated nodes (i.e. unrelated to any other node) appear at the bottom of the data structure.
Data frame of 4 variables:
\describe{
  \item{\code{endpoint.1}, \code{endpoint.2}}{Identifiers of two nodes from the attribute
                                              \code{nodes}.}
  \item{\code{items}}{Items in common between the two nodes.}
  \item{\code{weight}}{Number of items in common between the two nodes.}
}}

\item{\code{nodes_patterns}}{Set of associations between patterns and nodes determining whether a pattern
is included in a node.}

\item{\code{patterns}}{Set of mined patterns and characteristics of these patterns.
Data frame of 8 variables:
\describe{
  \item{\code{pattern}}{Set of items composing the pattern.}
  \item{\code{year}}{Year of appearance of the pattern among the transactions.}
  \item{\code{length}}{Number of items composing the pattern.}
  \item{\code{support}}{Proportion of transactions containing the set of items of the pattern.}
  \item{\code{frequency}}{Number of transactions containing the set of items of the pattern.}
  \item{\code{weight}}{Number of nodes containing the set of items of the pattern.}
  \item{\code{specificity}}{Specificity of the information conveyed by the pattern. It corresponds
    to the nature of a pattern of being specific of a particular combination or ubiquitous and
    allowing the formation of numerous combinations (with regard to the transactions).}
  \item{\code{status}}{Dynamic status of the pattern: persistent, declining, emergent or latent.}
}}

\item{\code{patterns_per_year}}{Number of occurrences of each pattern in the transactions, per year.}

\item{\code{p_links}}{Set of weights of the links between the patterns. Adjacency matrix containing the number
of items in common between each pair of patterns.}

\item{\code{pattern_links}}{Set of links between the patterns and characteristics of these links.
Isolated patterns (i.e. unrelated to any other pattern) appear at the bottom of the data structure.
Data frame of 5 variables:
\describe{
  \item{\code{endpoint.1}, \code{endpoint.2}}{Identifiers of two patterns from the attribute
                                              \code{patterns}.}
  \item{\code{items}}{Items in common between the two patterns.}
  \item{\code{weight}}{Number of items in common between the two patterns.}
  \item{\code{year}}{Year of appearance of the link between the two patterns.}
}}
}}

\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
            The spectrosome of occupational health problems.
            \emph{PLoS ONE} 13(1): e0190196.
            \url{https://doi.org/10.1371/journal.pone.0190196}.
            
            Bosson-Rieutort D, Sarazin P, Bicout DJ, Ho V, Lavoué J (2020).
            Occupational Co-exposures to Multiple Chemical Agents from Workplace Measurements by the US Occupational Safety and Health Administration.
            \emph{Annals of Work Exposures and Health}, Volume 64, Issue 4, May 2020, Pages 402–415.
            \url{https://doi.org/10.1093/annweh/wxaa008}.
}
\seealso{
The \code{TransactionAnalyzer} constructor: \code{\link{transaction.analyzer}}.

An example object of class \code{TransactionAnalyzer}: \code{\link{TA_instance}}.
}
\author{
Gauthier Magnin
}
