% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\docType{data}
\name{TRANSACTIONS}
\alias{TRANSACTIONS}
\title{TRANSACTIONS}
\format{
An object of class \code{character} of length 1.
}
\usage{
TRANSACTIONS
}
\description{
Reference value for naming entities: transactions.
}
\keyword{internal}
