% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_set.R
\docType{class}
\name{TransactionSet-class}
\alias{TransactionSet-class}
\alias{TransactionSet}
\title{Transaction Set}
\description{
S4 object class for representing transactions.
A transaction is defined as an itemset associated with any other data.
}
\details{
Since \code{year_key} is not required, transactions do not need to contain temporal data. In such a case,
\code{year_key} is \code{NA}.
}
\section{Slots}{

\describe{
\item{\code{data}}{List of transactions containing items and any additional data.
Each transaction is itself a list. All transactions share the same element names (corresponding to
the names of the various data).}

\item{\code{names}}{Names of the elements present in each transaction (i.e. names of the elements contained in
the list corresponding to a transaction).}

\item{\code{item_key}}{Name of the element containing the items of a transaction.}

\item{\code{year_key}}{Name of the element containing the year in which a transaction was made.}
}}

\section{Coercion}{

\describe{
\item{\code{as(trx, "data.frame")}}{Convert a \code{TransactionSet} object \code{trx} to \code{data.frame} (considering
the parameter \code{stringAsFactors} as \code{FALSE}).}
\item{\code{as(trx, "transactions")}}{Convert a \code{TransactionSet} object \code{trx} to a set of
\code{\link[arules:transactions-class]{transactions}} from the \code{\link{arules}} package. Only the items of
the transactions are considered. Other data are ignored.}
}
}

\seealso{
The \code{TransactionSet} constructor: \code{\link{transaction.set}}.

An example object of class \code{TransactionSet}: \code{\link{TS_instance}}.
}
\author{
Gauthier Magnin
}
