% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations_maker.R
\name{turn_interventions_into_observations}
\alias{turn_interventions_into_observations}
\title{Regroupe des prélèvements sous forme d'observations.}
\usage{
turn_interventions_into_observations(measures, additional = NULL)
}
\arguments{
\item{measures}{Ensemble de mesures de prélèvements.}

\item{additional}{Noms des informations à conserver lors de la constitution des observations.}
}
\value{
Liste des observations identifiées.
}
\description{
Les substances retrouvées dans une même intervention sont regroupées en une même observation.
}
\author{
Gauthier Magnin
}
