% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{turn_into_path}
\alias{turn_into_path}
\title{Chemin de dossier}
\usage{
turn_into_path(path)
}
\arguments{
\item{path}{Chaîne de caractères devant être utilisée comme chemin d'un dossier.}
}
\value{
Chaîne de caractères, terminant par le caractère \code{"/"}.
}
\description{
Vérifie si une chaîne de caractères peut être utilisée comme chemin d'un dossier.
Ajoute un caractère \code{"/"} si ce n'est pas le cas.
}
\keyword{internal}
