% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{turn_list_into_char}
\alias{turn_list_into_char}
\title{Turn a list into a character vector}
\usage{
turn_list_into_char(x)
}
\arguments{
\item{x}{List to turn into a character vector.}
}
\value{
Character vector corresponding to the input list.
}
\description{
Each element of the list is turned into a character value so the list is turned into a character vector.
}
\keyword{internal}
