% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_manager.R
\name{turn_list_into_data_frame}
\alias{turn_list_into_data_frame}
\title{Turn a list into a data frame}
\usage{
turn_list_into_data_frame(
  x,
  by_name = FALSE,
  stringsAsFactors = getOption("stringsAsFactors")
)
}
\arguments{
\item{x}{List of vectors to turn into a data frame.}

\item{by_name}{If \code{FALSE}, use the values of the list. If \code{TRUE}, use the names of these values.}

\item{stringsAsFactors}{\code{TRUE} or \code{FALSE} whether to convert character vectors to factors.}
}
\value{
Data frame in which the first column contains the names of the list and the second one
contains the related values or the names of these values (according to \code{by_name}).
}
\description{
Turn a list into a data frame associating names of the list with its values or with the names of
these values.
}
\note{
This function is not exported from the package.
}
\examples{
l <- list(e1 = c(v1 = "E", v2 = "F", v3 = "I"),
          e2 = c(v1 = "I"),
          e3 = c(v1 = "C", v2 = "I"),
          e4 = c(v1 = "A", v2 = "C", v3 = "D", v4 = "F"),
          e5 = c(v1 = "B", v2 = "D", v3 = "E", v4 = "I", v5 = "G"))

turn_list_into_data_frame(l)
turn_list_into_data_frame(l, by_name = TRUE)

}
\seealso{
\code{\link{coerce_list}}, \code{\link{turn_list_into_logical_matrix}}, \code{\link{turn_list_into_char_matrix}},
\code{\link{turn_data_frame_into_list}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
