% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{turn_list_into_logical_matrix}
\alias{turn_list_into_logical_matrix}
\title{Turn a list into a logical matrix}
\usage{
turn_list_into_logical_matrix(x, by_name = FALSE)
}
\arguments{
\item{x}{List to turn into a logical matrix.}

\item{by_name}{If \code{FALSE}, use the values of the elements of the list. If \code{TRUE}, use the names of
these values.}
}
\value{
Logical matrix in which row names are the names of the list and column names are the values
found in the elements of the list or the names of these values (according to \code{by_name}).
}
\description{
Turn a list into a logical matrix using the values of the list or the names of these values.
}
\examples{
l <- list(e1 = c(v1 = "E", v2 = "F", v3 = "I"),
          e2 = c(v1 = "I"),
          e3 = c(v1 = "C", v2 = "I"),
          e4 = c(v1 = "A", v2 = "C", v3 = "D", v4 = "F"),
          e5 = c(v1 = "B", v2 = "D", v3 = "E", v4 = "I", v5 = "G"))

turn_list_into_logical_matrix(l, by_name = FALSE)
turn_list_into_logical_matrix(l, by_name = TRUE)

}
\seealso{
\code{\link{turn_logical_matrix_into_list}}, \code{\link{invert_list}}.
}
\author{
Gauthier Magnin
}
