% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_manager.R
\name{turn_logical_matrix_into_list}
\alias{turn_logical_matrix_into_list}
\title{Turn a logical matrix into a list}
\usage{
turn_logical_matrix_into_list(x, by_row = TRUE)
}
\arguments{
\item{x}{Logical matrix to turn into a list.}

\item{by_row}{If \code{TRUE}, use the names of the rows as names of the list. If \code{FALSE}, use the names of
the columns as names of the list.}
}
\value{
List whose values correspond to the values \code{TRUE} of each row or of each column (according to
\code{by_row}).
}
\description{
Turn a logical matrix into a named list using the names of the columns and rows of the matrix.
}
\note{
This function is not exported from the package.
}
\examples{
m <- matrix(c(FALSE, FALSE, FALSE, FALSE, TRUE,  TRUE,  FALSE, TRUE,
              FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, TRUE,
              FALSE, FALSE, TRUE,  FALSE, FALSE, FALSE, FALSE, TRUE,
              TRUE,  FALSE, TRUE,  TRUE,  FALSE, TRUE,  FALSE, FALSE,
              FALSE, TRUE,  FALSE, TRUE,  TRUE,  FALSE, TRUE,  TRUE),
            nrow = 5, byrow = TRUE,
            dimnames = list(c("e1", "e2", "e3", "e4", "e5"),
                            c("A", "B", "C", "D", "E", "F", "G", "I")))

turn_logical_matrix_into_list(m)
turn_logical_matrix_into_list(m, by_row = FALSE)

}
\seealso{
\code{\link{coerce_to_list}}, \code{\link{turn_char_matrix_into_list}}, \code{\link{turn_data_frame_into_list}},
\code{\link{turn_list_into_logical_matrix}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
