% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_maker.R
\name{turn_obs_into_transactions}
\alias{turn_obs_into_transactions}
\title{Turn observations into arules transactions}
\usage{
turn_obs_into_transactions(observations, key = "CODE")
}
\arguments{
\item{observations}{List of observations to turn into \code{arules} transactions.}

\item{key}{Access key to the items in an observation.}
}
\value{
Set of transactions: \code{\link[arules:transactions-class]{transactions}} class object from
 the package \code{arules}.
}
\description{
Turn a set of observations into a set of \code{\link[arules:transactions-class]{transactions}} from
 the package \code{arules}.
Only the items of the observations are considered. Other data are ignored.
}
\author{
Gauthier Magnin
}
\keyword{internal}
