% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixer.R
\name{validate_classes}
\alias{validate_classes}
\title{Validate classes for the MCR approach}
\usage{
validate_classes(classes)
}
\arguments{
\item{classes}{List or logical matrix associating value names with classes.
If list, its names are the value names and the elements are vectors of associated classes.
If logical matrix, its columns are named according to the classes and the row names contain the
value names. A \code{TRUE} value indicates that a specific name is part of a specific class.}
}
\value{
Classes, as a logical matrix.
}
\description{
Check that the classes are given as a list or as a logical matrix. Stop the execution and print an
error message if not. If they are given as a list, they are converted into a logical matrix.
}
\note{
This function is not exported from the package.
}
\seealso{
\code{\link{check_data_for_mcr_by_class}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
