% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vector_notation}
\alias{vector_notation}
\title{Turn a set notation into a vector notation}
\usage{
vector_notation(sets)
}
\arguments{
\item{sets}{Itemsets written in mathematical notation. Must be of vector type or of factor type.}
}
\value{
List of character vectors corresponding to the itemsets \code{sets}.
}
\description{
Convert itemsets written in mathematical notation into a list of character vectors.
Example of conversion: \code{{item i,item j}} becomes \code{c("item i", "item j")}.
}
\examples{
vector_notation(as.factor(c("{A}", "{B,C}", "{A,B}", "{A}")))

}
\seealso{
\code{\link{set_notation}}.
}
\author{
Gauthier Magnin
}
