% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{weight_by_node_complexity,TransactionAnalyzer-method}
\alias{weight_by_node_complexity,TransactionAnalyzer-method}
\alias{weight_by_node_complexity}
\title{Pattern weight by node complexity}
\usage{
\S4method{weight_by_node_complexity}{TransactionAnalyzer}(object, patterns)
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}

\item{patterns}{Patterns whose weights according to the complexity of the nodes containing them
are to be computed. Any subset of \code{object["patterns"]$pattern}.\cr
\code{"patterns"} and \code{"p"} are special values for \code{object["patterns"]$pattern}.}
}
\value{
Two-column matrix containing, for each pattern, its weight related to complex nodes and
its weight related to simple nodes.
}
\description{
For each pattern, compute its weight related to complex nodes (i.e. nodes containing more than one
item and containing the pattern) and its weight related to simple node (i.e. nodes containing
only one item and containing the pattern).
}
\examples{
weight_by_node_complexity(TA_instance, "patterns")
weight_by_node_complexity(TA_instance, TA_instance["patterns"]$pattern[1:15])

}
\seealso{
\code{\link{get_complexes}}
}
\author{
Gauthier Magnin
}
