% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{which_associated_links,SpectralAnalyzer-method}
\alias{which_associated_links,SpectralAnalyzer-method}
\alias{which_associated_links}
\title{Type of links relating to a type of entities}
\usage{
\S4method{which_associated_links}{SpectralAnalyzer}(object, entities)
}
\arguments{
\item{object}{S4 object of class \code{SpectralAnalyzer}.}

\item{entities}{Type of entities for which to give the type of links.
Character corresponding to \code{NODES}, \code{PATTERNS} or their simplifications (see \code{\link{first_characters}}).}
}
\value{
Character corresponding to \code{NODE_LINKS} or \code{PATTERN_LINKS}.
}
\description{
Give a type of links given a type of entities (nodes or patterns).
}
\seealso{
\code{\link{which_entities}}, \code{\link{which_name}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
