% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{which_name,SpectralAnalyzer-method}
\alias{which_name,SpectralAnalyzer-method}
\alias{which_name}
\title{Type of entities or links corresponding to a name}
\usage{
\S4method{which_name}{SpectralAnalyzer}(object, name)
}
\arguments{
\item{object}{S4 object of class \code{SpectralAnalyzer}.}

\item{name}{Type of entities or links.
Character corresponding to \code{OBSERVATIONS}, \code{NODES}, \code{PATTERNS}, \code{RULES}, \code{NODE_LINKS}, \code{PATTERN_LINKS}
or their simplifications (see \code{\link{first_characters}}). One or more.}
}
\value{
Vector of characters corresponding to \code{OBSERVATIONS}, \code{NODES}, \code{PATTERNS}, \code{RULES},
\code{NODE_LINKS} or \code{PATTERN_LINKS}. Same size as the argument \code{name}.
}
\description{
Give the type of entities or links given a name referring to it.
}
\seealso{
\code{\link{which_entities}}, \code{\link{which_associated_links}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
