% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cap}
\alias{cap}
\title{Capitalize a character string}
\usage{
cap(x)
}
\arguments{
\item{x}{Character string to be capitalized.}
}
\value{
The character string corresponding to the argument \code{x}, starting with a capital letter.
}
\description{
Change the first letter of a string to an uppercase.
}
\seealso{
\code{\link{pluralize}}, \code{\link{first_characters}}, \code{\link{substr2}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
