% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{cap}
\alias{cap}
\title{Change la première lettre d'une chaîne de caractères en une majuscule.}
\usage{
cap(s)
}
\arguments{
\item{s}{Chaîne de caractères à modifier.}
}
\value{
La chaîne de caractères fournie par l'argument \code{s}, commençant par une majuscule.
}
\description{
Change la première lettre d'une chaîne de caractères en une majuscule.
}
