% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{check_access_for_category,SpectralAnalyzer-method}
\alias{check_access_for_category,SpectralAnalyzer-method}
\alias{check_access_for_category}
\title{Validation of parameters for search by category}
\usage{
\S4method{check_access_for_category}{SpectralAnalyzer}(object, category, value, stop = TRUE)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{category}{Name or number of the category to access (numbering according to the order of the
columns of \code{object["items_categories"]}).}

\item{value}{Sought value for the category specified by the argument \code{category}, or NA.}

\item{stop}{If \code{TRUE}, stop the execution and print an error message if the parameters do not
allow access to a category. If \code{FALSE}, see 'Value' section.}
}
\value{
\code{TRUE} or \code{FALSE} whether the parameters allow access to a category.
}
\description{
Check that the parameters provided match an existing category.
Print an error message if not.
}
\details{
If \code{value = NA}, only the parameter \code{category} is checked.
}
\seealso{
\code{\link{get_patterns}}, \code{\link{get_nodes}}.
         \code{\link{get_patterns_from_category}}, \code{\link{get_nodes_from_category}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
