% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{check_params_for_RI,SpectralAnalyzer-method}
\alias{check_params_for_RI,SpectralAnalyzer-method}
\title{Vérifie la validité des valeurs des paramètres donnés en arguments pour le calcul d'indices de recrutement.}
\usage{
\S4method{check_params_for_RI}{SpectralAnalyzer}(object, t, period)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{t}{Année de fin de la période, c'est-à-dire, date à laquelle caractériser le motif.
\code{NULL} indique que la caractérisation doit se faire par rapport à la dernière année
couverte par les observations.}

\item{period}{Intervalle de temps sur lequel calculer l'indice de recrutement (nombre d'années).
Si \code{t = 2015} et \code{period = 3}, alors le calcul est fait sur la période [2013 - 2015].
\code{Inf} indique que la période considérée couvre un intervalle commençant de la date de la 
plus ancienne observation et termine à l'année t.}
}
\value{
Liste contenant la valeur finale de \code{t} et celle de \code{period}.
}
\description{
Adapte leurs valeurs si elles n'entrent pas dans l'intervalle adéquate et affiche un message d'information.
}
\author{
Gauthier Magnin
}
