% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{cluster_chart,SpectralAnalyzer-method}
\alias{cluster_chart,SpectralAnalyzer-method}
\alias{cluster_chart}
\title{Cluster: subgraph of a spectrosome}
\usage{
\S4method{cluster_chart}{SpectralAnalyzer}(
  object,
  entities,
  characteristics,
  item,
  identifiers = "original",
  use_name = TRUE,
  n.cutoff = NULL,
  vertex_size = "relative",
  vertex_col = "status",
  c.cutoff = NULL,
  display_mixt = TRUE,
  path = getwd(),
  name = paste0(substr(entities, 1, nchar(entities) - 1), "_cluster_of_", item, ".png"),
  title = paste(cap(substr(entities, 1, nchar(entities) - 1)), "cluster of", item),
  ...
)
}
\arguments{
\item{object}{SpectralAnalyzer class object.}

\item{entities}{Type of entities for which to plot one of the clusters (nodes or patterns).
One of \code{"nodes"}, \code{"patterns"}.}

\item{characteristics}{Characteristics of nodes or patterns of which one of the clusters is to be
plotted.}

\item{item}{Identification code of the item whose cluster is to be plotted.}

\item{identifiers}{Which IDs to use to identify the nodes or patterns on the chart and in the
return data frames? One of \code{"original"}, \code{"new"}. \cr
\code{"original"} to use the original identifiers.
\code{"new"} to use new identifiers ordered according to \code{characteristics}.}

\item{use_name}{If \code{TRUE}, display the item name if it is defined. Display its identification
code otherwise.}

\item{n.cutoff}{If \code{use_names = TRUE}, limit number of characters to display concerning the name
of the item.}

\item{vertex_size}{Way how the sizes of the vertices of the graph should be defined.
One of \code{"relative"}, \code{"grouped"}, \code{"absolute"}, \code{"equal"}.
\describe{
  \item{\code{"relative"}}{The size of a vertex depends on the range of values of the weights.}
  \item{\code{"grouped"}}{The weights of the entities are grouped according to intervals.
                          One size is defined for each interval.}
  \item{\code{"absolute"}}{The size of a vertex is defined directly according to the weight of
                           the entity.}
  \item{\code{"equal"}}{The vertices are all the same size.}
}}

\item{vertex_col}{Way how the colors of the vertices of the graph should be defined.
One of \code{"status"}, \code{"categories"}, \code{"none"}.
If \code{"status"} and \code{entities = "patterns"}, coloring according to the status of the patterns.
If \code{"categories"}, coloring according to the categories associated with the items of the
entities represented.
In all other cases, the vertices are colored gray.}

\item{c.cutoff}{Limit number of characters to display in the legend for the categories represented.}

\item{display_mixt}{If \code{TRUE}, display in the legend the category values included only in mixed
links (or in mixed vertices, if \code{vertex_col = "categories"}).}

\item{path}{Path of the directory in which to save the charts.
By default, the charts are saved in the working directory.}

\item{name}{Name of the file in which to save the chart.
By default, the name depends on the arguments \code{entities} and \code{item}.
Example of default name: \code{"node_cluster_of_25.png"} if \code{entities = "nodes"} and
\code{item = 25}.}

\item{title}{Chart title.
By default, the title depends on the arguments \code{entities} and \code{item}.
Example of default title: \code{"Node cluster of 25"} if \code{entities = "nodes"} and
\code{item = 25}.}

\item{...}{Additional arguments to the function \code{\link[sna:gplot]{sna::gplot}} for plotting
the graph. See Details section.}
}
\value{
\code{NULL} if none or only one node or pattern contains the sought item. \cr
 Otherwise:
 \describe{
   \item{\code{vertices}}{Data frame of the nodes or patterns and characteristics used,
                          associated with the identifiers of the vertices of the graph and their
                          degrees in the graph.}
   \item{\code{edges}}{Data frame of information relating to the edges of the graph.}
   \item{\code{coords}}{Matrix of the coordinates of the vertices of the graph.
                        Can be reused via the parameter \code{coord} (see \code{...}).}
 }
}
\description{
Identify the cluster associated with one specific item and plot a spectrosome of this cluster.
}
\details{
If categories are associated with the items, each category generates a spectrosome.
 The category name is appended to the end of the file name.

If mixed links are relative to category values that are not represented by single links, these
 values are present in the legend below "Mixt", with no associated color.
The same is true for mixed vertices if the parameter \code{vertex_col} is \code{"categories"}.

If the chart is to be plotted from a subset of all nodes or patterns and some become isolated
 because the other entities to which they are normally linked are not part of \code{characteristics},
 these nodes or patterns are placed at the end of the return data frame \code{edges}.
These possible \code{n} additional lines are numbered \code{"A1"..."An"}.

Additional arguments can be supplied to the function in charge of plotting the graph.
 See the list of parameters: \code{\link[sna:gplot]{sna::gplot}}.
Among them, the following parameters are already defined and cannot be modified: \code{dat},
 \code{gmode}, \code{vertex.sides}, \code{vertex.cex}, \code{vertex.col}, \code{edge.col}.
The following parameters, which can be redefined, have the following values:
 \itemize{
   \item{\code{mode = "fruchtermanreingold"}}
   \item{\code{layout.par = list(repulse.rad = 4 ^ (log(nrow(nop_links), 10)))},
     where \code{nrow(nop_links)} is the sum of the number of links and the number of isolated
     entities.
     Can also be an object of type \code{expression} or \code{NULL}.}
   \item{\code{displaylabels = TRUE}}
   \item{\code{label.pos = 0}}
   \item{\code{boxed.labels = TRUE}}
   \item{\code{displayisolates = TRUE}}
 }
}
\seealso{
\code{\link{spectrosome_chart}}, \code{\link{degree}}, \code{\link[sna:gplot]{sna::gplot}}.
}
\author{
Gauthier Magnin
}
