% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{cluster_chart,SpectralAnalyzer-method}
\alias{cluster_chart,SpectralAnalyzer-method}
\alias{cluster_chart}
\title{Cluster : sous-graphe du spectrosome}
\usage{
\S4method{cluster_chart}{SpectralAnalyzer}(
  object,
  entities,
  item,
  vertex_size = "relative",
  path = getwd(),
  name = paste0(substr(entities, 1, nchar(entities) - 1), "_cluster_of_", item, ".png"),
  title = paste(cap(substr(entities, 1, nchar(entities) - 1)), "cluster of", item)
)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{entities}{Type d'élément pour rechercher un cluster (nœuds ou motifs).
Choix parmi \code{"nodes"}, \code{"patterns"}.}

\item{item}{Code de l'item dont le cluster est à mettre en évidence.}

\item{vertex_size}{Façon dont les tailles des sommets du graphe doivent être définies.
Choix parmi \code{"relative"}, \code{"grouped"}, \code{"absolute"}, \code{"equal"}.
\describe{
  \item{\code{"relative"}}{La taille d'un sommet dépend de l'intervalle de valeurs des différents poids.}
  \item{\code{"grouped"}}{Les poids des motifs sont regroupés selon des intervalles. À chaque intervalle correspond une taille.}
  \item{\code{"absolute"}}{La taille d'un sommet est définie directement en fonction du poids du motif.}
  \item{\code{"equal"}}{Les sommets ont tous la même taille.}
}}

\item{path}{Chemin du dossier dans lequel enregistrer les graphiques.
Par défaut, les graphiques sont enregistrés dans le répertoire de travail.}

\item{name}{Nom (avec extension) du fichier dans lequel enregistrer le graphique.
Par défaut, le nom dépend des arguments \code{entities} et \code{item}.
Exemple de nom par défaut : \code{"node_cluster_of_25"}.}

\item{title}{Titre du graphique.
Par défaut, le titre dépend des arguments \code{entities} et \code{item}.
Exemple de titre par défaut : \code{"Node cluster of 25"}.}
}
\value{
\code{NULL} si aucun ou un seul nœud ou motif contient l'item recherché. \cr
 Sinon, liste contenant :
 \describe{
   \item{\code{vertices}}{Data frame des nœuds ou motifs et caractéristiques utilisées,
                          associés aux identifiants des sommets du graphe et à leurs degrés dans le graphe.}
   \item{\code{edges}}{Data frame des informations relatives aux arêtes du graphe.}
 }
}
\description{
Identifie le cluster associé à l'item fourni en argument et en dessine un spectrosome.
}
\seealso{
\code{\link{spectrosome_chart}}, \code{\link{degree}}.
}
\author{
Gauthier Magnin
}
