% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{cluster_chart,SpectralAnalyzer-method}
\alias{cluster_chart,SpectralAnalyzer-method}
\alias{cluster_chart}
\title{Cluster: subgraph of a spectrosome}
\usage{
\S4method{cluster_chart}{SpectralAnalyzer}(
  object,
  nopc,
  item,
  identifiers = "original",
  use_name = TRUE,
  n.cutoff = NULL,
  vertex_size = "relative",
  size_range = c(0.5, 2.5),
  vertex_col = "status",
  c.cutoff = NULL,
  display_mixt = TRUE,
  title = paste("Cluster of", item),
  path = getwd(),
  name = paste0("cluster_of_", item, ".png"),
  ...
)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{nopc}{Data frame of \strong{n}odes \strong{o}r \strong{p}atterns and their
\strong{c}haracteristics. Nodes or patterns of which one of the clusters is to be plotted. Any subset
of \code{object["nodes"]} or \code{object["patterns"]}.\cr
\code{"nodes"}, \code{"n"}, \code{"patterns"} and \code{"p"} are specific values for
\code{object["nodes"]} and \code{object["patterns"]}.}

\item{item}{Identification code of the item whose cluster is to be plotted.}

\item{identifiers}{Which IDs to use to identify the nodes or patterns on the chart and in the
return data frames. One of \code{"original"}, \code{"new"}.
\describe{
  \item{\code{"original"}}{Use of the original identifiers.}
  \item{\code{"new"}}{Use of new identifiers ordered according to the subset corresponding to
                      \code{nopc}.}
}}

\item{use_name}{If \code{TRUE}, display the item name if it is defined. Display its identification
code otherwise.}

\item{n.cutoff}{If \code{use_names = TRUE}, limit number of characters to display concerning the name
of the item.}

\item{vertex_size}{Way how the sizes of the vertices of the graph should be defined.
One of \code{"relative"}, \code{"grouped"}, \code{"absolute"}, \code{"equal"}.
\describe{
  \item{\code{"relative"}}{The sizes are defined by a linear interpolation of the weights of the
                           entities in \code{size_range}.}
  \item{\code{"grouped"}}{The weights of the entities are grouped according to 5 intervals
                          defined by quantiles. The 5 corresponding size values are taken in a
                          regular sequence bounded by \code{size_range}.}
  \item{\code{"absolute"}}{The size of a vertex is defined directly according to the weight of
                           the entity.}
  \item{\code{"equal"}}{The vertices are all the same size of 1.}
}}

\item{size_range}{If \code{vertex_size} is \code{"relative"} or \code{"grouped"}.
Range of vertex size values (given as expansion factors).}

\item{vertex_col}{Way how the colors of the vertices of the graph should be defined.
One of \code{"status"}, \code{"categories"}, \code{"none"}.
If \code{"status"} and \code{nopc} refers to patterns, coloring according to the status of the
patterns. If \code{"categories"}, coloring according to the categories associated with the items of
the entities represented. In all other cases, the vertices are colored gray.}

\item{c.cutoff}{Limit number of characters to display in the legend for the categories represented.}

\item{display_mixt}{If \code{TRUE}, display in the legend the category values included only in mixed
links (or in mixed vertices, if \code{vertex_col = "categories"}).}

\item{title}{Chart title.
By default, the title depends on the argument \code{item}.
Example of default title: \code{"Cluster of 25"} if \code{item = 25}.}

\item{path}{Path of the directory in which to save the charts.
By default, the charts are saved in the working directory.}

\item{name}{Name of the file in which to save the chart.
By default, the name depends on the argument \code{item}.
Example of default name: \code{"cluster_of_25.png"} if \code{item = 25}.}

\item{...}{Additional arguments to the function \code{\link[sna:gplot]{gplot}} from the package
\code{sna} for plotting the graph. See Details section.}
}
\value{
\code{NULL} if none or only one node or pattern contains the sought item. \cr
 Otherwise:
 \describe{
   \item{\code{vertices}}{Data frame of the nodes or patterns and characteristics used,
                          associated with the identifiers of the vertices of the graph and their
                          degrees in the graph.}
   \item{\code{edges}}{Data frame of information relating to the edges of the graph.}
   \item{\code{coords}}{Matrix of the coordinates of the vertices of the graph.
                        Can be reused via the parameter \code{coord} (see \code{...}).}
 }
}
\description{
Identify the cluster associated with one specific item and plot a spectrosome of this cluster.
}
\details{
If categories are associated with the items, each category generates a spectrosome.
 The category name is appended to the end of the file name.

If mixed links are relative to category values that are not represented by single links, these
 values are present in the legend below "Mixt", with no associated color.
The same is true for mixed vertices if the parameter \code{vertex_col} is \code{"categories"}.

If the chart is to be plotted from a subset of all nodes or patterns and some become isolated
 because the other entities to which they are normally linked are not part of the subset \code{nopc},
 these nodes or patterns are placed at the end of the return data frame \code{edges}.
These possible \code{n} additional lines are numbered \code{"A1"..."An"}.

Additional arguments can be supplied to the function in charge of plotting the graph.
 See the list of parameters: \code{\link[sna:gplot]{sna::gplot}}.
Among them, the following parameters are already defined and cannot be modified: \code{dat},
 \code{gmode}, \code{vertex.sides}, \code{vertex.cex}, \code{vertex.col}, \code{edge.col}.
The following parameters, which can be redefined, have the following values:
 \itemize{
   \item{\code{mode = "fruchtermanreingold"}}
   \item{\code{displaylabels = TRUE}}
   \item{\code{label.pos = 0}}
   \item{\code{boxed.labels = TRUE}}
   \item{\code{displayisolates = TRUE}}
 }
}
\examples{
cluster_1 <- cluster_chart(SA_instance, "nodes", item = 3146,
                           name = "node_cluster_of_3146")
cluster_2 <- cluster_chart(SA_instance, SA_instance["patterns"], item = 3146)

}
\seealso{
\code{\link{spectrosome_chart}}, \code{\link{degree}}, \code{\link[sna:gplot]{sna::gplot}}.
}
\author{
Gauthier Magnin
}
