% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{cluster_chart,SpectralAnalyzer-method}
\alias{cluster_chart,SpectralAnalyzer-method}
\alias{cluster_chart}
\title{Cluster : sous-graphe du spectrosome}
\usage{
\S4method{cluster_chart}{SpectralAnalyzer}(
  object,
  entities,
  characteristics,
  item,
  use_name = TRUE,
  n.cutoff = NULL,
  vertex_size = "relative",
  vertex_col = "status",
  c.cutoff = NULL,
  display_mixt = TRUE,
  path = getwd(),
  name = paste0(substr(entities, 1, nchar(entities) - 1), "_cluster_of_", item, ".png"),
  title = paste(cap(substr(entities, 1, nchar(entities) - 1)), "cluster of", item),
  ...
)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{entities}{Type d'élément pour rechercher un cluster (nœuds ou motifs).
Choix parmi \code{"nodes"}, \code{"patterns"}.}

\item{characteristics}{Ensemble des caractéristiques des nœuds ou motifs dont l'un des clusters est à tracer.}

\item{item}{Code de l'item dont le cluster est à mettre en évidence.}

\item{use_name}{Si \code{TRUE}, affiche le nom de l'item s'il est défini. Affiche son code sinon.}

\item{n.cutoff}{Si \code{use_names = TRUE}, nombre limite de caractères à afficher concernant le nom de l'item.}

\item{vertex_size}{Façon dont les tailles des sommets du graphe doivent être définies.
Choix parmi \code{"relative"}, \code{"grouped"}, \code{"absolute"}, \code{"equal"}.
\describe{
  \item{\code{"relative"}}{La taille d'un sommet dépend de l'intervalle de valeurs des différents poids.}
  \item{\code{"grouped"}}{Les poids des motifs sont regroupés selon des intervalles. À chaque intervalle correspond une taille.}
  \item{\code{"absolute"}}{La taille d'un sommet est définie directement en fonction du poids du motif.}
  \item{\code{"equal"}}{Les sommets ont tous la même taille.}
}}

\item{vertex_col}{Façon dont les couleurs des sommets du graphe doivent être définies.
Choix parmi \code{"status"}, \code{"categories"}, \code{"none"}.
Si \code{"status"} et \code{entities = "patterns"}, coloration selon les statuts des motifs.
Si \code{"categories"}, coloration selon les catégories associées aux items des entités représentées.
Dans tous les autres cas, le sommets sont de couleur grise.}

\item{c.cutoff}{Nombre limite de caractères à afficher dans la légende concernant les catégories représentées.}

\item{display_mixt}{Si \code{TRUE}, affiche dans la légende les valeurs de catégorie incluses
uniquement dans des liens mixtes (ou dans des sommets mixtes, si \code{vertex_col = "categories"}).}

\item{path}{Chemin du dossier dans lequel enregistrer les graphiques.
Par défaut, les graphiques sont enregistrés dans le répertoire de travail.}

\item{name}{Nom du fichier dans lequel enregistrer le graphique.
Par défaut, le nom dépend des arguments \code{entities} et \code{item}.
Exemple de nom par défaut : \code{"node_cluster_of_25"}.}

\item{title}{Titre du graphique.
Par défaut, le titre dépend des arguments \code{entities} et \code{item}.
Exemple de titre par défaut : \code{"Node cluster of 25"}.}

\item{...}{Arguments supplémentaires fournis à la fonction \code{\link[sna:gplot]{sna::gplot}}
pour le traçage du graphe. Cf. section Details.}
}
\value{
\code{NULL} si aucun ou un seul nœud ou motif contient l'item recherché. \cr
 Sinon, liste contenant :
 \describe{
   \item{\code{vertices}}{Data frame des nœuds ou motifs et caractéristiques utilisées,
                          associés aux identifiants des sommets du graphe et à leurs degrés dans le graphe.}
   \item{\code{edges}}{Data frame des informations relatives aux arêtes du graphe.}
   \item{\code{coords}}{Matrice des coordonnées des sommets du graphe.}
 }
}
\description{
Identifie le cluster associé à l'item fourni en argument et en dessine un spectrosome.
}
\details{
Si des catégories sont associées aux items, chaque catégorie génère un spectrosome.
 Le nom de la catégorie est ajouté à la fin du nom du fichier.

Si des liens mixtes sont relatifs à des valeurs de catégorie qui ne sont pas représentées par des
 liens simples, ces valeurs apparaîssent dans la légende en dessous de "Mixt", sans couleur associée.
Il en est de même concernant les sommets mixtes si l'argument \code{vertex_col} vaut
 \code{"categories"}.

Des arguments supplémentaires peuvent être fournis à la fonction en charge du traçage du graphe.
 Voir la liste des paramètres : \code{\link[sna:gplot]{sna::gplot}}.
Parmi eux, les paramètres suivants sont déjà définis et ne peuvent pas être modifiés : \code{dat},
 \code{gmode}, \code{vertex.sides}, \code{vertex.cex}, \code{vertex.col}, \code{edge.col}.
Les paramètres suivants, pouvant être redéfinis, ont pour valeurs :
 \itemize{
   \item{\code{mode = "fruchtermanreingold"}}
   \item{\code{layout.par = list(repulse.rad = 4 ^ (log(nrow(nop_links), 10)))},
     où \code{nrow(nop_links)} correspond à la somme du nombre de liens et du nombre d'éléments isolés.
     Peut aussi être un objet de type \code{expression} ou \code{NULL}.}
   \item{\code{displaylabels = TRUE}}
   \item{\code{label.pos = 0}}
   \item{\code{boxed.labels = TRUE}}
   \item{\code{displayisolates = TRUE}}
 }
}
\seealso{
\code{\link{spectrosome_chart}}, \code{\link{degree}}, \code{\link[sna:gplot]{sna::gplot}}.
}
\author{
Gauthier Magnin
}
