% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{cluster_text,SpectralAnalyzer-method}
\alias{cluster_text,SpectralAnalyzer-method}
\alias{cluster_text}
\title{Affichage des noms des clusters}
\usage{
\S4method{cluster_text}{SpectralAnalyzer}(object, graph, links)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{graph}{Graphe généré par la fonction \code{\link[sna:gplot]{sna::gplot}} :
"A two-column matrix containing the vertex positions as x,y coordinates.".}

\item{links}{Liens des motifs utilisés pour générer \code{graph}.}
}
\description{
Identifie et affiche les noms des clusters sur le graphe fourni en argument.
}
\seealso{
\code{\link{spectrosome_chart}}, \code{\link[sna:gplot]{sna::gplot}}.
}
\author{
Delphine Bosson-Rieutort, Gauthier Magnin
}
\keyword{internal}
