% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_set.R
\name{co_occurrence_chart,TransactionSet-method}
\alias{co_occurrence_chart,TransactionSet-method}
\alias{co_occurrence_chart}
\alias{co_occurrence_chart,TransactionSet}
\title{Co-occurrence chart, for TransactionSet}
\usage{
\S4method{co_occurrence_chart}{TransactionSet}(
  object,
  items = NULL,
  co_occ = NULL,
  proportions = FALSE,
  min_occ = if (proportions) .Machine$double.xmin else 1,
  max_occ = Inf,
  vertex_size = 3,
  vertex_alpha = 1,
  vertex_margin = 0.05,
  label_size = 3,
  label_margin = 0.05,
  edge_looseness = 0.8,
  edge_alpha = 1,
  palette = "Blues",
  palette_direction = 1
)
}
\arguments{
\item{object}{S4 object of class \code{TransactionSet}.}

\item{items}{Items for which to plot co-occurrences between pairs.
The default \code{NULL} means to consider each existing item.}

\item{co_occ}{Matrix containing the co-occurrences (or their proportions) for at least the items
specified by the argument \code{items}. Is computed if \code{NULL}.}

\item{proportions}{\code{TRUE} if the proportions of co-occurrences are to be plotted (and computed, if
\code{co_occ} is \code{NULL}) instead of the co-occurrences themselves.}

\item{min_occ}{Minimum number of co-occurrences (or minimum proportion) to consider to plot a link
between two items.
Default value depends on the argument \code{proportions} and allows not to plot links between items
that never co-occur.}

\item{max_occ}{Maximum number of co-occurrences (or maximum proportion) to consider to plot a link
between two items.}

\item{vertex_size}{Size of the vertices.}

\item{vertex_alpha}{Opacity of the vertices (from 0 to 1).}

\item{vertex_margin}{Margin before the vertices (i.e. distance between the ends of the edges and the
centers of the vertices).}

\item{label_size}{Size of the labels associated with the vertices.}

\item{label_margin}{Margin before the labels (i.e. distance between the centers of the vertices and
the labels).}

\item{edge_looseness}{Looseness of the connecting lines (from 0 to 1).
The closer the value is to 0, the straighter the lines will be.
The closer the value is to 1, the more the lines will be curved.}

\item{edge_alpha}{Opacity of the lines connecting vertices (from 0 to 1).}

\item{palette}{Name (or number) of the sequential palette to use for coloring the edges.
One of \code{"Blues"}, \code{"BuGn"}, \code{"BuPu"}, \code{"GnBu"}, \code{"Greens"}, \code{"Greys"}, \code{"Oranges"}, \code{"OrRd"},
\code{"PuBu"}, \code{"PuBuGn"}, \code{"PuRd"}, \code{"Purples"}, \code{"RdPu"}, \code{"Reds"}, \code{"YlGn"}, \code{"YlGnBu"}, \code{"YlOrBr"},
\code{"YlOrRd"}.}

\item{palette_direction}{Direction in which to use the color palette.
If \code{1}, colors are in original order: from the lightest to the darkest.
If \code{-1}, color order is reversed: from the darkest to the lightest.}
}
\value{
Graph created with the packages \code{ggraph} and \code{ggplot2}.
}
\description{
Plot a graph in which vertices are items and edges are their co-occurences in transactions (i.e. for
each pair of items, the number of transactions containing it). Edges can also represent the
proportions of these co-occurrences (i.e., the ratio between the number of transactions containing a
pair of items and the number of transactions containing at least one of them).
}
\details{
The chart being plotted with the packages \code{ggraph} and \code{ggplot2}, it can be modified or completed
afterwards using \code{\link[ggplot2:last_plot]{ggplot2::last_plot}} or the returned object.

Items are ordered according to the order they are given. If the default value is given,
they are ordered alphanumerically.
}
\note{
If using the RStudio IDE and the value of the argument \code{edge_alpha} is not \eqn{1}, edges may not
be displayed in the RStudio "Plots" pane. However, they will be actually displayed in the "Plot Zoom"
window; while exporting the plot; or by using another graphics device.
}
\examples{
co_occurrence_chart(TS_instance)
co_occurrence_chart(TS_instance) +
  ggplot2::expand_limits(x = c(-1.2, 1.2), y = c(-1.2, 1.2))
co_occurrence_chart(TS_instance, min_occ = 2, palette = "OrRd")

co_occurrence_chart(TS_instance, items = c(497, 931, 3157, 3350))
co_occurrence_chart(TS_instance, items = c(497, 931, 3157, 3350),
                    proportions = TRUE)

}
\seealso{
\code{\link{co_occurrence_matrix}}.

Method for signature \code{TransactionAnalyzer}:
\code{\link[=co_occurrence_chart,TransactionAnalyzer-method]{co_occurrence_chart,TransactionAnalyzer}}.
}
\author{
Gauthier Magnin
}
