% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{compute_ksi_threshold,SpectralAnalyzer-method}
\alias{compute_ksi_threshold,SpectralAnalyzer-method}
\title{Calcule le nombre de motifs permettant d'expliquer l'essentiel des indices de recrutement.}
\usage{
\S4method{compute_ksi_threshold}{SpectralAnalyzer}(object, reporting_indexes)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{reporting_indexes}{Indices de recrutement associées aux motifs.}
}
\value{
Seuil calculé.
}
\description{
Calcule le nombre de motifs permettant d'expliquer l'essentiel des indices de recrutement.
}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
            \emph{The spectrosome of occupational health problems}. PLoS ONE 13(1): e0190196.
            \url{https://doi.org/10.1371/journal.pone.0190196}.
}
\author{
Gauthier Magnin
}
