% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{compute_pattern_distribution_in_nodes,SpectralAnalyzer-method}
\alias{compute_pattern_distribution_in_nodes,SpectralAnalyzer-method}
\alias{compute_pattern_distribution_in_nodes}
\title{Distribution des motifs parmi les nœuds}
\usage{
\S4method{compute_pattern_distribution_in_nodes}{SpectralAnalyzer}(object, patterns)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{patterns}{Liste des motifs dont les distributions sont à calculer.}
}
\value{
Liste contenant :
 \describe{
   \item{\code{weight_distribution}}{La distribution, pour chaque motif, des poids des nœuds dans lequel ils sont inclus.}
   \item{\code{length_distribution}}{La distribution, pour chaque motif, des longueurs des nœuds dans lequel ils sont inclus.}
 }
}
\description{
Calcule les distributions des poids et longueurs des nœuds dans lesquels sont inclus chaque motif.
}
\author{
Gauthier Magnin
}
\keyword{internal}
