% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{compute_pattern_distribution_in_nodes,SpectralAnalyzer-method}
\alias{compute_pattern_distribution_in_nodes,SpectralAnalyzer-method}
\title{Calcule les distributions des poids et longueurs des noeuds contenant un motif, pour chaque motif.}
\usage{
\S4method{compute_pattern_distribution_in_nodes}{SpectralAnalyzer}(object, patterns)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{patterns}{Liste des motifs dont les distributions sont à calculer.}
}
\value{
Liste contenant :
 \itemize{
   \item{} {La distribution, pour chaque motif, des poids des noeuds dans lequel il est inclus.}
   \item{} {La distribution, pour chaque motif, des longueurs des noeuds dans lequel il est inclus.}
 }
}
\description{
Calcule les distributions des poids et longueurs des noeuds contenant un motif, pour chaque motif.
}
\author{
Gauthier Magnin
}
