% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{compute_patterns_characteristics,SpectralAnalyzer-method}
\alias{compute_patterns_characteristics,SpectralAnalyzer-method}
\title{Calcule les caractéristiques des différents motifs (fréquence, poids, ordre, spécificté, caractère dynamique).}
\usage{
\S4method{compute_patterns_characteristics}{SpectralAnalyzer}(object)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}
}
\value{
Data.frame dans laquelle une ligne est une association entre un motif et ses caractéristiques.
}
\description{
La data.frame résultante est associée à l'attribut \code{patterns}.
}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
            \emph{The spectrosome of occupational health problems}. PLoS ONE 13(1): e0190196.
            \url{https://doi.org/10.1371/journal.pone.0190196}.
}
\author{
Gauthier Magnin
}
