% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{compute_patterns_characteristics,SpectralAnalyzer-method}
\alias{compute_patterns_characteristics,SpectralAnalyzer-method}
\alias{compute_patterns_characteristics}
\title{Computation of pattern characteristics}
\usage{
\S4method{compute_patterns_characteristics}{SpectralAnalyzer}(object)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}
}
\value{
Invisible. Data frame in which a line is an association between a pattern and its
 characteristics.
}
\description{
Compute the characteristics of the patterns (frequency, weight, order, specificity, dynamic status).
The resulting data frame is assigned to the attribute \code{patterns} of \code{object}.
}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
            The spectrosome of occupational health problems. \emph{PLoS ONE} 13(1): e0190196.
            \url{https://doi.org/10.1371/journal.pone.0190196}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
