% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{compute_patterns_characteristics,SpectralAnalyzer-method}
\alias{compute_patterns_characteristics,SpectralAnalyzer-method}
\alias{compute_patterns_characteristics}
\title{Computation of pattern characteristics}
\usage{
\S4method{compute_patterns_characteristics}{SpectralAnalyzer}(object)
}
\arguments{
\item{object}{S4 object of class \code{SpectralAnalyzer}.}
}
\value{
Invisible. Data frame in which a line is an association between a pattern and its
characteristics.
}
\description{
Compute the characteristics of the patterns (frequency, weight, length, specificity, dynamic status).
The resulting data frame is assigned to the attribute \code{patterns} of \code{object}.
}
\details{
The frequency and the weight of a pattern is the number of nodes and the number of transactions
containing it, respectively. The length of a pattern is the number of items composing it.

The specificity of a pattern is the information it conveys. It corresponds to the nature of the
pattern of being specific of a particular combination or ubiquitous and allowing the formation of
numerous combinations (with regard to the transactions).

The dynamic status is one of persistent, declining, emergent or latent.

\loadmathjax

\subsection{Specificity}{
}

The specificy of the pattern \eqn{p} is given by:
 \mjdeqn{S_p = \begin{cases} 1 & \mbox{if } U_p = 1 \cr 0 & \mbox{if } A_{p,1} = A_{p,2} = \ldots = A_{p,U_p} \cr \frac{H_{max,p} - H_p}{H_{max,p} - H_{min,p}} & \mbox{otherwise} \end{cases}}{S_p = 1 if U_p = 1, 0 if A_p1 = A_p2 = ... = A_{p,U_p}, (H_max,p - H_p) / (H_max,p - H_min,p) otherwise}
where \mjseqn{U} is the set of frequencies of the patterns, \mjseqn{A_p} is the
set of weights of the nodes containing the pattern \eqn{p} and:
\describe{
 \item{}{\mjeqn{H_p = - \sum_{i = 1}^{U_p} \frac{A_{p,i}}{W_p} \times ln \left( \frac{A_{p,i}}{W_p} \right)}{H_p = - sum(A_pi / W_p * ln(A_pi / W_p)) from i = 1 to U_p}}
 \item{}{\mjeqn{H_{max,p} = ln(U_p)}{H_max,p = ln(U_p)}}
 \item{}{\mjeqn{H_{min,p} = ln \left( \frac{W_p}{W_p - U_p + 1} \right) + \left( \frac{U_p - 1}{W_p} \times ln(W_p - U_p + 1) \right)}{H_min,p = ln(W_p / (W_p - U_p + 1)) + (U_p - 1) / W_p * ln(W_p - U_p + 1)}}
 \item{}{where \mjseqn{W} is the set of weights of the patterns.}
}

\subsection{Dynamic status}{
}

For each pattern, two reporting index values are computed and compared to two thresholds in order
to define its status.

A pattern is:
\itemize{
\item Persistent when it appears substantially during both the entire period of the
transactions (defined by the arguments \code{period} and \code{t}) and a shorter period defined by the
arguments \code{short_limit} and \code{t}.
\item Declining when it appears substantially during the entire period but much less during
the shorter period.
\item Emergent when it appears insignificantly during the entire period but much more during
the shorter period.
\item Latent when it appears insignificantly during both the entire period and the shorter period.
}

The reporting indexes of the pattern \eqn{p} which are computed are those at the two temporal limits
given by:
\mjdeqn{RI_{\infty,p} \; = lim_{t_0 \to -\infty} RI_p(t_1,t_0)}{RI_inf,p = lim of RI_p(t_1,t_0) as t approaches -inf}
\mjdeqn{RI_{l,p} = RI_p(t_1, t_1 - l + 1)}{RI_lp = RI_p(t_1, t_1 - l + 1)}
where \eqn{l} is the shorter period on which to compute a reporting index and \mjseqn{RI_p(t_1,t_0)}
is the reporting index of the pattern \eqn{p} given by:
\mjdeqn{RI_p(t_1,t_0) = \frac{\sum_{t = t_0}^{t_1} W_{p,t}}{\sum_{q \in P} \sum_{t = t_0}^{t_1} W_{q,t}}}{RI_p(t_1,t_0) = sum W_pt from t = t_0 to t_1 / sum W_{q,t} for q in P and from t = t_0 to t_1}
where \eqn{P} is the set of patterns, \mjeqn{W_{p,t}}{W_pt} is the weight of the pattern \eqn{p} in
the transactions of the year \eqn{t}, \mjseqn{t_0} and \mjseqn{t_1} are the first and last years
defining the period on which to compute the reporting index.

For each set of reporting indexes:
\itemize{
\item One threshold \mjseqn{\xi} is computed as follows:
\mjdeqn{\xi = \left\lceil \frac{1}{\sum_{p \in P} RI_p(t_1,t_0)^2} \right\rceil}{xi = ceiling(1 / sum(RI_p(t_1,t_0)^2) for p in P)}
\item The patterns are ordered in descending order of their reporting index value and separated by
this threshold.
\item The reporting index of the \mjseqn{\xi}\out{<sup>th</sup>} pattern is the \eqn{RI} threshold used to
classify the patterns.
}

The patterns are then classified as follows:
\itemize{
\item Patterns for which \mjeqn{RI_{\infty,p}}{RI_inf,p} and \mjeqn{RI_{l,p}}{RI_lp} are greater than or
equal to the \eqn{RI} thresholds are classified as "Persistent".
\item Patterns for which \mjeqn{RI_{\infty,p}}{RI_inf,p} is greater than or equal to the related threshold
and \mjeqn{RI_{l,p}}{RI_lp} is lower than the related threshold are classified as "Declining".
\item Patterns for which \mjeqn{RI_{\infty,p}}{RI_inf,p} is lower than the related threshold and
\mjeqn{RI_{l,p}}{RI_lp} is greater than or equal to the related threshold are classified as "Emergent".
\item Patterns for which \mjeqn{RI_{\infty,p}}{RI_inf,p} and \mjeqn{RI_{l,p}}{RI_lp} are lower than the
thresholds are classified as "Latent".
}
}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
The spectrosome of occupational health problems. \emph{PLoS ONE} 13(1): e0190196.
\url{https://doi.org/10.1371/journal.pone.0190196}.
}
\seealso{
\code{\link{compute_specificity}}, \code{\link{define_dynamic_status}},
}
\author{
Gauthier Magnin
}
\keyword{internal}
