% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{compute_reporting_indexes,SpectralAnalyzer-method}
\alias{compute_reporting_indexes,SpectralAnalyzer-method}
\alias{compute_reporting_indexes}
\title{Calcul d'indice de recrutement}
\usage{
\S4method{compute_reporting_indexes}{SpectralAnalyzer}(object, patterns, t = NULL, period = Inf)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{patterns}{Liste des motifs dont l'indice de recrutement est à calculer.}

\item{t}{Année de fin de la période, c'est-à-dire, date à laquelle caractériser le motif.
\code{NULL} indique que la caractérisation doit se faire par rapport à la dernière année
couverte par les observations.}

\item{period}{Intervalle de temps sur lequel calculer l'indice de recrutement (nombre d'années).
Si \code{t = 2015} et \code{period = 2}, alors le calcul est fait sur la période [2014 - 2015].
\code{Inf} indique que la période considérée couvre un intervalle commençant à la date de la 
plus ancienne observation et terminant à l'année \code{t}.}
}
\value{
Data frame associant un indice de recrutement à chaque motif.
}
\description{
Calcule l'indice de recrutement de chaque motif pour une période donnée.
Cet indice informe sur la proportion et l'importance du recrutement d'un motif en tenant compte
 du recrutement des autres motifs.
}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
            \emph{The spectrosome of occupational health problems}. PLoS ONE 13(1): e0190196.
            \url{https://doi.org/10.1371/journal.pone.0190196}.
}
\seealso{
\code{\link{compute_reporting_indexes_limits}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
