% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{compute_reporting_indexes,SpectralAnalyzer-method}
\alias{compute_reporting_indexes,SpectralAnalyzer-method}
\alias{compute_reporting_indexes}
\title{Compute the reporting index (RI)}
\usage{
\S4method{compute_reporting_indexes}{SpectralAnalyzer}(object, patterns, t = NULL, period = Inf)
}
\arguments{
\item{object}{SpectralAnalyzer class object.}

\item{patterns}{Patterns whose reporting indexes are to be computed.}

\item{t}{Year of the end of the period, i.e. the date on which to characterize the patterns.
\code{NULL} specifies that the characterization must be done in relation to the last year covered
by the observations.}

\item{period}{Time interval over which to compute the reporting indexes (number of years).
For example, if \code{t = 2015} and \code{period = 2} then the computation is made over the
period [2014 - 2015].
\code{Inf} specifies that the period considered covers an interval starting on the date of the
oldest observation and ending in the year \code{t}.}
}
\value{
Data frame associating one reporting index with each pattern.
}
\description{
Compute the reporting index of each pattern for a given period.
This index provides information on the proportion and importance of the occurrences of a pattern,
 taking into account the occurrences of the other patterns.
}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
            The spectrosome of occupational health problems. \emph{PLoS ONE} 13(1): e0190196.
            \url{https://doi.org/10.1371/journal.pone.0190196}.
}
\seealso{
\code{\link{compute_reporting_indexes_limits}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
