% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{compute_reporting_indexes_limits,SpectralAnalyzer-method}
\alias{compute_reporting_indexes_limits,SpectralAnalyzer-method}
\title{Calcule les limites temporelles utilisées pour caractériser les motifs.}
\usage{
\S4method{compute_reporting_indexes_limits}{SpectralAnalyzer}(
  object,
  patterns,
  first_limit,
  t = NULL,
  period = Inf
)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{patterns}{Liste des motifs dont les limites sont à calculer.}

\item{first_limit}{Intervalle de temps sur lequel calculer la première limite (nombre d'années).}

\item{t}{Année de fin de la période, c'est-à-dire, date à laquelle caractériser le motif.
\code{NULL} indique que le calcul doit se faire par rapport à la dernière année
couverte par les observations.}

\item{period}{Intervalle de temps sur lequel faire le calcul (nombre d'années).
Si \code{t = 2015} et \code{period = 3}, alors le calcul est fait sur la période [2013 - 2015].
\code{Inf} indique que la période considérée couvre un intervalle commençant de la date de la 
plus ancienne observation et termine à l'année t.}
}
\value{
Data.frame associant à chaque motif ses deux limites.
}
\description{
La première correspond à l'indice de recrutement calculé sur \code{first_limit} années.
La seconde correspond à l'indice de recrutement calculé sur la période définie par les paramètres \code{t} et \code{period}.
}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
            \emph{The spectrosome of occupational health problems}. PLoS ONE 13(1): e0190196.
            \url{https://doi.org/10.1371/journal.pone.0190196}.
}
\author{
Gauthier Magnin
}
