% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{compute_ri_threshold,SpectralAnalyzer-method}
\alias{compute_ri_threshold,SpectralAnalyzer-method}
\alias{compute_ri_threshold}
\title{RI threshold computation}
\usage{
\S4method{compute_ri_threshold}{SpectralAnalyzer}(object, reporting_indexes, xi = NULL)
}
\arguments{
\item{object}{S4 object of class \code{SpectralAnalyzer}.}

\item{reporting_indexes}{Reporting indexes associated with the patterns.}

\item{xi}{Number of patterns to consider before setting the RI threshold.
Is computed if \code{NULL}.}
}
\value{
Computed threshold.
}
\description{
\loadmathjax
Compute the limit value separating two dynamic profiles with respect to the reporting indexes.
The patterns are ordered in descending order of their reporting index value and separated by
 the threshold \mjseqn{\xi}.
}
\details{
\loadmathjax
The threshold \mjseqn{\xi} is given by:
 \mjdeqn{\xi = \left\lceil \frac{1}{\sum_{p \in P} RI_p(t_1,t_0)^2} \right\rceil}{xi = ceiling(1 / sum(RI_p(t_1,t_0)^2) for p in P)}
where \mjseqn{RI_p(t_1,t_0)} is the reporting index of the pattern \eqn{p} given by:
 \mjdeqn{RI_p(t_1,t_0) = \frac{\sum_{t = t_0}^{t_1} W_{p,t}}{\sum_{q \in P} \sum_{t = t_0}^{t_1} W_{q,t}}}{RI_p(t_1,t_0) = sum W_pt from t = t_0 to t_1 / sum W_qt for q in P and from t = t_0 to t_1}
where \eqn{P} is the set of patterns, \mjeqn{W_{p,t}}{W_pt} is the weight of the pattern \eqn{p} in
 the observations of the year \eqn{t}, \mjseqn{t_0} and \mjseqn{t_1} are the first and last years
 defining the period on which to compute the reporting index.
}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
            The spectrosome of occupational health problems. \emph{PLoS ONE} 13(1): e0190196.
            \url{https://doi.org/10.1371/journal.pone.0190196}.
}
\seealso{
\code{\link{compute_reporting_indexes}}, \code{\link{compute_reporting_indexes_limits}},
         \code{\link{compute_xi_threshold}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
