% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{compute_ri_threshold,SpectralAnalyzer-method}
\alias{compute_ri_threshold,SpectralAnalyzer-method}
\alias{compute_ri_threshold}
\title{Calcul de seuil de RI}
\usage{
\S4method{compute_ri_threshold}{SpectralAnalyzer}(object, reporting_indexes, ksi = NULL)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{reporting_indexes}{Indices de recrutement associées aux motifs.}

\item{ksi}{Nombre de motifs à considérer avant de fixer le seuil.
Est calculé si valeur \code{NULL}.}
}
\value{
Seuil calculé.
}
\description{
Calcule la valeur limite séparant deux statuts dynamiques par rapport aux indices de recrutement.
Les motifs sont ordonnés par ordre décroissant de leur valeur d'indice de recrutement et séparé 
par le seuil \code{ksi}.
}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
            \emph{The spectrosome of occupational health problems}. PLoS ONE 13(1): e0190196.
            \url{https://doi.org/10.1371/journal.pone.0190196}.
}
\seealso{
\code{\link{compute_reporting_indexes}}, \code{\link{compute_reporting_indexes_limits}},
         \code{\link{compute_ksi_threshold}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
