% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{compute_specificity,SpectralAnalyzer-method}
\alias{compute_specificity,SpectralAnalyzer-method}
\alias{compute_specificity}
\title{Specificity computation}
\usage{
\S4method{compute_specificity}{SpectralAnalyzer}(object, patterns, frequencies, weights)
}
\arguments{
\item{object}{S4 object of class \code{SpectralAnalyzer}.}

\item{patterns}{Patterns whose specificity is to be computed.}

\item{frequencies}{Vector of frequencies associated with the patterns contained in \code{patterns}.}

\item{weights}{Vector of weights associated with the patterns contained in \code{patterns}.}
}
\value{
Vector containing the specificity of each pattern contained in \code{patterns}.
}
\description{
Compute the specificity of the information conveyed by each pattern.
The specitificity corresponds to the nature of a pattern of being specific of a particular
combination or ubiquitous and allowing the formation of numerous combinations (with regard to the
observations).
}
\details{
\loadmathjax

The specificy of the pattern \eqn{p} is given by:
 \mjdeqn{S_p = \frac{H_{max} - H_p}{H_{max} - H_{min}}}{S_p = (H_max - H_p) / (H_max - H_min)}
where:
\describe{
 \item{}{\mjeqn{H_p = - \sum_{v = 1}^{U_p} \frac{A_{p,v}}{W_p} \times log \left( \frac{A_{p,v}}{W_p} \right)}{H_p = - sum(A_pv / W_p * log(A_pv / W_p)) from v = 1 to U_p}}
 \item{}{\mjeqn{H_{max} = \max log(U)}{H_max = max log(U)}}
 \item{}{\mjeqn{H_{min} = log \left( \frac{W_p}{W_p - U_p + 1} \right) + \left( \frac{U_p - 1}{W_p} \times log(W_p - U_p + 1) \right)}{H_min = log(W_p / (W_p - U_p + 1)) + (U_p - 1) / W_p * log(W_p - U_p + 1)}}
 \item{}{where \mjseqn{A_p} is the set of weights of the nodes containing the pattern \eqn{p},
         \eqn{W} is the set of weights of the patterns and
         \eqn{U} is the set of frequencies of the patterns.}
}
}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
The spectrosome of occupational health problems. \emph{PLoS ONE} 13(1): e0190196.
\url{https://doi.org/10.1371/journal.pone.0190196}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
