% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{compute_specificity,SpectralAnalyzer-method}
\alias{compute_specificity,SpectralAnalyzer-method}
\alias{compute_specificity}
\title{Specificity computation}
\usage{
\S4method{compute_specificity}{SpectralAnalyzer}(object, patterns, frequencies, weights)
}
\arguments{
\item{object}{S4 object of class \code{SpectralAnalyzer}.}

\item{patterns}{Patterns whose specificity is to be computed.}

\item{frequencies}{Vector of frequencies associated with the patterns contained in \code{patterns}.}

\item{weights}{Vector of weights associated with the patterns contained in \code{patterns}.}
}
\value{
Vector containing the specificity of each pattern contained in \code{patterns}.
}
\description{
Compute the specificity of the information conveyed by each pattern.
The specitificity corresponds to the nature of a pattern of being specific of a particular
combination or ubiquitous and allowing the formation of numerous combinations (with regard to the
transactions).
}
\details{
\loadmathjax

The specificy of the pattern \eqn{p} is given by:
 \mjdeqn{S_p = \begin{cases} 1 & \mbox{if } U_p = 1 \cr 0 & \mbox{if } A_{p,1} = A_{p,2} = \ldots = A_{p,U_p} \cr \frac{H_{max,p} - H_p}{H_{max,p} - H_{min,p}} & \mbox{otherwise} \end{cases}}{S_p = 1 if U_p = 1, 0 if A_p1 = A_p2 = ... = A_{p,U_p}, (H_max,p - H_p) / (H_max,p - H_min,p) otherwise}
where \mjseqn{U} is the set of frequencies of the patterns, \mjseqn{A_p} is the
set of weights of the nodes containing the pattern \eqn{p} and:
\describe{
 \item{}{\mjeqn{H_p = - \sum_{i = 1}^{U_p} \frac{A_{p,i}}{W_p} \times ln \left( \frac{A_{p,i}}{W_p} \right)}{H_p = - sum(A_pi / W_p * ln(A_pi / W_p)) from i = 1 to U_p}}
 \item{}{\mjeqn{H_{max,p} = ln(U_p)}{H_max,p = ln(U_p)}}
 \item{}{\mjeqn{H_{min,p} = ln \left( \frac{W_p}{W_p - U_p + 1} \right) + \left( \frac{U_p - 1}{W_p} \times ln(W_p - U_p + 1) \right)}{H_min,p = ln(W_p / (W_p - U_p + 1)) + (U_p - 1) / W_p * ln(W_p - U_p + 1)}}
 \item{}{where \mjseqn{W} is the set of weights of the patterns.}
}
}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
The spectrosome of occupational health problems. \emph{PLoS ONE} 13(1): e0190196.
\url{https://doi.org/10.1371/journal.pone.0190196}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
