% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{compute_specificity,SpectralAnalyzer-method}
\alias{compute_specificity,SpectralAnalyzer-method}
\alias{compute_specificity}
\title{Calcul de spécificité}
\usage{
\S4method{compute_specificity}{SpectralAnalyzer}(object, patterns, frequencies, weights)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{patterns}{Liste des motifs dont la spécificité est à calculer.}

\item{frequencies}{Vecteur des fréquences associées aux motifs contenus dans \code{patterns}.}

\item{weights}{Vecteur des poids associées aux motifs contenus dans \code{patterns}.}
}
\value{
Vecteur contenant la spécificité de chaque motif contenu dans \code{patterns}.
}
\description{
Calcule la spécificité de l'information portée par chaque motif.
La spécificité correspond au caractère d'un motif spécifique d'une combinaison particulière
 ou ubiquitaire et permettant la formation de nombreux agrégats.
}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
            \emph{The spectrosome of occupational health problems}. PLoS ONE 13(1): e0190196.
            \url{https://doi.org/10.1371/journal.pone.0190196}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
