% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{compute_specificity,SpectralAnalyzer-method}
\alias{compute_specificity,SpectralAnalyzer-method}
\alias{compute_specificity}
\title{Specificity computation}
\usage{
\S4method{compute_specificity}{SpectralAnalyzer}(object, patterns, frequencies, weights)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{patterns}{Patterns whose specificity is to be computed.}

\item{frequencies}{Vector of frequencies associated with the patterns contained in \code{patterns}.}

\item{weights}{Vector of weights associated with the patterns contained in \code{patterns}.}
}
\value{
Vector containing the specificity of each pattern contained in \code{patterns}.
}
\description{
Compute the specificity of the information conveyed by each pattern.
The specitificity corresponds to the character of a pattern of being specific of a particular
 combination or ubiquitous and allowing the formation of numerous aggregates.
}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
            The spectrosome of occupational health problems. \emph{PLoS ONE} 13(1): e0190196.
            \url{https://doi.org/10.1371/journal.pone.0190196}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
