% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{count_links,SpectralAnalyzer-method}
\alias{count_links,SpectralAnalyzer-method}
\title{Compte le nombre d'éléments en commun entre chacun des noeuds ou des motifs.}
\usage{
\S4method{count_links}{SpectralAnalyzer}(object, entities)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{entities}{Type d'élément pour lequel compter les liens (noeuds ou motifs).
Choix parmi \code{c("nodes", "patterns")}.}
}
\value{
Matrice du nombre de liens entre chaque noeud ou motif (matrice d'adjacence).
}
\description{
La matrice résultante est associée à l'attribut \code{n_links} ou \code{p_links} respectivement.
}
\author{
Gauthier Magnin
}
