% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{count_links,SpectralAnalyzer-method}
\alias{count_links,SpectralAnalyzer-method}
\alias{count_links}
\title{Counting of links}
\usage{
\S4method{count_links}{SpectralAnalyzer}(object, entities)
}
\arguments{
\item{object}{S4 object of class \code{SpectralAnalyzer}.}

\item{entities}{Type of entities for which to count links (nodes or patterns).
\code{NODES} or \code{PATTERNS}.}
}
\value{
Invisible. Adjacency matrix: matrix of the number of links between each pair of nodes
 or patterns.
}
\description{
Count the number of items in common between each pair of nodes or patterns.
The resulting matrix is assigned respectively to the attribute \code{n_links} or \code{p_links}
 of \code{object}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
