% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{define_dynamic_status,SpectralAnalyzer-method}
\alias{define_dynamic_status,SpectralAnalyzer-method}
\alias{define_dynamic_status}
\title{Dynamic status assignment}
\usage{
\S4method{define_dynamic_status}{SpectralAnalyzer}(object, patterns, status_limit, t = NULL, period = Inf)
}
\arguments{
\item{object}{S4 object of class \code{SpectralAnalyzer}.}

\item{patterns}{Patterns whose dynamic status are to be defined.}

\item{status_limit}{Time interval over which to characterize the status of the patterns in relation
to the period defined by the arguments \code{t} and \code{period}.}

\item{t}{Year of the end of the period, i.e. the date on which to characterize the pattern.
\code{NULL} specifies that the characterization must be done in relation to the last year covered
by the observations.}

\item{period}{Time interval over which to characterize the patterns (number of years).
For example, if \code{t = 2015} and \code{period = 2} then the computation is made over the
period [2014 - 2015].
\code{Inf} specifies that the period considered covers an interval starting on the date of the
oldest observation and ending in the year \code{t}.}
}
\value{
Data frame associating each pattern with its dynamic status.
}
\description{
Define the dynamic status of each pattern: persistent, declining, emergent or latent.
}
\details{
\loadmathjax

For each pattern, two reporting index values are computed and compared to two thresholds in order
to define its status.

A pattern is:
\itemize{
\item Persistent when it appears substantially during both the entire period of the
observations (defined by the arguments \code{period} and \code{t}) and a shorter period defined by the
arguments \code{status_limit} and \code{t}.
\item Declining when it appears substantially during the entire period but much less during
the shorter period.
\item Emergent when it appears insignificantly during the entire period but much more during
the shorter period.
\item Latent when it appears insignificantly during both the entire period and the shorter period.
}

The reporting indexes of the pattern \eqn{p} which are computed are those at the two temporal limits,
given by:
\mjdeqn{RI_{\infty,p} \; = lim_{t_0 \to -\infty} RI_p(t_1,t_0)}{RI_inf,p = lim of RI_p(t_1,t_0) as t approaches -inf}
\mjdeqn{RI_{l,p} = RI_p(t_1, t_1 - l + 1)}{RI_lp = RI_p(t_1, t_1 - l + 1)}
where \eqn{l} is the shorter period on which to compute a reporting index and \mjseqn{RI_p(t_1,t_0)}
is the reporting index of the pattern \eqn{p} given by:
\mjdeqn{RI_p(t_1,t_0) = \frac{\sum_{t = t_0}^{t_1} W_{p,t}}{\sum_{q \in P} \sum_{t = t_0}^{t_1} W_{q,t}}}{RI_p(t_1,t_0) = sum W_pt from t = t_0 to t_1 / sum W_{q,t} for q in P and from t = t_0 to t_1}
where \eqn{P} is the set of patterns, \mjeqn{W_{p,t}}{W_pt} is the weight of the pattern \eqn{p} in
the observations of the year \eqn{t}, \mjseqn{t_0} and \mjseqn{t_1} are the first and last years
defining the period on which to compute the reporting index.

For each set of reporting indexes:
\itemize{
\item One threshold \mjseqn{\xi} is computed as follows:
\mjdeqn{\xi = \frac{1}{\sum_{p \in P} RI_p(t_1,t_0)^2}}{ksi = 1 / sum(RI_p(t_1,t_0)^2) for p in P}
\item The patterns are ordered in descending order of their reporting index value and separated by
this threshold.
}

The patterns are then classified as follows:
\itemize{
\item Patterns for which \mjeqn{RI_{l,p}}{RI_lp} and \mjeqn{RI_{\infty,p}}{RI_inf,p} are before the
thresholds are classed as "Persistent".
\item Patterns for which \mjeqn{RI_{l,p}}{RI_lp} is after the related threshold and
\mjeqn{RI_{\infty,p}}{RI_inf,p} is before the related threshold are classed as "Declining".
\item Patterns for which \mjeqn{RI_{l,p}}{RI_lp} is before the related threshold and
\mjeqn{RI_{\infty,p}}{RI_inf,p} is after the related threshold are classed as "Emergent".
\item Patterns for which \mjeqn{RI_{l,p}}{RI_lp} and \mjeqn{RI_{\infty,p}}{RI_inf,p} are after the
thresholds are classed as "Latent".
}
}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
The spectrosome of occupational health problems. \emph{PLoS ONE} 13(1): e0190196.
\url{https://doi.org/10.1371/journal.pone.0190196}.
}
\seealso{
\code{\link{compute_reporting_indexes}}, \code{\link{compute_reporting_indexes_limits}},
\code{\link{compute_ksi_threshold}}, \code{\link{compute_ri_threshold}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
