% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{define_dynamic_status,SpectralAnalyzer-method}
\alias{define_dynamic_status,SpectralAnalyzer-method}
\alias{define_dynamic_status}
\title{Dynamic status assignment}
\usage{
\S4method{define_dynamic_status}{SpectralAnalyzer}(object, patterns, status_limit, t = NULL, period = Inf)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{patterns}{Patterns whose dynamic status are to be defined.}

\item{status_limit}{Time interval over which to characterize the status of the patterns in relation
to the period defined by the arguments \code{t} and \code{period}.}

\item{t}{Year of the end of the period, i.e. the date on which to characterize the pattern.
\code{NULL} specifies that the characterization must be done in relation to the last year covered
by the observations.}

\item{period}{Time interval over which to characterize the patterns (number of years).
For example, if \code{t = 2015} and \code{period = 2} then the computation is made over the
period [2014 - 2015].
\code{Inf} specifies that the period considered covers an interval starting on the date of the
oldest observation and ending in the year \code{t}.}
}
\value{
Data frame associating each pattern with its dynamic status.
}
\description{
Define the dynamic status of each pattern.
}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
            The spectrosome of occupational health problems. \emph{PLoS ONE} 13(1): e0190196.
            \url{https://doi.org/10.1371/journal.pone.0190196}.
}
\seealso{
\code{\link{compute_reporting_indexes}}, \code{\link{compute_reporting_indexes_limits}},
         \code{\link{compute_ksi_threshold}}, \code{\link{compute_ri_threshold}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
