% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{degree,SpectralAnalyzer-method}
\alias{degree,SpectralAnalyzer-method}
\alias{degree}
\title{Degree of a vertex}
\usage{
\S4method{degree}{SpectralAnalyzer}(object, ID, links)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{ID}{Identifier of the vertex (node or pattern) whose degree is to be calculated.}

\item{links}{Data frame of the links (or edges) of a spectrosome graph.}
}
\value{
Degree of the vertex.
}
\description{
Compute the degree of a vertex in a graph, i.e. the number of vertices to which it is adjacent.
}
\examples{
spectrosome <- spectrosome_chart(SA_instance, "patterns",
                                 SA_instance["patterns"][1:15, ])
degree(SA_instance, 7, spectrosome[["edges"]])

}
\author{
Gauthier Magnin
}
