% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dhms}
\alias{dhms}
\title{Convertit une période en secondes en une chaîne de caractères.}
\usage{
dhms(t)
}
\arguments{
\item{t}{Numeric - Temps à convertir en secondes.}
}
\value{
Chaîne de caractère attendue.
}
\description{
Exemples du format de la châine retournée : 1d 02h 25m 47s ; 01m 05s ; 19s.
}
