% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dhms}
\alias{dhms}
\title{Decomposition of a time}
\usage{
dhms(t)
}
\arguments{
\item{t}{numeric. Time to convert, in seconds.}
}
\value{
Character string corresponding to the converted time.
}
\description{
Convert a period defined in seconds to a character string.
Examples of the format of the returned string: 1d 02h 25m 47s; 01m 05s; 19s.
}
\references{
Stack Overflow topic :
 \href{https://stackoverflow.com/questions/27312292/convert-seconds-to-days-hoursminutesseconds}{Convert seconds to days:hours:minutes:seconds}
}
\keyword{internal}
