% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{display_time}
\alias{display_time}
\title{Affiche un temps d'exécution}
\usage{
display_time(expr)
}
\arguments{
\item{expr}{Expression R valide à chronométrer.}
}
\description{
Évalue une expression et affiche le temps écoulé selon le format défini par la fonction \code{dhms}.
Exemples du format de la chaîne retournée : [1d 02h 25m 47s] ; [01m 05s] ; [19s].
}
\details{
Un appel au garbage collector est effectué avant le chronométrage.
}
\seealso{
\code{\link{dhms}}.
}
\keyword{internal}
