% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{display_time}
\alias{display_time}
\title{Print an execution time}
\usage{
display_time(expr)
}
\arguments{
\item{expr}{Valid \code{R} expression to be timed.}
}
\description{
Evaluate an expression and print the elapsed time according to the format defined by the
 \code{dhms} function.
Examples of the display format: [1d 02h 25m 47s]; [01m 05s]; [19s].
}
\details{
A call to the garbage collector is made beforme timing.
}
\seealso{
\code{\link{dhms}}, \code{\link{system.time}}.
}
\keyword{internal}
