% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{display_time}
\alias{display_time}
\title{Évalue une expression et affiche le temps écoulé selon le format défini par la fonction \code{dhms}.}
\usage{
display_time(expr)
}
\arguments{
\item{expr}{Expression R valide à chronométrer.}
}
\description{
Un appel au garbage collector est effectué avant le chronométrage.
}
\seealso{
\code{\link{dhms}}.
}
