% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{extract_links,SpectralAnalyzer-method}
\alias{extract_links,SpectralAnalyzer-method}
\alias{extract_links}
\title{Extraction de liens}
\usage{
\S4method{extract_links}{SpectralAnalyzer}(object, entities, characteristics)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{entities}{Type d'élément pour lequel rechercher les liens.
Choix parmi \code{"nodes"}, \code{"patterns"}.}

\item{characteristics}{Data frame des caractéristiques des nœuds ou motifs dont les liens sont à rechercher.}
}
\value{
Data frame associant les nœuds ou motifs liés.
}
\description{
Extrait de l'ensemble des liens ceux correspondant aux nœuds ou motifs recherchés.
}
\details{
Si parmi les nœuds ou motifs, certains deviennent isolés du fait que les autres éléments
 auxquels ils sont normalement liés ne font pas partie de \code{characteristics}, ces nœuds ou
 motifs sont ajoutés à la fin de la data frame de retour.
Ces éventuelles \code{n} lignes additionnelles sont numérotées \code{"A1"..."An"}.
}
\author{
Gauthier Magnin
}
