% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{extract_nodes_from_category,SpectralAnalyzer-method}
\alias{extract_nodes_from_category,SpectralAnalyzer-method}
\alias{extract_nodes_from_category}
\title{Search for nodes by category}
\usage{
\S4method{extract_nodes_from_category}{SpectralAnalyzer}(
  object,
  nodes_characteristics,
  category,
  value,
  target
)
}
\arguments{
\item{object}{SpectralAnalyzer class object.}

\item{nodes_characteristics}{Data frame of nodes and their characteristics.}

\item{category}{Name or number of the category on which to search (numbering according to the order
of the columns of \code{object["items_categories"]}).}

\item{value}{Sought value for the category specified by the parameter \code{category}.}

\item{target}{Condition for a node to be extracted. One of \code{"vertices"}, \code{"edges"}.
\describe{
  \item{\code{"vertices"}}{Search for nodes containing an item associated with the sought category
                           value.}
  \item{\code{"edges"}}{Search for nodes generating links corresponding to the sought category
                        value.}
}}
}
\value{
Subset of the data frame of nodes that match the search criteria.
}
\description{
Extract the nodes corresponding to a sought category value.
}
\seealso{
\code{\link{extract_nodes_from_items}}, \code{\link{extract_nodes_from_characteristic}}.
}
\author{
Gauthier Magnin
}
