% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{extract_nodes_from_characteristic,SpectralAnalyzer-method}
\alias{extract_nodes_from_characteristic,SpectralAnalyzer-method}
\title{Extrait les noeuds satisfaisant un critère de recherche.}
\usage{
\S4method{extract_nodes_from_characteristic}{SpectralAnalyzer}(
  object,
  nodes_characteristics,
  characteristic,
  value,
  condition = "EQ"
)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{nodes_characteristics}{Data.frame des caractéristiques des noeuds.}

\item{characteristic}{Nom de la caractéristique sur laquelle faire la recherche.
Choix parmi \code{c("length", "weight")}.}

\item{value}{Valeur recherchée pour la caractéristique spécifiée par l'argument \code{characteristic}.}

\item{condition}{Condition de la recherche.
Choix parmi \code{c("EQ", "NE", "LT", "GT", "LE", "GE")}.
\code{"EQ"} : La valeur de la caractéristique du noeud doit être égale à la valeur recherchée (Equal).
\code{"NE"} : La valeur de la caractéristique du noeud doit être différente de la valeur recherchée (Not Equal).
\code{"LT"} : La valeur de la caractéristique du noeud doit être inférieure à la valeur recherchée (Less Than).
\code{"GT"} : La valeur de la caractéristique du noeud doit être supérieure à la valeur recherchée (Greater Than).
\code{"LE"} : La valeur de la caractéristique du noeud doit être inférieure ou égale à la valeur recherchée (Less than or Equal).
\code{"GE"} : La valeur de la caractéristique du noeud doit être supérieure ou égale à la valeur recherchée (Greater than or Equal).}
}
\value{
Sous-ensemble de la data.frame fournie en argument pour les noeuds satisfaisant le critère de recherche.
}
\description{
Extrait les noeuds satisfaisant un critère de recherche.
}
\author{
Gauthier Magnin
}
