% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{extract_nodes_from_items,SpectralAnalyzer-method}
\alias{extract_nodes_from_items,SpectralAnalyzer-method}
\alias{extract_nodes_from_items}
\title{Search for nodes by item}
\usage{
\S4method{extract_nodes_from_items}{SpectralAnalyzer}(object, nodes_characteristics, items, target = "all")
}
\arguments{
\item{object}{SpectralAnalyzer class object.}

\item{nodes_characteristics}{Data frame of nodes and their characteristics.}

\item{items}{Sought items (one or more).}

\item{target}{Condition for a node to be extracted. One of \code{"all"}, \code{"any"}.
\describe{
  \item{\code{"all"}}{All the sought items must be part of a node for this node to be extracted.}
  \item{\code{"any"}}{At least one of the sought items must be part of a node for this node to be
                      extracted.}
}}
}
\value{
Subset of the data frame of nodes that match the search criteria.
}
\description{
Extract the nodes containing one or more sought items.
}
\seealso{
\code{\link{extract_nodes_from_characteristic}}, \code{\link{extract_nodes_from_category}}.
}
\author{
Gauthier Magnin
}
