% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{extract_patterns_from_category,SpectralAnalyzer-method}
\alias{extract_patterns_from_category,SpectralAnalyzer-method}
\alias{extract_patterns_from_category}
\title{Search for patterns by category}
\usage{
\S4method{extract_patterns_from_category}{SpectralAnalyzer}(
  object,
  patterns_characteristics,
  category,
  value,
  target
)
}
\arguments{
\item{object}{SpectralAnalyzer class object.}

\item{patterns_characteristics}{Data frame of patterns and their characteristics.}

\item{category}{Name or number of the category on which to search (numbering according to the order
of the columns of \code{object["items_categories"]}).}

\item{value}{Sought value for the category specified by the argument \code{category}.}

\item{target}{Condition for a pattern to be extracted. One of \code{"vertices"}, \code{"edges"}.
\describe{
  \item{\code{"vertices"}}{Search for patterns containing an item associated with the sought
                           category value.}
  \item{\code{"edges"}}{Search for patterns generating links corresponding to the sought category
                        value.}
}}
}
\value{
Subset of the data frame of patterns that match the search criteria.
}
\description{
Extract the patterns corresponding to a sought category value.
}
\seealso{
\code{\link{extract_patterns_from_items}}, \code{\link{extract_patterns_from_characteristic}},
         \code{\link{extract_patterns_from_status}}.
}
\author{
Gauthier Magnin
}
