% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{extract_patterns_from_category,SpectralAnalyzer-method}
\alias{extract_patterns_from_category,SpectralAnalyzer-method}
\title{Extrait les motifs correspondant à une valeur de catégorie recherchée}
\usage{
\S4method{extract_patterns_from_category}{SpectralAnalyzer}(object, category, value, target)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{category}{Nom ou numéro de la catégorie à laquelle accéder (numérotation selon l'ordre des colonnes de \code{object["items_categories"]}).}

\item{value}{Valeur recherchée pour la catégorie spécifiée par l'argument \code{category}.}

\item{target}{Condition pour qu'un motif soit extrait. Choix parmi \code{c("vertices", "edges")}.
\code{"vertices"} : recherche des motifs contenant un item associé à la catégorie recherchée.
\code{"edges"} : recherche de motifs générant un lien correspondant à la catégorie recherchée.}
}
\value{
Data.frame des motifs correspondant aux critères de recherche.
}
\description{
Extrait les motifs correspondant à une valeur de catégorie recherchée
}
\author{
Gauthier Magnin
}
