% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{extract_patterns_from_characteristic,SpectralAnalyzer-method}
\alias{extract_patterns_from_characteristic,SpectralAnalyzer-method}
\alias{extract_patterns_from_characteristic}
\title{Recherche de motifs par caractéristique}
\usage{
\S4method{extract_patterns_from_characteristic}{SpectralAnalyzer}(
  object,
  patterns_characteristics,
  characteristic,
  value,
  condition = "EQ"
)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{patterns_characteristics}{Data frame des motifs et de leurs caractéristiques.}

\item{characteristic}{Nom de la caractéristique sur laquelle faire la recherche.
Choix parmi \code{"year"}, \code{"frequency"}, \code{"weight"}, \code{"order"}, \code{"specificity"}.}

\item{value}{Valeur recherchée pour la caractéristique spécifiée par l'argument \code{characteristic}.}

\item{condition}{Condition de la recherche.
Choix parmi \code{"EQ"}, \code{"NE"}, \code{"LT"}, \code{"GT"}, \code{"LE"}, \code{"GE"}.
\describe{
  \item{\code{"EQ"}}{\strong{EQ}ual : la valeur de la caractéristique doit être égale à celle recherchée.}
  \item{\code{"NE"}}{\strong{N}ot \strong{E}qual : la valeur de la caractéristique doit être différente de celle recherchée.}
  \item{\code{"LT"}}{\strong{L}ess \strong{T}han : la valeur de la caractéristique doit être inférieure à celle recherchée.}
  \item{\code{"GT"}}{\strong{G}reater \strong{T}han : la valeur de la caractéristique doit être supérieure à celle recherchée.}
  \item{\code{"LE"}}{\strong{L}ess than or \strong{E}qual : la valeur de la caractéristique doit être inférieure ou égale à celle recherchée.}
  \item{\code{"GE"}}{\strong{G}reater than or \strong{E}qual : la valeur de la caractéristique doit être supérieure ou égale à celle recherchée.}
}}
}
\value{
Sous-ensemble de la data frame fournie en argument contant les motifs satisfaisant le critère de recherche.
}
\description{
Extrait les motifs satisfaisant un critère de recherche relatif à une caractéristique.
}
\seealso{
\code{\link{extract_patterns_from_items}}, \code{\link{extract_patterns_from_status}},
         \code{\link{extract_patterns_from_category}}.
}
\author{
Gauthier Magnin
}
