% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{extract_patterns_from_characteristic,SpectralAnalyzer-method}
\alias{extract_patterns_from_characteristic,SpectralAnalyzer-method}
\alias{extract_patterns_from_characteristic}
\title{Search for patterns by characteristic}
\usage{
\S4method{extract_patterns_from_characteristic}{SpectralAnalyzer}(
  object,
  patterns_characteristics,
  characteristic,
  value,
  condition = "EQ"
)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{patterns_characteristics}{Data frame of patterns and their characteristics.
Any subset of \code{object["patterns"]}.}

\item{characteristic}{Name of the characteristic on which to do the search.
One of \code{"year"}, \code{"frequency"}, \code{"weight"}, \code{"order"}, \code{"specificity"}
See \code{\link{extract_patterns_from_status}} to search by \code{"status"}.}

\item{value}{Sought value for the characteristic specified by the parameter \code{characteristic}.}

\item{condition}{Search condition.
One of \code{"EQ"}, \code{"NE"}, \code{"LT"}, \code{"GT"}, \code{"LE"}, \code{"GE"}.
\describe{
  \item{\code{"EQ"}}{\strong{EQ}ual: the value of the characteristic must be equal to that sought.}
  \item{\code{"NE"}}{\strong{N}ot \strong{E}qual: the value of the characteristic must be different
                     from that sought.}
  \item{\code{"LT"}}{\strong{L}ess \strong{T}han: the value of the characteristic must be less
                     than that sought.}
  \item{\code{"GT"}}{\strong{G}reater \strong{T}han: the value of the characteristic must be greater
                     than that sought.}
  \item{\code{"LE"}}{\strong{L}ess than or \strong{E}qual: the value of the caracteristic must be
                     less than or equal to that sought.}
  \item{\code{"GE"}}{\strong{G}reater than or \strong{E}qual: the value of the characteristic must
                     be greater than or equal to that sought.}
}}
}
\value{
Subset of the data frame of patterns that match the search criteria.
}
\description{
Extract the patterns satisfying a search criterion according to one characteristic.
}
\examples{
extract_patterns_from_characteristic(SA_instance, SA_instance["patterns"],
                                     characteristic = "weight",
                                     value = 3, condition = "GE")

}
\seealso{
\code{\link{extract_patterns_from_items}}, \code{\link{extract_patterns_from_status}},
         \code{\link{extract_patterns_from_category}}.
}
\author{
Gauthier Magnin
}
