% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{extract_patterns_from_items,SpectralAnalyzer-method}
\alias{extract_patterns_from_items,SpectralAnalyzer-method}
\title{Extrait les motifs contenant un ou plusieurs items recherchés.}
\usage{
\S4method{extract_patterns_from_items}{SpectralAnalyzer}(
  object,
  patterns_characteristics,
  items,
  target = "all"
)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{patterns_characteristics}{Data.frame des caractéristiques des motifs.}

\item{items}{Éléments recherchés (un ou plusieurs).}

\item{target}{Condition pour qu'un motif soit extrait. Choix parmi \code{c("all", "any")}.
\code{"all"} exprime que l'intégralité des éléments recherchés doivent faire partie d'un motif pour que ce motif soit extrait.
\code{"any"} exprime qu'au moins un des éléments recherchés doit faire partie d'un motif pour que ce motif soit extrait.}
}
\value{
Sous-ensemble de la data.frame fournie en argument pour les motifs correspondant au critère de recherche.
}
\description{
Extrait les motifs contenant un ou plusieurs items recherchés.
}
\author{
Gauthier Magnin
}
