% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{extract_patterns_from_items,SpectralAnalyzer-method}
\alias{extract_patterns_from_items,SpectralAnalyzer-method}
\alias{extract_patterns_from_items}
\title{Search for patterns by item}
\usage{
\S4method{extract_patterns_from_items}{SpectralAnalyzer}(
  object,
  patterns_characteristics,
  items,
  presence = "all"
)
}
\arguments{
\item{object}{SpectralAnalyzer class object.}

\item{patterns_characteristics}{Data frame of patterns and their characteristics.}

\item{items}{Sought items (one or more).}

\item{presence}{Item presence condition for a pattern to be extracted.
One of \code{"all"}, \code{"any"}.
\describe{
  \item{\code{"all"}}{All the sought items must be part of a pattern for this pattern to be
                      extracted.}
  \item{\code{"any"}}{At least one of the sought items must be part of a pattern for this pattern
                      to be extracted.}
}}
}
\value{
Subset of the data frame of patterns that match the search criteria.
}
\description{
Extract the patterns containing one or more sought items.
}
\seealso{
\code{\link{extract_patterns_from_characteristic}}, \code{\link{extract_patterns_from_status}},
         \code{\link{extract_patterns_from_category}}.
}
\author{
Gauthier Magnin
}
