% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{extract_patterns_from_status,SpectralAnalyzer-method}
\alias{extract_patterns_from_status,SpectralAnalyzer-method}
\alias{extract_patterns_from_status}
\title{Search for patterns by status}
\usage{
\S4method{extract_patterns_from_status}{SpectralAnalyzer}(
  object,
  patterns_characteristics,
  value,
  condition = "EQ"
)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{patterns_characteristics}{Data frame of patterns and their characteristics.}

\item{value}{Status value sought (one or more)}

\item{condition}{Search condition. One of \code{"EQ"}, \code{"NE"}.
\describe{
  \item{\code{"EQ"}}{\strong{EQ}ual: the status of the pattern must be one of the sought values.}
  \item{\code{"NE"}}{\strong{N}ot \strong{E}qual: the status of the pattern must be different from
                     the sought values.}
}}
}
\value{
Subset of the data frame of patterns that match the search criteria.
}
\description{
Extract the patterns whose status match one or more sought values.
}
\examples{
extract_patterns_from_status(SA_instance, SA_instance["patterns"],
                             value = SA_instance["Class"]$STATUS_PERSISTENT,
                             condition = "EQ")

extract_patterns_from_status(SA_instance, SA_instance["patterns"],
                             value = c("Persistent", "Declining"),
                             condition = "NE")

}
\seealso{
\code{\link{extract_patterns_from_items}}, \code{\link{extract_patterns_from_characteristic}},
         \code{\link{extract_patterns_from_category}}.
}
\author{
Gauthier Magnin
}
