% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_maker.R
\name{get_all_items}
\alias{get_all_items}
\title{Search all items}
\usage{
get_all_items(observations, key = "CODE")
}
\arguments{
\item{observations}{List of observations on which to do the search.}

\item{key}{Access key to the items in an observation.}
}
\value{
Vector of all unique items.
}
\description{
Extract all the items contained in the observations.
}
\examples{
obs <- make_observations(oedb_sample, by = "ID", additional = c("CODE", "NAME"))
get_all_items(obs)

}
\author{
Gauthier Magnin
}
