% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations_maker.R
\name{get_all_items}
\alias{get_all_items}
\title{Recherche l'ensemble des items}
\usage{
get_all_items(observations, key = "CODE")
}
\arguments{
\item{observations}{Liste des observations sur lesquels effectuer la recherche.}

\item{key}{Clé d'accès aux items de chaque observation.}
}
\value{
Vecteur des identifiants de l'ensemble des items.
}
\description{
Extrait l'intégralité des items contenus dans les observations.
}
\author{
Gauthier Magnin
}
