% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{get_complexes,SpectralAnalyzer-method}
\alias{get_complexes,SpectralAnalyzer-method}
\alias{get_complexes}
\title{Search for complex nodes or patterns}
\usage{
\S4method{get_complexes}{SpectralAnalyzer}(
  object,
  entities,
  characteristics,
  category,
  target,
  min_nb_values = 2
)
}
\arguments{
\item{object}{SpectralAnalyzer class object.}

\item{entities}{Type of entities to search for complexes.
One of \code{"nodes"}, \code{"patterns"}.}

\item{characteristics}{Data frame of the characteristics of the nodes or patterns whose complexes
are to be sought.}

\item{category}{Name or number of the category on which to search (numbering according to the order
of the columns of \code{object["items_categories"]}).}

\item{target}{Condition for a node or a pattern to be extracted.
One of \code{"items"}, \code{"links"}, \code{"vertices"}, \code{"edges"}.
\describe{
  \item{\code{"items"}, \code{"vertices"}}{Search for nodes or patterns associated (via their items)
        with several values of the category \code{category}.}
  \item{\code{"links"}, \code{"edges"}}{Search for nodes or patterns generating links corresponding
        with several values of the category \code{category}.}
}}

\item{min_nb_values}{Minimum number of different values of the category \code{category} a node, a
pattern or a link must have to extract the related entity.}
}
\value{
Subset of the data frame that corresponds to the complex entities sought.
}
\description{
Extract from the given nodes or patterns those which are complexes by the number of values with
 which they are associated, with regard to one category.
}
\seealso{
\code{\link{get_isolates}}, \code{\link{get_non_isolates}}, \code{\link{get_links}}.
}
\author{
Gauthier Magnin
}
