% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_maker.R
\name{get_info_from_items}
\alias{get_info_from_items}
\title{Search for information by item}
\usage{
get_info_from_items(
  observations,
  items,
  info_names,
  presence = "all",
  key = "CODE"
)
}
\arguments{
\item{observations}{List of observations on which to do the search.}

\item{items}{Sought items.}

\item{info_names}{Names of information to extract from observations.}

\item{presence}{Item presence condition for information to be extracted from an observation.
One of \code{"all"}, \code{"any"}, \code{"exact"}.
 \describe{
  \item{\code{"all"}}{All the sought items must be part of an observation for its information
                      to be extracted.}
  \item{\code{"any"}}{At least one of the sought items must be part of an observation for its
                      information to be extracted.}
   \item{\code{"exact"}}{The item set contained in an observation must be exactly the same as the
                         sought item set for this observation to be extracted.}
 }}

\item{key}{Access key to the items in an observation.}
}
\value{
Vector or list of information corresponding to the search.
 Vector if only one type of information is to be extracted. List otherwise.
}
\description{
Retrieve information associated with observations that contain a set of sought items.
}
\examples{
obs <- make_observations(oedb_sample, by = "ID",
                         additional = c("CODE", "NAME", "JOB.TITLE", "JOB.TASK"))
get_info_from_items(obs, items = 3146, info_names = c("JOB.TITLE", "JOB.TASK"))

}
\seealso{
\code{\link{get_items_from_info}}, \code{\link{get_all_items}}.
}
\author{
Gauthier Magnin
}
