% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{get_isolates,SpectralAnalyzer-method}
\alias{get_isolates,SpectralAnalyzer-method}
\alias{get_isolates}
\title{Search for isolated nodes or patterns}
\usage{
\S4method{get_isolates}{SpectralAnalyzer}(object, nopc)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{nopc}{Data frame of \strong{n}odes \strong{o}r \strong{p}atterns and their
\strong{c}haracteristics. Nodes or patterns whose isolated are to be sought. Any subset of
\code{object["nodes"]} or \code{object["patterns"]}.\cr
\code{"nodes"}, \code{"n"}, \code{"patterns"} and \code{"p"} are special values for
\code{object["nodes"]} and \code{object["patterns"]}.}
}
\value{
Subset of the data frame that corresponds to isolated entities.
}
\description{
Extract from the given nodes or patterns those which are isolated (i.e. those which are not linked
 to any other entity).
}
\examples{
get_isolates(SA_instance, "patterns")
get_isolates(SA_instance, SA_instance["patterns"][1:10, ])

}
\seealso{
\code{\link{get_non_isolates}}, \code{\link{get_complexes}}, \code{\link{get_links}}.
}
\author{
Gauthier Magnin
}
