% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_maker.R
\name{get_items_from_info}
\alias{get_items_from_info}
\title{Search for items by specific information}
\usage{
get_items_from_info(
  observations,
  info,
  presence = "all",
  additional = NULL,
  key = "CODE"
)
}
\arguments{
\item{observations}{List of observations on which to do the search.}

\item{info}{Named list of sought information. Element names must refer to the names of variables
contained in the observations and values must correspond to the sought values for these variables.}

\item{presence}{Information presence condition for an item to be extracted from an observation.
One of \code{"all"}, \code{"any"}.
\describe{
 \item{\code{"all"}}{All the sought information must be part of an observation for its items to
                     be extracted.}
 \item{\code{"any"}}{At least one of the sought information must be part of an observation for
                     its items to be extracted.}
}}

\item{additional}{Names of additional information to extract during the search.}

\item{key}{Access key to the items in an observation.}
}
\value{
Vector of the item codes corresponding to the search if \code{additional} is
 equal to \code{NULL}. List of codes and information requested otherwise.
}
\description{
Retrieve the items associated with observations whose information matches to one or more sought
 information.
}
\examples{
obs <- make_observations(oedb_sample, by = "ID",
                         additional = c("CODE", "NAME",
                                        "JOB.TITLE", "JOB.TASK", "SAMPLE.ID"))
get_items_from_info(obs, info = list(JOB.TITLE = 44132017))
get_items_from_info(obs, info = list(JOB.TITLE = 44132017), additional = "SAMPLE.ID")

}
\seealso{
\code{\link{get_all_items}}, \code{\link{get_info_from_items}}.
}
\author{
Gauthier Magnin
}
