% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{get_links,SpectralAnalyzer-method}
\alias{get_links,SpectralAnalyzer-method}
\alias{get_links}
\title{Get links between nodes or patterns}
\usage{
\S4method{get_links}{SpectralAnalyzer}(object, entities, characteristics)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{entities}{Type of entities to search for links.
One of \code{"nodes"}, \code{"patterns"}.}

\item{characteristics}{Data frame of the characteristics of the nodes or patterns whose links are
to be sought.}
}
\value{
Data frame associating the linked nodes or linked patterns.
}
\description{
Extract from the links those corresponding to the desired nodes or patterns.
}
\details{
If among the nodes or patterns some become isolated because the other entities to which they
 are normally linked are not part of \code{characteristics}, these nodes or patterns are placed
 at the end of the return data frame.
These possible \code{n} additional lines are numbered \code{"A1"..."An"}.
}
\examples{
get_links(SA_instance, "patterns", SA_instance["patterns"])
get_links(SA_instance, "patterns", SA_instance["patterns"][1:10, ])

}
\author{
Gauthier Magnin
}
