% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{get_non_isolates,SpectralAnalyzer-method}
\alias{get_non_isolates,SpectralAnalyzer-method}
\alias{get_non_isolates}
\title{Search for non-isolated nodes or patterns}
\usage{
\S4method{get_non_isolates}{SpectralAnalyzer}(object, entities, characteristics)
}
\arguments{
\item{object}{SpectralAnalyzer class object.}

\item{entities}{Type of entities to search for non-isolated.
One of \code{"nodes"}, \code{"patterns"}.}

\item{characteristics}{Data frame of the characteristics of the nodes or patterns whose non-isolated
are to be sought.}
}
\value{
Subset of the data frame that corresponds to non-isolated entities.
}
\description{
Extract from the given nodes or patterns those which are not isolated.
}
\seealso{
\code{\link{get_isolates}}, \code{\link{get_complexes}}, \code{\link{get_links}}.
}
\author{
Gauthier Magnin
}
