% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations_maker.R
\name{get_obs_from_info}
\alias{get_obs_from_info}
\title{Recherche d'observations par information spécifique}
\usage{
get_obs_from_info(observations, ..., target = "all")
}
\arguments{
\item{observations}{Liste des observations sur lesquels effectuer la recherche.}

\item{...}{Arguments de type \code{clé = valeur} où \code{clé} fait référence au nom d'une variable
contenue dans les observations et \code{valeur} correspond à la valeur recherchée pour cette variable.}

\item{target}{Condition pour qu'une observation soit extraite. Choix parmi \code{"all"}, \code{"any"}.
\describe{
 \item{\code{"all"}}{L'intégralité des informations recherchées doivent faire partie d'une observation
                     pour que cette observation soit extraite.}
 \item{\code{"any"}}{Au moins une des informations recherchées doit faire partie d'une observation pour
                     que cette observation soit extraite.}
}}
}
\value{
Sous-ensemble de la liste d'observations correspondant aux critères de recherche.
}
\description{
Extrait les observations correspondant à un ou plusieurs critères de recherche.
}
\seealso{
\code{\link{get_obs_from_items}}.
}
\author{
Gauthier Magnin
}
