% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_maker.R
\name{get_obs_from_items}
\alias{get_obs_from_items}
\title{Search for observations by item}
\usage{
get_obs_from_items(observations, items, target = "all", key = "CODE")
}
\arguments{
\item{observations}{List of observations on which to do the search.}

\item{items}{Sought items.}

\item{target}{Condition for an observation to be extracted. One of \code{"all"}, \code{"any"}.
\describe{
  \item{\code{"all"}}{All the sought items must be part of an observation for this
                      observation to be extracted.}
  \item{\code{"any"}}{At least one of the sought items must be part of an observation for
                      this observation to be extracted.}
}}

\item{key}{Access key to the items in an observation.}
}
\value{
Subset of the list of observations that match the search criteria.
}
\description{
Extract the observations containing one or more sought items.
}
\seealso{
\code{\link{get_complex_obs}}, \code{\link{get_simple_obs}}, \code{\link{get_obs_from_info}}.
}
\author{
Gauthier Magnin
}
