% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_set.R
\name{get_simple_trx,TransactionSet-method}
\alias{get_simple_trx,TransactionSet-method}
\alias{get_simple_trx}
\title{Search for simple transactions}
\usage{
\S4method{get_simple_trx}{TransactionSet}(object)
}
\arguments{
\item{object}{S4 object of class \code{TransactionSet}.}
}
\value{
S4 object of class \code{TransactionSet} containing the subset of transactions that contain exactly
one item.
}
\description{
Extract the transactions containing exactly one item.
}
\examples{
get_simple_trx(TS_instance)

}
\seealso{
\code{\link{get_complex_trx}}, \code{\link{get_trx_from_items}}, \code{\link{get_trx_from_info}}.
}
\author{
Gauthier Magnin
}
