% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_analyzer.R
\name{get_trx_from_category,TransactionAnalyzer-method}
\alias{get_trx_from_category,TransactionAnalyzer-method}
\alias{get_trx_from_category}
\title{Search for transactions by category}
\usage{
\S4method{get_trx_from_category}{TransactionAnalyzer}(object, trx, category, value)
}
\arguments{
\item{object}{S4 object of class \code{TransactionAnalyzer}.}

\item{trx}{S4 object of class \code{TransactionSet}. Any subset of \code{object["transactions"]}.}

\item{category}{Name or number of the category on which to search (numbering according to the order
of the columns of \code{object["items_categories"]}.}

\item{value}{Sought value for the category specified by the argument \code{category}.}
}
\value{
Subset of the transaction set that match the search criteria.
}
\description{
Extract the transactions corresponding to a sought category value.
}
\examples{
get_trx_from_category(TA_instance, TA_instance["transactions"], "family", "Chrome")

}
\seealso{
About transactions: \code{\link{get_trx_from_items}}, \code{\link{get_trx_from_info}}, \code{\link{get_complex_trx}},
\code{\link{get_simple_trx}}.

About nodes and patterns: \code{\link{get_nodes}}, \code{\link{get_patterns}}, \code{\link{get_complexes}}, \code{\link{get_isolates}},
\code{\link{get_non_isolates}}.
}
\author{
Gauthier Magnin
}
