% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_set.R
\name{get_trx_from_items,TransactionSet-method}
\alias{get_trx_from_items,TransactionSet-method}
\alias{get_trx_from_items}
\title{Search for transactions by item}
\usage{
\S4method{get_trx_from_items}{TransactionSet}(object, items, presence = "all", as_indices = FALSE)
}
\arguments{
\item{object}{S4 object of class \code{TransactionSet}.}

\item{items}{Sought items.}

\item{presence}{Item presence condition for a transaction to be extracted.
One of \code{"all"}, \code{"any"}, \code{"exact"}, \code{"only"}.
\describe{
\item{\code{"all"}}{All the sought items must be part of a transaction for this transaction to be
extracted.}
\item{\code{"any"}}{At least one of the sought items must be part of a transaction for this transaction
to be extracted.}
\item{\code{"exact"}}{The item set contained in a transaction must be exactly the same as the sought
item set for this transaction to be extracted.}
\item{\code{"only"}}{A transaction must contain only the sought items (any of them) for this
transaction to be extracted.}
}}

\item{as_indices}{\code{TRUE} or \code{FALSE} whether to return transactions or only
their indices.}
}
\value{
S4 object of class \code{TransactionSet} containing the subset of
transactions that match the search criteria, or indices of these
transactions (according to the argument \code{as_indices}). If the given
transactions are named (and \code{as_indices} is \code{TRUE}), the returned indices
are named as well.
}
\description{
Extract the transactions containing one or more sought items.
}
\examples{
get_trx_from_items(TS_instance, items = c(25, 192), presence = "all")
get_trx_from_items(TS_instance, items = c(25, 192), presence = "any")
get_trx_from_items(TS_instance, items = c(25, 192), presence = "exact")

get_trx_from_items(TS_instance, items = c(192, 3146), presence = "any")
get_trx_from_items(TS_instance, items = c(192, 3146), presence = "only")

get_trx_from_items(TS_instance,
                   items = c(25, 192),
                   presence = "any",
                   as_indices = TRUE)

}
\seealso{
\code{\link{get_complex_trx}}, \code{\link{get_simple_trx}}, \code{\link{get_trx_from_info}}.
}
\author{
Gauthier Magnin
}
