% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_helper.R
\name{is_color}
\alias{is_color}
\title{Check color representation}
\usage{
is_color(x, int = TRUE)
}
\arguments{
\item{x}{Character or numeric vector.}

\item{int}{Logical indicating whether to consider integer values as valid representations.}
}
\value{
Logical vector giving for each value of \code{x} whether it is a valid color representation.
}
\description{
Check if a value is a valid color representation.
The three kinds of R color specifications are considered as valid representations (color name,
hexadecimal string and positive integer).
}
\note{
This function is not exported from the package.
}
\seealso{
Color names: \code{\link[grDevices:colors]{grDevices::colors}}, hexadecimal values: \code{\link[grDevices:rgb]{grDevices::rgb}},
graphics palette: \code{\link[grDevices:palette]{grDevices::palette}}.

Color to RGB conversion: \code{\link[grDevices:col2rgb]{grDevices::col2rgb}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
