% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{list_obs_per_year,SpectralAnalyzer-method}
\alias{list_obs_per_year,SpectralAnalyzer-method}
\title{Identifie les observations distinctes par année et calcule le nombre de recrutement
de chacune de ces observations.}
\usage{
\S4method{list_obs_per_year}{SpectralAnalyzer}(object)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}
}
\value{
Data.frame des observations distinctes, par année, et de leurs caractéristiques (longueur et poids).
}
\description{
La data.frame résultante est associée à l'attribut \code{nodes_per_year}.
}
\author{
Gauthier Magnin
}
