% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{list_obs_per_year,SpectralAnalyzer-method}
\alias{list_obs_per_year,SpectralAnalyzer-method}
\alias{list_obs_per_year}
\title{Enumeration of the observations per year}
\usage{
\S4method{list_obs_per_year}{SpectralAnalyzer}(object)
}
\arguments{
\item{object}{S4 object of class \code{SpectralAnalyzer}.}
}
\value{
Invisible. Matrix of the number of occurrences of each separate observation, per year.
 The lines correspond to the observations. The columns correspond to the years.
}
\description{
Identify the separate observations per year (considering only their respective items) and count their
 number of occurrences for each one.
The resulting matrix is assigned to the attribute \code{nodes_per_year} of \code{object}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
