% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{list_patterns_per_year,SpectralAnalyzer-method}
\alias{list_patterns_per_year,SpectralAnalyzer-method}
\alias{list_patterns_per_year}
\title{Enumeration of patterns per year}
\usage{
\S4method{list_patterns_per_year}{SpectralAnalyzer}(object)
}
\arguments{
\item{object}{S4 object of class \code{SpectralAnalyzer}.}
}
\value{
Invisible. Matrix of the weights of each pattern, per year.
 The lines correspond to the patterns. The column correspond to the years.
}
\description{
Count the number of occurrences of each pattern per year.
The resulting matrix is assigned to the attribute \code{patterns_per_year} of \code{object}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
