% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{list_patterns_per_year,SpectralAnalyzer-method}
\alias{list_patterns_per_year,SpectralAnalyzer-method}
\title{Compte le nombre d'apparition de chaque motif selon l'année.}
\usage{
\S4method{list_patterns_per_year}{SpectralAnalyzer}(object)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}
}
\value{
Matrice des poids de chaque motif dans laquelle les lignes correspondent aux motifs et les colonnes aux années.
}
\description{
La matrice résultante est associée à l'attribut \code{patterns_per_year}.
}
\author{
Gauthier Magnin
}
