% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{list_separate_obs,SpectralAnalyzer-method}
\alias{list_separate_obs,SpectralAnalyzer-method}
\title{Identifie les observations distinctes et calcule le nombre de recrutement
de chacune de ces observations.}
\usage{
\S4method{list_separate_obs}{SpectralAnalyzer}(object)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}
}
\value{
Data.frame des observations distinctes et de leurs caractéristiques (longueur et poids).
}
\description{
La data.frame résultante est associée à l'attribut \code{nodes}.
}
\author{
Gauthier Magnin
}
