% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{list_separate_obs,SpectralAnalyzer-method}
\alias{list_separate_obs,SpectralAnalyzer-method}
\alias{list_separate_obs}
\title{Enumeration of nodes}
\usage{
\S4method{list_separate_obs}{SpectralAnalyzer}(object)
}
\arguments{
\item{object}{S4 object of class \code{SpectralAnalyzer}.}
}
\value{
Invisible. Data frame of the separate observations and their characteristics (length and
 weight).
}
\description{
Identify the separate observations (considering only their respective items) and compute their size
 and number of occurrences.
The resulting data frame is assigned to the attribute \code{nodes} of \code{object}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
