% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{list_separate_patterns,SpectralAnalyzer-method}
\alias{list_separate_patterns,SpectralAnalyzer-method}
\title{Identifie l'ensemble des motifs distincts générés à partir des observations.}
\usage{
\S4method{list_separate_patterns}{SpectralAnalyzer}(
  object,
  target,
  count = 1,
  min_length = 1,
  max_length = Inf
)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{target}{Type de motifs à énumérer.
Choix parmi \code{c("frequent itemsets", "closed frequent itemsets", "maximally frequent itemsets")}.}

\item{count}{Nombre minimal d'apparition d'un motif pour être considéré comme "fréquent", et par conséquent, conservé.}

\item{min_length}{Taille minimale qu'un motif doit avoir pour être conservé lors de l'énumération.}

\item{max_length}{Taille maximale qu'un motif doit avoir pour être conservé lors de l'énumération.
La valeur \code{Inf} par défaut correspond à une recherche de motifs sans limite maximale de taille.}
}
\value{
Data.frame dans laquelle une ligne est une association entre un motif et sa fréquence dans l'ensemble des observations.
}
\description{
La data.frame résultante est associée à l'attribut \code{patterns}.
}
\seealso{
\code{\link[arules]{APparameter}}
}
\author{
Gauthier Magnin
}
