% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{list_separate_patterns,SpectralAnalyzer-method}
\alias{list_separate_patterns,SpectralAnalyzer-method}
\alias{list_separate_patterns}
\title{Enumeration of patterns}
\usage{
\S4method{list_separate_patterns}{SpectralAnalyzer}(
  object,
  target,
  count = 1,
  min_length = 1,
  max_length = Inf
)
}
\arguments{
\item{object}{SpectralAnalyzer class object.}

\item{target}{Type of patterns to enumerate. One of \code{"frequent itemsets"},
\code{"closed frequent itemsets"}, \code{"maximally frequent itemsets"}.}

\item{count}{Minimum number of occurrences of a pattern to be considered as "frequent".}

\item{min_length}{Minimum number of items that a pattern must have to be kept when enumerating.}

\item{max_length}{Maximum number of items that a pattern must have to be kept when enumerating.
The default \code{Inf} corresponds to a pattern search without maximum size limit.}
}
\value{
Invisible. Data frame in which a line is an association between a pattern and its
 frequency in the set of observations.
}
\description{
Identify the patterns generated from the observations.
The resulting data frame is assigned to the attribute \code{patterns} of \code{object}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
