% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_maker.R
\name{make_INRS_observations}
\alias{make_INRS_observations}
\title{Grouping samples}
\usage{
make_INRS_observations(
  measures,
  mode,
  work_situations = NULL,
  variable_names = NULL,
  additional = NULL,
  unique_values = TRUE
)
}
\arguments{
\item{measures}{Set of sampling measures. Data frame that must contain at least the following three
variables:
\describe{
  \item{\code{ID}}{Inspection identifier. The measures resulting from the same inspection must
        have the same inspection identifier.}
  \item{\code{YEAR}}{Year of inspection.}
  \item{\code{CODE}}{Code identifying the substance being sampled.}
}}

\item{mode}{Mode of creation of observations. One of \code{1}, \code{2}, \code{3}.
\describe{
  \item{\code{1}}{Construction of observations from the situations described by \code{work_situations}
   and existing situations from combinations of values of variables defined by \code{variable_names}
   that would not be described in \code{work_situations}.}
  \item{\code{2}}{Construction of observations only from the situations described by
   \code{work_situations}. Remaining samples are ignored.}
  \item{\code{3}}{Construction of observations only from combinations of values of variables
   defined by \code{variable_names}.}
}}

\item{work_situations}{Description of work situations. Data frame that must contain the values
associated with certain variables from \code{measures} forming one work situation.
A \code{WS_ID} (\strong{W}ork \strong{S}ituation \strong{ID}entifier) column must make possible
to identify the different associations of values considered to be the same work situation.}

\item{variable_names}{Names of variables to be taken into account to consider a work situation,
ie. vector of names of variables contained in \code{measures} whose different combinations of
values form the different work situations not described in \code{work_situations}.}

\item{additional}{Names of information to keep when creating observations: vector included in
\code{colnames(measures)}.}

\item{unique_values}{logical or character.
\itemize{
  \item{If \code{TRUE}, simplification of the values associated with the variables defined in
        \code{additional} in order to remove duplicates.}
  \item{If \code{FALSE}, keep the duplicates and the correspondence between the values of
        these variables for the same sample (as many values as there are samples).}
  \item{Otherwise, vector of variable names included in \code{colnames(measures)} for which the
        removal of duplicates must be performed.}
}}
}
\value{
List of observations identified. Each observation is a list containing:
 \describe{
   \item{\code{CODE}}{The codes identifying the substances grouped together.}
   \item{\code{YEAR}}{The year of the inspection from which the samples are grouped.}
   \item{\code{...}}{Data related to the samples grouped, concerning the variables defined in
                     the argument \code{additional}.}
 }
}
\description{
Grouping samples in the form of observations.
An observation corresponds to a set of substances found in the same work situation during the same
 inspection.
Specific situations are defined by the argument \code{work_situations}.
Non-specific situations are described by existing combinations of variables whose names are defnied
 by the argument \code{variable_names}.
}
\examples{
to_keep <- c("CODE", "NAME", "ACTIVITY", "JOB.TITLE", "SAMPLE.ID")
ws <- data.frame(WS_ID = c(1, 2, 2, 3, 3),
                 JOB.TITLE = c(44121004, 44142001, 44132032, 44132019, 44132030),
                 JOB.TASK = c("A5440", "A6410", "A5110", "A5260", "A5240"),
                 stringsAsFactors = FALSE)
ws_vars <- c("JOB.TITLE", "JOB.TASK")

## Make observations from 3 specific work situations and as many other
## situations as JOB.TITLE, JOB.TASK pairs
obs_1 <- make_INRS_observations(oedb_sample, mode = 1,
                                work_situations = ws,
                                variable_names = ws_vars,
                                additional = to_keep)

## Make observations from specific work situations only
obs_2 <- make_INRS_observations(oedb_sample, mode = 2,
                                work_situations = ws,
                                additional = to_keep)

## Make observations where each pair of JOB.TITLE, JOB.TASK is one situation
obs_3 <- make_INRS_observations(oedb_sample, mode = 3,
                                variable_names = ws_vars,
                                additional = to_keep)

}
\seealso{
\code{\link{make_observations}}.
}
\author{
Gauthier Magnin
}
