% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations_maker.R
\name{make_INRS_observations}
\alias{make_INRS_observations}
\title{Regroupement de prélèvements}
\usage{
make_INRS_observations(
  measures,
  mode,
  work_situations = NULL,
  variable_names = NULL,
  additional = NULL,
  unique_values = TRUE
)
}
\arguments{
\item{measures}{Ensemble de mesures de prélèvements. Data frame devant contenir au moins les
trois variables suivantes :
\describe{
  \item{\code{ID}}{Identifiant de l'intervention. Les mesures issues d'une même intervention
        doivent avoir le même identifiant d'intervention.}
  \item{\code{YEAR}}{Année de réalisation de l'intervention.}
  \item{\code{CODE}}{Code identifiant la substance faisant l'objet du prélèvement.}
}}

\item{mode}{Mode de création des observations. Choix parmi \code{1}, \code{2}, \code{3}.
\describe{
 \item{\code{1}}{Construction d'observations à partir des situations décrites par \code{work_situations}
  et des situations existantes à partir de combinaisons de \code{variable_names} qui ne seraient pas
  décrites dans \code{work_situations}.}
\item{\code{2}}{Construction d'observations uniquement à partir des situations décrites par \code{work_situations}.
  Les prélèvements restants sont ignorés.}
\item{\code{3}}{Construction d'observations uniquement à partir des combinaisons de \code{variable_names}.}
}}

\item{work_situations}{Description de situations de travail. Data frame devant contenir les
valeurs associées à certaines variables de \code{measures} formant une situation de travail.
Une colonne \code{WS_ID} (\strong{W}ork \strong{S}ituation \strong{ID}entifier) doit permettre
d'identifier les différentes associations de valeurs considérées comme une même situation de travail.}

\item{variable_names}{Noms des variables à prendre en compte pour considérer une situation de travail,
c'est-à-dire, vecteur de noms de variables contenues dans \code{measures} dont les différentes
combinaisons forment les différentes situations de travail non décrites dans \code{work_situations}.}

\item{additional}{Noms d'informations à conserver lors de la constitution des observations,
c'est-à-dire, vecteur de noms de variables contenues dans \code{measures}.}

\item{unique_values}{logical ou character.
\itemize{
  \item{Si \code{TRUE}, simplification des valeurs associées aux variables définies dans
        \code{additional} de manière à supprimer les doublons.}
  \item{Si \code{FALSE}, conserve les doublons et la correspondance entre les valeurs de ces
        variables pour un même prélèvement (autant de valeurs que de prélèvements).}
  \item{Sinon, vecteur des variables inclus dans \code{colnames(measures)} pour lesquelles la
        suppression des doublons doit être réalisée.}
}}
}
\value{
Liste des observations identifiées. Chaque observation est une liste contenant :
   \describe{
     \item{CODE}{Les codes identifiant les substances regroupées.}
     \item{YEAR}{L'année de réalisation de l'intervention de laquelle les prélèvements sont regroupées.}
     \item{ID}{L'identifiant dede l'intervention de laquelle les prélèvements sont regroupées.}
     \item{...}{Les données associées aux prélèvements, concernant les variables définies
                dans l'argument \code{additional}.}
   }
}
\description{
Regroupe des prélèvements sous forme d'observations.
Une observation correspond à un ensemble de substances retrouvées dans une même situation de travail
 lors d'une même intervention.
Les situations spécifiques sont définies par l'argument \code{work_situations}.
Les situations non spécifiques sont décrites par les combinaisons existantes des variables dont les
 noms sont définis par l'argument \code{variable_names}.
}
\seealso{
\code{\link{make_observations}}.
}
\author{
Gauthier Magnin
}
