% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_maker.R
\name{make_obs_from_unspecified_ws}
\alias{make_obs_from_unspecified_ws}
\title{Regroupement de prélèvements selon des situations de travail non spécifiques}
\usage{
make_obs_from_unspecified_ws(
  measures,
  variable_names,
  additional = NULL,
  unique_values = TRUE
)
}
\arguments{
\item{measures}{Ensemble de mesures de prélèvements. Data frame devant contenir au moins les
deux variables suivantes :
\describe{
  \item{\code{YEAR}}{Année de réalisation de l'intervention.}
  \item{\code{CODE}}{Code identifiant la substance faisant l'objet du prélèvement.}
}}

\item{variable_names}{Noms des variables à prendre en compte pour considérer une situation de travail,
c'est-à-dire, vecteur de noms de variables contenues dans \code{measures} dont les différentes
combinaisons forment les différentes situations de travail.}

\item{additional}{Noms d'informations à conserver lors de la constitution des observations,
c'est-à-dire, vecteur de noms de variables contenues dans \code{measures}.}

\item{unique_values}{logical ou character.
\itemize{
  \item{Si \code{TRUE}, simplification des valeurs associées aux variables définies dans
        \code{additional} de manière à supprimer les doublons.}
  \item{Si \code{FALSE}, conserve les doublons et la correspondance entre les valeurs de ces
        variables pour un même prélèvement (autant de valeurs que de prélèvements).}
  \item{Sinon, vecteur des variables inclus dans \code{colnames(measures)} pour lesquelles la
        suppression des doublons doit être réalisée.}
}}
}
\value{
Liste des observations identifiées. Chaque observation est une liste contenant :
   \describe{
     \item{CODE}{Les codes identifiant les substances regroupées.}
     \item{YEAR}{L'année de réalisation de l'intervention de laquelle les prélèvements sont regroupées.}
     \item{...}{Les données associées aux prélèvements, concernant les variables définies
                dans l'argument \code{additional}.}
   }
}
\description{
Regroupe des prélèvements sous forme d'observations (make observations from unspecified Work Situation).
Une observation correspond à un ensemble de substances retrouvées dans une même situation de travail.
Ces situations sont décrites par les combinaisons de valeurs possibles des variables dont les noms
 sont définis en paramètre.
}
\seealso{
\code{\link{make_observations}}, \code{\link{make_obs_from_ws}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
