% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_maker.R
\name{make_observations}
\alias{make_observations}
\title{Grouping data}
\usage{
make_observations(data, by, additional = NULL, unique_values = TRUE)
}
\arguments{
\item{data}{Data frame containing a set of data to group together.}

\item{by}{Variable names included in \code{colnames(data)} used to group the data.
Each combination of values of these different variables generates one observation.}

\item{additional}{Names of information to keep when creating observations: vector included in
\code{colnames(data)}.}

\item{unique_values}{logical or character.
\itemize{
  \item{If \code{TRUE}, simplification of the values associated with the variables defined in
        \code{additional} in order to remove duplicates.}
  \item{If \code{FALSE}, keep the duplicates and the correspondence between the values of
        these variables for the same line in \code{data} (as many values as there are rows
        grouped).}
  \item{Otherwise, vector of variable names included in \code{colnames(data)} for which the
        removal of duplicates must be performed.}
}}
}
\value{
List of observations identified. Each observation is named by the combination of values
 of the variables named in \code{by} that generated it. Each observation is a list whose elements
 correspond to the values of the variables named in \code{by} and \code{additional} which have
 been grouped together.
}
\description{
Grouping data in the form of observations.
}
\details{
Variables of type \code{factor} are cast to \code{character}.
}
\seealso{
\code{\link{make_INRS_observations}} for the construction of well-targeted observations
 specific to the field of occupational exposure in relation to inspections and similar work
 situations.
}
\author{
Gauthier Magnin
}
