% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations_maker.R
\name{make_observations}
\alias{make_observations}
\title{Regroupement de prélèvements}
\usage{
make_observations(
  measures,
  mode,
  work_situations = NULL,
  variable_names = NULL,
  additional = NULL
)
}
\arguments{
\item{measures}{Ensemble de mesures de prélèvements.}

\item{mode}{Mode de création des observations. Choix parmi \code{1}, \code{2}, \code{3}.
\describe{
 \item{\code{1}}{Construction d'observations à partir des situations décrites par \code{work_situations}
  et des situations existantes à partir de combinaisons de \code{variable_names} qui ne seraient pas
  décrites dans \code{work_situations}.}
\item{\code{2}}{Construction d'observations uniquement à partir des situations décrites par \code{work_situations}.
  Les prélèvements restants sont ignorés.}
\item{\code{3}}{Construction d'observations uniquement à partir des combinaisons de \code{variable_names}.}
}}

\item{work_situations}{Description de situations de travail.}

\item{variable_names}{Noms des variables à prendre en compte pour considérer une situation de travail.}

\item{additional}{Noms des informations à conserver lors de la constitution des observations.}
}
\value{
Liste des observations identifiées.
}
\description{
Regroupe des prélèvements sous forme d'observations.
Une observation correspond à un ensemble de substances retrouvées dans une même situation de travail
 lors d'une même intervention.
Les situations connues sont décrites par l'argument \code{work_situations}.
Les situations non connues sont décrites par les combinaisons existantes des variables dont les noms sont
 définis par l'argument \code{variable_names}.
}
\author{
Gauthier Magnin
}
