% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_maker.R
\name{make_observations}
\alias{make_observations}
\title{Regroupement de données}
\usage{
make_observations(data, by, additional = NULL, unique_values = TRUE)
}
\arguments{
\item{data}{Data frame contenant un ensemble de données à regrouper.}

\item{by}{Noms de variables inclus dans \code{colnames(data)} servant à regrouper les données
de \code{data}.
Chaque combinaison de valeurs de ces différentes variables engendre une observation.}

\item{additional}{Noms d'informations à conserver lors de la constitution des observations
(vecteur inclus dans \code{colnames(data)}).}

\item{unique_values}{logical ou character.
\itemize{
  \item{Si \code{TRUE}, simplification des valeurs associées aux variables définies dans
        \code{additional} de manière à supprimer les doublons.}
  \item{Si \code{FALSE}, conserve les doublons et la correspondance entre les valeurs de ces
        variables pour une même ligne de \code{data} (autant de valeurs que de lignes
        regroupées).}
  \item{Sinon, vecteur des variables inclus dans \code{colnames(data)} pour lesquelles la
        suppression des doublons doit être réalisée.}
}}
}
\value{
Liste des observations identifiées. Chaque observation est nommée par la combinaison de
 valeurs des variables nommées dans \code{by} qui l'a générée. Chaque observation est une liste
 dont les éléments correspondent aux valeurs des variables nommées dans \code{by} et
 \code{additional} qui ont été regroupées
}
\description{
Regroupe des données sous forme d'observations.
}
\details{
Les variables de type \code{factor} sont converties en type \code{character}.
}
\seealso{
\code{\link{make_INRS_observations}} pour la construction d'observations bien ciblées
 par rapport à des interventions et à des situations de travail similaires.
}
\author{
Gauthier Magnin
}
