% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixer.R
\name{mcr_chart}
\alias{mcr_chart}
\title{MCR approach scatter plot}
\usage{
mcr_chart(values, references,
          thq_col = NULL,
          regions = FALSE,
          regions_col = c("#b3cde3", "#edf8fb", "#8c96c6", "#88419d"),
          regions_alpha = 0.2,
          regions_lab = !regions,
          regression = FALSE,
          log_transform = TRUE,
          plot = TRUE)
mcr_chart(hi, mcr, thq,
          thq_col = NULL,
          regions = FALSE,
          regions_col = c("#b3cde3", "#edf8fb", "#8c96c6", "#88419d"),
          regions_alpha = 0.2,
          regions_lab = !regions,
          regression = FALSE,
          log_transform = TRUE,
          plot = TRUE)
}
\arguments{
\item{values}{Numeric named matrix or list of numeric named vectors. Vectors of values for which the
chart is to be plotted.}

\item{references}{Numeric vector or list of numeric vectors. Reference values associated with the
\code{values}. See 'Details' to know the way it is associated with \code{values}.}

\item{hi}{Numeric vector. \strong{H}azard \strong{i}ndexes for which the chart is to be plotted.}

\item{mcr}{Numeric vector. \strong{M}aximum \strong{c}umulative \strong{r}atios associated with the hazard indexes
\code{hi}.}

\item{thq}{Numeric named vector or list of numeric named vectors. \strong{T}op \strong{h}azard \strong{q}uotients
associated with the hazard indexes \code{hi}. If list, only the first named value of each element of the
list is considered.}

\item{thq_col}{Character named vector. Colors to assign to the \strong{t}op \strong{h}azard \strong{q}uotients
elements.}

\item{regions}{If \code{TRUE}, the regions corresponding to the MIAT groups are filled with the colors
defined by \code{regions_col}.}

\item{regions_col}{Character vector of length 4. Define the colors for the regions of the MIAT groups
(in order: I, II, IIIA and IIIB).}

\item{regions_alpha}{Value between 0 and 1. Opacity of the regions filled with \code{regions_col}.}

\item{regions_lab}{Logical value or vector of length 4. Define if labels for the MIAT groups should
be displayed (in order: I, II, IIIA, IIIB). \code{TRUE} and \code{FALSE} are special values for all \code{TRUE} or
all \code{FALSE}.}

\item{regression}{If \code{TRUE}, the linear regression between the X and Y coordinates of the points
is represented with 95\% confidence interval.}

\item{log_transform}{If \code{TRUE}, the log version of the chart is plotted (i.e.
\mjeqn{log_{10}(HI)}{log10(HI)} versus \mjeqn{log_{10}(MCR - 1)}{log10(MCR - 1)}). If \code{FALSE},
the standard version of the chart is plotted (i.e. \eqn{HI} versus \eqn{MCR}).}

\item{plot}{If \code{FALSE} and the returned object is assigned, the chart is not plotted.
Otherwise, the chart is plotted in the active graphics device.}
}
\value{
Chart created with the \code{ggplot2} package (invisible) or \code{NULL} if no points can be plotted
(see 'Details').
}
\description{
Plot \mjeqn{log_{10}(HI)}{log10(HI)} versus \mjeqn{log_{10}(MCR - 1)}{log10(MCR - 1)} with the
percentage of the reciprocal of the maximum cumulative ratio, the elements producing the top hazard
quotients and the associated MIAT groups.
}
\details{
If \code{values} is a matrix, the reference values are applied once on each set of values, i.e. on each row.
Therefore, there must be one reference value for each column of the matrix.

If \code{values} is a list, the reference values can be a vector of named values or a list. In this case,
if \code{references} is a vector, there must be one reference for each name present in \code{values}.
Otherwise, \code{references} is a list of vectors having the same lengths as those present in \code{values}
so that \code{values} and \code{references} can be matched.

The chart being plotted with the \code{ggplot2} package, it can be modified or completed afterwards using
\code{\link[ggplot2:last_plot]{ggplot2::last_plot}} or the returned object.

Color specification can be done using the R predefined color names or hexadecimal values.

In the standard version of the chart, the grey area represents the region in which no point can be
plotted because \eqn{MCR} cannot be lower than 1. In the log version, such a region does not exist.
However, in the latter, points having \eqn{MCR} equal to 1 have an ordinate equal to \code{-Inf} and
therefore cannot be plotted and generate a warning message.

Points that are on the boundaries of the region of the group I belong to this group.
Points that are on the boundaries of the region of the group II belong to this group.
Points that are on the boundary between the regions of the groups IIIA and IIIB belong to the
group IIIB.

Arguments \code{values} and \code{references} are used to compute the hazard quotients and the hazard indexes
before searching for the top and maximum hazard quotients, computing the maximum cumulative ratios
then plot the chart. Thus, call the function with the arguments \code{hi}, \code{mcr} and \code{thq} is faster
(if they are already computed).

\loadmathjax
The mixtures are assigned to the groups according the following conditions:
\itemize{
\item Group I: \mjeqn{MHQ_i \ge 1}{MHQ_i >= 1}
\item Group II: \mjeqn{MHQ_i < 1, HI_i \le 1}{MHQ_i < 1, HI_i <= 1}
\item Group IIIA: \mjeqn{MHQ_i < 1, HI_i > 1, MCR_i < 2}{MHQ_i < 1, HI_i > 1, MCR_i < 2}
\item Group IIIB: \mjeqn{MHQ_i < 1, HI_i > 1, MCR_i \ge 2}{MHQ_i < 1, HI_i > 1, MCR_i >= 2}
}

The maximum cumulative ratio of the vector \eqn{i} is given by:
\mjdeqn{MCR_i = \frac{HI_i}{MHQ_i}}{MCR_i = HI_i / MHQ_i}
where \eqn{HI} denotes the hazard index and \eqn{MHQ} denotes the maximum hazard quotient.

The reciprocal of the maximum cumulative ratio of the vector \eqn{i} is given by:
\mjdeqn{Reciprocal~of~MCR_i = \frac{1}{MCR_i} = \frac{MHQ_i}{HI_i}}{Reciprocal of MCR_i = 1 / MCR_i = MHQ_i / HI_i}
where \eqn{MCR}, \eqn{MHQ} and {HI} denotes the maximum cumulative ratio, the maximum hazard
quotient and the hazard index respectively.

The hazard index of the vector \eqn{i} is given by:
\mjdeqn{HI_i = \sum_{j = 1}^n HQ_{i,j}}{HI_i = sum(HQ_ij) from j = 1 to n}
where \eqn{HQ} denotes the hazard quotients and \eqn{n} denotes the number of hazard quotients.

The maximum hazard quotient of the vector \eqn{i} is given by:
\mjdeqn{MHQ_i = HQ_{M,i} = \max_{j \in \lbrace 1,...,n\rbrace} HQ_{i,j}}{MHQ_i = HQ_Mi = max HQ_i}
where \eqn{HQ} denotes the hazard quotients and \eqn{n} denotes the number of hazard quotients.

The hazard quotient of the value \eqn{j} in the vector \eqn{i} is given by:
\mjdeqn{HQ_{i,j} = \frac{V_{i,j}}{RV_j}}{HQ_ij = V_ij / RV_j}
where \eqn{V} denotes the \code{values} and \eqn{RV} denotes the \code{references}.

If several values are equal to the maximum hazard quotient, the name retained as the top hazard
quotient is the first one considering the given order.
}
\note{
Due to the multiple possible usages, all arguments except \code{values} and \code{references} must be explicitly
named in the function call.
}
\examples{
## Creating a matrix of 50*5 values and one reference value for each of the 5
## elements (A, B, C, D and E).
v <- matrix(sample(seq(0.1, 1.1, by = 0.1), 250, replace = TRUE),
            nrow = 50, byrow = TRUE,
            dimnames = list(paste0("set.", 1:50), LETTERS[1:5]))
r <- sample(seq(1,5), 5, replace = TRUE)

mcr_chart(v, r, regions = TRUE)
mcr_chart(v, r,
          thq_col = c(A = "blue", B = "green", C = "red", D = "yellow3", E = "grey"),
          regions = FALSE,
          regions_lab = c(TRUE, TRUE, FALSE, TRUE),
          regression = TRUE)
mcr_chart(hi = hazard_index(v, r),
          mcr = maximum_cumulative_ratio(v, r),
          thq = top_hazard_quotient(v, r))

mcr_chart(v, r, regions = TRUE, log_transform = FALSE)
mcr_chart(v, r,
          thq_col = c(A = "blue", B = "green", C = "red", D = "yellow3", E = "grey"),
          regions = FALSE,
          regions_lab = c(TRUE, TRUE, FALSE, TRUE),
          regression = TRUE,
          log_transform = FALSE)
mcr_chart(hi = hazard_index(v, r),
          mcr = maximum_cumulative_ratio(v, r),
          thq = top_hazard_quotient(v, r),
          log_transform = FALSE)

## MCR chart on list
mcr_chart(values = list(set.1 = c(A = 0.1, B = 0.5),
                        set.2 = c(A = 0.2),
                        set.3 = c(B = 0.3, C = 0.4)),
          references = c(A = 1, B = 2, C = 3),
          log_transform = TRUE)
mcr_chart(values = list(set.1 = c(A = 0.1, B = 0.5),
                        set.2 = c(A = 0.2),
                        set.3 = c(B = 0.3, C = 0.4)),
          references = list(c(1, 2),
                            1,
                            c(2, 3)),
          log_transform = FALSE)

}
\references{
Reyes JM, Price PS (2018).
An analysis of cumulative risks based on biomonitoring data for six phthalates using the Maximum Cumulative Ratio.
\emph{Environment International}, 112, 77-84.
\url{https://doi.org/10.1016/j.envint.2017.12.008}.

De Brouwere K, et al. (2014).
Application of the maximum cumulative ratio (MCR) as a screening tool for the evaluation of mixtures in residential indoor air.
\emph{The Science of the Total Environment}, 479-480, 267-276.
\url{https://doi.org/10.1016/j.scitotenv.2014.01.083}.
}
\seealso{
\code{\link{mcr_summary}}, \code{\link{maximum_cumulative_ratio}}, \code{\link{hazard_index}}, \code{\link{reciprocal_of_mcr}},
\code{\link{top_hazard_quotient}}, \code{\link{classify_mixture}}.
}
\author{
Gauthier Magnin
}
