% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{network_density,SpectralAnalyzer-method}
\alias{network_density,SpectralAnalyzer-method}
\alias{network_density}
\title{Network density}
\usage{
\S4method{network_density}{SpectralAnalyzer}(object, links)
}
\arguments{
\item{object}{S4 object of class \code{SpectralAnalyzer}.}

\item{links}{Data frame of the links (or edges) of a spectrosome graph.}
}
\value{
Density of the network.
}
\description{
Compute the density of a graph as the ratio between the number of links identified and the
 maximum number of possible links (i.e. if all the vertices of the graph were connected to each other).
}
\examples{
spectrosome <- spectrosome_chart(SA_instance, SA_instance["patterns"][1:15, ])
network_density(SA_instance, spectrosome[["edges"]])

}
\author{
Gauthier Magnin
}
