% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{network_density,SpectralAnalyzer-method}
\alias{network_density,SpectralAnalyzer-method}
\alias{network_density}
\title{Network density}
\usage{
\S4method{network_density}{SpectralAnalyzer}(object, links)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{links}{Data frame of the links (or edges) of a spectrosome graph.}
}
\value{
Density of the network.
}
\description{
Compute the density of the graph as the ratio between the number of links identified and
 the maximum number of possible links.
}
\examples{
spectrosome <- spectrosome_chart(SA_instance, SA_instance["patterns"][1:15, ])
network_density(SA_instance, spectrosome[["edges"]])

}
\author{
Gauthier Magnin
}
