% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oedb_sample}
\alias{oedb_sample}
\title{Example of occupational exposure data}
\format{
A data frame with 131 rows and 9 variables:
\describe{
\item{\code{ID}}{\emph{Character}. Identifier of an inspection in a specific
business establishment.}
\item{\code{YEAR}}{\emph{Integer}. Year of the inspection.}
\item{\code{CODE}}{\emph{Integer}. Code identifying the substance sampled.}
\item{\code{NAME}}{\emph{Character}. Name of the substance sampled.}
\item{\code{SAMPLE.ID}}{\emph{Integer}. Identifier of the sample. Several samples
may refer to the same inspection.}
\item{\code{ACTIVITY}}{\emph{Character}. Economic activity of the business
establishment inspected.}
\item{\code{JOB.TITLE}}{\emph{Integer}. Specific profession inspected.}
\item{\code{JOB.TASK}}{\emph{Character}. Specific job task inspected.}
\item{\code{CONCENTRATION}}{\emph{Numeric}. Measured concentration value.}
}
}
\usage{
oedb_sample
}
\description{
An illustrative dataset containing identified substances in specific
situations.
}
\keyword{datasets}
