% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{plot_spectrum_chart,SpectralAnalyzer-method}
\alias{plot_spectrum_chart,SpectralAnalyzer-method}
\alias{plot_spectrum_chart}
\title{Pattern spectrum plotting}
\usage{
\S4method{plot_spectrum_chart}{SpectralAnalyzer}(
  object,
  patterns_characteristics,
  weights_by_node_type,
  title = "Spectrum of patterns"
)
}
\arguments{
\item{object}{SpectralAnalyzer class object.}

\item{patterns_characteristics}{Patterns (and their characteristics) whose spectrum is to be plotted.}

\item{weights_by_node_type}{Data frame containing for each pattern, its wieght in complexe nodes
and its weight in simple nodes.}

\item{title}{Chart title.}
}
\description{
Plot a spectrum chart.
}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
            The spectrosome of occupational health problems. \emph{PLoS ONE} 13(1): e0190196.
            \url{https://doi.org/10.1371/journal.pone.0190196}.
}
\seealso{
\code{\link{spectrum_chart}}, \code{\link{compute_pattern_distribution_in_nodes}}.
}
\author{
Delphine Bosson-Rieutort, Gauthier Magnin
}
\keyword{internal}
