% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{plot_spectrum_chart,SpectralAnalyzer-method}
\alias{plot_spectrum_chart,SpectralAnalyzer-method}
\alias{plot_spectrum_chart}
\title{Spectre des motifs}
\usage{
\S4method{plot_spectrum_chart}{SpectralAnalyzer}(
  object,
  patterns_characteristics,
  weights_by_node_type,
  title = "Spectrum of patterns"
)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{patterns_characteristics}{Ensemble des caractéristiques des motifs dont le spectre est à tracer.}

\item{weights_by_node_type}{Data frame contenant, pour chaque motif, son poids dans des nœuds complexes
et son poids dans des nœuds simples.}

\item{title}{Titre du graphique.}
}
\description{
Dessine un graphique de type spectre.
}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
            \emph{The spectrosome of occupational health problems}. PLoS ONE 13(1): e0190196.
            \url{https://doi.org/10.1371/journal.pone.0190196}.
}
\seealso{
\code{\link{spectrum_chart}}, \code{\link{compute_pattern_distribution_in_nodes}}.
}
\author{
Delphine Bosson-Rieutort, Gauthier Magnin
}
\keyword{internal}
