% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pluralize}
\alias{pluralize}
\title{Pluzalize a character string}
\usage{
pluralize(x, var)
}
\arguments{
\item{x}{Character string to be pluralized.}

\item{var}{Variable to consider whether to add an "s" to the string.}
}
\value{
The character string corresponding to the argument \code{x}, ending with an "s" if \code{length(var) > 1}.
}
\description{
Add an "s" at the end of the string if the given variable has a length greater than \eqn{1}.
}
\note{
This function is not exported from the package.
}
\seealso{
\code{\link{cap}}, \code{\link{first_characters}}, \code{\link{substr2}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
