% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{reset,SpectralAnalyzer-method}
\alias{reset,SpectralAnalyzer-method}
\alias{reset}
\title{Partial reset of a spectral analyzer}
\usage{
\S4method{reset}{SpectralAnalyzer}(object, from = 1)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{from}{Step from which to recompute the attributes.}
}
\description{
Redefine the attributes of a spectral analyzer from a specific step.
}
\examples{
## Change one attribute (for instance, the pattern enumeration target) and
## enumerate the patterns again for a new analysis
SA_instance["target"] <- "frequent itemsets"
reset(SA_instance, from = 5)

}
\author{
Gauthier Magnin
}
