% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{reset,SpectralAnalyzer-method}
\alias{reset,SpectralAnalyzer-method}
\alias{reset}
\title{Partial reset of a spectral analyzer}
\usage{
\S4method{reset}{SpectralAnalyzer}(object, from = 1, verbose = TRUE)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{from}{Step from which to recompute the attributes.}

\item{verbose}{Logical value indicating whether to report progress.}
}
\description{
Reset the attributes of a spectral analyzer from a specific step.
}
\details{
The steps for initializing a spectral analyzer are:
\enumerate{
  \item{Enumeration of the observations per year.}
  \item{Enumeration of the nodes and calculation of the number of occurrences.}
  \item{Counting links between nodes.}
  \item{Elaboration of links between nodes.}
  \item{Mining for itemsets.}
  \item{Linking nodes to patterns.}
  \item{Characterization of patterns per year.}
  \item{Computation of pattern characteristics.}
  \item{Counting links between patterns.}
  \item{Elaboration of links between patterns.}
}
}
\note{
The following steps may be quite long (depending on the case):
 \itemize{
   \item{Step 4: elaboration of links between nodes (depending on the amount of data).}
   \item{Steps 5 and 6: mining for itemsets and linking nodes to patterns (depending on the
     mining parameters and the amount of data).}
   \item{Step 10: elaboration of links between patterns (depending on the amount of patterns,
     resulting from the mining parameters and the amount of data).}
 }
}
\examples{
## Change one attribute (for instance, the pattern enumeration target) and
## enumerate the patterns again for a new analysis
SA_instance["target"] <- "frequent itemsets"
reset(SA_instance, from = 5)

}
\seealso{
The \code{\link{SpectralAnalyzer}} constructor: \code{\link{spectral.analyzer}}.

Initialization: \code{\link{init}}, \code{\link{is_init}}.
}
\author{
Gauthier Magnin
}
