% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{search_links,SpectralAnalyzer-method}
\alias{search_links,SpectralAnalyzer-method}
\title{Identifie les liens selon les éléments en commun, entre les noeuds ou les motifs.}
\usage{
\S4method{search_links}{SpectralAnalyzer}(object, entities)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{entities}{Type d'élément pour lequel compter les liens (noeuds ou motifs).
Choix parmi \code{c("nodes", "patterns")}.}
}
\value{
Data.frame associant certaines caractéristiques aux paires de noeuds ou motifs liés.
}
\description{
La data.frame résultante est associée à l'attribut \code{nodes_links} ou \code{patterns_links} respectivement.
}
\author{
Delphine Bosson-Rieutort

Gauthier Magnin
}
