% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{search_links,SpectralAnalyzer-method}
\alias{search_links,SpectralAnalyzer-method}
\alias{search_links}
\title{Elaboration of links}
\usage{
\S4method{search_links}{SpectralAnalyzer}(object, entities)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{entities}{Type of entities for which to elaborate links (nodes or patterns).
\code{SpectralAnalyzer.NODES} or \code{SpectralAnalyzer.PATTERNS}.}
}
\value{
Invisible. Data frame detailing the links between pairs of nodes or patterns.
 Isolated nodes or patterns (i.e. unrelated to any other entity) appear at the bottom of the data
 frame.
}
\description{
Identify the links according to items in common between nodes or patterns.
The resulting data frame is assigned respectively to the attribute \code{nodes_links} or
 \code{patterns_links} of \code{object}.
}
\author{
Delphine Bosson-Rieutort, Gauthier Magnin
}
\keyword{internal}
