% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_helper.R
\name{shadowtext}
\alias{shadowtext}
\title{Text shading}
\usage{
shadowtext(
  x,
  y = NULL,
  labels,
  col = "black",
  bg = "white",
  theta = seq(0, 2 * pi, length.out = 32),
  r = 0.1,
  ...
)
}
\arguments{
\item{x, y}{Coordinates where the text \code{labels} should be written.}

\item{labels}{Text to be written.}

\item{col}{Text color.}

\item{bg}{Outline color.}

\item{theta}{Angles to be used to create the outline.}

\item{r}{Outline size.}
}
\description{
Draw text with an outline on a plot.
The outline is created by first drawing the text with the second color serveral times
 and with slight offsets.
}
\note{
This function is not exported from the package.
}
\references{
Greg Snow. Stack Overflow topic:
 \href{https://stackoverflow.com/questions/25631216/r-plots-is-there-any-way-to-draw-border-shadow-or-buffer-around-text-labels}{Any way to draw border, shadow or buffer around text labels}.
}
\keyword{internal}
