% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{spectral.analyzer}
\alias{spectral.analyzer}
\title{Constructeur d'objet SpectralAnalyzer.}
\usage{
spectral.analyzer(
  observations,
  items = NULL,
  target = "closed frequent itemsets",
  count = 1,
  min_length = 1,
  max_length = Inf,
  status_limit = 2
)
}
\arguments{
\item{observations}{Liste des éléments retrouvés pour chaque observation.
Liste sous la forme \code{list( list( CODE = character(), NAME = character(), YEAR = numeric ) )}.
Chaque observation est alors une liste sous la forme \code{list( CODE = character(), NAME = character(), YEAR = numeric )}
où \code{CODE} et \code{NAME} sont associés pour identifier un élément.}

\item{items}{Data.frame associant une ou plusieurs catégories à un élément (\code{items$item}).
La valeur \code{NULL} par défaut précise qu'aucune catégorie n'est définie.}

\item{target}{Type de motifs à énumérer.
Choix parmi \code{c("frequent itemsets", "closed frequent itemsets", "maximally frequent itemsets")}.
Par défaut, \code{"closed frequent itemsets"} fournissant une synthèse des motifs fréquents
de sorte à économiser l'espace mémoire nécessaire.
Pour énumérer l'intégralité des motifs possibles, utiliser "frequent itemsets".}

\item{count}{Nombre minimal d'apparition d'un motif pour être considéré comme "fréquent", et par conséquent, conservé.}

\item{min_length}{Taille minimale qu'un motif doit avoir pour être conservé lors de l'énumération.}

\item{max_length}{Taille maximale qu'un motif doit avoir pour être conservé lors de l'énumération.
La valeur \code{Inf} par défaut correspond à une recherche de motifs sans limite maximale de taille.}

\item{status_limit}{Intervalle de temps pour lequel caractériser le statut des motifs par rapport à la période totale d'observations (nombre d'années).}
}
\value{
Nouvel objet de classe SpectralAnalyzer.
}
\description{
Constructeur d'objet SpectralAnalyzer.
}
\seealso{
\code{\link[arules]{APparameter}}
}
\author{
Gauthier Magnin
}
