% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{spectral.analyzer}
\alias{spectral.analyzer}
\title{Spectral Analyzer constructor}
\usage{
spectral.analyzer(
  observations,
  items = NULL,
  target = "closed frequent itemsets",
  count = 1,
  min_length = 1,
  max_length = Inf,
  status_limit = 2,
  init = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{observations}{\code{ObservationSet} class object: list of observations containing the items
corresponding to each one. Each observation is itself a list containing at least two elements
representing items and temporal data. It can contain any additional information but such data will
be ignored.

Items must be character or numeric values and must not contain the character \code{"/"}.
Temporal data must correspond to the years in which the observations were made and must be numeric
values.}

\item{items}{Data frame associating a name (column \code{name}) and possibly one or more categories
(additional columns) to each item (column \code{item}). Each category must be of type \code{factor}.
The columns \code{item} and \code{name} must be of type \code{character}. The default value
(\code{NULL}) specifies that no name or category is defined.}

\item{target}{Type of patterns to mine. One of \code{"frequent itemsets"},
\code{"closed frequent itemsets"}, \code{"maximally frequent itemsets"} (see 'Details').}

\item{count}{Minimum number of occurrences that a pattern must appear to be considered as "frequent".}

\item{min_length}{Minimum number of items that a pattern must have to be kept when mining patterns.}

\item{max_length}{Maximum number of items that a pattern must have to be kept when mining patterns.
The default \code{Inf} corresponds to a pattern search without maximum size limit.}

\item{status_limit}{Time interval for which to characterize the status of the patterns in relation
to the total period of observations (number of years).}

\item{init}{If \code{TRUE}, attributes relating to nodes, links between nodes, patterns and links
between patterns are initialized.
If \code{FALSE}, only attributes relating to observations, items and categories are initialized.}

\item{verbose}{Logical value indicating whether to report progress.}
}
\value{
New object of class \code{SpectralAnalyzer}.
}
\description{
Create and initialize an object of class \code{SpectralAnalyzer}.
}
\details{
If items are not specified using the argument \code{items}, they are automatically listed from
 the \code{observations} without any categorization or specific denomination.

The type of patterns mined can be:
 \itemize{
   \item{\code{"frequent itemsets"}: itemsets appearing in the observations according to an
         occurrence threshold defined by the parameter \code{count}.}
   \item{\code{"closed frequent itemsets"}: maximal itemsets of equivalence classes. An equivalence
         class is defined as the set of itemsets appearing in the same observations.}
   \item{\code{"maximally frequent itemsets"}: frequent itemsets which do not have any frequent
         superset. Also named as the maximals by inclusion of the frequent itemsets
         or the positive boundary of the frequent itemsets.}
 }

The \strong{frequent itemsets} are an exhaustive list of frequent itemsets.
 The \strong{closed frequent itemsets} synthesize the information so as to reduce the memory space
 required without real loss of information thanks to the equivalence classes.
 The \strong{maximal frequent itemsets} required even less memory since they are a subset of the
 closed frequent itemsets.
 
For an explanation with illustrated examples about the different types of itemsets, read the
 vignette titled "\emph{Mining itemsets}".
 
The steps for initializing a spectral analyzer are:
 \enumerate{
   \item{Enumeration of the observations per year.}
   \item{Enumeration of the nodes and calculation of the number of occurrences.}
   \item{Counting links between nodes.}
   \item{Elaboration of links between nodes.}
   \item{Mining for itemsets.}
   \item{Linking nodes to patterns.}
   \item{Characterization of patterns per year.}
   \item{Computation of pattern characteristics.}
   \item{Counting links between patterns.}
   \item{Elaboration of links between patterns.}
 }
 
The argument \code{init} and the method \code{\link{init}} allow to skip initialization steps.
}
\note{
The following steps may be quite long (depending on the case):
 \itemize{
   \item{Step 4: elaboration of links between nodes (depending on the amount of data).}
   \item{Steps 5 and 6: mining for itemsets and linking nodes to patterns (depending on the
     mining parameters and the amount of data).}
   \item{Step 10: elaboration of links between patterns (depending on the amount of patterns,
     resulting from the mining parameters and the amount of data).}
 }
}
\examples{
## Creating a SpectralAnalyzer from a list of observations
obs <- make_observations(oedb_sample, by = "ID",
                         additional = c("CODE", "NAME", "YEAR"))
obs_object <- observation.set(data = obs, item_key = "CODE", year_key = "YEAR")

sa_object_1 <- spectral.analyzer(obs_object)

## Creating a SpectralAnalyzer after associating item identifiers with
## names and one category
items_ids <- get_all_items(obs_object)
category_1 <- substances_information[match(items_ids,
                                           substances_information$CODE),
                                     "SUBFAMILY"]
category_1[is.na(category_1)] <- "Unknown"
names <- substances_information[match(items_ids,
                                      substances_information$CODE),
                                "NAME"]

items <- data.frame(item = items_ids, name = names, family = category_1)
sa_object_2 <- spectral.analyzer(obs_object, items)

}
\seealso{
The class: \code{\link{SpectralAnalyzer}}.

Initialization: \code{\link{init}}, \code{\link{is_init}}, \code{\link{reset}}.
}
\author{
Gauthier Magnin
}
