% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{spectral.analyzer}
\alias{spectral.analyzer}
\title{Spectral Analyzer constructor}
\usage{
spectral.analyzer(
  observations,
  items = NULL,
  target = "closed frequent itemsets",
  count = 1,
  min_length = 1,
  max_length = Inf,
  status_limit = 2
)
}
\arguments{
\item{observations}{List of observations containing the items corresponding to each observation.
Each observation is itself a list in the form \code{list( CODE = character(), YEAR = numeric )}.
Values of \code{CODE} must not contain the character \code{"/"}.
An observation can contain any additional information in its list.}

\item{items}{Data frame associating a name (column \code{name}) and one or more categories
(additional columns) to each item (column \code{item}). Each category must be of type \code{factor}.
The columns \code{item} and \code{name} must be of type \code{character}. The default value,
\code{NULL}, specifies that no name or category is defined.}

\item{target}{Type of patterns to enumerate. One of \code{"frequent itemsets"},
\code{"closed frequent itemsets"}, \code{"maximally frequent itemsets"}.
By default, \code{"closed frequent itemsets"} provide a summary of frequent patterns in order to
save the necessary memory space. To use even less memory, use \code{"maximally frequent itemsets"}.
To enumerate all possible patterns, use \code{"frequent itemsets"}.}

\item{count}{Minimum number of occurrences of a pattern to be considered as "frequent".}

\item{min_length}{Minimum number of items that a pattern must have to be kept when enumerating.}

\item{max_length}{Maximum number of items that a pattern must have to be kept when enumerating.
The default \code{Inf} corresponds to a pattern search without maximum size limit.}

\item{status_limit}{Time interval for which to characterize the status of the patterns in relation
to the total period of observations (number of years).}
}
\value{
New object of class \code{SpectralAnalyzer}.
}
\description{
Create and initialize an object of class \code{SpectralAnalyzer}.
}
\details{
If items are not specified using the argument \code{items}, they are automatically listed from
 the values of \code{CODE} in the argument \code{observations} without any categorization or
 specific denomination.
}
\examples{
## Creating a SpectralAnalyzer from a list of observations
obs <- make_observations(oedb_sample, by = "ID",
                         additional = c("CODE", "NAME", "YEAR"))

sa_object_1 <- spectral.analyzer(obs)

## Creating a SpectralAnalyzer after associating item identifiers with
## names and one category
items_ids <- get_all_items(obs)
category_1 <- substances_classification[match(items_ids,
                                              substances_classification$CODE),
                                        "SUBFAMILY"]
category_1[is.na(category_1)] <- "Unknown"
names <- substances_classification[match(items_ids,
                                         substances_classification$CODE),
                                   "NAME"]

items <- data.frame(item = items_ids, name = names, family = category_1)
sa_object_2 <- spectral.analyzer(obs, items)

}
\seealso{
\code{\link{SpectralAnalyzer}}, \code{\link{reset}}.
}
\author{
Gauthier Magnin
}
