% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\docType{class}
\name{SpectralAnalyzer-class}
\alias{SpectralAnalyzer-class}
\title{Classe d'objet S4 permettant une analyse spectrale.}
\description{
Classe d'objet S4 permettant une analyse spectrale.
}
\section{Slots}{

\describe{
\item{\code{observations}}{Liste des éléments retrouvés pour chaque observation.}

\item{\code{items}}{Ensemble des codes des différents éléments retrouvés dans les observations.}

\item{\code{items_categories}}{Catégories associées aux différents éléments observés.}

\item{\code{target}}{Type de motifs à énumérer lors de l'analyse.}

\item{\code{count}}{Nombre minimal d'apparition d'un motif pour être conservé lors de l'énumération des motifs.}

\item{\code{min_length}}{Taille minimale qu'un motif doit avoir pour être conservé lors de l'énumération des motifs.}

\item{\code{max_length}}{Taille maximale qu'un motif doit avoir pour être conservé lors de l'énumération des motifs.}

\item{\code{status_limit}}{Intervalle de temps pour lequel caractériser le statut des motifs par rapport à la période totale d'observations.}

\item{\code{nodes_per_year}}{Nombre de recrutements de chaque observation par année.}

\item{\code{nodes}}{Ensemble des noeuds, c'est-à-dire ensemble des observations distinctes.}

\item{\code{n_links}}{Ensemble des poids des liens entre les noeuds.}

\item{\code{nodes_links}}{Ensemble des liens entre les noeuds et caractéristiques de ces liens.}

\item{\code{obs_patterns}}{Ensemble des motifs associés à chaque observation disctincte.}

\item{\code{patterns_per_year}}{Ensemble des motifs, par année.}

\item{\code{patterns}}{Ensemble des motifs distincts et de leurs caractéristiques.}

\item{\code{p_links}}{Ensemble des poids des liens entre les motifs.}

\item{\code{patterns_links}}{Ensemble des liens entre les motifs et caractéristiques de ces liens.}

\item{\code{Class}}{Liste des attributs de classe (indépendants de l'instance).}
}}

\author{
Gauthier Magnin
}
