% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\docType{class}
\name{SpectralAnalyzer-class}
\alias{SpectralAnalyzer-class}
\alias{SpectralAnalyzer}
\title{Spectral Analyzer}
\description{
S4 object class allowing spectral analyzes.
}
\details{
Default colors are assigned to the values of each category associated with the items.
 These colors are selected circularly among the 20 colors of the palette \code{category20}
 from D3 (see \code{ggsci::pal_d3("category20")}).
Therefore, if the number of values exceeds \code{20}, some colors will be used more than once.
 For example, the \out{22<sup>nd</sup>} value will share the color of the \out{2<sup>nd</sup>}
 value.
See the attribute \code{categories_colors} to reassign colors to the category values.
}
\section{Slots}{

\describe{
\item{\code{observations}}{List of observations containing the items corresponding to each one.}

\item{\code{items}}{Set of codes identifying the items found in the observations.}

\item{\code{items_categories}}{Categories associated with the items.}

\item{\code{categories_colors}}{Colors associated with the values of the categories associated with the items.}

\item{\code{status_colors}}{Colors associated with the values of the status the patterns can have.}

\item{\code{target}}{Type of patterns to enumerate during the analysis.}

\item{\code{count}}{Minimum number of occurrences of a pattern to be kept when enumerating patterns.}

\item{\code{min_length}}{Minimum number of items that a pattern must have to be kept when enumerating patterns.}

\item{\code{max_length}}{Maximum number of items that a pattern must have to be kept when enumerating patterns.}

\item{\code{status_limit}}{Time interval for which to characterize the status of the patterns in relation to the total period of observations.}

\item{\code{nodes}}{Set of nodes: set of separate observations, and characteristics of these nodes.}

\item{\code{nodes_per_year}}{Number of occurrences of each node per year.}

\item{\code{n_links}}{Set of weights of the links between the nodes.}

\item{\code{nodes_links}}{Set of links between the nodes and characteristics of these links.}

\item{\code{nodes_patterns}}{Set of associations between patterns and nodes.}

\item{\code{patterns}}{Set of separate patterns and their characteristics.}

\item{\code{patterns_per_year}}{Number of occurrences of each pattern per year.}

\item{\code{p_links}}{Set of weights of the links between the patterns.}

\item{\code{patterns_links}}{Set of links between the patterns and characteristics of these links.}
}}

\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
            The spectrosome of occupational health problems.
            \emph{PLoS ONE} 13(1): e0190196.
            \url{https://doi.org/10.1371/journal.pone.0190196}.
            
            Bosson-Rieutort D, Sarazin P, Bicout DJ, Ho V, Lavoué J (2020).
            Occupational Co-exposures to Multiple Chemical Agents from Workplace Measurements by the US Occupational Safety and Health Administration.
            \emph{Annals of Work Exposures and Health}, Volume 64, Issue 4, May 2020, Pages 402–415.
            \url{https://doi.org/10.1093/annweh/wxaa008}.
}
\seealso{
The \code{SpectralAnalyzer} constructor: \code{\link{spectral.analyzer}}.

An example object of class \code{SpectralAnalyzer}: \code{\link{SA_instance}}.
}
\author{
Gauthier Magnin
}
