% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{spectrosome_chart,SpectralAnalyzer-method}
\alias{spectrosome_chart,SpectralAnalyzer-method}
\alias{spectrosome_chart}
\title{Spectrosome}
\usage{
\S4method{spectrosome_chart}{SpectralAnalyzer}(
  object,
  entities,
  characteristics,
  nb_graph = 1,
  min_link_weight = 1,
  vertex_size = "relative",
  path = getwd(),
  name = paste0("spectrosome_of_", entities, ".png"),
  title = paste0("Network of ", entities),
  ...
)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{entities}{Type d'élément pour lequel construire le spectrosome (nœuds ou motifs).
Choix parmi \code{"nodes"}, \code{"patterns"}.}

\item{characteristics}{Ensemble des caractéristiques des nœuds ou motifs dont le spectrosome est à tracer.}

\item{nb_graph}{Nombre de graphiques à générer et enregistrer. Le placement des sommets diffère entre chaque exemplaire.}

\item{min_link_weight}{Nombre minimum d'items en commun entre deux entités pour afficher le lien sur le graphe.}

\item{vertex_size}{Façon dont les tailles des sommets du graphe doivent être définies.
Choix parmi \code{"relative"}, \code{"grouped"}, \code{"absolute"}, \code{"equal"}.
\describe{
  \item{\code{"relative"}}{La taille d'un sommet dépend de l'intervalle de valeurs des différents poids.}
  \item{\code{"grouped"}}{Les poids des motifs sont regroupés selon des intervalles. À chaque intervalle correspond une taille.}
  \item{\code{"absolute"}}{La taille d'un sommet est définie directement en fonction du poids du motif.}
  \item{\code{"equal"}}{Les sommets ont tous la même taille.}
}}

\item{path}{Chemin du dossier dans lequel enregistrer les graphiques.
Par défaut, les graphiques sont enregistrés dans le répertoire de travail.}

\item{name}{Nom (avec extension) du fichier dans lequel enregistrer le graphique.
Si \code{nb_graph} est supérieur à \code{1}, un numéro est ajouté automatiquement à la fin du nom du fichier.}

\item{title}{Titre du graphique.}

\item{...}{Arguments supplémentaires fournis à la fonction \code{\link[sna:gplot]{sna::gplot}}
pour le traçage du graphe. Cf. section Details.}
}
\value{
Liste contenant :
 \describe{
   \item{\code{vertices}}{Data frame des nœuds ou motifs et caractéristiques utilisées,
                          associés aux identifiants des sommets du graphe et à leurs degrés dans le graphe.}
   \item{\code{edges}}{Data frame des informations relatives aux arêtes du graphe.}
   \item{\code{coords}}{Liste contenant les matrices des coordonnées des sommets du graphe.
                        Autant de matrices que de graphiques (\code{nb_graph}), pouvant être
                        réutilisées via l'argument \code{coord} (cf. \code{...}).}
 }
}
\description{
Construit un ou plusieurs graphiques de type spectrosome et les enregistre au format PNG.
}
\details{
Si des catégories sont associées aux items, chaque catégorie génère un spectrosome.
 Le nom de la catégorie est ajouté à la fin du nom du fichier.

Si des liens mixtes sont relatifs à des valeurs de catégorie qui ne sont pas représentées par des
 liens simples, ces valeurs apparaîssent dans la légende en dessous de "Mixt", sans couleur associée.
 
Si \code{min_link_weight} est supérieur à 1, certains nœuds ou motifs peuvent devenir isolés
 du fait que leurs liens avec les autres éléments peuvent ne plus être considérés. Ces nouveaux
 sommets isolés sont déplacés à la fin de la data frame de retour \code{edges}.
Les \code{n} lignes additionnelles sont numérotées \code{"A1"..."An"}.

Des arguments supplémentaires peuvent être fournis à la fonction en charge du traçage du graphe.
 Voir la liste des paramètres : \code{\link[sna:gplot]{sna::gplot}}.
Parmi eux, les paramètres suivants sont déjà définis et ne peuvent pas être modifiés : \code{dat},
 \code{gmode}, \code{coord}, \code{vertex.sides}, \code{vertex.cex}, \code{vertex.col}, \code{edge.col}.
Les paramètres suivants, pouvant être redéfinis, ont pour valeurs :
 \itemize{
   \item{\code{mode = "fruchtermanreingold"}}
   \item{\code{layout.par = list(repulse.rad = 4 ^ (log(nrow(nop_links), 10)))},
     où \code{nrow(nop_links)} correspond à la somme du nombre de liens et du nombre d'éléments isolés.}
   \item{\code{displaylabels = TRUE}}
   \item{\code{label.pos = 0}}
   \item{\code{boxed.labels = TRUE}}
   \item{\code{displayisolates = TRUE}}
 }
}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
            \emph{The spectrosome of occupational health problems}. PLoS ONE 13(1): e0190196.
            \url{https://doi.org/10.1371/journal.pone.0190196}.
}
\seealso{
\code{\link{degree}}, \code{\link[sna:gplot]{sna::gplot}}.
}
\author{
Delphine Bosson-Rieutort, Gauthier Magnin
}
