% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{spectrosome_chart,SpectralAnalyzer-method}
\alias{spectrosome_chart,SpectralAnalyzer-method}
\alias{spectrosome_chart}
\title{Spectrosome}
\usage{
\S4method{spectrosome_chart}{SpectralAnalyzer}(
  object,
  entities,
  characteristics,
  identifiers = "original",
  nb_graphs = 1,
  min_link_weight = 1,
  vertex_size = "relative",
  size_range = c(0.5, 2.5),
  vertex_col = "status",
  clusters = Inf,
  highlight = 3,
  use_names = TRUE,
  n.cutoff = NULL,
  c.cutoff = NULL,
  display_mixt = TRUE,
  path = getwd(),
  name = paste0("spectrosome_of_", entities, ".png"),
  title = paste0("Network of ", entities),
  ...
)
}
\arguments{
\item{object}{SpectralAnalyzer class object.}

\item{entities}{Type of entities for which to plot the spectrosome (nodes or patterns).
One of \code{"nodes"}, \code{"patterns"}.}

\item{characteristics}{Characteristics of nodes or patterns whose spectrosome is to be plotted.}

\item{identifiers}{Which IDs to use to identify the nodes or patterns on the chart and in the
return data frames? One of \code{"original"}, \code{"new"}. \cr
\code{"original"} to use the original identifiers.
\code{"new"} to use new identifiers ordered according to \code{characteristics}.}

\item{nb_graphs}{Number of graphics to generate and save. The place of the vertices differs between
each copy.}

\item{min_link_weight}{Minimum number of items in common between two entities to plot their link on
the chart.}

\item{vertex_size}{Way how the sizes of the vertices of the graph should be defined.
One of \code{"relative"}, \code{"grouped"}, \code{"absolute"}, \code{"equal"}.
\describe{
  \item{\code{"relative"}}{The sizes are defined by a linear interpolation of the weights of the
                           entities in \code{size_range}.}
  \item{\code{"grouped"}}{The weights of the entities are grouped according to 5 intervals
                          defined by quantiles. The 5 corresponding size values are taken in a
                          regular sequence bounded by \code{size_range}.}
  \item{\code{"absolute"}}{The size of a vertex is defined directly according to the weight of
                           the entity.}
  \item{\code{"equal"}}{The vertices are all the same size of 1.}
}}

\item{size_range}{If \code{vertex_size} is \code{"relative"} or \code{"grouped"}.
Range of vertex size values (given as expansion factors).}

\item{vertex_col}{Way how the colors of the vertices of the graph should be defined.
One of \code{"status"}, \code{"categories"}, \code{"none"}.
If \code{"status"} and \code{entities = "patterns"}, coloring according to the status of the patterns.
If \code{"categories"}, coloring according to the categories associated with the items of the
entities represented.
In all other cases, the vertices are colored gray.}

\item{clusters}{Maximum number of clusters to name on the graph.
If the actual number of clusters is greater, the names of the smaller clusters are not displayed.}

\item{highlight}{Number of clusters to highlight among those named on the graph.
The names of the largest clusters are displayed in bold.}

\item{use_names}{If \code{TRUE}, display item names if they are defined. Display their identification
codes otherwise.}

\item{n.cutoff}{If \code{use_names = TRUE}, limit number of characters to display concerning the names
of the represented items.}

\item{c.cutoff}{Limit number of characters to display in the legend for the categories represented.}

\item{display_mixt}{If \code{TRUE}, display in the legend the category values included only in mixed
links (or in mixed vertices, if \code{vertex_col = "categories"}).}

\item{path}{Path of the directory in which to save the charts.
By default, the charts are saved in the working directory.}

\item{name}{Name of the file in which to save the chart.
If \code{nb_graphs} is greater than \code{1}, a number is automatically added to the end of the
file name.}

\item{title}{Chart title.}

\item{...}{Additional arguments to the function \code{\link[sna:gplot]{sna::gplot}} for plotting
the graph. See Details section.}
}
\value{
\describe{
   \item{\code{vertices}}{Data frame of the nodes or patterns and characteristics used,
                          associated with the identifiers of the vertices of the graph and their
                          degrees in the graph.}
   \item{\code{edges}}{Data frame of information relating to the edges of the graph.}
   \item{\code{coords}}{List containing the coordinate matrices of the vertices of the graph.
                        As many matrices as there are charts (\code{nb_graphs}), which can be
                        reused via the argument \code{coord} (see \code{...}).}
 }
}
\description{
Plot one or more spectrosome charts and save them in PNG format.
}
\details{
If categories are associated with the items, each category generates a spectrosome.
 The category name is appended to the end of the file name.

If mixed links are relative to category values that are not represented by single links, these
 values are present in the legend below "Mixt", with no associated color.
The same is true for mixed vertices if the argument \code{vertex_col} is \code{"categories"}.

If \code{min_link_weight} is greater than 1 or the chart is to be plotted from a subset of
 all nodes or patterns, some of them may become isolated because their links to the other
 entities may no longer be considered.
These new isolated vertices are moved to the end of the return data frame \code{edges}.
The \code{n} related lines are numbered \code{"A1"..."An"}.

The colors associated with the values of each category represented are selected circularly
 among the 20 colors of the palette \code{category20} from D3 (see \code{ggsci::pal_d3("category20")}).
Therefore, if the number of values exceeds \code{20}, some colors will be used more than once.
 For example, the \out{22<sup>nd</sup>} value will share the color of the \out{2<sup>nd</sup>}
 value.
See the attribute \code{categories_colors} of \code{object} to reassign colors to the category values.

The names of clusters confused because all of their links are mixed links, are not displayed.

Additional arguments can be supplied to the function in charge of plotting the graph.
 See the list of parameters: \code{\link[sna:gplot]{sna::gplot}}.
Among them, the following parameters are already defined and cannot be modified: \code{dat},
 \code{gmode}, \code{vertex.sides}, \code{vertex.cex}, \code{vertex.col}, \code{edge.col}.
The following parameters, which can be redefined, have the following values:
 \itemize{
   \item{\code{mode = "fruchtermanreingold"}}
   \item{\code{displaylabels = TRUE}}
   \item{\code{label.pos = 0}}
   \item{\code{boxed.labels = TRUE}}
   \item{\code{displayisolates = TRUE}}
 }
}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
            The spectrosome of occupational health problems. \emph{PLoS ONE} 13(1): e0190196.
            \url{https://doi.org/10.1371/journal.pone.0190196}.
}
\seealso{
\code{\link{cluster_chart}}, \code{\link{degree}}, \code{\link[sna:gplot]{sna::gplot}}.
}
\author{
Delphine Bosson-Rieutort, Gauthier Magnin
}
