% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{spectrum_chart,SpectralAnalyzer-method}
\alias{spectrum_chart,SpectralAnalyzer-method}
\alias{spectrum_chart}
\title{Pattern spectrum}
\usage{
\S4method{spectrum_chart}{SpectralAnalyzer}(
  object,
  pc,
  identifiers = "original",
  sort = TRUE,
  title = "Spectrum of patterns",
  path = getwd(),
  name = "spectrum_of_patterns.pdf"
)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{pc}{Data frame of \strong{p}atterns and their \strong{c}haracteristics. Patterns whose spectrum
is to be plotted. Any subset of \code{object["patterns"]}.\cr
\code{"patterns"} and \code{"p"} are specific values for \code{object["patterns"]}.}

\item{identifiers}{Which IDs to use to identify the patterns on the chart and in the return data frame.
One of \code{"original"}, \code{"new"}.
\describe{
  \item{\code{"original"}}{Use of the original identifiers.}
  \item{\code{"new"}}{Use of new identifiers based on the sorting of patterns or on the order in
                      \code{pc}.}
}}

\item{sort}{If \code{TRUE}, the patterns are sorted on the chart as described in 'Details' section.
Otherwise, they are ordered on the chart in the order in which they are given.}

\item{title}{Chart title.}

\item{path}{Path of the directory in which to save the chart.
By default, the chart is saved in the working directory.}

\item{name}{Name of the file in which to save the chart.}
}
\value{
Data frame of the patterns and characteristics used, associated with the identifiers visible
 on the chart.
}
\description{
Plot a spectrum chart and save it as a PDF file.
}
\details{
The patterns are sorted according to their specificities (desc.), status (in order of
 \code{object["status_colors"]}, default is  \code{"Persistent"}, \code{"Declining"},
 \code{"Emergent"}, \code{"Latent"}), weights (desc.) and sizes (asc.).
 If two patterns have the same characteristics concerning these ones, they are ordered relative to
 each other in the order they are given.
}
\examples{
spectrum_1 <- spectrum_chart(SA_instance, "patterns")
spectrum_2 <- spectrum_chart(SA_instance, SA_instance["patterns"][1:15, ],
                             name = "spectrum_of_patterns_1-15")

}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
            The spectrosome of occupational health problems. \emph{PLoS ONE} 13(1): e0190196.
            \url{https://doi.org/10.1371/journal.pone.0190196}.
}
\author{
Delphine Bosson-Rieutort, Gauthier Magnin
}
