% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{spectrum_chart,SpectralAnalyzer-method}
\alias{spectrum_chart,SpectralAnalyzer-method}
\alias{spectrum_chart}
\title{Pattern spectrum}
\usage{
\S4method{spectrum_chart}{SpectralAnalyzer}(
  object,
  pc,
  identifiers = "original",
  sort = TRUE,
  title = "Spectrum of patterns",
  path = NULL,
  name = NULL
)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{pc}{Data frame of \strong{p}atterns and their \strong{c}haracteristics. Patterns whose spectrum
is to be plotted. Any subset of \code{object["patterns"]}.\cr
\code{"patterns"} and \code{"p"} are special values for \code{object["patterns"]}.}

\item{identifiers}{Which IDs to use to identify the patterns on the chart and in the return data frame.
One of \code{"original"}, \code{"new"}.
\describe{
  \item{\code{"original"}}{Use of the original identifiers.}
  \item{\code{"new"}}{Use of new identifiers based on the sorting of patterns or on the order in
                      \code{pc}.}
}}

\item{sort}{If \code{TRUE}, the patterns are sorted on the chart as described in 'Details' section.
Otherwise, they are ordered on the chart in the order in which they are given.}

\item{title}{Chart title.}

\item{path}{Path of the directory in which to save the chart as a PDF file. To be ignored to plot
the chart in the active device.}

\item{name}{Name of the PDF file in which to save the chart. To be ignored to plot the chart in the
active device.}
}
\value{
Data frame of the patterns and characteristics used, associated with the identifiers visible
 on the chart.
}
\description{
Plot a spectrum chart: bar chart of the given patterns showing their characteristics.
 It can be automatically saved as a PDF file.
}
\details{
The patterns are sorted according to their specificities (desc.), status (in order of
 \code{object["status_colors"]}, default is  \code{"Persistent"}, \code{"Declining"},
 \code{"Emergent"}, \code{"Latent"}), weights (desc.) and sizes (asc.).
 If two patterns have the same characteristics concerning these ones, they are ordered relative to
 each other in the order they are given.

If the arguments \code{path} and \code{name} are not \code{NULL}, the chart is plotted in a PDF file
 of A4 landscape paper size. If they are \code{NULL}, the chart is plotted in the active device.
}
\examples{
spectrum_1 <- spectrum_chart(SA_instance, "patterns")
spectrum_2 <- spectrum_chart(SA_instance, SA_instance["patterns"][1:15, ])

spectrum_1 <- spectrum_chart(SA_instance, "patterns",
                             path = getwd(), name = "spectrum_of_patterns")

}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
            The spectrosome of occupational health problems. \emph{PLoS ONE} 13(1): e0190196.
            \url{https://doi.org/10.1371/journal.pone.0190196}.
}
\seealso{
\code{\link{weight_by_node_complexity}}.
}
\author{
Delphine Bosson-Rieutort, Gauthier Magnin
}
