% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{spectrum_chart,SpectralAnalyzer-method}
\alias{spectrum_chart,SpectralAnalyzer-method}
\alias{spectrum_chart}
\title{Pattern spectrum}
\usage{
\S4method{spectrum_chart}{SpectralAnalyzer}(
  object,
  patterns_characteristics,
  identifiers = "original",
  path = getwd(),
  name = "spectrum_of_patterns.pdf",
  title = "Spectrum of patterns"
)
}
\arguments{
\item{object}{SpectralAnalyzer class object.}

\item{patterns_characteristics}{Patterns (and their characteristics) whose spectrum is to be plotted.}

\item{identifiers}{Which IDs to use to identify the patterns on the chart and in the return data frame?
One of \code{"original"}, \code{"new"}. \cr
\code{"original"} to use the original identifiers.
\code{"new"} to use new identifiers based on pattern sorting (see Details section to learn more
about the sort that is performed).}

\item{path}{Path of the directory in which to save the chart.
By default, the chart is saved in the working directory.}

\item{name}{Name of the file in which to save the chart.}

\item{title}{Chart title.}
}
\value{
Data frame of the patterns and characteristics used, associated with the identifiers visible
 on the chart.
}
\description{
Plot a spectrum chart and save it as a PDF file.
}
\details{
The patterns are sorted according to their specificities (desc.), status (\code{"Persistent"},
 \code{"Declining"}, \code{"Emergent"}, \code{"Latent"}), weights (desc.) and sizes (asc.).
}
\references{
Bosson-Rieutort D, de Gaudemaris R, Bicout DJ (2018).
            \emph{The spectrosome of occupational health problems}. PLoS ONE 13(1): e0190196.
            \url{https://doi.org/10.1371/journal.pone.0190196}.
}
\author{
Delphine Bosson-Rieutort, Gauthier Magnin
}
