% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_set.R
\name{[,TransactionSet,ANY,ANY,ANY-method}
\alias{[,TransactionSet,ANY,ANY,ANY-method}
\alias{[,TransactionSet-method}
\alias{[<-,TransactionSet,ANY,ANY,ANY-method}
\alias{[<-,TransactionSet-method}
\title{Extract or replace parts of an object of class TransactionSet}
\usage{
\S4method{[}{TransactionSet,ANY,ANY,ANY}(x, i, j, drop)

\S4method{[}{TransactionSet,ANY,ANY,ANY}(x, i, j) <- value
}
\arguments{
\item{x}{Object from which to extract element(s) or in which to replace
element(s).}

\item{i}{Numeric or character values. Indices specifying elements to extract
or replace. See 'Details' section.}

\item{j, drop}{Unused.}

\item{value}{Value of type similar to the element to be replaced.}
}
\description{
General selector and mutator to access the attributes of an object of class
\code{TransactionSet}.
}
\details{
Character values can be used to access attributes as well as sub-elements of
the attribute \code{data}.

Numeric values can be used to access elements of the attribute \code{data}.
Replacing one element of \code{data} this way will not change its name.

Replacing the attribute \code{names}:
\itemize{
\item reorders the elements contained in each transaction if \code{value} is a
reordered equivalent of \code{names};
\item renames the elements contained in each transaction if \code{value} is the same
length as \code{names} but is not equivalent. In this case, the attributes
\code{item_key} and \code{year_key} are updated (unless \code{year_key} is \code{NA});
\item removes missing elements of \code{value} in each transaction if it is smaller
than \code{names}.
}
}
\examples{
TS_instance["data"]

TS_instance["item_key"]
TS_instance["CODE"]

TS_instance[3]
TS_instance[3:5]

TS_instance["year_key"] <- NA
TS_instance[3] <- list(CODE = 1,
                       YEAR = 2000,
                       JOB.TITLE = 45454545,
                       JOB.TASK = "A3000",
                       SAMPLE.ID = c(1, 3, 5, 8))
TS_instance["names"] <- c("CODE", "YEAR")

}
\author{
Gauthier Magnin
}
