% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixer.R
\name{subset_from_class}
\alias{subset_from_class}
\title{Subsets of values and references by class}
\usage{
subset_from_class(values, references = NULL, classes, class_name)
}
\arguments{
\item{values}{Numeric named vector or matrix, or list of numeric named vectors.
Values from which to extract a subset according to one specific class.}

\item{references}{Numeric vector or list of numeric vectors. Reference values associated with the
\code{values}. See 'Details' to know the way it is associated with \code{values}.}

\item{classes}{List or logical matrix associating the \code{values} names with classes.
If list, its names are those present in \code{values} and the elements are vectors of associated classes.
If logical matrix, its columns are named according to the classes and the row names
contain the names associated with the \code{values}. A \code{TRUE} value indicates that a specific name
is part of a specific class.}

\item{class_name}{Name of the class of interest, the one for which to extract the corresponding
subsets of \code{values} and \code{references}.}
}
\value{
If \code{references} is \code{NULL}, subset of values that correspond to the class of interest.\cr
If not, list containing:
\describe{
\item{\code{values}}{Subset of values that correspond to the class of interest.}
\item{\code{references}}{Subset of references that correspond to the class of interest.}
}
}
\description{
Extract from values and references those corresponding to one specific class.
}
\details{
If \code{values} is a vector, it must have one reference value for each value of the vector.

If \code{values} is a matrix, it must have one reference value for each column.

If \code{values} is a list, the reference values can be a vector of named values or a list. In this case,
if \code{references} is a vector, there must be one reference for each name present in \code{values}.
Otherwise, \code{references} is a list of vectors having the same lengths as those present in \code{values}
so that \code{values} and \code{references} can be matched.

If \code{classes} is a list, it will be turned into a logical matrix before processing. Thus, call the
function with such a matrix is slightly faster.
}
\examples{
## Association of classes (C1 to C8) with elements A, B, C, D and E
classes <- list(A = c("C5", "C6", "C8"),
                B = "C8",
                C = c("C3", "C8"),
                D = c("C1", "C3", "C4", "C6"),
                E = c("C2", "C4", "C5", "C7", "C8"))

## Subsets with values as a vector
subset_from_class(values = c(A = 1, B = 2, C = 3, D = 4, E = 5),
                  classes = classes,
                  class_name = "C8")

## Subsets with values as a matrix
subset_from_class(
  values = matrix(c(0.1,0.2,0.3,0.4,0.5,
                    0.6,0.7,0.8,0.9,1.),
                  nrow = 2, byrow = 2,
                  dimnames = list(c("set.1", "set.2"), LETTERS[1:5])),
  references = c(1,2,3,4,5),
  classes = classes,
  class_name = "C8"
)

## Subsets with values as a list
v <- list(set.1 = c(A = 0.1, D = 0.5),
          set.2 = c(D = 0.2),
          set.3 = c(A = 0.3, C = 0.4))

subset_from_class(values = v,
                  classes = classes,
                  class_name = "C8")

subset_from_class(values = v,
                  references = c(A = 1, C = 3, D = 4),
                  classes = classes,
                  class_name = "C8")

subset_from_class(values = v,
                  references = list(c(1, 4),
                                    4,
                                    c(1, 3)),
                  classes = classes,
                  class_name = "C8")

}
\seealso{
Generic function to apply the MCR approach according to classes: \code{\link{mcr_approach_by_class}}.

Specific functions of the MCR approach applying according to classes: \code{\link{mcr_summary_by_class}},
\code{\link{mcr_chart_by_class}}, \code{\link{thq_pairs_by_class}}, \code{\link{thq_by_group_by_class}}.
}
\author{
Gauthier Magnin
}
