% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixer.R
\name{thq_pairs_for_matrix}
\alias{thq_pairs_for_matrix}
\title{Top Hazard Quotient pairs, in matrix}
\usage{
thq_pairs_for_matrix(values, references,
                     threshold = TRUE,
                     alone = FALSE,
                     ignore_zero = TRUE)
thq_pairs_for_matrix(hq, hi,
                     threshold = TRUE,
                     alone = FALSE,
                     ignore_zero = TRUE)
}
\arguments{
\item{values}{Numeric named matrix. Vectors of values for which the top two hazard quotients are to
be identified.}

\item{references}{Numeric vector. Reference values associated with the \code{values}. See 'Details' to
know the way it is associated with \code{values}.}

\item{hq}{Numeric named matrix. \strong{H}azard \strong{q}uotients for which the top two pairs are to be
identified.}

\item{hi}{Numeric vector. \strong{H}azard \strong{i}ndexes associated with the hazard quotients \code{hq}.}

\item{threshold}{If \code{TRUE}, only values or hazard quotients associated with hazard indices greater
than 1 are considered.}

\item{alone}{If \code{TRUE}, take into account single top hazard quotients (i.e. sets of values of length
1). If so, a level named \code{"NULL"} is added as combination with such top hazard quotients.}

\item{ignore_zero}{\code{TRUE} or \code{FALSE} whether to ignore values equal to 0.}
}
\value{
\code{NULL} if (1) \code{alone = FALSE} and no set of \code{values} (or of \code{hq}) has more than one element
considered;
or (2) \code{threshold = TRUE} and no related hazard index is greater than 1;
or (3) \code{threshold = TRUE}, \code{alone = FALSE} and no set of values meets the two conditions.

Pairs that produce the top two hazard quotients in the sets of values selected accordingly to the
arguments \code{threshold} and \code{alone}. Vector if only one pair; two-row matrix otherwise.
}
\description{
Identify the top two hazard quotients pairs for which the associated hazard
indices are greater than 1.
}
\details{
The reference values are applied once on each set of values, i.e. on each row.
Therefore, there must be one reference value for each column of the matrix.

Arguments \code{values} and \code{references} are used to compute the hazard quotients and the hazard indices
before identifying the highest hazard quotients then building the contingency table. Thus, call the
function with the arguments \code{hq} and \code{hi} is faster (if they are already computed).

\loadmathjax
The hazard index of the vector \eqn{i} is given by:
\mjdeqn{HI_i = \sum_{j = 1}^n HQ_{i,j}}{HI_i = sum(HQ_ij) from j = 1 to n}
where \eqn{HQ} denotes the hazard quotients and \eqn{n} denotes the number of hazard quotients.

The hazard quotient of the value \eqn{j} in the vector \eqn{i} is given by:
\mjdeqn{HQ_{i,j} = \frac{V_{i,j}}{RV_j}}{HQ_ij = V_ij / RV_j}
where \eqn{V} denotes the \code{values} and \eqn{RV} denotes the \code{references}.

If the number of hazard quotients that are greater than or equal to the second greater hazard
quotient is greater than 2, only the first two values are considered and in the order given.
For example, if the hazard quotients are \code{D = 5}, \code{B = 1}, \code{C = 3} and \code{A = 3}, the top two hazard
quotients considered are \code{D} and \code{C}.
}
\note{
Due to the multiple possible usages, all arguments except \code{values} and \code{references} must be explicitly
named in the function call.

This function is not exported from the package.
}
\references{
Reyes JM, Price PS (2018).
An analysis of cumulative risks based on biomonitoring data for six phthalates using the Maximum Cumulative Ratio.
\emph{Environment International}, 112, 77-84.
\url{https://doi.org/10.1016/j.envint.2017.12.008}.
}
\seealso{
\code{\link{thq_pairs}}, \code{\link{thq_pairs_for_list}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
