% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_set.R
\docType{class}
\name{TransactionSet-class}
\alias{TransactionSet-class}
\alias{TransactionSet}
\alias{print,TransactionSet-method}
\alias{length,TransactionSet-method}
\title{Transaction Set}
\description{
S4 object class for representing transactions.
A transaction is defined as an itemset associated with any other data.
}
\details{
The length of a \code{TransactionSet} object is the number of transactions it contains (i.e. is equal to
the length of the attribute \code{data}).

If the transactions do not contain temporal data, then the attribute \code{year_key} is \code{NA}.
}
\section{Slots}{

\describe{
\item{\code{data}}{List of transactions containing items and any additional data.
Each transaction is itself a list. All transactions share the same element names (corresponding to
the names of the various data).}

\item{\code{names}}{Names of the elements present in each transaction (i.e. names of the elements contained in
the list corresponding to a transaction).}

\item{\code{item_key}}{Name of the element of a transaction containing its items.}

\item{\code{year_key}}{Name of the element of a transaction containing the year in which it was made.}
}}

\section{Coercion}{

\describe{
\item{\code{as(object, Class = "data.frame")}}{Convert a \code{TransactionSet} object to \code{data.frame}
(considering the parameter \code{stringAsFactors} as \code{FALSE}).}
\item{\code{as(object, Class = "transactions")}}{Convert a \code{TransactionSet} object to a set of
\code{\link[arules:transactions-class]{transactions}} from the package \code{arules}. Only the items of
the transactions are considered. Other data are ignored.}
}
}

\seealso{
The \code{TransactionSet} constructor: \code{\link{transaction.set}}.

An example object of class \code{TransactionSet}: \code{\link{TS_instance}}.

Selectors and mutators to access the attributes of a \code{TransactionSet} object:
\code{\link[=[,TransactionSet-method]{[}},
\code{\link[=[,TransactionSet-method]{[[}},
\code{\link[=[,TransactionSet-method]{$}}.
}
\author{
Gauthier Magnin
}
