% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{tree_chart,SpectralAnalyzer-method}
\alias{tree_chart,SpectralAnalyzer-method}
\alias{tree_chart}
\title{Arbre de la multi-association}
\usage{
\S4method{tree_chart}{SpectralAnalyzer}(
  object,
  patterns_characteristics,
  display_text = NULL,
  cutoff = NULL,
  path = getwd(),
  name = "multi-association_tree.pdf",
  title = "Multi-association tree"
)
}
\arguments{
\item{object}{Objet de classe SpectralAnalyzer.}

\item{patterns_characteristics}{Ensemble des caractéristiques des motifs dont l'arbre est à tracer.}

\item{display_text}{Texte à afficher sur le graphique à côté des motifs.
Choix entre les identifiants \code{"ID"} des motifs ou l'une des caractéristiques (\code{"weight"},
\code{"frequency"}, \code{"specificity"}, \code{"year"}, \code{"status"}).
Par défaut (valeur \code{NULL}), aucun texte n'est affiché.}

\item{cutoff}{Nombre limite de caractères à afficher dans la légende concernant les catégories représentées.}

\item{path}{Chemin du dossier dans lequel enregistrer le graphique.
Par défaut, le graphique est enregistré dans le répertoire de travail.}

\item{name}{Nom du fichier dans lequel enregistrer le graphique.}

\item{title}{Titre du graphique.}
}
\value{
Data frame des motifs représentées sur le graphique, associés à leurs caractéristiques et
 identifiants (visibles sur le graphique si \code{display_text = "ID"}).
}
\description{
Construit un graphique de type arbre de la multi-association et l'enregistre dans un fichier au
format PDF.
}
\details{
Si des catégories sont associées aux items, chaque catégorie génère un arbre.
 Le nom de la catégorie est ajouté à la fin du nom du fichier.

Les motifs d'ordre 1 ne sont pas dessinés. Seuls les items inclus dans les motifs d'ordre supérieur
 le sont.

Les motifs sont triés selon leurs poids.

Les couleurs associées aux valeurs de l'éventuelle catégorie représentée sont sélectionnées
 de manière circulaire parmi les 20 couleurs de la palette \code{category20} de D3 (cf.
 \code{ggsci::pal_d3("category20")}).
Par conséquent, si le nombre de valeurs dépasse \code{20}, certaines couleurs seront utilisées
 plusieurs fois. Par exemple, la \out{22<sup>e</sup>} valeur partagera la couleur de la
 \out{2<sup>e</sup>} valeur.
}
\author{
Delphine Bosson-Rieutort, Gauthier Magnin
}
