% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{tree_chart,SpectralAnalyzer-method}
\alias{tree_chart,SpectralAnalyzer-method}
\alias{tree_chart}
\title{Multi-association tree}
\usage{
\S4method{tree_chart}{SpectralAnalyzer}(
  object,
  patterns_characteristics,
  identifiers = "original",
  use_names = TRUE,
  n.cutoff = NULL,
  display_status = TRUE,
  display_text = "ID",
  c.cutoff = NULL,
  sort_by = "category",
  path = getwd(),
  name = "multi-association_tree.pdf",
  title = "Multi-association tree"
)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{patterns_characteristics}{Patterns (and their characteristics) whose tree is to be plotted.
Any subset of \code{object["patterns"]}.}

\item{identifiers}{Which IDs to use to identify the patterns on the chart and in the return data frame?
One of \code{"original"}, \code{"new"}. \cr
\code{"original"} to use the original identifiers.
\code{"new"} to use new identifiers based on pattern sorting (see Details section to learn more
about the sort that is performed).}

\item{use_names}{If \code{TRUE}, display item names if they are defined. Display their identification
codes otherwise.}

\item{n.cutoff}{If \code{use_names = TRUE}, limit number of characters to display concerning the names
of the represented items.}

\item{display_status}{If \code{TRUE}, display pattern status.}

\item{display_text}{Text to display on the chart next to the patterns.
Pattern identifiers (\code{"ID"}) or one of the other characteristics (\code{"weight"},
\code{"frequency"}, \code{"specificity"}, \code{"year"}).
The \code{NULL} value specifies that none of this information should be displayed.}

\item{c.cutoff}{Limit number of characters to display in the legend for the categories represented.}

\item{sort_by}{Sorting method of displayed items. One of \code{"category"}, \code{"item"}.}

\item{path}{Path of the directory in which to save the chart.
By default, the chart is saved in the working directory.}

\item{name}{Name of the file in which to save the chart.}

\item{title}{Chart title.}
}
\value{
Data frame of the patterns represented on the chart, associated with their characteristics
 and identifiers (visible on the chart if \code{display_text = "ID"}).
}
\description{
Plot a multi-association tree chart and save it as a PDF file.
}
\details{
If categories are associated with items, each category generates a tree.
 The category name is appended to the end of the file name.

Patterns of order 1 are not drawn. Only items included in higher-order patterns are.

The patterns are sorted according to their order values, then to their weights.

The colors associated with the values of the possible category represented are selected circularly
 among the 20 colors of the palette \code{category20} from D3 (see \code{ggsci::pal_d3("category20")}).
Therefore, if the number of values exceeds \code{20}, some colors will be used more than once.
 For example, the \out{22<sup>nd</sup>} value will share the color of the \out{2<sup>nd</sup>}
 value.
See the attribute \code{categories_colors} of \code{object} to reassign colors to the category values.
}
\examples{
tree_1 <- tree_chart(SA_instance, SA_instance["patterns"],
                     n.cutoff = 20, c.cutoff = 17)
tree_2 <- tree_chart(SA_instance, SA_instance["patterns"][1:15, ],
                     c.cutoff = 17, name = "multi-association_tree_1-15")

}
\references{
Bosson-Rieutort D, Sarazin P, Bicout DJ, Ho V, Lavoué J (2020).
            Occupational Co-exposures to Multiple Chemical Agents from Workplace Measurements by the US Occupational Safety and Health Administration.
            \emph{Annals of Work Exposures and Health}, Volume 64, Issue 4, May 2020, Pages 402–415.
            \url{https://doi.org/10.1093/annweh/wxaa008}.
}
\author{
Delphine Bosson-Rieutort, Gauthier Magnin
}
