% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TS_instance}
\alias{TS_instance}
\title{Example of a TransactionSet object}
\format{
An object of class \code{TransactionSet} containing 14 transactions of 5
elements:
\describe{
\item{\code{CODE}}{\emph{Integer}. Codes identifying the items corresponding to the
transaction.}
\item{\code{YEAR}}{\emph{Integer}. Year in which the transaction was made.}
\item{\code{JOB.TITLE}, \code{JOB.TASK}, \code{SAMPLE.ID}}{\emph{Integer}, \emph{character},
\emph{integer}. Additional data related to the transaction.}
}

For more about the attributes, see \code{\link{TransactionSet}}.
}
\usage{
TS_instance
}
\description{
An example of an object of class \code{TransactionSet}.
}
\details{
Here is the way it was created using the dataset \code{\link{oedb_sample}}:
\preformatted{
## Making a list of transactions by grouping data
trx <- make_transactions(oedb_sample,
                         by = "ID",
                         additional = c("CODE", "YEAR",
                                        "JOB.TITLE", "JOB.TASK", "SAMPLE.ID"))

## Creation of the TransactionSet
TS_instance <- transaction.set(data = trx, item_key = "CODE", year_key = "YEAR")
TS_instance["names"] <- TS_instance["names"][-1]
}
}
\keyword{datasets}
