% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_manager.R
\name{turn_data_frame_into_list}
\alias{turn_data_frame_into_list}
\title{Turn a data frame into a list}
\usage{
turn_data_frame_into_list(x, indices = c(1, 2))
}
\arguments{
\item{x}{Data frame to turn into a list.}

\item{indices}{Two-element vector indicating the names of the associated columns to use to create
the list.}
}
\value{
List in which the names correspond to the column of \code{x} corresponding to the first element
of \code{indices} and the values correspond to the column of \code{x} corresponding to the second element of
\code{indices}.
}
\description{
Turn two columns of a data frame into a list.
}
\examples{
df <- data.frame(V1 = c("e1", "e1", "e1", "e2", "e3", "e3", "e4",
                        "e4", "e4", "e4", "e5", "e5", "e5", "e5", "e5"),
                 V2 = c("E", "F", "I", "I", "C", "I", "A",
                        "C", "D", "F", "B", "D", "E", "I", "G"),
                 stringsAsFactors = FALSE)

turn_data_frame_into_list(df)

}
\seealso{
\code{\link{turn_list_into_data_frame}}, \code{\link{turn_logical_matrix_into_list}},
\code{\link{turn_char_matrix_into_list}}, \code{\link{invert_list}}.
}
\author{
Gauthier Magnin
}
