% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{turn_into_data_frame}
\alias{turn_into_data_frame}
\title{Turn into data frame}
\usage{
turn_into_data_frame(x)
}
\arguments{
\item{x}{Table, matrix or vector to turn into a data frame.}
}
\value{
Data frame corresponding to the input variable.
}
\description{
Turn a table, a matrix or a vector into a data frame.
}
\details{
If \code{x} is a vector, the output is a one-line data frame.
Otherwise, the output has the same dimensions as \code{x}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
