% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{turn_into_path}
\alias{turn_into_path}
\title{Directory path}
\usage{
turn_into_path(path)
}
\arguments{
\item{path}{Character string to be used as a directory path.}
}
\value{
Character string \code{path}, ending with \code{"/"}.
}
\description{
Check if a string can be used as a directory path.
Add a \code{"/"} character if it cannot.
}
\author{
Gauthier Magnin
}
\keyword{internal}
