% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_manager.R
\name{turn_list_into_logical_matrix}
\alias{turn_list_into_logical_matrix}
\title{Turn a list into a logical matrix}
\usage{
turn_list_into_logical_matrix(x, by_name = FALSE)
}
\arguments{
\item{x}{List of vectors to turn into a logical matrix.}

\item{by_name}{If \code{FALSE}, use the values of the list. If \code{TRUE}, use the names of these values.}
}
\value{
Logical matrix in which row names are the names of the list and column names are the values
of the list or the names of these values (according to \code{by_name}).
}
\description{
Turn a list into a logical matrix associating names of the list with its values or with the names of
these values.
}
\examples{
l <- list(e1 = c(v1 = "E", v2 = "F", v3 = "I"),
          e2 = c(v1 = "I"),
          e3 = c(v1 = "C", v2 = "I"),
          e4 = c(v1 = "A", v2 = "C", v3 = "D", v4 = "F"),
          e5 = c(v1 = "B", v2 = "D", v3 = "E", v4 = "I", v5 = "G"))

turn_list_into_logical_matrix(l)
turn_list_into_logical_matrix(l, by_name = TRUE)

}
\seealso{
\code{\link{coerce_list}}, \code{\link{turn_list_into_char_matrix}}, \code{\link{turn_list_into_data_frame}},
\code{\link{turn_logical_matrix_into_list}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
