% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uom.dat-package.R
\docType{package}
\name{uom.dat-package}
\alias{uom.dat}
\alias{uom.dat-package}
\title{uom.dat: University of Montreal - Data Analysis Tools}
\description{
R processing, analysis and visualization tools initially designed
    for processing data from occupational exposure databases, especially for the
    concept of multiexposure. Main analytical approaches are Frequent Itemset
    Mining, Spectrosome, Association Rules and Maximum Cumulative Ratio.
}
\details{
For an organized list of datasets, functions, classes and methods, use
 \code{vignette("list_of_help_pages", "uom.dat")}

For an alphabetical list and for the package description, use \code{library(help = "uom.dat")}

For package news information, use \code{news(package = "uom.dat")}

For information on how to cite the package, use \code{citation("uom.dat")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UofMontreal-Multiexpo/uom.dat}
  \item Report bugs at \url{https://github.com/UofMontreal-Multiexpo/uom.dat/issues}
}

}
\author{
\strong{Maintainer}: Jérôme Lavoué \email{jerome.lavoue@umontreal.ca}

Authors:
\itemize{
  \item Gauthier Magnin \email{gauthier.magnin@gmail.com} [contractor]
  \item Delphine Bosson-Rieutort \email{delphine.bosson-rieutort@umontreal.ca}
}

Other contributors:
\itemize{
  \item CHUM - Centre Hospitalier de l'Université de Montréal [copyright holder, funder]
}

}
