% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uom.dat-package.R
\docType{package}
\name{uom.dat-package}
\alias{uom.dat}
\alias{uom.dat-package}
\title{uom.dat: University of Montreal - Data Analysis Tools}
\description{
R processing, analysis and visualization tools initially designed for processing data from
    occupational exposure databases, especially for the concept of multiexposure.
    Main analytical approaches are Frequent Itemset Mining, Spectrosome, Association Rules and Maximum
    Cumulative Ratio.
}
\details{
For an organized list of datasets, functions, classes and methods, use
 \code{vignette("list_of_help_pages")}

For an alphabetical list and for the package description, use \code{library(help = "uom.dat")}

For information on how to cite the package, use \code{citation("uom.dat")}
}
\author{
Gauthier Magnin, Delphine Bosson-Rieutort
 
 Maintainer: Gauthier Magnin \email{gauthier.magnin@gmail.com}
}
