% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{which_entities,SpectralAnalyzer-method}
\alias{which_entities,SpectralAnalyzer-method}
\alias{which_entities}
\title{Detect the type of entities}
\usage{
\S4method{which_entities}{SpectralAnalyzer}(object, npr, entities = SpectralAnalyzer.NODES_OR_PATTERNS)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{npr}{Data frame of \strong{n}odes, \strong{p}atterns or association \strong{r}ules and their characteristics.\cr
\code{"nodes"}, \code{"n"}, \code{"patterns"}, \code{"p"}, \code{"rules"} and \code{"r"} are specific values.}

\item{entities}{Define if the data frame is either a data frame of nodes or a data frame of patterns
(\code{NODES_OR_PATTERNS}), or if it can also be a data frame of rules (\code{NODES_PATTERNS_OR_RULES}).}
}
\value{
Character corresponding to \code{NODES}, \code{PATTERNS} or \code{RULES}.
}
\description{
Detect the type of entities contained in a data frame among nodes, patterns and association rules.
}
\details{
The detection uses the column names of the data frame and search for \code{"node"}, \code{"pattern"} or
\code{"antecedent"} for a data frame of nodes, patterns or rules, respectively.

The argument \code{entities} is only used to adapt a possible error message.
}
\seealso{
\code{\link{get_nopc}}, \code{\link{which_associated_links}}, \code{\link{which_name}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
