% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_analyzer.R
\name{which_name,SpectralAnalyzer-method}
\alias{which_name,SpectralAnalyzer-method}
\alias{which_name}
\title{Type of entities or links corresponding to a name}
\usage{
\S4method{which_name}{SpectralAnalyzer}(object, name)
}
\arguments{
\item{object}{\code{SpectralAnalyzer} class object.}

\item{name}{Type of entities or links.
Character corresponding to \code{NODES}, \code{PATTERNS}, \code{RULES}, \code{NODES_LINKS}, \code{PATTERN_LINKS} or
their simplifications. One or more.}
}
\value{
Character corresponding to \code{NODES}, \code{PATTERNS}, \code{RULES}, \code{NODES_LINKS} or \code{PATTERN_LINKS}
}
\description{
Give the type of entities or links given a name referring to it.
}
\seealso{
\code{\link{which_entities}}, \code{\link{which_associated_links}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
